serenity

Module builder

Available on crate feature builder only.

A set of builders used to make using methods on certain structs simpler to use.

These are used when not all parameters are required, all parameters are optional, and/or sane default values for required parameters can be applied by a builder.

Re-exports

pub use create_poll::CreatePoll;

pub use create_poll::CreatePollAnswer;

Modules

create_poll

Structs

AddMember

A builder to add parameters when using GuildId::add_member.

AutocompleteChoice

CreateAllowedMentions

A builder to manage the allowed mentions on a message.

CreateAttachment

A builder for creating a new attachment from a file path, file data, or URL.

CreateAutocompleteResponse

Discord docs

CreateBotAuthParameters

A builder for constructing an invite link with custom OAuth2 scopes.

CreateButton

A builder for creating a button component in a message

CreateChannel

A builder for creating a new GuildChannel in a Guild.

CreateCheckbox

A builder for creating a checkbox.

CreateCheckboxGroup

A builder for creating a checkbox group.

CreateCheckboxGroupOption

A builder for creating an option of a checkbox group component in a message.

CreateCommand

A builder for creating a new Command.

CreateCommandOption

A builder for creating a new CommandOption.

CreateCommandPermission

A builder for creating an CommandPermission.

CreateContainer

A builder to create a container, which acts similarly to embeds.

CreateEmbed

A builder to create an embed in a message

CreateEmbedAuthor

A builder to create the author data of an embed. See CreateEmbed::author

CreateEmbedFooter

A builder to create the footer data for an embed. See CreateEmbed::footer

CreateFile

A builder for specifying a file to be uploaded.

CreateFileUpload

A builder for creating a file upload in a modal.

CreateForumPost

Discord docs.

CreateForumTag

Discord docs

CreateGuildWelcomeChannel

A builder for creating a GuildWelcomeChannel.

CreateInputText

A builder for creating an input text component in a modal

CreateInteractionResponseFollowup

Discord docs

CreateInteractionResponseMessage

Discord docs.

CreateInvite

A builder to create an Invite for use via ChannelId::create_invite.

CreateLabel

A builder for creating a label, a top-level layout component that wraps modal components with a text label and optional description.

CreateMediaGallery

A builder to create a media gallery, a component that can contain multiple pieces of media.

CreateMediaGalleryItem

Builder to create individual media gallery items.

CreateMessage

A builder to specify the contents of an send message request, primarily meant for use through GenericChannelId::send_message.

CreateModal

Discord docs.

CreateRadioGroup

A builder for creating a radio group.

CreateRadioGroupOption

A builder for creating an option of a radio group component in a message.

CreateRoleColours

The colours of a Discord role, secondary_colour and tertiary_colour may only be set if the Guild has the ENHANCED_ROLE_COLORS feature.

CreateScheduledEvent

Discord docs

CreateSection

A builder to create a section component, supports up to a max of 3 components with an accessory.

CreateSelectMenu

A builder for creating a select menu component in a message

CreateSelectMenuOption

A builder for creating an option of a select menu component in a message.

CreateSeparator

A builder for creating a separator.

CreateSoundboard

A builder to create a soundboard sound.

CreateStageInstance

Builder for creating a stage instance

CreateSticker

A builder to create a guild sticker

CreateTextDisplay

A builder to create a text display component.

CreateThread

Discord docs:

CreateThumbnail

A builder to create a thumbnail for a section.

CreateUnfurledMediaItem

A builder to create a media item.

CreateWebhook

Discord docs

DataUri

A wrapper around some base64-encoded data. Used when an endpoint expects a base64 payload directly as part of the JSON body, instead of as a multipart upload.

EditAttachments

You can add new attachments and edit existing ones using this builder.

EditAutoModRule

A builder for creating or editing guild AutoMod rules.

EditChannel

A builder to edit a GuildChannel for use via GuildChannel::edit.

EditCommand

A builder for editing an existing Command.

EditCommandPermissions

A builder for creating several CommandPermission.

EditCurrentMember

A builder which edits the properties of the bot’s Member, to be used in conjunction with GuildId::edit_current_member.

EditGuild

A builder to optionally edit certain fields of a guild

EditGuildIncidentActions

A builder for editing guild incident actions.

EditGuildWelcomeScreen

A builder to edit the welcome screen of a guild

EditGuildWidget

A builder to specify the fields to edit in a GuildWidget.

EditInteractionResponse

Discord docs

EditMember

A builder which edits the properties of a Member, to be used in conjunction with Member::edit.

EditMessage

A builder to specify the fields to edit in an existing message.

EditProfile

A builder to edit the current user’s settings, to be used in conjunction with CurrentUser::edit.

EditRole

A builder to create or edit a Role for use via a number of model methods.

EditScheduledEvent

Discord docs

EditSoundboard

A builder to create or edit a Soundboard for use with GuildId::edit_soundboard.

EditStageInstance

Edits a StageInstance.

EditSticker

A builder to create or edit a Sticker for use via a number of model methods.

EditThread

Discord docs.

EditVoiceState

A builder which edits a user’s voice state, to be used in conjunction with GuildChannel::edit_voice_state.

EditWebhook

Discord docs

EditWebhookMessage

A builder to specify the fields to edit in an existing Webhook’s message.

ExecuteWebhook

A builder to create the content for a Webhook’s execution.

GetEntitlements

Builds a request to fetch active and ended Entitlements.

GetMessages

Builds a request to the API to retrieve messages.

Enums

AttachmentData

AutocompleteValue

CreateActionRow

A builder for creating a components action row in a message.

CreateComponent

A builder for creating components in a structured way.

CreateContainerComponent

An enum of all valid container components.

CreateInteractionResponse

Discord docs.

CreateModalComponent

CreateSectionAccessory

An enum of all valid section accessories.

CreateSectionComponent

An enum of all valid section components.

CreateSelectMenuKind

Discord docs: String Select, User Select, Role Select, Mentionable Select, Channel Select.