Struct serenity::builder::CreateApplicationCommand

pub struct CreateApplicationCommand(pub HashMap<&'static str, Value>);

A builder for creating a new Command.

Self::name and Self::description are required fields.

Tuple Fields
0: HashMap<&'static str, Value>
Implementations

impl CreateApplicationCommand

pub fn name<D: ToString>(&mut self, name: D) -> &mut Self

Specifies the name of the application command.

Note: Must be between 1 and 32 lowercase characters, matching r"^[\w-]{1,32}$". Two global commands of the same app cannot have the same name. Two guild-specific commands of the same app cannot have the same name.

pub fn name_localized<D: ToString, E: ToString>( &mut self, locale: E, name: D ) -> &mut Self

Specifies a localized name of the application command.

.name("birthday")
.name_localized("zh-CN", "็ๆฅ")
.name_localized("el", "ฮณฮตฮฝฮญฮธฮปฮนฮฑ")

pub fn kind(&mut self, kind: CommandType) -> &mut Self

Specifies the type of the application command.

pub fn default_member_permissions( &mut self, permissions: Permissions ) -> &mut Self

Specifies the default permissions required to execute the command.

pub fn dm_permission(&mut self, enabled: bool) -> &mut Self

Specifies if the command is available in DMs.

pub fn default_permission(&mut self, default_permission: bool) -> &mut Self

๐Deprecated: replaced by default_member_permissions

Specifies if the command should not be usable by default

Note: Setting it to false will disable it for anyone, including administrators and guild owners.

pub fn description<D: ToString>(&mut self, description: D) -> &mut Self

Specifies the description of the application command.

Note: Must be between 1 and 100 characters long.

pub fn description_localized<D: ToString, E: ToString>( &mut self, locale: E, description: D ) -> &mut Self

Specifies a localized description of the application command.

.description("Wish a friend a happy birthday")
.description_localized("zh-CN", "็ฅไฝๆๅ็ๆฅๅฟซไน")

pub fn create_option<F>(&mut self, f: F) -> &mut Selfwhere F: FnOnce(&mut CreateApplicationCommandOption) -> &mut CreateApplicationCommandOption,

Creates an application command option for the application command.

Note: Application commands can have up to 25 options.

pub fn add_option( &mut self, option: CreateApplicationCommandOption ) -> &mut Self

Adds an application command option for the application command.

Note: Application commands can have up to 25 options.

pub fn set_options( &mut self, options: Vec<CreateApplicationCommandOption> ) -> &mut Self

Sets all the application command options for the application command.

Note: Application commands can have up to 25 options.