serenity::builder

Struct CreateChannel

pub struct CreateChannel<'a> { /* private fields */ }

Available on crate feature builder only.

A builder for creating a new GuildChannel in a Guild.

Except Self::name, all fields are optional.

Discord docs.

Implementations

impl<'a> CreateChannel<'a>

pub fn new(name: impl Into<Cow<'a, str>>) -> Self

Creates a builder with the given name, setting Self::kind to ChannelType::Text and leaving all other fields empty.

pub fn name(self, name: impl Into<Cow<'a, str>>) -> Self

Specify the channel name (1-100 characters).

Replaces the current value set in Self::new.

pub fn kind(self, kind: ChannelType) -> Self

Specify the type of channel.

pub fn category(self, id: ChannelId) -> Self

Specify the parent category for the channel.

Only for ChannelType::Text, ChannelType::Voice, ChannelType::News, ChannelType::Stage, ChannelType::Forum

pub fn topic(self, topic: impl Into<Cow<'a, str>>) -> Self

Specify the channel topic (0-1024 characters).

Only for ChannelType::Text, ChannelType::News, ChannelType::Forum

pub fn nsfw(self, b: bool) -> Self

Specify whether the channel is age-restricted (18+).

Only for ChannelType::Text, ChannelType::Voice, ChannelType::News, ChannelType::Stage, ChannelType::Forum

pub fn bitrate(self, rate: u32) -> Self

Specify the bitrate (in bits per second; minimum 8000) of the voice or stage channel.

For voice channels, normal servers can set bitrate up to 96000, servers with Boost level 1 can set up to 128000, servers with Boost level 2 can set up to 256000, and servers with Boost level 3 or the VIP_REGIONS guild feature can set up to 384000. For stage channels, bitrate can be set up to 64000.

Only for ChannelType::Voice and ChannelType::Stage

pub fn user_limit(self, limit: NonMaxU16) -> Self

Specify the user limit of the voice channel.

Only for ChannelType::Voice and ChannelType::Stage

pub fn rate_limit_per_user(self, seconds: NonMaxU16) -> Self

Specify the amount of seconds a user has to wait before sending another message.

Bots and users with the BYPASS_SLOWMODE permission are unaffected.

Note: Must be between 0 and 21600 seconds (360 minutes, or 6 hours).

pub fn position(self, pos: u16) -> Self

Specify the sorting position of the channel.

Channels with the same position are sorted by id.

pub fn permissions( self, perms: impl Into<Cow<'a, [PermissionOverwrite]>>, ) -> Self

A set of overwrites defining what a user or a user carrying a certain role can and cannot do.

Example

Inheriting permissions from an existing channel:

use serenity::builder::CreateChannel;
use serenity::model::channel::{PermissionOverwrite, PermissionOverwriteType};
use serenity::model::id::UserId;
use serenity::model::permissions::Permissions;

// Assuming a guild has already been bound.
let permissions = vec![PermissionOverwrite {
    allow: Permissions::VIEW_CHANNEL,
    deny: Permissions::SEND_TTS_MESSAGES,
    kind: PermissionOverwriteType::Member(UserId::new(1234)),
}];

let builder = CreateChannel::new("my_new_cool_channel").permissions(permissions);
guild_id.create_channel(&http, builder).await?;

pub fn rtc_region(self, rtc_region: Cow<'a, str>) -> Self

Specify the channel voice region id of the voice or stage channel.

Automatic when left unset.

Only for ChannelType::Voice and ChannelType::Stage

pub fn video_quality_mode(self, video_quality_mode: VideoQualityMode) -> Self

Specify the camera video quality mode of the voice channel.

Only for ChannelType::Voice and ChannelType::Stage

pub fn default_auto_archive_duration( self, default_auto_archive_duration: AutoArchiveDuration, ) -> Self

Specify the default duration, in minutes, that the clients use (not the API) for newly created threads in the channel to automatically archive the thread after recent activity.

Only for ChannelType::Text, ChannelType::News, ChannelType::Forum

pub fn default_reaction_emoji(self, default_reaction_emoji: ForumEmoji) -> Self

Specify the emoji to show in the add reaction button on a thread in a forum channel.

Only for ChannelType::Forum

pub fn available_tags( self, available_tags: impl Into<Cow<'a, [CreateForumTag<'a>]>>, ) -> Self

Specify the set of tags that can be used in a forum channel.

Only for ChannelType::Forum

pub fn default_sort_order(self, default_sort_order: SortOrder) -> Self

Specify the default sort order type used to order posts in forum channels.

Only for ChannelType::Forum

pub fn default_forum_layout(self, default_forum_layout: ForumLayoutType) -> Self

Specify the default forum layout view used to display posts in forum channels.

pub fn default_thread_rate_limit_per_user( self, default_thread_rate_limit_per_user: NonMaxU16, ) -> Self

Specify the initial rate_limit_per_user to set on newly created threads in a channel. This field is copied to the thread at creation time and does not live update.

pub fn flags(self, flags: ChannelFlags) -> Self

Channel flags combined as a bitfield. Currently only ChannelFlags::REQUIRE_TAG (forum only) and ChannelFlags::IS_SPOILER_CHANNEL (text-based and voice only) are supported.

pub fn audit_log_reason(self, reason: &'a str) -> Self

Sets the request’s audit log reason.

pub async fn execute( self, http: &Http, guild_id: GuildId, ) -> Result<GuildChannel>

Available on crate feature http only.

Creates a new Channel in the guild.

Note: Requires the Manage Channels permission. If setting permission overwrites, only permissions your bot has in the guild can be allowed/denied. Setting Manage Roles permission in channels is only possible for guild administrators.

Fires a ChannelCreateEvent.

Errors

Returns Error::Http if the current user lacks permission or if invalid data is given.