Struct serenity::builder::CreateCommandOption

pub struct CreateCommandOption(_);

Available on crate feature builder only.

A builder for creating a new CommandOption.

Self::kind, Self::name, and Self::description are required fields.

Discord docs.

Implementations

impl CreateCommandOption

pub fn new( kind: CommandOptionType, name: impl Into<String>, description: impl Into<String> ) -> Self

Creates a new builder with the given option type, name, and description, leaving all other fields empty.

pub fn kind(self, kind: CommandOptionType) -> Self

Sets the CommandOptionType, replacing the current value as set in Self::new.

pub fn name(self, name: impl Into<String>) -> Self

Sets the name of the option, replacing the current value as set in Self::new.

Note: Must be between 1 and 32 lowercase characters, matching r"^[\w-]{1,32}$".

pub fn name_localized( self, locale: impl Into<String>, name: impl Into<String> ) -> Self

Specifies a localized name of the option.

.name("age")
.name_localized("zh-CN", "岁数")

pub fn description(self, description: impl Into<String>) -> Self

Sets the description for the option, replacing the current value as set in [Self::new].

Note: Must be between 1 and 100 characters.

pub fn description_localized( self, locale: impl Into<String>, description: impl Into<String> ) -> Self

Specifies a localized description of the option.

.description("Wish a friend a happy birthday")
.description_localized("zh-CN", "祝你朋友生日快乐")

pub fn required(self, required: bool) -> Self

Sets if this option is required or optional.

Note: This defaults to false.

pub fn add_int_choice(self, name: impl Into<String>, value: i32) -> Self

Adds an optional int-choice.

Note: There can be no more than 25 choices set. Name must be between 1 and 100 characters. Value must be between -2^53 and 2^53.

pub fn add_int_choice_localized( self, name: impl Into<String>, value: i32, locales: impl IntoIterator<Item = (impl Into<String>, impl Into<String>)> ) -> Self

Adds a localized optional int-choice. See Self::add_int_choice for more info.

pub fn add_string_choice( self, name: impl Into<String>, value: impl Into<String> ) -> Self

Adds an optional string-choice.

Note: There can be no more than 25 choices set. Name must be between 1 and 100 characters. Value must be up to 100 characters.

pub fn add_string_choice_localized( self, name: impl Into<String>, value: impl Into<String>, locales: impl IntoIterator<Item = (impl Into<String>, impl Into<String>)> ) -> Self

Adds a localized optional string-choice. See Self::add_string_choice for more info.

pub fn add_number_choice(self, name: impl Into<String>, value: f64) -> Self

Adds an optional number-choice.

Note: There can be no more than 25 choices set. Name must be between 1 and 100 characters. Value must be between -2^53 and 2^53.

pub fn add_number_choice_localized( self, name: impl Into<String>, value: f64, locales: impl IntoIterator<Item = (impl Into<String>, impl Into<String>)> ) -> Self

Adds a localized optional number-choice. See Self::add_number_choice for more info.

pub fn set_autocomplete(self, value: bool) -> Self

Optionally enable/disable autocomplete interactions for this option.

Notes:

May not be set to true if choices are set

Options using autocomplete are not confined to only use given choices

pub fn add_sub_option(self, sub_option: CreateCommandOption) -> Self

If the option is a SubCommandGroup or SubCommand, nested options are its parameters.

Note: A command can have up to 25 subcommand groups, or subcommands. A subcommand group can have up to 25 subcommands. A subcommand can have up to 25 options.

pub fn channel_types(self, channel_types: Vec<ChannelType>) -> Self

If the option is a Channel, it will only be able to show these types.

pub fn min_int_value(self, value: u64) -> Self

Sets the minimum permitted value for this integer option

pub fn max_int_value(self, value: u64) -> Self

Sets the maximum permitted value for this integer option

pub fn min_number_value(self, value: f64) -> Self

Sets the minimum permitted value for this number option

pub fn max_number_value(self, value: f64) -> Self

Sets the maximum permitted value for this number option

pub fn min_length(&mut self, value: u16) -> &mut Self

Sets the minimum permitted length for this string option.

The value of min_length must be greater or equal to 0.

pub fn max_length(&mut self, value: u16) -> &mut Self

Sets the maximum permitted length for this string option.

The value of max_length must be greater or equal to 1.