Struct serenity::builder::CreateInteractionResponseFollowup

pub struct CreateInteractionResponseFollowup<'a>(pub HashMap<&'static str, Value>, pub Vec<AttachmentType<'a>>);

Tuple Fields
0: HashMap<&'static str, Value>1: Vec<AttachmentType<'a>>
Implementations

impl<'a> CreateInteractionResponseFollowup<'a>

pub fn content<D: ToString>(&mut self, content: D) -> &mut Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn username<D: ToString>(&mut self, username: D) -> &mut Self

Override the default username of the webhook

pub fn avatar<D: ToString>(&mut self, avatar_url: D) -> &mut Self

Override the default avatar of the webhook

pub fn tts(&mut self, tts: bool) -> &mut Self

Set whether the message is text-to-speech.

Think carefully before setting this to true.

Defaults to false.

pub fn add_file<T: Into<AttachmentType<'a>>>(&mut self, file: T) -> &mut Self

Appends a file to the message.

pub fn add_files<T: Into<AttachmentType<'a>>, It: IntoIterator<Item = T>>( &mut self, files: It ) -> &mut Self

Appends a list of files to the message.

pub fn files<T: Into<AttachmentType<'a>>, It: IntoIterator<Item = T>>( &mut self, files: It ) -> &mut Self

Sets a list of files to include in the message.

Calling this multiple times will overwrite the file list. To append files, call Self::add_file or Self::add_files instead.

pub fn embed<F>(&mut self, f: F) -> &mut Selfwhere F: FnOnce(&mut CreateEmbed) -> &mut CreateEmbed,

Create an embed for the message.

pub fn add_embed(&mut self, embed: CreateEmbed) -> &mut Self

Adds an embed to the message.

pub fn add_embeds(&mut self, embeds: Vec<CreateEmbed>) -> &mut Self

Adds multiple embeds to the message.

pub fn set_embed(&mut self, embed: CreateEmbed) -> &mut Self

Sets a single embed to include in the message

Calling this will overwrite the embed list. To append embeds, call Self::add_embed instead.

pub fn set_embeds( &mut self, embeds: impl IntoIterator<Item = CreateEmbed> ) -> &mut Self

Sets a list of embeds to include in the message.

Calling this multiple times will overwrite the embed list. To append embeds, call Self::add_embed instead.

pub fn allowed_mentions<F>(&mut self, f: F) -> &mut Selfwhere F: FnOnce(&mut CreateAllowedMentions) -> &mut CreateAllowedMentions,

Set the allowed mentions for the message.

pub fn flags(&mut self, flags: MessageFlags) -> &mut Self

Sets the flags for the response.

pub fn ephemeral(&mut self, ephemeral: bool) -> &mut Self

Adds or removes the ephemeral flag

pub fn components<F>(&mut self, f: F) -> &mut Selfwhere F: FnOnce(&mut CreateComponents) -> &mut CreateComponents,

Creates components for this message.

pub fn set_components(&mut self, components: CreateComponents) -> &mut Self

Sets the components of this message.