Struct serenity::builder::CreateThread

pub struct CreateThread<'a> { /* private fields */ }

Available on crate feature builder only.

Implementations

impl<'a> CreateThread<'a>

pub fn new(name: impl Into<String>) -> Self

Creates a builder with the given thread name, leaving all other fields empty.

pub async fn execute(
self,
http: impl AsRef<Http>,
channel_id: ChannelId,
message_id: Option<MessageId>
) -> Result<GuildChannel>

Available on crate feature http only.

Creates a thread, either private or public. Public threads require a message to connect the thread to.

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub fn name(self, name: impl Into<String>) -> Self

The name of the thread. Replaces the current value as set in Self::new.

Note: Must be between 2 and 100 characters long.

pub fn auto_archive_duration(self, duration: u16) -> Self

Duration in minutes to automatically archive the thread after recent activity.

Note: Can only be set to 60, 1440, 4320, 10080 currently.

pub fn rate_limit_per_user(self, seconds: u16) -> Self

How many seconds must a user wait before sending another message.

Bots, or users with the MANAGE_MESSAGES and/or MANAGE_CHANNELS permissions are exempt from this restriction.

Note: Must be between 0 and 21600 seconds (360 minutes or 6 hours).

pub fn kind(self, kind: ChannelType) -> Self

The thread type, either ChannelType::PublicThread or ChannelType::PrivateThread.

Note: This defaults to ChannelType::PrivateThread in order to match the behavior when thread documentation was first published. This is a bit of a weird default though, and thus is highly likely to change in the future, so it is recommended to always explicitly setting it to avoid any breaking change.

pub fn audit_log_reason(self, reason: &'a str) -> Self

Sets the request’s audit log reason.