Struct serenity::builder::EditChannel

pub struct EditChannel<'a> { /* private fields */ }

Available on crate feature builder only.

A builder to edit a GuildChannel for use via GuildChannel::edit.

Discord docs.

Examples

Edit a channel, providing a new name and topic:

let builder = EditChannel::new().name("new name").topic("a test topic");
if let Err(why) = channel.edit(&http, builder).await {
    // properly handle the error
}

Implementations

impl<'a> EditChannel<'a>

pub fn new() -> Self

Equivalent to Self::default.

pub fn bitrate(self, bitrate: u32) -> Self

The bitrate of the channel in bits.

This is for voice channels only.

pub fn video_quality_mode(self, quality: VideoQualityMode) -> Self

The camera video quality mode of the channel.

This is for voice channels only.

pub fn voice_region(self, id: Option<String>) -> Self

The voice region of the channel. It is automatic when None.

This is for voice channels only.

pub fn name(self, name: impl Into<String>) -> Self

The name of the channel.

Must be between 2 and 100 characters long.

pub fn position(self, position: u16) -> Self

The position of the channel in the channel list.

pub fn topic(self, topic: impl Into<String>) -> Self

The topic of the channel. Can be empty.

Must be between 0 and 1024 characters long.

This is for text channels only.

pub fn nsfw(self, nsfw: bool) -> Self

Is the channel inappropriate for work?

This is for text channels only.

pub fn user_limit(self, user_limit: u32) -> Self

The number of users that may be in the channel simultaneously.

This is for voice channels only.

pub fn category<C: Into<Option<ChannelId>>>(self, category: C) -> Self

The parent category of the channel.

This is for text and voice channels only.

pub fn rate_limit_per_user(self, seconds: u16) -> Self

How many seconds must a user wait before sending another message.

Bots, or users with the MANAGE_MESSAGES and/or MANAGE_CHANNELS permissions are exempt from this restriction.

Note: Must be between 0 and 21600 seconds (360 minutes or 6 hours).

pub fn permissions( self, perms: impl IntoIterator<Item = PermissionOverwrite> ) -> Self

A set of overwrites defining what a user or a user carrying a certain role can or can’t do.

Example

Inheriting permissions from an existing channel:

use serenity::model::channel::{PermissionOverwrite, PermissionOverwriteType};
use serenity::model::id::UserId;
use serenity::model::permissions::Permissions;

// Assuming a channel has already been bound.
let permissions = vec![PermissionOverwrite {
    allow: Permissions::VIEW_CHANNEL,
    deny: Permissions::SEND_TTS_MESSAGES,
    kind: PermissionOverwriteType::Member(UserId::new(1234)),
}];

let builder = EditChannel::new().name("my_edited_cool_channel").permissions(permissions);
channel.edit(http, builder).await?;

pub fn available_tags( self, tags: impl IntoIterator<Item = CreateForumTag> ) -> Self

If this is a forum channel, sets the tags that can be assigned to forum posts.

pub fn audit_log_reason(self, reason: &'a str) -> Self

Sets the request’s audit log reason.

pub fn kind(self, kind: ChannelType) -> Self

The type of channel; only conversion between text and announcement is supported and only in guilds with the “NEWS” feature

pub fn default_auto_archive_duration( self, default_auto_archive_duration: AutoArchiveDuration ) -> Self

The default duration that the clients use (not the API) for newly created threads in the channel, in minutes, to automatically archive the thread after recent activity

pub fn flags(self, flags: ChannelFlags) -> Self

Channel flags combined as a bitfield. Currently only ChannelFlags::REQUIRE_TAG is supported.

pub fn default_reaction_emoji( self, default_reaction_emoji: Option<ForumEmoji> ) -> Self

The emoji to show in the add reaction button on a thread in a forum channel

pub fn default_thread_rate_limit_per_user( self, default_thread_rate_limit_per_user: u16 ) -> Self

The initial rate_limit_per_user to set on newly created threads in a channel. This field is copied to the thread at creation time and does not live update.

pub fn default_sort_order(self, default_sort_order: SortOrder) -> Self

The default sort order type used to order posts in forum channels

pub fn default_forum_layout(self, default_forum_layout: ForumLayoutType) -> Self

The default forum layout type used to display posts in forum channels