Struct serenity::builder::EditMember

pub struct EditMember(pub Map<String, Value>);

A builder which edits the properties of a Member, to be used in conjunction with Member::edit.

Methods

impl EditMember

fn deafen(self, deafen: bool) -> Self

Whether to deafen the member.

Requires the Deafen Members permission.

fn mute(self, mute: bool) -> Self

Whether to mute the member.

Requires the Mute Members permission.

fn nickname(self, nickname: &str) -> Self

Changes the member's nickname. Pass an empty string to reset the nickname.

Requires the Manage Nicknames permission.

fn roles(self, roles: &[RoleId]) -> Self

Set the list of roles that the member should have.

Requires the Manage Roles permission to modify.

fn voice_channel<C: Into<ChannelId>>(self, channel_id: C) -> Self

The Id of the voice channel to move the member to.

Requires the Move Members permission.