Struct serenity::builder::EditMessage

pub struct EditMessage { /* private fields */ }

Available on crate feature builder only.

A builder to specify the fields to edit in an existing message.

Examples

Editing the content of a Message to "hello":

let builder = EditMessage::new().content("hello");
message.edit(ctx, builder).await?;

Implementations

impl EditMessage

pub fn new() -> Self

Equivalent to Self::default.

pub async fn execute(
self,
http: impl AsRef<Http>,
channel_id: ChannelId,
message_id: MessageId
) -> Result<Message>

Available on crate feature http only.

Edits a message in the channel.

Note: Message contents must be under 2000 unicode code points, and embeds must be under 6000 code points.

Note: Requires that the current user be the author of the message. Other users can only call Self::suppress_embeds, but additionally require the Manage Messages permission to do so.

Note: If any embeds or attachments are set, they will overwrite the existing contents of the message, deleting existing embeds and attachments. Preserving them requires calling Self::add_existing_attachment in the case of attachments. In the case of embeds, duplicate copies of the existing embeds must be sent. Luckily, CreateEmbed implements From<Embed>, so one can simply call embed.into().

Errors

Returns a ModelError::MessageTooLong if the message contents are over the above limits.

Returns Error::Http if the user lacks permission, as well as if invalid data is given.

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn add_embed(self, embed: CreateEmbed) -> Self

Add an embed for the message.

Note: This will keep all existing embeds. Use Self::embed() to replace existing embeds.

pub fn add_embeds(self, embeds: Vec<CreateEmbed>) -> Self

Add multiple embeds for the message.

Note: This will keep all existing embeds. Use Self::embeds() to replace existing embeds.

pub fn embed(self, embed: CreateEmbed) -> Self

Set an embed for the message.

Note: This will replace all existing embeds. Use Self::add_embed() to keep existing embeds.

pub fn embeds(self, embeds: Vec<CreateEmbed>) -> Self

Set multiple embeds for the message.

Note: This will replace all existing embeds. Use Self::add_embeds() to keep existing embeds.

pub fn suppress_embeds(self, suppress: bool) -> Self

Suppress or unsuppress embeds in the message, this includes those generated by Discord themselves.

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Sets the components of this message.

pub fn button(self, button: CreateButton) -> Self

Adds a clickable button to this message.

Convenience method that wraps Self::components. Arranges buttons in action rows automatically.

pub fn select_menu(self, select_menu: CreateSelectMenu) -> Self

Adds an interactive select menu to this message.

Convenience method that wraps Self::components.

pub fn flags(self, flags: MessageFlags) -> Self

Sets the flags for the message.

pub fn attachment(self, attachment: CreateAttachment) -> Self

Add a new attachment for the message.

This can be called multiple times.

pub fn add_existing_attachment(self, id: AttachmentId) -> Self

Add an existing attachment by id.

pub fn remove_existing_attachment(self, id: AttachmentId) -> Self

Remove an existing attachment by id.