Struct serenity::builder::EditProfile

pub struct EditProfile(pub Map<String, Value>);

A builder to edit the current user's settings, to be used in conjunction with CurrentUser::edit.

Methods

impl EditProfile

fn avatar(self, avatar: Option<&str>) -> Self

Sets the avatar of the current user. None can be passed to remove an avatar.

A base64-encoded string is accepted as the avatar content.

Examples

A utility method - utils::read_image - is provided to read an image from a file and return its contents in base64-encoded form:

        use serenity::utils;

        // assuming a `context` has been bound

        let base64 = utils::read_image("./my_image.jpg")
        .expect("Failed to read image");

        let _ = context.edit_profile(|profile| {
            profile.avatar(Some(&base64))
        });
}

fn email(self, email: &str) -> Self

Modifies the current user's email address.

Note that when modifying the email address, the current password must also be provided.

No validation is performed on this by the library.

Note: This can only be used by user accounts.

fn new_password(self, new_password: &str) -> Self

Modifies the current user's password.

Note that when modifying the password, the current password must also be provided.

fn password(self, password: &str) -> Self

Used for providing the current password as verification when modifying the password or modifying the associated email address.

fn username(self, username: &str) -> Self

Modifies the current user's username.

When modifying the username, if another user has the same new username and current discriminator, a new unique discriminator will be assigned. If there are no available discriminators with the requested username, an error will occur.