Struct serenity::builder::EditRole

pub struct EditRole(pub Map<String, Value>);

A builer to create or edit a Role for use via a number of model methods.

These are:

PartialGuild::create_role

Guild::create_role

Guild::edit_role

GuildId::create_role

GuildId::edit_role

Role::edit

Defaults are provided for each parameter on role creation.

Examples

Create a hoisted, mentionable role named "a test role":

// assuming a `channel_id` and `guild_id` has been bound

let role = guild_id.create_role(|r| r
    .hoist(true)
    .mentionable(true)
    .name("a test role"));

Methods

impl EditRole

fn new(role: &Role) -> Self

Creates a new builder with the values of the given Role.

fn colour(self, colour: u64) -> Self

Sets the colour of the role.

fn hoist(self, hoist: bool) -> Self

Whether or not to hoist the role above lower-positioned role in the user list.

fn mentionable(self, mentionable: bool) -> Self

Whether or not to make the role mentionable, notifying its users.

fn name(self, name: &str) -> Self

The name of the role to set.

fn permissions(self, permissions: Permissions) -> Self

The set of permissions to assign the role.

fn position(self, position: u8) -> Self

The position to assign the role in the role list. This correlates to the role's position in the user list.