Struct serenity::builder::EditWebhookMessage

pub struct EditWebhookMessage { /* private fields */ }

Available on crate feature builder only.

A builder to specify the fields to edit in an existing Webhook’s message.

Discord docs

Implementations

impl EditWebhookMessage

pub fn new() -> Self

Equivalent to Self::default.

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn in_thread(self, thread_id: impl Into<ChannelId>) -> Self

Edits a message within a given thread. If the provided thread Id doesn’t belong to the current webhook, the API will return an error.

pub fn add_embed(self, embed: CreateEmbed) -> Self

Adds an embed for the message.

Embeds from the original message are reset when adding new embeds and must be re-added.

pub fn add_embeds(self, embeds: Vec<CreateEmbed>) -> Self

Adds multiple embeds to the message.

Embeds from the original message are reset when adding new embeds and must be re-added.

pub fn embed(self, embed: CreateEmbed) -> Self

Sets a single embed to include in the message

Calling this will overwrite the embed list. To append embeds, call Self::add_embed instead.

pub fn embeds(self, embeds: Vec<CreateEmbed>) -> Self

Sets the embeds for the message.

Note: You can only have up to 10 embeds per message.

Calling this will overwrite the embed list. To append embeds, call Self::add_embeds instead.

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Sets the components for this message. Requires an application-owned webhook, meaning either the webhook’s kind field is set to WebhookType::Application, or it was created by an application (and has kind WebhookType::Incoming).

pub fn button(self, button: CreateButton) -> Self

Adds a clickable button to this message.

Convenience method that wraps Self::components. Arranges buttons in action rows automatically.

pub fn select_menu(self, select_menu: CreateSelectMenu) -> Self

Adds an interactive select menu to this message.

Convenience method that wraps Self::components.

pub fn new_attachment(self, attachment: CreateAttachment) -> Self

Add a new attachment for the message.

This can be called multiple times.

If this is called one or more times, existing attachments will reset. To keep them, provide their IDs to Self::keep_existing_attachment.

pub fn keep_existing_attachment(self, id: AttachmentId) -> Self

Keeps an existing attachment by id.

To be used after Self::new_attachment or Self::clear_existing_attachments.

pub fn clear_existing_attachments(self) -> Self

Clears existing attachments.

In combination with Self::keep_existing_attachment, this can be used to selectively keep only some existing attachments.