Struct serenity::builder::GetMessages

pub struct GetMessages(pub BTreeMap<String, u64>);

Builds a request for a request to the API to retrieve messages.

This can have 2 different sets of parameters. The first set is around where to get the messages:

after

around

before

most_recent

These can not be mixed, and the first in the list alphabetically will be used. If one is not specified, most_recent will be used.

The fourth parameter is to specify the number of messages to retrieve. This does not need to be called and defaults to a value of 50.

This should be used only for retrieving messages; see GuildChannel::messages for examples.

Examples

Creating a GetMessages builder to retrieve the first 25 messages after the message with an Id of 158339864557912064:

use serenity::model::{ChannelId, MessageId};
use serenity::builder::GetMessages;

let retriever = GetMessages::default()
    .after(MessageId(158339864557912064))
    .limit(25);

// you can then pass it into a function which retrieves messages:
let channel_id = ChannelId(81384788765712384);

let _messages = channel_id.messages(|_| retriever)?;

Methods

impl GetMessages

fn after<M: Into<MessageId>>(self, message_id: M) -> Self

Indicates to retrieve the messages after a specific message, given by its Id.

fn around<M: Into<MessageId>>(self, message_id: M) -> Self

Indicates to retrieve the messages around a specific message in either direction (before+after) the given message.

fn before<M: Into<MessageId>>(self, message_id: M) -> Self

Indicates to retrieve the messages before a specific message, given by its Id.

fn limit(self, limit: u64) -> Self

The maximum number of messages to retrieve for the query.

If this is not specified, a default value of 50 is used.

Note: This field is capped to 100 messages due to a Discord limitation. If an amount larger than 100 is supplied, it will be reduced.

fn most_recent(self) -> Self

This is a function that is here for completeness. You do not need to call this - except to clear previous calls to after, around, and before - as it is the default value.