Trait serenity::builder::Builder

pub trait Builder {
    type Context<'ctx>;
    type Built;

    // Required method
    fn execute<'life0, 'async_trait>(
        self,
        cache_http: impl 'async_trait + CacheHttp,
        ctx: Self::Context<'life0>
    ) -> Pin<Box<dyn Future<Output = Result<Self::Built>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Available on crate features builder and http only.

Common trait for all HTTP request builders in this module.

Required Associated Types

type Context<'ctx>

Additional data that’s only required when sending a request off to the API.

type Built

Required Methods

fn execute<'life0, 'async_trait>( self, cache_http: impl 'async_trait + CacheHttp, ctx: Self::Context<'life0> ) -> Pin<Box<dyn Future<Output = Result<Self::Built>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Serializes a builder’s fields and sends the request off the API, returning the response.

Object Safety

This trait is not object safe.