serenity::builder

Trait ToArrayString

pub trait ToArrayString: Copy {
    type ArrayString: ArrayStringErased;

    const MAX_LENGTH: usize;

    // Required method
    fn to_arraystring(self) -> Self::ArrayString;
}

Available on crate feature builder only.

A no-alloc version of [ToString] implemented for bool/integer/float types formatting into an ArrayString.

Required Associated Constants

const MAX_LENGTH: usize

The maximum length that Self can be formatted into. This is used for the capacity generic of ArrayString.

Note for implementors

This must match the capacity generic used in ArrayString, otherwise logic bugs and panics may occur.

Required Associated Types

type ArrayString: ArrayStringErased

An associated type to turn ArrayString’s const generic into a type generic, working around limitations of the current type system.

This is always ArrayString, but in generic code this only usable as impl Deref<Target = str>.

Required Methods

fn to_arraystring(self) -> Self::ArrayString

Returns the value’s formatted representation in an appropriately sized ArrayString.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".