serenity::cache

Struct Cache

#[non_exhaustive]
pub struct Cache { /* private fields */ }

Available on crate feature cache only.

A cache containing data received from Shards.

Using the cache allows to avoid REST API requests via the http module where possible. Issuing too many requests will lead to ratelimits.

This is the list of cached resources and the events that populate them:

channels: ChannelCreateEvent, ChannelUpdateEvent, GuildCreateEvent

guilds: GuildCreateEvent

unavailable_guilds: ReadyEvent, GuildDeleteEvent

users: GuildMemberAddEvent, GuildMemberRemoveEvent, GuildMembersChunkEvent, PresenceUpdateEvent, ReadyEvent

presences: PresenceUpdateEvent, ReadyEvent

messages: MessageCreateEvent

The documentation of each event contains the required gateway intents.

Implementations

impl Cache

pub fn new() -> Self

Creates a new cache.

pub fn new_with_settings(settings: Settings) -> Self

Creates a new cache instance with settings applied.

Examples

use serenity::cache::{Cache, Settings};

let mut settings = Settings::default();
settings.max_messages = 10;

let cache = Cache::new_with_settings(settings);

pub fn unknown_members(&self) -> u32

Fetches the number of Members that have not had data received.

The important detail to note here is that this is the number of _member_s that have not had data received. A single User may have multiple associated member objects that have not been received.

This can be used in combination with Shard::chunk_guild, and can be used to determine how many members have not yet been received.

pub fn guilds(&self) -> Vec<GuildId>

Fetches a vector of all Guilds’ Ids that are stored in the cache.

Note that if you are utilizing multiple Shards, then the guilds retrieved over all shards are included in this count – not just the current Context’s shard, if accessing from one.

Examples

Print all of the Ids of guilds in the Cache:

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn dispatch(&self, ctx: &Context, event: &FullEvent) {
        match event {
            FullEvent::Ready {
                data_about_bot, ..
            } => {
                let guilds = ctx.cache.guilds().len();
                println!("Guilds in the Cache: {guilds}");
            },
            _ => {},
        }
    }
}

pub fn channel_messages( &self, channel_id: GenericChannelId, ) -> Option<ChannelMessagesRef<'_>>

Get a reference to the cached messages for a channel based on the given Id.

Examples

Find all messages by user ID 8 in channel ID 7:

if let Some(messages_in_channel) = cache.channel_messages(GenericChannelId::new(7)) {
    let messages_by_user: Vec<_> =
        messages_in_channel.iter().filter(|m| m.author.id == 8).collect();
}

pub fn guild(&self, id: GuildId) -> Option<GuildRef<'_>>

Gets a reference to a guild from the cache based on the given id.

Examples

Retrieve a guild from the cache and print its name:

// assuming the cache is in scope, e.g. via `Context`
if let Some(guild) = cache.guild(GuildId::new(7)) {
    println!("Guild name: {}", guild.name);
};

pub fn guild_count(&self) -> usize

Returns the number of cached guilds.

pub fn unavailable_guilds(&self) -> ReadOnlyMapRef<'_, GuildId, ()>

This method clones and returns all unavailable guilds.

pub fn shard_count(&self) -> NonZeroU16

Returns the number of shards.

pub fn message( &self, channel_id: GenericChannelId, message_id: MessageId, ) -> Option<MessageRef<'_>>

Retrieves a Channel’s message from the cache based on the channel’s and message’s given Ids.

Note: This will clone the entire message.

Examples

Retrieving the message object from a channel.

match cache.message(message.channel_id, message.id) {
    Some(m) => assert_eq!(message.content, m.content),
    None => println!("No message found in cache."),
};

pub fn settings(&self) -> SettingsRef<'_>

Returns the settings.

Examples

Printing the maximum number of messages in a channel to be cached:

use serenity::cache::Cache;

let mut cache = Cache::new();
println!("Max settings: {}", cache.settings().max_messages);

pub fn set_max_messages(&self, max: usize)

Sets the maximum amount of messages per channel to cache.

By default, no messages will be cached.

pub fn current_user(&self) -> CurrentUserRef<'_>

This method provides a reference to the user used by the bot.

pub fn guild_categories( &self, guild_id: GuildId, ) -> Option<ExtractMap<ChannelId, GuildChannel>>

Clones all channel categories in the given guild and returns them.

pub fn update<E: CacheUpdate>(&self, e: &mut E) -> Option<E::Output>

Updates the cache with the update implementation for an event or other custom update implementation.

Refer to the documentation for CacheUpdate for more information.

Examples

Refer to the CacheUpdate examples.