Module serenity::client::bridge

Available on crate feature client only.

A collection of bridged support between the client module and other modules.

Warning: You likely do not need to mess with anything in here. Beware. This is lower-level functionality abstracted by the Client.

Modules

gatewaygateway

The client gateway bridge is support essential for the client module.

voicevoice