Struct serenity::client::bridge::gateway::GatewayIntents

pub struct GatewayIntents {
    pub bits: u64,
}

Gateway Intents will limit the events your bot will receive via the gateway. By default, no intents are specified by Serenity.

Fields
bits: u64

The flags composing gateway intents.

Note

Do not modify this yourself; use the provided methods. Do the same when creating, unless you’re absolutely certain that you’re giving valid intents flags.

Implementations

impl GatewayIntents

pub const GUILDS: GatewayIntents

Enables following gateway events:

GUILD_CREATE

GUILD_DELETE

GUILD_ROLE_CREATE

GUILD_ROLE_UPDATE

GUILD_ROLE_DELETE

CHANNEL_CREATE

CHANNEL_UPDATE

CHANNEL_DELETE

CHANNEL_PINS_UPDATE

pub const GUILD_MEMBERS: GatewayIntents

Enables following gateway events:

GUILD_MEMBER_ADD

GUILD_MEMBER_UPDATE

GUILD_MEMBER_REMOVE

Info: This intent is privileged. In order to use it, you must head to your application in the Developer Portal and enable the toggle for Privileged Intents.

This intent is also necessary to even recieve the events in contains.

pub const GUILD_BANS: GatewayIntents

Enables following gateway events:

GUILD_BAN_ADD

GUILD_BAN_REMOVE

pub const GUILD_EMOJIS: GatewayIntents

Enables following gateway event:

GUILD_EMOJIS_UPDATE

pub const GUILD_INTEGRATIONS: GatewayIntents

Enables following gateway event:

GUILD_INTEGRATIONS_UPDATE

pub const GUILD_WEBHOOKS: GatewayIntents

Enables following gateway event:

WEBHOOKS_UPDATE

pub const GUILD_INVITES: GatewayIntents

Enables following gateway events:

INVITE_CREATE

INVITE_DELETE

pub const GUILD_VOICE_STATES: GatewayIntents

Enables following gateway event:

VOICE_STATE_UPDATE

pub const GUILD_PRESENCES: GatewayIntents

Enables following gateway event:

PRESENCE_UPDATE

Info: This intent is privileged. In order to use it, you must head to your application in the Developer Portal and enable the toggle for Privileged Intents.

This intent is also necessary to even recieve the events in contains.

pub const GUILD_MESSAGES: GatewayIntents

Enables following gateway events:

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

pub const GUILD_MESSAGE_REACTIONS: GatewayIntents

Enables following gateway events:

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOJI

pub const GUILD_MESSAGE_TYPING: GatewayIntents

Enable following gateway event:

TYPING_START

pub const DIRECT_MESSAGES: GatewayIntents

Enable following gateway events:

CHANNEL_CREATE

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

CHANNEL_PINS_UPDATE

pub const DIRECT_MESSAGE_REACTIONS: GatewayIntents

Enable following gateway events:

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOJI

pub const DIRECT_MESSAGE_TYPING: GatewayIntents

Enable following gateway event:

TYPING_START

pub const fn empty() -> GatewayIntents

Returns an empty set of flags

pub const fn all() -> GatewayIntents

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u64) -> Option<GatewayIntents>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> GatewayIntents

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> GatewayIntents

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: GatewayIntents) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: GatewayIntents) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: GatewayIntents)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: GatewayIntents)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: GatewayIntents)

Toggles the specified flags in-place.

pub fn set(&mut self, other: GatewayIntents, value: bool)

Inserts or removes the specified flags depending on the passed value.

impl GatewayIntents

pub const fn non_privileged() -> GatewayIntents

Gets all of the intents that don’t are considered privileged by Discord.

pub const fn privileged() -> GatewayIntents

Gets all of the intents that are considered privileged by Discord. Use of these intents will require explicitly whitelisting the bot.

pub fn is_privileged(self) -> bool

Checks if any of the included intents are privileged

pub fn guilds(self) -> bool

Shorthand for checking that the set of intents contains the GUILDS intent.

pub fn guild_members(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_MEMBERS intent.

pub fn guild_bans(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_BANS intent.

pub fn guild_emojis(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_EMOJIS intent.

pub fn guild_integrations(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_INTEGRATIONS intent.

pub fn guild_webhooks(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_WEBHOOKS intent.

pub fn guild_invites(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_INVITES intent.

pub fn guild_voice_states(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_VOICE_STATES intent.

pub fn guild_presences(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_PRESENCES intent.

pub fn guild_message_reactions(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_MESSAGE_REACTIONS intent.

pub fn guild_message_typing(self) -> bool

Shorthand for checking that the set of intents contains the GUILD_MESSAGE_TYPING intent.

pub fn direct_messages(self) -> bool

Shorthand for checking that the set of intents contains the DIRECT_MESSAGES intent.

pub fn direct_message_reactions(self) -> bool

Shorthand for checking that the set of intents contains the DIRECT_MESSAGE_REACTIONS intent.

pub fn direct_message_typing(self) -> bool

Shorthand for checking that the set of intents contains the DIRECT_MESSAGE_TYPING intent.