Enum serenity::client::FullEvent

#[non_exhaustive]
pub enum FullEvent {

    CommandPermissionsUpdate {
        ctx: Context,
        permission: CommandPermission,
    },
    AutoModRuleCreate {
        ctx: Context,
        rule: Rule,
    },
    AutoModRuleUpdate {
        ctx: Context,
        rule: Rule,
    },
    AutoModRuleDelete {
        ctx: Context,
        rule: Rule,
    },
    AutoModActionExecution {
        ctx: Context,
        execution: ActionExecution,
    },
    CacheReady {
        ctx: Context,
        guilds: Vec<GuildId>,
    },
    ShardsReady {
        ctx: Context,
        total_shards: u32,
    },
    ChannelCreate {
        ctx: Context,
        channel: GuildChannel,
    },
    CategoryCreate {
        ctx: Context,
        category: GuildChannel,
    },
    CategoryDelete {
        ctx: Context,
        category: GuildChannel,
    },
    ChannelDelete {
        ctx: Context,
        channel: GuildChannel,
        messages: Option<Vec<Message>>,
    },
    ChannelPinsUpdate {
        ctx: Context,
        pin: ChannelPinsUpdateEvent,
    },
    ChannelUpdate {
        ctx: Context,
        old: Option<Channel>,
        new: Channel,
    },
    GuildAuditLogEntryCreate {
        ctx: Context,
        entry: AuditLogEntry,
        guild_id: GuildId,
    },
    GuildBanAddition {
        ctx: Context,
        guild_id: GuildId,
        banned_user: User,
    },
    GuildBanRemoval {
        ctx: Context,
        guild_id: GuildId,
        unbanned_user: User,
    },
    GuildCreate {
        ctx: Context,
        guild: Guild,
        is_new: Option<bool>,
    },
    GuildDelete {
        ctx: Context,
        incomplete: UnavailableGuild,
        full: Option<Guild>,
    },
    GuildEmojisUpdate {
        ctx: Context,
        guild_id: GuildId,
        current_state: HashMap<EmojiId, Emoji>,
    },
    GuildIntegrationsUpdate {
        ctx: Context,
        guild_id: GuildId,
    },
    GuildMemberAddition {
        ctx: Context,
        new_member: Member,
    },
    GuildMemberRemoval {
        ctx: Context,
        guild_id: GuildId,
        user: User,
        member_data_if_available: Option<Member>,
    },
    GuildMemberUpdate {
        ctx: Context,
        old_if_available: Option<Member>,
        new: Option<Member>,
        event: GuildMemberUpdateEvent,
    },
    GuildMembersChunk {
        ctx: Context,
        chunk: GuildMembersChunkEvent,
    },
    GuildRoleCreate {
        ctx: Context,
        new: Role,
    },
    GuildRoleDelete {
        ctx: Context,
        guild_id: GuildId,
        removed_role_id: RoleId,
        removed_role_data_if_available: Option<Role>,
    },
    GuildRoleUpdate {
        ctx: Context,
        old_data_if_available: Option<Role>,
        new: Role,
    },
    GuildStickersUpdate {
        ctx: Context,
        guild_id: GuildId,
        current_state: HashMap<StickerId, Sticker>,
    },
    GuildUpdate {
        ctx: Context,
        old_data_if_available: Option<Guild>,
        new_but_incomplete: PartialGuild,
    },
    InviteCreate {
        ctx: Context,
        data: InviteCreateEvent,
    },
    InviteDelete {
        ctx: Context,
        data: InviteDeleteEvent,
    },
    Message {
        ctx: Context,
        new_message: Message,
    },
    MessageDelete {
        ctx: Context,
        channel_id: ChannelId,
        deleted_message_id: MessageId,
        guild_id: Option<GuildId>,
    },
    MessageDeleteBulk {
        ctx: Context,
        channel_id: ChannelId,
        multiple_deleted_messages_ids: Vec<MessageId>,
        guild_id: Option<GuildId>,
    },
    MessageUpdate {
        ctx: Context,
        old_if_available: Option<Message>,
        new: Option<Message>,
        event: MessageUpdateEvent,
    },
    ReactionAdd {
        ctx: Context,
        add_reaction: Reaction,
    },
    ReactionRemove {
        ctx: Context,
        removed_reaction: Reaction,
    },
    ReactionRemoveAll {
        ctx: Context,
        channel_id: ChannelId,
        removed_from_message_id: MessageId,
    },
    ReactionRemoveEmoji {
        ctx: Context,
        removed_reactions: Reaction,
    },
    PresenceReplace {
        ctx: Context,
        presences: Vec<Presence>,
    },
    PresenceUpdate {
        ctx: Context,
        new_data: Presence,
    },
    Ready {
        ctx: Context,
        data_about_bot: Ready,
    },
    Resume {
        ctx: Context,
        event: ResumedEvent,
    },
    ShardStageUpdate {
        ctx: Context,
        event: ShardStageUpdateEvent,
    },
    TypingStart {
        ctx: Context,
        event: TypingStartEvent,
    },
    UserUpdate {
        ctx: Context,
        old_data: Option<CurrentUser>,
        new: CurrentUser,
    },
    VoiceServerUpdate {
        ctx: Context,
        event: VoiceServerUpdateEvent,
    },
    VoiceStateUpdate {
        ctx: Context,
        old: Option<VoiceState>,
        new: VoiceState,
    },
    WebhookUpdate {
        ctx: Context,
        guild_id: GuildId,
        belongs_to_channel_id: ChannelId,
    },
    InteractionCreate {
        ctx: Context,
        interaction: Interaction,
    },
    IntegrationCreate {
        ctx: Context,
        integration: Integration,
    },
    IntegrationUpdate {
        ctx: Context,
        integration: Integration,
    },
    IntegrationDelete {
        ctx: Context,
        integration_id: IntegrationId,
        guild_id: GuildId,
        application_id: Option<ApplicationId>,
    },
    StageInstanceCreate {
        ctx: Context,
        stage_instance: StageInstance,
    },
    StageInstanceUpdate {
        ctx: Context,
        stage_instance: StageInstance,
    },
    StageInstanceDelete {
        ctx: Context,
        stage_instance: StageInstance,
    },
    ThreadCreate {
        ctx: Context,
        thread: GuildChannel,
    },
    ThreadUpdate {
        ctx: Context,
        thread: GuildChannel,
    },
    ThreadDelete {
        ctx: Context,
        thread: PartialGuildChannel,
    },
    ThreadListSync {
        ctx: Context,
        thread_list_sync: ThreadListSyncEvent,
    },
    ThreadMemberUpdate {
        ctx: Context,
        thread_member: ThreadMember,
    },
    ThreadMembersUpdate {
        ctx: Context,
        thread_members_update: ThreadMembersUpdateEvent,
    },
    GuildScheduledEventCreate {
        ctx: Context,
        event: ScheduledEvent,
    },
    GuildScheduledEventUpdate {
        ctx: Context,
        event: ScheduledEvent,
    },
    GuildScheduledEventDelete {
        ctx: Context,
        event: ScheduledEvent,
    },
    GuildScheduledEventUserAdd {
        ctx: Context,
        subscribed: GuildScheduledEventUserAddEvent,
    },
    GuildScheduledEventUserRemove {
        ctx: Context,
        unsubscribed: GuildScheduledEventUserRemoveEvent,
    },
    Ratelimit {
        data: RatelimitInfo,
    },

}

Available on crate features gateway and client only.

This enum stores every possible event that an EventHandler can receive.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CommandPermissionsUpdate

Fields

ctx: Context

permission: CommandPermission

Dispatched when the permissions of an application command was updated.

Provides said permission’s data.

AutoModRuleCreate

Fields

ctx: Context

rule: Rule

Dispatched when an auto moderation rule was created.

Provides said rule’s data.

AutoModRuleUpdate

Fields

ctx: Context

rule: Rule

Dispatched when an auto moderation rule was updated.

Provides said rule’s data.

AutoModRuleDelete

Fields

ctx: Context

rule: Rule

Dispatched when an auto moderation rule was deleted.

Provides said rule’s data.

AutoModActionExecution

Fields

ctx: Context

execution: ActionExecution

Dispatched when an auto moderation rule was triggered and an action was executed.

Provides said action execution’s data.

CacheReady

Fields

ctx: Context

guilds: Vec<GuildId>

Available on crate feature cache only.

Dispatched when the cache has received and inserted all data from guilds.

This process happens upon starting your bot and should be fairly quick. However, cache actions performed prior this event may fail as the data could be not inserted yet.

Provides the cached guilds’ ids.

ShardsReady

Fields

ctx: Context

total_shards: u32

Available on crate feature cache only.

Dispatched when every shard has received a Ready event

ChannelCreate

Fields

ctx: Context

channel: GuildChannel

Dispatched when a channel is created.

Provides said channel’s data.

CategoryCreate

Fields

ctx: Context

category: GuildChannel

Dispatched when a category is created.

Provides said category’s data.

CategoryDelete

Fields

ctx: Context

category: GuildChannel

Dispatched when a category is deleted.

Provides said category’s data.

ChannelDelete

Fields

ctx: Context

channel: GuildChannel

messages: Option<Vec<Message>>

Dispatched when a channel is deleted.

Provides said channel’s data.

ChannelPinsUpdate

Fields

ctx: Context

pin: ChannelPinsUpdateEvent

Dispatched when a pin is added, deleted.

Provides said pin’s data.

ChannelUpdate

Fields

ctx: Context

old: Option<Channel>

new: Channel

Dispatched when a channel is updated.

The old channel data is only provided when the cache feature is enabled.

GuildAuditLogEntryCreate

Fields

ctx: Context

entry: AuditLogEntry

guild_id: GuildId

Dispatched when a new audit log entry is created.

Provides said entry’s data and the id of the guild where it was created.

GuildBanAddition

Fields

ctx: Context

guild_id: GuildId

banned_user: User

Dispatched when a user is banned from a guild.

Provides the guild’s id and the banned user’s data.

GuildBanRemoval

Fields

ctx: Context

guild_id: GuildId

unbanned_user: User

Dispatched when a user’s ban is lifted from a guild.

Provides the guild’s id and the lifted user’s data.

GuildCreate

Fields

ctx: Context

guild: Guild

is_new: Option<bool>

Dispatched when a guild is created; or an existing guild’s data is sent to us.

Provides the guild’s data and whether the guild is new (only when cache feature is enabled).

GuildDelete

Fields

ctx: Context

incomplete: UnavailableGuild

full: Option<Guild>

Dispatched when a guild is deleted.

Provides the partial data of the guild sent by discord, and the full data from the cache, if cache feature is enabled and the data is available.

The unavailable flag in the partial data determines the status of the guild. If the flag is false, the bot was removed from the guild, either by being kicked or banned. If the flag is true, the guild went offline.

GuildEmojisUpdate

Fields

ctx: Context

guild_id: GuildId

current_state: HashMap<EmojiId, Emoji>

Dispatched when the emojis are updated.

Provides the guild’s id and the new state of the emojis in the guild.

GuildIntegrationsUpdate

Fields

ctx: Context

guild_id: GuildId

Dispatched when a guild’s integration is added, updated or removed.

Provides the guild’s id.

GuildMemberAddition

Fields

ctx: Context

new_member: Member

Dispatched when a user joins a guild.

Provides the guild’s id and the user’s member data.

Note: This event will not trigger unless the “guild members” privileged intent is enabled on the bot application page.

GuildMemberRemoval

Fields

ctx: Context

guild_id: GuildId

user: User

member_data_if_available: Option<Member>

Dispatched when a user’s membership ends by leaving, getting kicked, or being banned.

Provides the guild’s id, the user’s data, and the user’s member data if cache feature is enabled and the data is available.

Note: This event will not trigger unless the “guild members” privileged intent is enabled on the bot application page.

GuildMemberUpdate

Fields

ctx: Context

old_if_available: Option<Member>

new: Option<Member>

event: GuildMemberUpdateEvent

Dispatched when a member is updated (e.g their nickname is updated).

Provides the member’s old and new data (if cache feature is enabled and data is available) and the new raw data about updated fields.

Note: This event will not trigger unless the “guild members” privileged intent is enabled on the bot application page.

GuildMembersChunk

Fields

ctx: Context

chunk: GuildMembersChunkEvent

Dispatched when the data for offline members was requested.

Provides the guild’s id and the data.

GuildRoleCreate

Fields

ctx: Context

new: Role

Dispatched when a role is created.

Provides the guild’s id and the new role’s data.

GuildRoleDelete

Fields

ctx: Context

guild_id: GuildId

removed_role_id: RoleId

removed_role_data_if_available: Option<Role>

Dispatched when a role is deleted.

Provides the guild’s id, the role’s id and its data (if cache feature is enabled and the data is available).

GuildRoleUpdate

Fields

ctx: Context

old_data_if_available: Option<Role>

new: Role

Dispatched when a role is updated.

Provides the guild’s id, the role’s old (if cache feature is enabled and the data is available) and new data.

GuildStickersUpdate

Fields

ctx: Context

guild_id: GuildId

current_state: HashMap<StickerId, Sticker>

Dispatched when the stickers are updated.

Provides the guild’s id and the new state of the stickers in the guild.

GuildUpdate

Fields

ctx: Context

old_data_if_available: Option<Guild>

new_but_incomplete: PartialGuild

Dispatched when the guild is updated.

Provides the guild’s old full data (if cache feature is enabled and the data is available) and the new, albeit partial data.

InviteCreate

Fields

ctx: Context

data: InviteCreateEvent

Dispatched when a invite is created.

Provides data about the invite.

InviteDelete

Fields

ctx: Context

data: InviteDeleteEvent

Dispatched when a invite is deleted.

Provides data about the invite.

Message

Fields

ctx: Context

new_message: Message

Dispatched when a message is created.

Provides the message’s data.

MessageDelete

Fields

ctx: Context

channel_id: ChannelId

deleted_message_id: MessageId

guild_id: Option<GuildId>

Dispatched when a message is deleted.

Provides the guild’s id, the channel’s id and the message’s id.

MessageDeleteBulk

Fields

ctx: Context

channel_id: ChannelId

multiple_deleted_messages_ids: Vec<MessageId>

guild_id: Option<GuildId>

Dispatched when multiple messages were deleted at once.

Provides the guild’s id, channel’s id and the deleted messages’ ids.

MessageUpdate

Fields

ctx: Context

old_if_available: Option<Message>

new: Option<Message>

event: MessageUpdateEvent

Dispatched when a message is updated.

Provides the message update data, as well as the actual old and new message if cache feature is enabled and the data is available.

ReactionAdd

Fields

ctx: Context

add_reaction: Reaction

Dispatched when a new reaction is attached to a message.

Provides the reaction’s data.

ReactionRemove

Fields

ctx: Context

removed_reaction: Reaction

Dispatched when a reaction is detached from a message.

Provides the reaction’s data.

ReactionRemoveAll

Fields

ctx: Context

channel_id: ChannelId

removed_from_message_id: MessageId

Dispatched when all reactions of a message are detached from a message.

Provides the channel’s id and the message’s id.

ReactionRemoveEmoji

Fields

ctx: Context

removed_reactions: Reaction

Dispatched when all reactions of a message are detached from a message.

Provides the channel’s id and the message’s id.

PresenceReplace

Fields

ctx: Context

presences: Vec<Presence>

This event is legacy, and likely no longer sent by discord.

PresenceUpdate

Fields

ctx: Context

new_data: Presence

Dispatched when a user’s presence is updated (e.g off -> on).

Provides the presence’s new data.

Note: This event will not trigger unless the “guild presences” privileged intent is enabled on the bot application page.

Ready

Fields

ctx: Context

data_about_bot: Ready

Dispatched upon startup.

Provides data about the bot and the guilds it’s in.

Resume

Fields

ctx: Context

event: ResumedEvent

Dispatched upon reconnection.

ShardStageUpdate

Fields

ctx: Context

event: ShardStageUpdateEvent

Dispatched when a shard’s connection stage is updated

Provides the context of the shard and the event information about the update.

TypingStart

Fields

ctx: Context

event: TypingStartEvent

Dispatched when a user starts typing.

UserUpdate

Fields

ctx: Context

old_data: Option<CurrentUser>

new: CurrentUser

Dispatched when the bot’s data is updated.

Provides the old (if cache feature is enabled and the data is available) and new data.

VoiceServerUpdate

Fields

ctx: Context

event: VoiceServerUpdateEvent

Dispatched when a guild’s voice server was updated (or changed to another one).

Provides the voice server’s data.

VoiceStateUpdate

Fields

ctx: Context

old: Option<VoiceState>

new: VoiceState

Dispatched when a user joins, leaves or moves to a voice channel.

Provides the guild’s id (if available) and the old state (if cache feature is enabled and GatewayIntents::GUILDS is enabled) and the new state of the guild’s voice channels.

WebhookUpdate

Fields

ctx: Context

guild_id: GuildId

belongs_to_channel_id: ChannelId

Dispatched when a guild’s webhook is updated.

Provides the guild’s id and the channel’s id the webhook belongs in.

InteractionCreate

Fields

ctx: Context

interaction: Interaction

Dispatched when an interaction is created (e.g a slash command was used or a button was clicked).

Provides the created interaction.

IntegrationCreate

Fields

ctx: Context

integration: Integration

Dispatched when a guild integration is created.

Provides the created integration.

IntegrationUpdate

Fields

ctx: Context

integration: Integration

Dispatched when a guild integration is updated.

Provides the updated integration.

IntegrationDelete

Fields

ctx: Context

integration_id: IntegrationId

guild_id: GuildId

application_id: Option<ApplicationId>

Dispatched when a guild integration is deleted.

Provides the integration’s id, the id of the guild it belongs to, and its associated application id

StageInstanceCreate

Fields

ctx: Context

stage_instance: StageInstance

Dispatched when a stage instance is created.

Provides the created stage instance.

StageInstanceUpdate

Fields

ctx: Context

stage_instance: StageInstance

Dispatched when a stage instance is updated.

Provides the updated stage instance.

StageInstanceDelete

Fields

ctx: Context

stage_instance: StageInstance

Dispatched when a stage instance is deleted.

Provides the deleted stage instance.

ThreadCreate

Fields

ctx: Context

thread: GuildChannel

Dispatched when a thread is created or the current user is added to a private thread.

Provides the thread.

ThreadUpdate

Fields

ctx: Context

thread: GuildChannel

Dispatched when a thread is updated.

Provides the updated thread.

ThreadDelete

Fields

ctx: Context

thread: PartialGuildChannel

Dispatched when a thread is deleted.

Provides the partial deleted thread.

ThreadListSync

Fields

ctx: Context

thread_list_sync: ThreadListSyncEvent

Dispatched when the current user gains access to a channel.

Provides the threads the current user can access, the thread members, the guild Id, and the channel Ids of the parent channels being synced.

ThreadMemberUpdate

Fields

ctx: Context

thread_member: ThreadMember

Dispatched when the ThreadMember for the current user is updated.

Provides the updated thread member.

ThreadMembersUpdate

Fields

ctx: Context

thread_members_update: ThreadMembersUpdateEvent

Dispatched when anyone is added to or removed from a thread. If the current user does not have the GatewayIntents::GUILDS, then this event will only be sent if the current user was added to or removed from the thread.

Provides the added/removed members, the approximate member count of members in the thread, the thread Id and its guild Id.

GuildScheduledEventCreate

Fields

ctx: Context

event: ScheduledEvent

Dispatched when a scheduled event is created.

Provides data about the scheduled event.

GuildScheduledEventUpdate

Fields

ctx: Context

event: ScheduledEvent

Dispatched when a scheduled event is updated.

Provides data about the scheduled event.

GuildScheduledEventDelete

Fields

ctx: Context

event: ScheduledEvent

Dispatched when a scheduled event is deleted.

Provides data about the scheduled event.

GuildScheduledEventUserAdd

Fields

ctx: Context

subscribed: GuildScheduledEventUserAddEvent

Dispatched when a guild member has subscribed to a scheduled event.

Provides data about the subscription.

GuildScheduledEventUserRemove

Fields

ctx: Context

unsubscribed: GuildScheduledEventUserRemoveEvent

Dispatched when a guild member has unsubscribed from a scheduled event.

Provides data about the cancelled subscription.

Ratelimit

Fields

data: RatelimitInfo

Dispatched when an HTTP rate limit is hit

Implementations

impl FullEvent

pub fn snake_case_name(&self) -> &'static str

Returns the name of this event as a snake case string

if let FullEvent::Message { .. } = &event {
    assert_eq!(event.snake_case_name(), "message");
}

pub async fn dispatch(self, handler: &dyn EventHandler)

Runs the given EventHandler’s code for this event.