Struct serenity::client::Context

pub struct Context {
    pub shard: ShardMessenger,
    pub shard_id: ShardId,
    pub http: Arc<Http>,
    pub cache: Arc<Cache>,
    /* private fields */
}

Available on crate feature client only.

The context is a general utility struct provided on event dispatches, which helps with dealing with the current “context” of the event dispatch. The context also acts as a general high-level interface over the associated Shard which received the event, or the low-level http module.

The context contains “shortcuts”, like for interacting with the shard. Methods like Self::set_activity will unlock the shard and perform an update for you to save a bit of work.

A context will only live for the event it was dispatched for. After the event handler finished, it is destroyed and will not be re-used.

Fields
shard: ShardMessenger

The messenger to communicate with the shard runner.

shard_id: ShardId

The ID of the shard this context is related to.

http: Arc<Http>cache: Arc<Cache>
Available on crate feature cache only.

Implementations

impl Context

pub fn data<Data: Send + Sync + 'static>(&self) -> Arc<Data>

A container for a data type that can be used across contexts.

The purpose of the data field is to be accessible and persistent across contexts; that is, data can be modified by one context, and will persist through the future and be accessible through other contexts. This is useful for anything that should “live” through the program: counters, database connections, custom user caches, etc.

Panics

Panics if the generic provided is not equal to the type provided in ClientBuilder::data.

pub fn online(&self)

Available on crate feature gateway only.

Sets the current user as being Online. This maintains the current activity.

Examples

Set the current user to being online on the shard:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if msg.content == "!online" {
            ctx.online();
        }
    }
}

pub fn idle(&self)

Available on crate feature gateway only.

Sets the current user as being Idle. This maintains the current activity.

Examples

Set the current user to being idle on the shard:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if msg.content == "!idle" {
            ctx.idle();
        }
    }
}

pub fn dnd(&self)

Available on crate feature gateway only.

Sets the current user as being DoNotDisturb. This maintains the current activity.

Examples

Set the current user to being Do Not Disturb on the shard:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if msg.content == "!dnd" {
            ctx.dnd();
        }
    }
}

pub fn invisible(&self)

Available on crate feature gateway only.

Sets the current user as being Invisible. This maintains the current activity.

Examples

Set the current user to being invisible on the shard:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if msg.content == "!invisible" {
            ctx.invisible();
        }
    }
}

pub fn reset_presence(&self)

Available on crate feature gateway only.

“Resets” the current user’s presence, by setting the activity to None and the online status to Online.

Use Self::set_presence for fine-grained control over individual details.

Examples

Reset the current user’s presence on the shard:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if msg.content == "!reset_presence" {
            ctx.reset_presence();
        }
    }
}

pub fn set_activity(&self, activity: Option<ActivityData>)

Available on crate feature gateway only.

Sets the current activity.

Examples

Create a command named ~setgame that accepts a name of a game to be playing:

use serenity::gateway::ActivityData;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        let mut args = msg.content.splitn(2, ' ');

        if let (Some("~setgame"), Some(game)) = (args.next(), args.next()) {
            ctx.set_activity(Some(ActivityData::playing(game)));
        }
    }
}

pub fn set_presence(&self, activity: Option<ActivityData>, status: OnlineStatus)

Available on crate feature gateway only.

Sets the current user’s presence, providing all fields to be passed.

Examples

Setting the current user as having no activity and being Idle:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn ready(&self, ctx: Context, _: Ready) {
        use serenity::model::user::OnlineStatus;

        ctx.set_presence(None, OnlineStatus::Idle);
    }
}

Setting the current user as playing "Heroes of the Storm", while being DoNotDisturb:

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn ready(&self, context: Context, _: Ready) {
        use serenity::gateway::ActivityData;
        use serenity::model::user::OnlineStatus;

        let activity = ActivityData::playing("Heroes of the Storm");
        let status = OnlineStatus::DoNotDisturb;

        context.set_presence(Some(activity), status);
    }
}