serenity

Module collector

Available on crate feature collector only.

Structs

ComponentInteractionCollector

A ComponentInteractionCollector receives ComponentInteraction’s match the given filters for a set duration.

CreateQuickModal

Convenience builder to create a modal, wait for the user to submit and parse the response.

MessageCollector

A MessageCollector receives Message’s match the given filters for a set duration.

ModalInteractionCollector

A ModalInteractionCollector receives ModalInteraction’s match the given filters for a set duration.

QuickModalResponse

ReactionCollector

A ReactionCollector receives Reaction’s match the given filters for a set duration.

Traits

CollectComponentInteractions

CollectMessages

CollectModalInteractions

CollectReactions

QuickModal

Functions

collect

Fundamental collector function. All collector types in this module are just wrappers around this function.