Struct serenity::collector::component_interaction_collector::CollectComponentInteraction

pub struct CollectComponentInteraction { /* private fields */ }

Implementations

impl CollectComponentInteraction

pub fn filter_limit(self, limit: u32) -> Self

Limits how many interactions will attempt to be filtered.

The filter checks whether the message has been sent in the right guild, channel, and by the right author.

pub fn collect_limit(self, limit: u32) -> Self

Limits how many interactions can be collected.

An interaction is considered collected, if the interaction passes all the requirements.

pub fn filter<F: Fn(&Arc<MessageComponentInteraction>) -> bool + 'static + Send + Sync>( self, function: F ) -> Self

Sets a filter function where interactions passed to the function must return true, otherwise the interaction won’t be collected. This is the last instance to pass for an interaction to count as collected.

This function is intended to be an interaction filter.

pub fn author_id(self, author_id: impl Into<u64>) -> Self

Sets the required author ID of an interaction. If an interaction is not triggered by a user with this ID, it won’t be received.

pub fn message_id(self, message_id: impl Into<u64>) -> Self

Sets the message on which the interaction must occur. If an interaction is not on a message with this ID, it won’t be received.

pub fn guild_id(self, guild_id: impl Into<u64>) -> Self

Sets the guild in which the interaction must occur. If an interaction is not on a message with this guild ID, it won’t be received.

pub fn channel_id(self, channel_id: impl Into<u64>) -> Self

Sets the channel on which the interaction must occur. If an interaction is not on a message with this channel ID, it won’t be received.

pub fn timeout(self, duration: Duration) -> Self

Sets a duration for how long the collector shall receive interactions.

impl CollectComponentInteraction

pub fn new(shard_messenger: impl AsRef<ShardMessenger>) -> Self