Struct serenity::collector::event_collector::EventCollectorBuilder

pub struct EventCollectorBuilder<'a> { /* private fields */ }

Future building a stream of events.

Implementations

impl<'a> EventCollectorBuilder<'a>

pub fn new(shard_messenger: impl AsRef<ShardMessenger>) -> Self

A future that builds an EventCollector based on the settings.

pub fn filter_limit(self, limit: u32) -> Self

Limits how many events will attempt to be filtered.

The filter checks whether the event has the right related guild, channel, user, and message. Only events with types passed to Self::add_event_type as counted towards this limit.

pub fn collect_limit(self, limit: u32) -> Self

Limits how many events can be collected.

An event is considered collected, if the event passes all the requirements.

pub fn filter<F: Fn(&Arc<Event>) -> bool + 'static + Send + Sync>(
self,
function: F
) -> Self

Sets a filter function where events passed to the function must return true, otherwise the event won’t be collected and failed the filter process. This is the last step to pass for a event to count as collected.

pub fn add_event_type(self, event_type: EventType) -> Self

Adds an EventType that this collector will collect. If an event does not have one of these types, it won’t be received.

pub fn add_user_id(self, user_id: impl Into<UserId>) -> Self

Sets the required user ID of an event. If an event does not have this ID, it won’t be received.

pub fn add_channel_id(self, channel_id: impl Into<ChannelId>) -> Self

Sets the required channel ID of an event. If an event does not have this ID, it won’t be received.

pub fn add_guild_id(self, guild_id: impl Into<GuildId>) -> Self

Sets the required guild ID of an event. If an event does not have this ID, it won’t be received.

pub fn add_message_id(self, message_id: impl Into<MessageId>) -> Self

Sets the required message ID of an event. If an event does not have this ID, it won’t be received.

pub fn timeout(self, duration: Duration) -> Self

Sets a duration for how long the collector shall receive events.