Module serenity::collector::reaction_collector

Structs

CollectReaction

ReactionCollector

A reaction collector receives reactions matching a the given filter for a set duration.

ReactionCollectorBuilder

ReactionFilter

Filters events on the shard’s end and sends them to the collector.

Enums

ReactionAction

Marks whether the reaction has been added or removed.