Struct serenity::collector::CollectorBuilder

pub struct CollectorBuilder<'a, Item: Collectable> { /* private fields */ }

Available on crate feature collector only.

Implementations

impl CollectorBuilder<'_, ComponentInteraction>

pub fn channel_id(self, channel_id: ChannelId) -> Self

Sets the channel on which the interaction must occur. If an interaction is not on a message with this channel ID, it won’t be received.

pub fn guild_id(self, guild_id: GuildId) -> Self

Sets the guild in which the interaction must occur. If an interaction is not on a message with this guild ID, it won’t be received.

pub fn message_id(self, message_id: MessageId) -> Self

Sets the message on which the interaction must occur. If an interaction is not on a message with this ID, it won’t be received.

pub fn custom_ids(self, custom_ids: Vec<String>) -> Self

Sets acceptable custom IDs for the interaction. If an interaction does not contain one of the custom IDs, it won’t be received.

pub fn author_id(self, author_id: UserId) -> Self

Sets the required author ID of an interaction. If an interaction is not triggered by a user with this ID, it won’t be received.

impl CollectorBuilder<'_, Event>

pub fn add_event_type(self, event_type: EventType) -> Self

Adds an EventType that this collector will collect. If an event does not have one of these types, it won’t be received.

pub fn add_user_id(self, user_id: impl Into<UserId>) -> Result<Self>

Sets the required user ID of an event. If an event does not have this ID, it won’t be received.

Errors

Errors if a relevant EventType has not been added.

pub fn add_channel_id(self, channel_id: impl Into<ChannelId>) -> Result<Self>

Sets the required channel ID of an event. If an event does not have this ID, it won’t be received.

Errors

Errors if a relevant EventType has not been added.

pub fn add_guild_id(self, guild_id: impl Into<GuildId>) -> Result<Self>

Sets the required guild ID of an event. If an event does not have this ID, it won’t be received.

Errors

Errors if a relevant EventType has not been added.

pub fn add_message_id(self, message_id: impl Into<MessageId>) -> Result<Self>

Sets the required message ID of an event. If an event does not have this ID, it won’t be received.

Errors

Errors if a relevant EventType has not been added.

impl CollectorBuilder<'_, Message>

pub fn channel_id(self, channel_id: ChannelId) -> Self

Sets the required channel ID of a message. If a message does not meet this ID, it won’t be received.

pub fn author_id(self, author_id: UserId) -> Self

Sets the required author ID of a message. If a message does not meet this ID, it won’t be received.

pub fn guild_id(self, guild_id: GuildId) -> Self

Sets the required guild ID of a message. If a message does not meet this ID, it won’t be received.

impl CollectorBuilder<'_, ModalInteraction>

pub fn channel_id(self, channel_id: ChannelId) -> Self

Sets the channel on which the interaction must occur. If an interaction is not on a message with this channel ID, it won’t be received.

pub fn guild_id(self, guild_id: GuildId) -> Self

Sets the guild in which the interaction must occur. If an interaction is not on a message with this guild ID, it won’t be received.

pub fn message_id(self, message_id: MessageId) -> Self

Sets the message on which the interaction must occur. If an interaction is not on a message with this ID, it won’t be received.

pub fn custom_ids(self, custom_ids: Vec<String>) -> Self

Sets acceptable custom IDs for the interaction. If an interaction does not contain one of the custom IDs, it won’t be received.

pub fn author_id(self, author_id: UserId) -> Self

Sets the required author ID of an interaction. If an interaction is not triggered by a user with this ID, it won’t be received.

impl CollectorBuilder<'_, ReactionAction>

pub fn added(self, is_accepted: bool) -> Self

If set to true, added reactions will be collected.

Set to true by default.

pub fn removed(self, is_accepted: bool) -> Self

If set to true, removed reactions will be collected.

Set to false by default.

pub fn channel_id(self, channel_id: ChannelId) -> Self

Sets the channel on which the reaction must occur. If a reaction is not on a message with this channel ID, it won’t be received.

pub fn guild_id(self, guild_id: GuildId) -> Self

Sets the guild in which the reaction must occur. If a reaction is not on a message with this guild ID, it won’t be received.

pub fn message_id(self, message_id: MessageId) -> Self

Sets the message on which the reaction must occur. If a reaction is not on a message with this ID, it won’t be received.

pub fn author_id(self, author_id: UserId) -> Self

Sets the required author ID of a reaction. If a reaction is not issued by a user with this ID, it won’t be received.

impl<'a, Item: Collectable> CollectorBuilder<'a, Item>

pub fn new(shard_messenger: &'a ShardMessenger) -> Self

pub fn build(self) -> Collector<Item>where
Filter<Item>: FilterTrait<Item>,

pub fn filter(
self,
function: Arc<dyn Fn(&Item::FilterItem) -> bool + Send + Sync + 'static>
) -> Self

Sets a filter function where items passed to the function must return true, otherwise the item won’t be collected and failed the filter process.

This is the last instance to pass for an item to count as collected.

pub fn collect_limit(self, limit: u32) -> Self

Limits how many items can be collected.

An item is considered collected, if the message passes all the requirements.

pub fn filter_limit(self, limit: u32) -> Self

Limits how many events will attempt to be filtered.

pub fn timeout(self, duration: Duration) -> Self

Sets a Duration for how long the collector shall receive events.

impl<Item: Collectable + Send + Sync + 'static> CollectorBuilder<'_, Item>where
Filter<Item>: FilterTrait<Item>,

pub async fn collect_single(self) -> Option<Arc<Item>>