Struct serenity::collector::ComponentInteractionCollector

pub struct ComponentInteractionCollector { /* private fields */ }

Available on crate feature collector only.

A ComponentInteractionCollector receives ComponentInteraction’s match the given filters for a set duration.

Implementations

impl ComponentInteractionCollector

pub fn new(shard: impl AsRef<ShardMessenger>) -> Self

Creates a new collector without any filters configured.

pub fn timeout(self, duration: Duration) -> Self

Sets a duration for how long the collector shall receive interactions.

pub fn filter(
self,
filter: impl Fn(&ComponentInteraction) -> bool + Send + Sync + 'static
) -> Self

Sets a generic filter function.

pub fn author_id(self, author_id: UserId) -> Self

Filters ComponentInteraction’s by a specific UserId.

pub fn channel_id(self, channel_id: ChannelId) -> Self

Filters ComponentInteraction’s by a specific ChannelId.

pub fn guild_id(self, guild_id: GuildId) -> Self

Filters ComponentInteraction’s by a specific GuildId.

pub fn message_id(self, message_id: MessageId) -> Self

Filters ComponentInteraction’s by a specific MessageId.

pub fn custom_ids(self, custom_ids: Vec<String>) -> Self

Filters ComponentInteraction’s by a specific Vec<String>.

pub fn collect_stream(self) -> impl Stream<Item = ComponentInteraction>

Returns a [Stream] over all collected ComponentInteraction.

pub async fn collect_single(self) -> Option<ComponentInteraction>

Returns the next ComponentInteraction which passes the filters.