serenity::framework

Module standard

👎Deprecated:
The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

Available on crate features framework and standard_framework only.

Modules

bucketsDeprecated

help_commandsDeprecated

A collection of default help commands for the framework.

macrosDeprecated

Structs

ArgsDeprecated

A utility struct for handling “arguments” of a command.

BucketBuilderDeprecated

CheckDeprecated

A check can be part of a command or group and will be executed to determine whether a user is permitted to use related item.

CommandDeprecated

CommandGroupDeprecated

CommandOptionsDeprecated

ConfigurationDeprecated

The configuration to use for a StandardFramework associated with a Client instance.

GroupOptionsDeprecated

HelpCommandDeprecated

HelpOptionsDeprecated

IterDeprecated

Parse each argument individually, as an iterator.

RawArgumentsDeprecated

Access to all of the arguments, as an iterator.

StandardFrameworkDeprecated

A utility for easily managing dispatches to commands.

WithWhiteSpaceDeprecated

A configuration struct for deciding whether the framework should allow optional whitespace between prefixes, group prefixes and command names.

Enums

ArgErrorDeprecated

Defines how an operation on an Args method failed.

DelimiterDeprecated

Dictates how Args should split arguments, if by one character, or a string.

DispatchErrorDeprecated

An enum representing all possible fail conditions under which a command won’t be executed.

HelpBehaviourDeprecated

Describes the behaviour the help-command shall execute once it encounters a command which the user or command fails to meet following criteria :

OnlyInDeprecated

ReasonDeprecated

This type describes why a check has failed.

Traits

CommonOptionsDeprecated

Type Aliases

CheckFunctionDeprecated

CommandErrorDeprecated

CommandFnDeprecated

CommandResultDeprecated

HelpCommandFnDeprecated