Struct serenity::framework::standard::buckets::BucketBuilder

pub struct BucketBuilder { /* private fields */ }

This is supported on crate features framework and standard_framework only.

Implementations

impl BucketBuilder

pub fn new_global() -> Self

A bucket collecting tickets per command invocation.

pub fn new_user() -> Self

A bucket collecting tickets per user.

pub fn new_guild() -> Self

A bucket collecting tickets per guild.

pub fn new_channel() -> Self

A bucket collecting tickets per channel.

pub fn new_category() -> Self

This is supported on crate feature cache only.

A bucket collecting tickets per channel category.

This requires the cache, as messages do not contain their channel’s category.

pub fn delay(&mut self, secs: u64) -> &mut Self

The “break” time between invocations of a command.

Expressed in seconds.

pub fn time_span(&mut self, secs: u64) -> &mut Self

How long the bucket will apply for.

Expressed in seconds.

pub fn limit(&mut self, n: u32) -> &mut Self

Number of invocations allowed per Self::time_span.

pub fn check(
&mut self,
check: for<'fut> fn(_: &'fut Context, _: &'fut Message) -> BoxFuture<'fut, bool>
) -> &mut Self

Middleware confirming (or denying) that the bucket is eligible to apply. For instance, to limit the bucket to just one user.

pub fn delay_action(
&mut self,
action: for<'fut> fn(_: &'fut Context, _: &'fut Message) -> BoxFuture<'fut, ()>
) -> &mut Self

This function will be called once a user’s invocation has been delayed.

pub fn limit_for(&mut self, target: LimitedFor) -> &mut Self

Limit the bucket for a specific type of target.

pub fn await_ratelimits(&mut self, amount: u32) -> &mut Self

If this is set to an amount greater than 0, the invocation of the command will be delayed amount times instead of stopping command dispatch.

By default this value is 0 and rate limits will cancel instead.