Struct serenity::framework::standard::buckets::RateLimitInfo

pub struct RateLimitInfo {
    pub rate_limit: Duration,
    pub active_delays: u32,
    pub max_delays: u32,
    pub is_first_try: bool,
    pub action: RateLimitAction,
}

Available on crate features framework and standard_framework only.

Contains information about a rate limit.

Fields
rate_limit: Duration

Time to elapse in order to invoke a command again.

active_delays: u32

Amount of active delays by this target.

max_delays: u32

Maximum delays that this target can invoke.

is_first_try: bool

Whether this is the first time the rate limit info has been returned for this target without the rate limit to elapse.

action: RateLimitAction

How the command invocation has been treated by the framework.

Implementations

impl RateLimitInfo

pub fn as_secs(&self) -> u64

Gets the duration of the rate limit in seconds.

pub fn as_millis(&self) -> u128

Gets the duration of the rate limit in milliseconds.

pub fn as_micros(&self) -> u128

Gets the duration of the rate limit in microseconds.