Enum serenity::framework::standard::DispatchError

pub enum DispatchError {
    CheckFailed(Arc<Command>),
    CommandDisabled(String),
    BlockedUser,
    BlockedGuild,
    LackOfPermissions(Permissions),
    RateLimited(i64),
    OnlyForDM,
    OnlyForGuilds,
    OnlyForOwners,
    LackingRole,
    NotEnoughArguments {
        min: i32,
        given: usize,
    },
    TooManyArguments {
        max: i32,
        given: usize,
    },
    IgnoredBot,
    WebhookAuthor,
}

A enum representing all possible fail conditions under which a command won't be executed.

Variants
CheckFailed(Arc<Command>)

When a custom function check has failed.

CommandDisabled(String)

When the requested command is disabled in bot configuration.

BlockedUser

When the user is blocked in bot configuration.

BlockedGuild

When the guild or its owner is blocked in bot configuration.

LackOfPermissions(Permissions)

When the command requester lacks specific required permissions.

RateLimited(i64)

When the command requester has exceeded a ratelimit bucket. The attached value is the time a requester has to wait to run the command again.

OnlyForDM

When the requested command can only be used in a direct message or group channel.

OnlyForGuilds

When the requested command can only be ran in guilds, or the bot doesn't support DMs.

OnlyForOwners

When the requested command can only be used by bot owners.

LackingRole

When the requested command requires one role.

NotEnoughArguments

When there are too few arguments.

Fields of NotEnoughArguments

min: i32

given: usize

TooManyArguments

When there are too many arguments.

Fields of TooManyArguments

max: i32

given: usize

IgnoredBot

When the command was requested by a bot user when they are set to be ignored.

WebhookAuthor

When the bot ignores webhooks and a command was issued by one.