serenity::framework::standard::help_commands
Function create_customised_help_data

pub async fn create_customised_help_data<'a>(
    ctx: &Context,
    msg: &Message,
    args: &'a Args,
    groups: &[&'static CommandGroup],
    owners: &HashSet<UserId, impl BuildHasher + Send + Sync>,
    help_options: &'a HelpOptions,
) -> CustomisedHelpData<'a>

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

Available on crate features framework and standard_framework and cache only.

Iterates over all commands and forges them into a CustomisedHelpData, taking HelpOptions into consideration when deciding on whether a command shall be picked and in what textual format.