serenity::framework::standard::help_commands

Function has_all_requirements

pub fn has_all_requirements(
    cache: impl AsRef<Cache>,
    cmd: &CommandOptions,
    msg: &Message,
) -> bool

👎Deprecated:
The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

Available on crate features framework and standard_framework and cache only.

Checks whether a user is member of required roles and given the required permissions.