Function serenity::framework::standard::help_commands::plain

pub async fn plain(
    ctx: &Context, 
    msg: &Message, 
    args: Args, 
    help_options: &HelpOptions, 
    groups: &[&'static CommandGroup], 
    owners: HashSet<UserId>
) -> Result<Message, Error>

This is supported on crate features framework and standard_framework and cache and http only.

Posts formatted text displaying each individual command group and its commands.

Examples

Use the command with exec_help:

use std::collections::HashSet;
use std::hash::BuildHasher;

use serenity::framework::standard::help_commands::*;
use serenity::framework::standard::macros::help;
use serenity::framework::standard::{
    Args,
    CommandGroup,
    CommandResult,
    HelpOptions,
    StandardFramework,
};
use serenity::model::prelude::*;

#[help]
async fn my_help(
    context: &Context,
    msg: &Message,
    args: Args,
    help_options: &'static HelpOptions,
    groups: &[&'static CommandGroup],
    owners: HashSet<UserId>,
) -> CommandResult {
    let _ = plain(context, msg, args, &help_options, groups, owners).await?;
    Ok(())
}

let framework = StandardFramework::new().help(&MY_HELP);

Errors

Returns the same errors as ChannelId::send_message.