Struct serenity::framework::standard::Check

pub struct Check {
    pub name: &'static str,
    pub function: CheckFunction,
    pub check_in_help: bool,
    pub display_in_help: bool,
}

Available on crate features framework and standard_framework only.

A check can be part of a command or group and will be executed to determine whether a user is permitted to use related item.

Additionally, a check may hold additional settings.

Fields
name: &'static str

Name listed in help-system.

function: CheckFunction

Function that will be executed.

check_in_help: bool

Whether a check should be evaluated in the help-system. false will ignore check and won’t fail execution.

display_in_help: bool

Whether a check shall be listed in the help-system. false won’t affect whether the check will be evaluated help, solely Self::check_in_help sets this.