serenity::framework::standard
Trait CommonOptions

pub trait CommonOptions {
    // Required methods
    fn required_permissions(&self) -> &Permissions;
    fn allowed_roles(&self) -> &'static [&'static str];
    fn checks(&self) -> &'static [&'static Check];
    fn only_in(&self) -> OnlyIn;
    fn help_available(&self) -> bool;
    fn owners_only(&self) -> bool;
    fn owner_privilege(&self) -> bool;
}

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

Available on crate features framework and standard_framework only.

Required Methods

fn required_permissions(&self) -> &Permissions

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

fn allowed_roles(&self) -> &'static [&'static str]

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

fn checks(&self) -> &'static [&'static Check]

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

fn only_in(&self) -> OnlyIn

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

fn help_available(&self) -> bool

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

fn owners_only(&self) -> bool

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

fn owner_privilege(&self) -> bool

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling