serenity::gateway

Module client

Available on crate feature gateway only.

A module for Client and supporting types.

The Client contains information about a bot’s token, as well as event handlers. Dispatching events to handlers and starting sharded gateway connections is handled directly by the client. In addition, the client automatically handles caching via the Cache struct.

Sharding

If you do not require sharding - such as for a small bot - then use Client::start. If you don’t know what sharding is, refer to the sharding module documentation.

There are a few methods of sharding available:

Client::start_autosharded: retrieves the number of shards Discord recommends using from the API, and then automatically starts that number of shards.

Client::start_shard: starts a single shard for use in the instance, handled by the instance of the Client. Use this if you only want 1 shard handled by this instance.

Client::start_shards: starts all shards in this instance. This is best for when you want a completely shared State.

Client::start_shard_range: start a range of shards within this instance. This should be used when you, for example, want to split 10 shards across 3 instances.

Click here for an example on how to use a Client.

Structs

Client

A high-level client that abstracts over the REST API as well as Discord’s gateway.

ClientBuilder

A builder implementing IntoFuture building a Client to interact with Discord.

Context

A general utility struct provided on event dispatches.

Traits

EventHandler

RawEventHandler

An event handler that receives raw dispatch events.