serenity::gateway::client
Struct ClientBuilder

pub struct ClientBuilder { /* private fields */ }

Available on crate feature gateway only.

A builder implementing IntoFuture building a Client to interact with Discord.

Implementations

impl ClientBuilder

pub fn new(token: &str, intents: GatewayIntents) -> Self

Construct a new builder to call methods on for the client construction. The token will automatically be prefixed “Bot “ if not already.

Panic: If you have enabled the framework-feature (on by default), you must specify a framework via the Self::framework method, otherwise awaiting the builder will cause a panic.

pub fn new_with_http(http: Arc<Http>, intents: GatewayIntents) -> Self

Construct a new builder with a Http instance to calls methods on for the client construction.

Panic: If you have enabled the framework-feature (on by default), you must specify a framework via the Self::framework method, otherwise awaiting the builder will cause a panic.

pub fn data<D: Any + Send + Sync>(self, data: Arc<D>) -> Self

Sets the global data type that can be accessed from Context::data.

pub fn cache_settings(self, settings: CacheSettings) -> Self

Available on crate feature cache only.

Sets the settings of the cache. Refer to Settings for more information.

pub fn get_cache_settings(&self) -> &CacheSettings

Available on crate feature cache only.

Gets the cache settings. See Self::cache_settings for more info.

pub fn framework<F>(self, framework: F) -> Self
where F: Framework + 'static,

Available on crate feature framework only.

Sets the command framework to be used. It will receive messages sent over the gateway and then consider - based on its settings - whether to dispatch a command.

Info: If a reference to the framework is required for manual dispatch, you can implement Framework on Arc<YourFrameworkType> instead of YourFrameworkType.

pub fn get_framework(&self) -> Option<&dyn Framework>

Available on crate feature framework only.

Gets the framework, if already initialized. See Self::framework for more info.

pub fn voice_manager<V>(self, voice_manager: impl Into<Arc<V>>) -> Self
where V: VoiceGatewayManager + 'static,

Available on crate feature voice only.

Sets the voice gateway handler to be used. It will receive voice events sent over the gateway and then consider - based on its settings - whether to dispatch a command.

pub fn get_voice_manager(&self) -> Option<Arc<dyn VoiceGatewayManager>>

Available on crate feature voice only.

Gets the voice manager, if already initialized. See Self::voice_manager for more info.

pub fn intents(self, intents: GatewayIntents) -> Self

Sets all intents directly, replacing already set intents. Intents are a bitflag, you can combine them by performing the |-operator.

What are Intents

A gateway intent sets the types of gateway events (e.g. member joins, guild integrations, guild emoji updates, …) the bot shall receive. Carefully picking the needed intents greatly helps the bot to scale, as less intents will result in less events to be received hence less processed by the bot.

Privileged Intents

The intents GatewayIntents::GUILD_PRESENCES, GatewayIntents::GUILD_MEMBERS and GatewayIntents::MESSAGE_CONTENT are privileged. Privileged intents need to be enabled in the developer portal. Once the bot is in 100 guilds or more, the bot must be verified in order to use privileged intents.

pub fn get_intents(&self) -> GatewayIntents

Gets the intents. See Self::intents for more info.

pub fn event_handler<H>(self, event_handler: impl Into<Arc<H>>) -> Self
where H: EventHandler + 'static,

Adds an event handler with multiple methods for each possible event.

pub fn get_event_handler(&self) -> Option<&Arc<dyn EventHandler>>

Gets the added event handlers. See Self::event_handler for more info.

pub fn raw_event_handler<H>(self, raw_event_handler: impl Into<Arc<H>>) -> Self
where H: RawEventHandler + 'static,

Adds an event handler with a single method where all received gateway events will be dispatched.

pub fn get_raw_event_handler(&self) -> Option<&Arc<dyn RawEventHandler>>

Gets the added raw event handlers. See Self::raw_event_handler for more info.

pub fn activity(self, activity: ActivityData) -> Self

Sets the initial activity.

pub fn status(self, status: OnlineStatus) -> Self

Sets the initial status.

pub fn get_presence(&self) -> &PresenceData

Gets the initial presence. See Self::activity and Self::status for more info.