serenity::gateway::client

Trait EventHandler

pub trait EventHandler: Send + Sync {
    // Provided methods
    fn filter_event(
        &self,
        _context: &Context,
        event: Box<Event>,
    ) -> Option<Box<Event>> { ... }
    fn dispatch<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        _context: &'life1 Context,
        _event: &'life2 FullEvent,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn ratelimit<'life0, 'async_trait>(
        &'life0 self,
        _data: RatelimitInfo,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

Available on crate feature gateway only.

Provided Methods

fn filter_event( &self, _context: &Context, event: Box<Event>, ) -> Option<Box<Event>>

Checks if the event should be dispatched or ignored. Returns Some(event) by default.

Returning None will drop the event and never process it.

Returning a mutated Some is a supported use case, but may cause contradictions to documented behaviour or otherwise cause unexpected errors within serenity/library consumers. Mutate with extreme caution.

Warning

Similar to RawEventHandler, this method runs synchronously to the ShardRunner, keep runtime complexity low.

fn dispatch<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, _context: &'life1 Context, _event: &'life2 FullEvent, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Dispatches an event through this handler, allowing for event matching and handling based on individual event variants.

fn ratelimit<'life0, 'async_trait>( &'life0 self, _data: RatelimitInfo, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Dispatched when an HTTP rate limit is hit.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".