serenity::gateway

Module sharding

Available on crate feature gateway only.

Mechanisms for configuring and managing sharded gateway connections.

Sharding is a method for load-balancing bots across separate threads or processes. Sharding is enforced on bots by Discord once they reach a certain number of guilds (2500). Once this threshold is reached, a but must be sharded such that at most 2500 guilds are allocated per shard.

The “recommended” number of guilds per shard is around 1000. Sharding allows for bots to be distributed by handing shards off to separate processes or even separate machines in a distributed network (e.g. cloud workers). However, sometimes you may wish for all shards to share some global state. Serenity accomodates both of these usecases.

See Discord’s documentation for more information.

This module also provides some lower-level facilities for performing sharding manually:

ShardManager

The shard manager provides an interface for managing the starting, restarting, and shutdown of shards by spawning tasks, each containing a ShardRunner, and communicating with each task using message passing.

ShardRunner

The shard runner is responsible for directly running a single shard and communicating with the gateway through its respective WebSocket client. It performs actions such as identifying, reconnecting, resuming, and sending presence updates to the gateway.

Structs

CollectorCallbackcollector

Newtype around a callback that will be called on every incoming request. As long as this collector should still receive events, it should return true. Once it returns false, it is removed.

Shard

An abstract handler for a websocket connection to Discord’s gateway.

ShardManager

A manager for handling the status of shards by starting them, restarting them, and stopping them when required.

ShardManagerOptions

ShardQueue

A queue of ShardIds that is split up into multiple buckets according to the value of max_concurrency.

ShardRunner

A runner for managing a Shard and its respective WebSocket client.

ShardRunnerInfo

Information about a ShardRunner.

ShardRunnerOptions

Options to be passed to ShardRunner::new.

ShardStageUpdateEvent

An event denoting that a shard’s connection stage was changed.

Enums

ConnectionStage

Indicates the current connection stage of a Shard.

ShardAction

ShardManagerMessage

A message indicating what action the ShardManager should take.

ShardRunnerMessage

A message to send from a shard over a WebSocket.

TransportCompression

The transport compression method to use.

Constants

DEFAULT_WAIT_BETWEEN_SHARD_START

The default time to wait between starting each shard or set of shards.