serenity::gateway::sharding

Struct Shard

pub struct Shard {
    pub client: WsClient,
    pub intents: GatewayIntents,
    /* private fields */
}

Available on crate feature gateway only.

An abstract handler for a websocket connection to Discord’s gateway.

Allows a user to send and receive messages over said websocket, including:

setting the current activity

setting the current online status

receiving gateway events

connection management via heartbeating

Shard management (including starting, restarting, heartbeating), is performed by the Client automatically on the user’s behalf.

Stand-alone shards

You may instantiate a shard yourself - decoupled from the Client - by calling Shard::new. Most use cases will not necessitate this, and unless you’re doing something really weird you can just let the client do it for you.

Note: You really do not need to do this, especially if using multiple shards. Just call one of the appropriate methods on the Client.

Examples

See the documentation for Self::new on how to use this.

Fields
client: WsClientintents: GatewayIntents
Implementations

impl Shard

pub async fn new( ws_url: Arc<str>, token: Token, info: ShardInfo, intents: GatewayIntents, presence: Option<PresenceData>, compression: TransportCompression, ) -> Result<Shard>

Instantiates a new instance of a Shard, bypassing the client.

Note: You should likely never need to do this yourself.

Examples

Instantiating a new Shard manually for a bot with no shards, and then listening for events:

use std::num::NonZeroU16;
use std::sync::Arc;

use serenity::gateway::{Shard, TransportCompression};
use serenity::model::gateway::{GatewayIntents, ShardInfo};
use serenity::model::id::ShardId;
use serenity::secrets::Token;
use tokio::sync::Mutex;
let token = Token::from_env("DISCORD_TOKEN")?;
let shard_info = ShardInfo {
    id: ShardId(0),
    total: NonZeroU16::MIN,
};

// retrieve the gateway response, which contains the URL to connect to
let gateway = Arc::from(http.get_gateway().await?.url);
let shard = Shard::new(
    gateway,
    token,
    shard_info,
    GatewayIntents::all(),
    None,
    TransportCompression::None,
)
.await?;

// at this point, you can create a `loop`, and receive events and match
// their variants

Errors

On Error, will return either Error::Gateway, Error::Tungstenite or a Rustls/native TLS error.

pub fn set_application_id_callback( &mut self, callback: impl FnOnce(ApplicationId) + Send + Sync + 'static, )

Sets a callback to be called when the gateway receives the application’s ID from Discord.

Used internally by serenity to set the Http’s internal application ID automatically.

pub fn presence(&self) -> &PresenceData

Retrieves the current presence of the shard.

pub async fn heartbeat(&mut self) -> Result<()>

Sends a heartbeat to the gateway with the current sequence.

This sets the last heartbeat time to now.

Errors

Returns GatewayError::HeartbeatFailed if there was an error sending a heartbeat.

pub fn seq(&self) -> u64

pub fn session_id(&self) -> Option<&str>

pub fn set_activity(&mut self, activity: Option<ActivityData>)

pub fn set_status(&mut self, status: OnlineStatus)

pub fn shard_info(&self) -> ShardInfo

Retrieves a copy of the current shard information.

For example, if using 3 shards in total, and if this is shard 1, then it can be read as “the second of three shards”.

pub fn stage(&self) -> ConnectionStage

Returns the current connection stage of the shard.

pub fn handle_event( &mut self, event: Result<GatewayEvent>, ) -> Result<Option<ShardAction>>

Handles an event from the gateway over the receiver, requiring the receiver to be passed if a reconnect needs to occur.

The best case scenario is that one of two values is returned:

Ok(None): a heartbeat, late hello, or session invalidation was received;

Ok(Some((event, None))): an op0 dispatch was received, and the shard’s voice state will be updated, if the voice feature is enabled.

Errors

Returns a GatewayError::InvalidAuthentication if invalid authentication was sent in the IDENTIFY.

Returns a GatewayError::InvalidShardData if invalid shard data was sent in the IDENTIFY.

Returns a GatewayError::NoAuthentication if no authentication was sent in the IDENTIFY.

Returns a GatewayError::OverloadedShard if the shard would have too many guilds assigned to it.

pub async fn do_heartbeat(&mut self) -> bool

Does a heartbeat if needed. Returns false if something went wrong and the shard should be restarted.

true is returned under one of the following conditions:

the heartbeat interval has not elapsed

a heartbeat was successfully sent

there is no known heartbeat interval yet

false is returned under one of the following conditions:

a heartbeat acknowledgement was not received in time

an error occurred while heartbeating

pub fn heartbeat_latency(&self) -> Option<Duration>

Returns the heartbeat latency between the shard and the gateway.

pub async fn chunk_guild( &mut self, guild_id: GuildId, limit: Option<u16>, presences: bool, filter: ChunkGuildFilter, nonce: Option<&str>, ) -> Result<()>

Requests that one or multiple Guilds be chunked.

This will ask the gateway to start sending member chunks for large guilds. If a guild is large enough, then a full member list will not be provided upon connection, and must instead be requested directly. The full list will be sent in “chunks” until all members matching the request have been sent.

Member chunks are sent as the Event::GuildMembersChunk event. Each chunk only contains a partial amount of the total members.

Errors

Errors if there is a problem with the WS connection.

pub async fn request_soundboard_sounds( &mut self, guild_ids: &[GuildId], ) -> Result<()>

Requests Soundboard sounds to be fetched from one or multiple Guilds.

This will ask the gateway to start sending soundboard sounds.

Soundboard sounds are sent as the Event::SoundboardSounds event.

Errors

Errors if there is a problem with the WS connection.

pub async fn request_channel_info( &mut self, guild_id: GuildId, fields: &[&str], ) -> Result<()>

Requests ephemeral channel data for channels in a Guild.

Errors

Errors if there is a problem with the WS connection.

pub async fn update_voice_state( &mut self, guild_id: GuildId, channel_id: Option<ChannelId>, self_mute: bool, self_deaf: bool, ) -> Result<()>

Available on crate feature voice only.

Indicates to the gateway that the client wants to join, move, or disconnect from a voice channel.

Errors

Errors if there is a problem with the WS connection.

pub async fn identify(&mut self) -> Result<()>

Sets the shard as going into identifying stage, which sets:

the time that the last heartbeat sent as being now

the stage to ConnectionStage::Identifying

Errors

Errors if there is a problem with the WS connection.

pub async fn reinitialize(&mut self) -> Result<WsClient>

Reinitializes an existing WebSocket client, replacing it.

This will set the stage of the shard before and after instantiation of the client.

Errors

Errors if unable to establish a websocket connection.

pub async fn resume(&mut self) -> Result<()>

Errors

Errors if unable to re-establish a websocket connection.

pub async fn update_presence(&mut self) -> Result<()>

Errors

Errors if there is a problem with the WS connection.