serenity::gateway::sharding

Struct ShardManager

pub struct ShardManager {

    pub data: Arc<dyn Any + Send + Sync>,
    pub event_handler: Option<Arc<dyn EventHandler>>,
    pub raw_event_handler: Option<Arc<dyn RawEventHandler>>,
    pub framework: Arc<OnceLock<Arc<dyn Framework>>>,
    pub last_start: Option<Instant>,
    pub queue: ShardQueue,
    pub runners: Arc<DashMap<ShardId, ShardRunnerMetadata>>,
    pub voice_manager: Option<Arc<dyn VoiceGatewayManager + 'static>>,
    pub ws_url: Arc<str>,
    pub compression: TransportCompression,
    pub shard_total: NonZeroU16,
    pub wait_time_between_shard_start: Duration,
    pub cache: Arc<Cache>,
    pub http: Arc<Http>,
    pub intents: GatewayIntents,
    pub presence: Option<PresenceData>,
    /* private fields */

}

Available on crate feature gateway only.

A manager for handling the status of shards by starting them, restarting them, and stopping them when required.

Fields
data: Arc<dyn Any + Send + Sync>

A copy of Client::data to be given to runners for contextual dispatching.

event_handler: Option<Arc<dyn EventHandler>>

A reference to an EventHandler.

raw_event_handler: Option<Arc<dyn RawEventHandler>>

A reference to a RawEventHandler.

framework: Arc<OnceLock<Arc<dyn Framework>>>
Available on crate feature framework only.

A copy of the framework.

last_start: Option<Instant>

The instant that a shard was last started.

This is used to determine how long to wait between shard IDENTIFYs.

queue: ShardQueue

The shards that are queued for booting.

runners: Arc<DashMap<ShardId, ShardRunnerMetadata>>

The shard runners currently managed.

Note: It is highly recommended to not mutate this yourself unless you need to. Instead prefer to use methods on this struct that are provided where possible.

voice_manager: Option<Arc<dyn VoiceGatewayManager + 'static>>
Available on crate feature voice only.

A copy of the client’s voice manager.

ws_url: Arc<str>

A copy of the URL to use to connect to the gateway.

compression: TransportCompression

The compression method to use for the WebSocket connection.

shard_total: NonZeroU16

The total amount of shards to start.

wait_time_between_shard_start: Duration

Number of seconds to wait between each start.

cache: Arc<Cache>
Available on crate feature cache only.
http: Arc<Http>intents: GatewayIntentspresence: Option<PresenceData>
Implementations

impl ShardManager

pub fn new(opt: ShardManagerOptions) -> Self

pub fn get_shutdown_trigger(&self) -> impl FnOnce() -> bool + Send + use<>

Retrieves a function which can be used to shut down the ShardManager later.

This function will return true if the ShardManager has successfully been notified to shut down, or false if it has already shut down and been dropped.

pub async fn run( &mut self, shard_index: u16, shard_init: u16, shard_total: NonZeroU16, ) -> Result<(), GatewayError>

The main interface for starting the management of shards. Initializes the shards by queueing them for starting, and then listens for ShardManagerMessages in a loop.

This function takes care of:

Starting ShardRunners and spawning tasks.

Restarting shards (by shutting them down and spawning a new task) when requested.

Quitting if a shard runner encounters a fatal gateway error.

Errors

Returns a GatewayError if an unrecoverable error is received from the gateway. These include the following:

GatewayError::InvalidAuthentication

GatewayError::InvalidApiVersion

GatewayError::InvalidGatewayIntents

GatewayError::DisallowedGatewayIntents

pub fn intents(&self) -> GatewayIntents

Returns the gateway intents used for this gateway connection.