Module serenity::http::client

Available on crate feature http only.

Structs

Http

Note: For all member functions that return a Result, the Error kind will be either Error::Http or Error::Json.

HttpBuilder

A builder for the underlying Http client that performs requests to Discord’s HTTP API. If you do not need to use a proxy or do not need to disable the rate limiter, you can use Http::new or Http::new_with_application_id instead.