Enum serenity::http::error::Error

#[non_exhaustive]
pub enum Error {
    UnsuccessfulRequest(ErrorResponse),
    RateLimitI64F64,
    RateLimitUtf8,
    Url(UrlError),
    InvalidHeader(InvalidHeaderValue),
    Request(ReqwestError),
    InvalidScheme,
    InvalidPort,
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UnsuccessfulRequest(ErrorResponse)

When a non-successful status code was received for a request.

RateLimitI64F64

When the decoding of a ratelimit header could not be properly decoded into an i64 or f64.

RateLimitUtf8

When the decoding of a ratelimit header could not be properly decoded from UTF-8.

Url(UrlError)

When parsing an URL failed due to invalid input.

InvalidHeader(InvalidHeaderValue)

Header value contains invalid input.

Request(ReqwestError)

Reqwest’s Error contain information on why sending a request failed.

InvalidScheme

When using a proxy with an invalid scheme.

InvalidPort

When using a proxy with an invalid port.

Implementations

impl Error

pub async fn from_response(r: Response) -> Self

pub fn is_unsuccessful_request(&self) -> bool

Returns true when the error is caused by an unsuccessful request

pub fn is_url_error(&self) -> bool

Returns true when the error is caused by the url containing invalid input

pub fn is_invalid_header(&self) -> bool

Returns true when the error is caused by an invalid header

pub fn status_code(&self) -> Option<StatusCode>

Returns the status code if the error is an unsuccessful request