Function serenity::http::create_guild

pub fn create_guild(map: &Value) -> Result<PartialGuild>

Creates a guild with the data provided.

Only a PartialGuild will be immediately returned, and a full Guild will be received over a Shard, if at least one is running.

Note: This endpoint is currently limited to 10 active guilds. The limits are raised for whitelisted GameBridge applications. See the documentation on this endpoint for more info.

Examples

Create a guild called "test" in the US West region:

extern crate serde_json;

use serde_json::builder::ObjectBuilder;
use serde_json::Value;
use serenity::http;

let map = ObjectBuilder::new()
    .insert("name", "test")
    .insert("region", "us-west")
    .build();

let _result = http::create_guild(map);