Enum serenity::http::ratelimiting::Route

pub enum Route {
    ChannelsId(u64),
    ChannelsIdInvites(u64),
    ChannelsIdMessages(u64),
    ChannelsIdMessagesBulkDelete(u64),
    ChannelsIdMessagesId(LightMethod, u64),
    ChannelsIdMessagesIdAck(u64),
    ChannelsIdMessagesIdReactions(u64),
    ChannelsIdMessagesIdReactionsUserIdType(u64),
    ChannelsIdPermissionsOverwriteId(u64),
    ChannelsIdPins(u64),
    ChannelsIdPinsMessageId(u64),
    ChannelsIdTyping(u64),
    ChannelsIdWebhooks(u64),
    Gateway,
    GatewayBot,
    Guilds,
    GuildsId(u64),
    GuildsIdBans(u64),
    GuildsIdAuditLogs(u64),
    GuildsIdBansUserId(u64),
    GuildsIdChannels(u64),
    GuildsIdEmbed(u64),
    GuildsIdEmojis(u64),
    GuildsIdEmojisId(u64),
    GuildsIdIntegrations(u64),
    GuildsIdIntegrationsId(u64),
    GuildsIdIntegrationsIdSync(u64),
    GuildsIdInvites(u64),
    GuildsIdMembers(u64),
    GuildsIdMembersId(u64),
    GuildsIdMembersIdRolesId(u64),
    GuildsIdMembersMeNick(u64),
    GuildsIdPrune(u64),
    GuildsIdRegions(u64),
    GuildsIdRoles(u64),
    GuildsIdRolesId(u64),
    GuildsIdWebhooks(u64),
    InvitesCode,
    UsersId,
    UsersMe,
    UsersMeChannels,
    UsersMeGuilds,
    UsersMeGuildsId,
    VoiceRegions,
    WebhooksId,
    None,
}

A representation of all routes registered within the library. These are safe and memory-efficient representations of each path that functions exist for in the http module.

Variants
ChannelsId(u64)

Route for the /channels/:channel_id path.

The data is the relevant ChannelId.

ChannelsIdInvites(u64)

Route for the /channels/:channel_id/invites path.

The data is the relevant ChannelId.

ChannelsIdMessages(u64)

Route for the /channels/:channel_id/messages path.

The data is the relevant ChannelId.

ChannelsIdMessagesBulkDelete(u64)

Route for the /channels/:channel_id/messages/bulk-delete path.

The data is the relevant ChannelId.

ChannelsIdMessagesId(LightMethod, u64)

Route for the /channels/:channel_id/messages/:message_id path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdAck(u64)

Route for the /channels/:channel_id/messages/:message_id/ack path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdReactions(u64)

Route for the /channels/:channel_id/messages/:message_id/reactions path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdReactionsUserIdType(u64)

Route for the /channels/:channel_id/messages/:message_id/reactions/:reaction/@me path.

The data is the relevant ChannelId.

ChannelsIdPermissionsOverwriteId(u64)

Route for the /channels/:channel_id/permissions/:target_id path.

The data is the relevant ChannelId.

ChannelsIdPins(u64)

Route for the /channels/:channel_id/pins path.

The data is the relevant ChannelId.

ChannelsIdPinsMessageId(u64)

Route for the /channels/:channel_id/pins/:message_id path.

The data is the relevant ChannelId.

ChannelsIdTyping(u64)

Route for the /channels/:channel_id/typing path.

The data is the relevant ChannelId.

ChannelsIdWebhooks(u64)

Route for the /channels/:channel_id/webhooks path.

The data is the relevant ChannelId.

Gateway

Route for the /gateway path.

GatewayBot

Route for the /gateway/bot path.

Guilds

Route for the /guilds path.

GuildsId(u64)

Route for the /guilds/:guild_id path.

The data is the relevant GuildId.

GuildsIdBans(u64)

Route for the /guilds/:guild_id/bans path.

The data is the relevant GuildId.

GuildsIdAuditLogs(u64)

Route for the /guilds/:guild_id/audit-logs path. The data is the relevant GuildId.

GuildsIdBansUserId(u64)

Route for the /guilds/:guild_id/bans/:user_id path.

The data is the relevant GuildId.

GuildsIdChannels(u64)

Route for the /guilds/:guild_id/channels/:channel_id path.

The data is the relevant GuildId.

GuildsIdEmbed(u64)

Route for the /guilds/:guild_id/embed path.

The data is the relevant GuildId.

GuildsIdEmojis(u64)

Route for the /guilds/:guild_id/emojis path.

The data is the relevant GuildId.

GuildsIdEmojisId(u64)

Route for the /guilds/:guild_id/emojis/:emoji_id path.

The data is the relevant GuildId.

GuildsIdIntegrations(u64)

Route for the /guilds/:guild_id/integrations path.

The data is the relevant GuildId.

GuildsIdIntegrationsId(u64)

Route for the /guilds/:guild_id/integrations/:integration_id path.

The data is the relevant GuildId.

GuildsIdIntegrationsIdSync(u64)

Route for the /guilds/:guild_id/integrations/:integration_id/sync path.

The data is the relevant GuildId.

GuildsIdInvites(u64)

Route for the /guilds/:guild_id/invites path.

The data is the relevant GuildId.

GuildsIdMembers(u64)

Route for the /guilds/:guild_id/members path.

The data is the relevant GuildId.

GuildsIdMembersId(u64)

Route for the /guilds/:guild_id/members/:user_id path.

The data is the relevant GuildId.

GuildsIdMembersIdRolesId(u64)

Route for the /guilds/:guild_id/members/:user_id/roles/:role_id path.

The data is the relevant GuildId.

GuildsIdMembersMeNick(u64)

Route for the /guilds/:guild_id/members/@me/nick path.

The data is the relevant GuildId.

GuildsIdPrune(u64)

Route for the /guilds/:guild_id/prune path.

The data is the relevant GuildId.

GuildsIdRegions(u64)

Route for the /guilds/:guild_id/regions path.

The data is the relevant GuildId.

GuildsIdRoles(u64)

Route for the /guilds/:guild_id/roles path.

The data is the relevant GuildId.

GuildsIdRolesId(u64)

Route for the /guilds/:guild_id/roles/:role_id path.

The data is the relevant GuildId.

GuildsIdWebhooks(u64)

Route for the /guilds/:guild_id/webhooks path.

The data is the relevant GuildId.

InvitesCode

Route for the /invites/:code path.

UsersId

Route for the /users/:user_id path.

UsersMe

Route for the /users/@me path.

UsersMeChannels

Route for the /users/@me/channels path.

UsersMeGuilds

Route for the /users/@me/guilds path.

UsersMeGuildsId

Route for the /users/@me/guilds/:guild_id path.

VoiceRegions

Route for the /voice/regions path.

WebhooksId

Route for the /webhooks/:webhook_id path.

None

Route where no ratelimit headers are in place (i.e. user account-only routes).

This is a special case, in that if the route is None then pre- and post-hooks are not executed.