Module serenity::http::routing

Available on crate feature http only.

Enums

Route

A representation of all routes registered within the library. These are safe and memory-efficient representations of each path that functions exist for in the http module.

RouteInfo