Enum serenity::http::routing::Route

#[non_exhaustive]
pub enum Route {

    ChannelsId(ChannelId),
    ChannelsIdInvites(ChannelId),
    ChannelsIdMessages(ChannelId),
    ChannelsIdMessagesBulkDelete(ChannelId),
    ChannelsIdMessagesId(LightMethod, ChannelId),
    ChannelsIdMessagesIdAck(ChannelId),
    ChannelsIdMessagesIdReactions(ChannelId),
    ChannelsIdMessagesIdReactionsUserIdType(ChannelId),
    ChannelsIdPermissionsOverwriteId(ChannelId),
    ChannelsIdPins(ChannelId),
    ChannelsIdPinsMessageId(ChannelId),
    ChannelsIdCrosspostsMessageId(ChannelId),
    ChannelsIdTyping(ChannelId),
    ChannelsIdWebhooks(ChannelId),
    ChannelsIdMessagesIdThreads(ChannelId),
    ChannelsIdThreads(ChannelId),
    ChannelsIdThreadMembersMe(ChannelId),
    ChannelsIdThreadMembersUserId(ChannelId),
    ChannelsIdThreadMembers(ChannelId),
    ChannelsIdArchivedPublicThreads(ChannelId),
    ChannelsIdArchivedPrivateThreads(ChannelId),
    ChannelsIdMeJoindedArchivedPrivateThreads(ChannelId),
    FollowNewsChannel(ChannelId),
    Gateway,
    GatewayBot,
    Guilds,
    GuildsId(GuildId),
    GuildsIdAutoModRules(GuildId),
    GuildsIdAutoModRulesId(GuildId),
    GuildsIdBans(GuildId),
    GuildsIdAuditLogs(GuildId),
    GuildsIdBansUserId(GuildId),
    GuildsIdChannels(GuildId),
    GuildsIdWidget(GuildId),
    GuildsIdPreview(GuildId),
    GuildsIdEmojis(GuildId),
    GuildsIdEmojisId(GuildId),
    GuildsIdIntegrations(GuildId),
    GuildsIdIntegrationsId(GuildId),
    GuildsIdIntegrationsIdSync(GuildId),
    GuildsIdInvites(GuildId),
    GuildsIdMembers(GuildId),
    GuildsIdMembersId(GuildId),
    GuildsIdMembersIdRolesId(GuildId),
    GuildsIdMembersMe(GuildId),
    GuildsIdMembersMeNick(GuildId),
    GuildsIdMembersSearch(GuildId),
    GuildsIdPrune(GuildId),
    GuildsIdRegions(GuildId),
    GuildsIdRoles(GuildId),
    GuildsIdRolesId(GuildId),
    GuildsIdScheduledEvents(GuildId),
    GuildsIdScheduledEventsId(GuildId),
    GuildsIdScheduledEventsIdUsers(GuildId),
    GuildsIdStickers(GuildId),
    GuildsIdStickersId(GuildId),
    GuildsIdVanityUrl(GuildId),
    GuildsIdVoiceStates(GuildId),
    GuildsIdVoiceStatesMe(GuildId),
    GuildsIdWebhooks(GuildId),
    GuildsIdWelcomeScreen(GuildId),
    GuildsIdThreadsActive,
    InvitesCode,
    StickerPacks,
    StickersId,
    UsersId,
    UsersMe,
    UsersMeChannels,
    UsersMeConnections,
    UsersMeGuilds,
    UsersMeGuildsId,
    VoiceRegions,
    WebhooksId(WebhookId),
    WebhooksIdMessagesId(WebhookId),
    WebhooksApplicationId(ApplicationId),
    InteractionsId(InteractionId),
    ApplicationsIdCommands(ApplicationId),
    ApplicationsIdCommandsId(ApplicationId),
    ApplicationsIdGuildsIdCommands(ApplicationId),
    ApplicationsIdGuildsIdCommandsPermissions(ApplicationId),
    ApplicationsIdGuildsIdCommandIdPermissions(ApplicationId),
    ApplicationsIdGuildsIdCommandsId(ApplicationId),
    StageInstances,
    StageInstancesChannelId(ChannelId),
    None,

}

Available on crate feature http only.

A representation of all routes registered within the library. These are safe and memory-efficient representations of each path that functions exist for in the http module.

Used as ratelimit buckets.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ChannelsId(ChannelId)

Route for the /channels/:channel_id path.

ChannelsIdInvites(ChannelId)

Route for the /channels/:channel_id/invites path.

ChannelsIdMessages(ChannelId)

Route for the /channels/:channel_id/messages path.

ChannelsIdMessagesBulkDelete(ChannelId)

Route for the /channels/:channel_id/messages/bulk-delete path.

ChannelsIdMessagesId(LightMethod, ChannelId)

Route for the /channels/:channel_id/messages/:message_id path.

This route is a unique case. The ratelimit for message deletions is different than the overall route ratelimit.

Refer to the docs on [Rate Limits] in the yellow warning section. [Rate Limits]: https://discord.com/developers/docs/topics/rate-limits

ChannelsIdMessagesIdAck(ChannelId)

Route for the /channels/:channel_id/messages/:message_id/ack path.

ChannelsIdMessagesIdReactions(ChannelId)

Route for the /channels/:channel_id/messages/:message_id/reactions path.

ChannelsIdMessagesIdReactionsUserIdType(ChannelId)

Route for the /channels/:channel_id/messages/:message_id/reactions/:reaction/@me path.

ChannelsIdPermissionsOverwriteId(ChannelId)

Route for the /channels/:channel_id/permissions/:target_id path.

ChannelsIdPins(ChannelId)

Route for the /channels/:channel_id/pins path.

ChannelsIdPinsMessageId(ChannelId)

Route for the /channels/:channel_id/pins/:message_id path.

ChannelsIdCrosspostsMessageId(ChannelId)

Route for the /channels/:channel_id/message/:message_id/crosspost path.

ChannelsIdTyping(ChannelId)

Route for the /channels/:channel_id/typing path.

ChannelsIdWebhooks(ChannelId)

Route for the /channels/:channel_id/webhooks path.

ChannelsIdMessagesIdThreads(ChannelId)

Route for the /channels/:channel_id/messages/:message_id/threads path.

ChannelsIdThreads(ChannelId)

Route for the /channels/:channel_id/threads path.

ChannelsIdThreadMembersMe(ChannelId)

Route for the /channels/:channel_id/thread-members/@me path.

ChannelsIdThreadMembersUserId(ChannelId)

Route for the /channels/:channel_id/thread-members/:user_id path.

ChannelsIdThreadMembers(ChannelId)

Route for the /channels/channel_id/thread-members path.

ChannelsIdArchivedPublicThreads(ChannelId)

Route for the /channels/:channel_id/threads/archived/public path.

ChannelsIdArchivedPrivateThreads(ChannelId)

Route for the /channels/:channel_id/threads/archived/private path.

ChannelsIdMeJoindedArchivedPrivateThreads(ChannelId)

Route for the /channels/:channel_id/users/@me/threads/archived/private path.

FollowNewsChannel(ChannelId)

Route for the /channels/{channel.id}/followers path.

Gateway

Route for the /gateway path.

GatewayBot

Route for the /gateway/bot path.

Guilds

Route for the /guilds path.

GuildsId(GuildId)

Route for the /guilds/:guild_id path.

GuildsIdAutoModRules(GuildId)

Route for the /guilds/:guild_id/auto-moderation/rules path.

GuildsIdAutoModRulesId(GuildId)

Route for the /guilds/:guild_id/auto-moderation/rules/:rule_id path.

GuildsIdBans(GuildId)

Route for the /guilds/:guild_id/bans path.

GuildsIdAuditLogs(GuildId)

GuildsIdBansUserId(GuildId)

Route for the /guilds/:guild_id/bans/:user_id path.

GuildsIdChannels(GuildId)

Route for the /guilds/:guild_id/channels/:channel_id path.

GuildsIdWidget(GuildId)

Route for the /guilds/:guild_id/widget path.

GuildsIdPreview(GuildId)

Route for the /guilds/:guild_id/preview path.

GuildsIdEmojis(GuildId)

Route for the /guilds/:guild_id/emojis path.

GuildsIdEmojisId(GuildId)

Route for the /guilds/:guild_id/emojis/:emoji_id path.

GuildsIdIntegrations(GuildId)

Route for the /guilds/:guild_id/integrations path.

GuildsIdIntegrationsId(GuildId)

Route for the /guilds/:guild_id/integrations/:integration_id path.

GuildsIdIntegrationsIdSync(GuildId)

Route for the /guilds/:guild_id/integrations/:integration_id/sync path.

GuildsIdInvites(GuildId)

Route for the /guilds/:guild_id/invites path.

GuildsIdMembers(GuildId)

Route for the /guilds/:guild_id/members path.

GuildsIdMembersId(GuildId)

Route for the /guilds/:guild_id/members/:user_id path.

GuildsIdMembersIdRolesId(GuildId)

Route for the /guilds/:guild_id/members/:user_id/roles/:role_id path.

GuildsIdMembersMe(GuildId)

Route for the /guilds/:guild_id/members/@me path.

GuildsIdMembersMeNick(GuildId)

Route for the /guilds/:guild_id/members/@me/nick path.

GuildsIdMembersSearch(GuildId)

Route for the /guilds/:guild_id/members/search path.

GuildsIdPrune(GuildId)

Route for the /guilds/:guild_id/prune path.

GuildsIdRegions(GuildId)

Route for the /guilds/:guild_id/regions path.

GuildsIdRoles(GuildId)

Route for the /guilds/:guild_id/roles path.

GuildsIdRolesId(GuildId)

Route for the /guilds/:guild_id/roles/:role_id path.

GuildsIdScheduledEvents(GuildId)

Route for the /guilds/:guild_id/scheduled-events path.

GuildsIdScheduledEventsId(GuildId)

Route for the /guilds/:guild_id/scheduled-events/:event_id path.

GuildsIdScheduledEventsIdUsers(GuildId)

Route for the /guilds/:guild_id/scheduled-events/:event_id/users path.

GuildsIdStickers(GuildId)

Route for the /guilds/:guild_id/stickers path.

GuildsIdStickersId(GuildId)

Route for the /guilds/:guild_id/stickers/:sticker_id path.

GuildsIdVanityUrl(GuildId)

Route for the /guilds/:guild_id/vanity-url path.

GuildsIdVoiceStates(GuildId)

Route for the /guilds/:guild_id/voice-states/:user_id path.

GuildsIdVoiceStatesMe(GuildId)

Route for the /guilds/:guild_id/voice-states/@me path.

GuildsIdWebhooks(GuildId)

Route for the /guilds/:guild_id/webhooks path.

GuildsIdWelcomeScreen(GuildId)

Route for the /guilds/:guild_id/welcome-screen path.

GuildsIdThreadsActive

Route for the /guilds/:guild_id/threads/active path.

InvitesCode

Route for the /invites/:code path.

StickerPacks

Route for the /sticker-packs path.

StickersId

Route for the /stickers/:sticker_id path.

UsersId

Route for the /users/:user_id path.

UsersMe

Route for the /users/@me path.

UsersMeChannels

Route for the /users/@me/channels path.

UsersMeConnections

Route for the /users/@me/connections path.

UsersMeGuilds

Route for the /users/@me/guilds path.

UsersMeGuildsId

Route for the /users/@me/guilds/:guild_id path.

VoiceRegions

Route for the /voice/regions path.

WebhooksId(WebhookId)

Route for the /webhooks/:webhook_id path.

WebhooksIdMessagesId(WebhookId)

Route for the /webhooks/:webhook_id/:token/messages/:message_id path.

WebhooksApplicationId(ApplicationId)

Route for the /webhooks/:application_id path.

InteractionsId(InteractionId)

Route for the /interactions/:interaction_id path.

ApplicationsIdCommands(ApplicationId)

Route for the /applications/:application_id path.

ApplicationsIdCommandsId(ApplicationId)

Route for the /applications/:application_id/commands/:command_id path.

ApplicationsIdGuildsIdCommands(ApplicationId)

Route for the /applications/:application_id/guilds/:guild_id path.

ApplicationsIdGuildsIdCommandsPermissions(ApplicationId)

Route for the /applications/:application_id/guilds/:guild_id/commands/permissions path.

ApplicationsIdGuildsIdCommandIdPermissions(ApplicationId)

Route for the /applications/:application_id/guilds/:guild_id/commands/:command_id/permissions path.

ApplicationsIdGuildsIdCommandsId(ApplicationId)

Route for the /applications/:application_id/guilds/:guild_id path.

StageInstances

Route for the /stage-instances path.

StageInstancesChannelId(ChannelId)

Route for the /stage-instances/:channel_id path.

None

Route where no ratelimit headers are in place (i.e. user account-only routes).

This is a special case, in that if the route is None then pre- and post-hooks are not executed.

Implementations

impl Route

pub fn channel(channel_id: ChannelId) -> String

pub fn channel_invites(channel_id: ChannelId) -> String

pub fn channel_message(channel_id: ChannelId, message_id: MessageId) -> String

pub fn channel_message_crosspost(
channel_id: ChannelId,
message_id: MessageId
) -> String

pub fn channel_message_reaction<D, T>(
channel_id: ChannelId,
message_id: MessageId,
user_id: D,
reaction_type: T
) -> Stringwhere
D: Display,
T: Display,

pub fn channel_message_reaction_emoji<T>(
channel_id: ChannelId,
message_id: MessageId,
reaction_type: T
) -> Stringwhere
T: Display,

pub fn channel_message_reactions(
channel_id: ChannelId,
message_id: MessageId
) -> String

pub fn channel_message_reactions_list(
channel_id: ChannelId,
message_id: MessageId,
reaction: &str,
limit: u8,
after: Option<u64>
) -> String

pub fn channel_messages(channel_id: ChannelId, query: Option<&str>) -> String

pub fn channel_messages_bulk_delete(channel_id: ChannelId) -> String

pub fn channel_follow_news(channel_id: ChannelId) -> String

pub fn channel_permission(channel_id: ChannelId, target_id: TargetId) -> String

pub fn channel_pin(channel_id: ChannelId, message_id: MessageId) -> String

pub fn channel_pins(channel_id: ChannelId) -> String

pub fn channel_typing(channel_id: ChannelId) -> String

pub fn channel_webhooks(channel_id: ChannelId) -> String

pub fn channel_public_threads(
channel_id: ChannelId,
message_id: MessageId
) -> String

pub fn channel_private_threads(channel_id: ChannelId) -> String

pub fn channel_thread_member(channel_id: ChannelId, user_id: UserId) -> String

pub fn channel_thread_member_me(channel_id: ChannelId) -> String

pub fn channel_thread_members(channel_id: ChannelId) -> String

pub fn channel_archived_public_threads(
channel_id: ChannelId,
before: Option<u64>,
limit: Option<u64>
) -> String

pub fn channel_archived_private_threads(
channel_id: ChannelId,
before: Option<u64>,
limit: Option<u64>
) -> String

pub fn channel_joined_private_threads(
channel_id: ChannelId,
before: Option<u64>,
limit: Option<u64>
) -> String

pub fn gateway() -> &'static str

pub fn gateway_bot() -> &'static str

pub fn guild(guild_id: GuildId) -> String

pub fn guild_with_counts(guild_id: GuildId) -> String

pub fn guild_audit_logs(
guild_id: GuildId,
action_type: Option<u8>,
user_id: Option<UserId>,
before: Option<u64>,
limit: Option<u8>
) -> String

pub fn guild_automod_rule(guild_id: GuildId, rule_id: RuleId) -> String

pub fn guild_automod_rules(guild_id: GuildId) -> String

pub fn guild_ban(guild_id: GuildId, user_id: UserId) -> String

pub fn guild_ban_optioned(
guild_id: GuildId,
user_id: UserId,
delete_message_days: u8
) -> String

pub fn guild_kick_optioned(guild_id: GuildId, user_id: UserId) -> String

pub fn guild_bans(guild_id: GuildId) -> String

pub fn guild_channels(guild_id: GuildId) -> String

pub fn guild_widget(guild_id: GuildId) -> String

pub fn guild_preview(guild_id: GuildId) -> String

pub fn guild_emojis(guild_id: GuildId) -> String

pub fn guild_emoji(guild_id: GuildId, emoji_id: EmojiId) -> String

pub fn guild_integration(
guild_id: GuildId,
integration_id: IntegrationId
) -> String

pub fn guild_integration_sync(
guild_id: GuildId,
integration_id: IntegrationId
) -> String

pub fn guild_integrations(guild_id: GuildId) -> String

pub fn guild_invites(guild_id: GuildId) -> String

pub fn guild_member(guild_id: GuildId, user_id: UserId) -> String

pub fn guild_member_role(
guild_id: GuildId,
user_id: UserId,
role_id: RoleId
) -> String

pub fn guild_members(guild_id: GuildId) -> String

pub fn guild_members_search(
guild_id: GuildId,
query: &str,
limit: Option<u64>
) -> String

pub fn guild_members_optioned(
guild_id: GuildId,
after: Option<u64>,
limit: Option<u64>
) -> String

pub fn guild_member_me(guild_id: GuildId) -> String

pub fn guild_nickname(guild_id: GuildId) -> String

pub fn guild_prune(guild_id: GuildId, days: u8) -> String

pub fn guild_regions(guild_id: GuildId) -> String

pub fn guild_role(guild_id: GuildId, role_id: RoleId) -> String

pub fn guild_roles(guild_id: GuildId) -> String

pub fn guild_scheduled_event(
guild_id: GuildId,
event_id: ScheduledEventId,
with_user_count: Option<bool>
) -> String

pub fn guild_scheduled_events(
guild_id: GuildId,
with_user_count: Option<bool>
) -> String

pub fn guild_scheduled_event_users(
guild_id: GuildId,
event_id: ScheduledEventId,
after: Option<u64>,
before: Option<u64>,
limit: Option<u64>,
with_member: Option<bool>
) -> String

pub fn guild_sticker(guild_id: GuildId, sticker_id: StickerId) -> String

pub fn guild_stickers(guild_id: GuildId) -> String

pub fn guild_vanity_url(guild_id: GuildId) -> String

pub fn guild_voice_states(guild_id: GuildId, user_id: UserId) -> String

pub fn guild_voice_states_me(guild_id: GuildId) -> String

pub fn guild_webhooks(guild_id: GuildId) -> String

pub fn guild_welcome_screen(guild_id: GuildId) -> String

pub fn guild_threads_active(guild_id: GuildId) -> String

pub fn guilds() -> &'static str

pub fn invite(code: &str) -> String

pub fn invite_optioned(
code: &str,
member_counts: bool,
expiration: bool,
event_id: Option<ScheduledEventId>
) -> String

pub fn oauth2_application_current() -> &'static str

pub fn private_channel() -> &'static str

pub fn status_incidents_unresolved() -> &'static str

pub fn status_maintenances_active() -> &'static str

pub fn status_maintenances_upcoming() -> &'static str

pub fn sticker(sticker_id: StickerId) -> String

pub fn sticker_packs() -> &'static str

pub fn user<D: Display>(target: D) -> String

pub fn user_me_connections() -> &'static str

pub fn user_dm_channels<D: Display>(target: D) -> String

pub fn user_guild<D: Display>(target: D, guild_id: GuildId) -> String

pub fn user_guilds<D: Display>(target: D) -> String

pub fn user_guilds_optioned<D: Display>(
target: D,
after: Option<u64>,
before: Option<u64>,
limit: Option<u64>
) -> String

pub fn voice_regions() -> &'static str

pub fn webhook(webhook_id: WebhookId) -> String

pub fn webhook_with_token<D>(webhook_id: WebhookId, token: D) -> Stringwhere
D: Display,

pub fn webhook_with_token_optioned<D>(
webhook_id: WebhookId,
thread_id: Option<ChannelId>,
token: D,
wait: bool
) -> Stringwhere
D: Display,

pub fn webhook_message<D>(
webhook_id: WebhookId,
token: D,
message_id: MessageId
) -> Stringwhere
D: Display,

pub fn webhook_original_interaction_response<D: Display>(
application_id: ApplicationId,
token: D
) -> String

pub fn webhook_followup_message<D: Display>(
application_id: ApplicationId,
token: D,
message_id: MessageId
) -> String

pub fn webhook_followup_messages<D: Display>(
application_id: ApplicationId,
token: D
) -> String

pub fn interaction_response<D: Display>(
interaction_id: InteractionId,
token: D
) -> String

pub fn application_command(
application_id: ApplicationId,
command_id: CommandId
) -> String

pub fn application_commands(application_id: ApplicationId) -> String

pub fn application_guild_command(
application_id: ApplicationId,
guild_id: GuildId,
command_id: CommandId
) -> String

pub fn application_guild_command_permissions(
application_id: ApplicationId,
guild_id: GuildId,
command_id: CommandId
) -> String

pub fn application_guild_commands(
application_id: ApplicationId,
guild_id: GuildId
) -> String

pub fn application_guild_commands_permissions(
application_id: ApplicationId,
guild_id: GuildId
) -> String

pub fn stage_instances() -> &'static str

pub fn stage_instance(channel_id: ChannelId) -> String