serenity::http

Struct HttpBuilder

pub struct HttpBuilder { /* private fields */ }

Available on crate feature http only.

A builder for the underlying Http client.

If you do not need to use a proxy or do not need to disable the rate limiter, you can use Http::new instead.

Example

Create an instance of Http with a proxy and rate limiter disabled

let http = HttpBuilder::without_token()
    .proxy("http://127.0.0.1:3000")
    .ratelimiter_disabled(true)
    .build();

Implementations

impl HttpBuilder

pub fn new(token: Token) -> Self

Construct a new builder.

pub fn without_token() -> Self

Construct a new builder without a token set.

Most Discord functionality requires a logged-in Bot token, but there are some exceptions such as webhook endpoints.

pub fn application_id(self, application_id: ApplicationId) -> Self

Sets the application_id to use interactions.

pub fn client(self, client: Client) -> Self

Sets the [reqwest::Client]. If one isn’t provided, a default one will be used.

pub fn ratelimiter(self, ratelimiter: Ratelimiter) -> Self

Sets the ratelimiter to be used. If one isn’t provided, a default one will be used.

pub fn ratelimiter_disabled(self, ratelimiter_disabled: bool) -> Self

Sets whether or not the ratelimiter is disabled. By default if this this not used, it is enabled. In most cases, this should be used in conjunction with Self::proxy.

Note: You should not disable the ratelimiter unless you have another form of rate limiting. Disabling the ratelimiter has the main purpose of delegating rate limiting to an API proxy via Self::proxy instead of the current process.

pub fn proxy<'a>(self, proxy: impl Into<Cow<'a, str>>) -> Self

Sets the proxy that Discord HTTP API requests will be passed to. This is mainly intended for something like twilight-http-proxy where multiple processes can make API requests while sharing a single ratelimiter.

The proxy should be in the form of the protocol and hostname, e.g. http://127.0.0.1:3000 or http://myproxy.example

This will simply send HTTP API requests to the proxy instead of Discord API to allow the proxy to intercept, rate limit, and forward requests. This is different than a native proxy’s behavior where it will tunnel requests that use TLS via HTTP CONNECT method (e.g. using [reqwest::Proxy]).

Panics

Panics if the proxy URL is larger than u16::MAX characters… what are you doing?

pub fn default_allowed_mentions( self, allowed_mentions: CreateAllowedMentions<'static>, ) -> Self

Sets the CreateAllowedMentions used by default for each request that would use it.

This only takes effect if you are calling through the model or builder methods, not directly calling Http methods, as Http is simply used as a convenient storage for these.

pub fn build(self) -> Http

Use the given configuration to build the Http client.