Module serenity::json::prelude

Re-exports

pub use super::*;

Functions

from_readerNon-simd-json and std

Deserialize an instance of type T from an IO stream of JSON.

from_sliceNon-simd-json

Deserialize an instance of type T from bytes of JSON text.

from_strNon-simd-json

Deserialize an instance of type T from a string of JSON text.

from_valueNon-simd-json

Interpret a serde_json::Value as an instance of type T.

to_stringNon-simd-json

Serialize the given data structure as a String of JSON.

to_string_prettyNon-simd-json

Serialize the given data structure as a pretty-printed String of JSON.

to_valueNon-simd-json

Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

to_vecNon-simd-json

Serialize the given data structure as a JSON byte vector.

to_vec_prettyNon-simd-json

Serialize the given data structure as a pretty-printed JSON byte vector.