Enum serenity::model::channel::Channel

#[non_exhaustive]
pub enum Channel {
    Guild(GuildChannel),
    Private(PrivateChannel),
    Category(ChannelCategory),
}

A container for any channel.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Guild(GuildChannel)

A text, voice, or stage channel within a Guild.

Private(PrivateChannel)

A private channel to another User. No other users may access the channel. For multi-user “private channels”, use a group.

Category(ChannelCategory)

A category of GuildChannels

Implementations

impl Channel

pub fn guild(self) -> Option<GuildChannel>

Converts from Channel to Option<GuildChannel>.

Converts self into an Option<GuildChannel>, consuming self, and discarding a PrivateChannel, or ChannelCategory, if any.

Examples

Basic usage:

match channel.guild() {
    Some(guild) => {
        println!("It's a guild named {}!", guild.name);
    },
    None => { println!("It's not a guild!"); },
}

pub fn private(self) -> Option<PrivateChannel>

Converts from Channel to Option<PrivateChannel>.

Converts self into an Option<PrivateChannel>, consuming self, and discarding a GuildChannel, or ChannelCategory, if any.

Examples

Basic usage:

match channel.private() {
    Some(private) => {
        println!("It's a private channel with {}!", &private.recipient);
    },
    None => { println!("It's not a private channel!"); },
}

pub fn category(self) -> Option<ChannelCategory>

Converts from Channel to Option<ChannelCategory>.

Converts self into an Option<ChannelCategory>, consuming self, and discarding a GuildChannel, or PrivateChannel, if any.

Examples

Basic usage:

match channel.category() {
    Some(category) => {
        println!("It's a category named {}!", category.name);
    },
    None => { println!("It's not a category!"); },
}

pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<()>

Deletes the inner channel.

Errors

If the cache is enabled, returns ModelError::InvalidPermissions, if the current user lacks permission.

Otherwise will return Error::Http if the current user does not have permission.

pub fn is_nsfw(&self) -> bool

Determines if the channel is NSFW.

pub fn id(&self) -> ChannelId

Retrieves the Id of the inner GuildChannel, or PrivateChannel.

pub fn position(&self) -> Option<i64>

Retrieves the position of the inner GuildChannel or ChannelCategory.

If other channel types are used it will return None.