Struct serenity::model::channel::ChannelCategory

#[non_exhaustive]
pub struct ChannelCategory {
    pub id: ChannelId,
    pub guild_id: GuildId,
    pub parent_id: Option<ChannelId>,
    pub position: i64,
    pub kind: ChannelType,
    pub name: String,
    pub nsfw: bool,
    pub permission_overwrites: Vec<PermissionOverwrite>,
}

A category of GuildChannels.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: ChannelId

Id of this category.

guild_id: GuildId

Guild Id this category belongs to.

parent_id: Option<ChannelId>

If this category belongs to another category.

position: i64

The position of this category.

kind: ChannelType

Indicator of the type of channel this is.

This should always be ChannelType::Category.

name: String

The name of the category.

nsfw: bool

Whether this category is nsfw. (This’ll be inherited by all channels in this category)

permission_overwrites: Vec<PermissionOverwrite>

Permission overwrites for the GuildChannels.

Implementations

impl ChannelCategory

pub async fn create_permission( &self, http: impl AsRef<Http>, target: &PermissionOverwrite ) -> Result<()>

Adds a permission overwrite to the category’s channels.

Note: Requires the Manage Channels permission.

Also requires the Manage Roles permission if not modifying the permissions for only the current user.

Errors

Returns Error::Http if the current user lacks permission, or if an invalid value was set.

pub async fn delete_permission( &self, http: impl AsRef<Http>, permission_type: PermissionOverwriteType ) -> Result<()>

Deletes all permission overrides in the category from the channels.

Note: Requires the Manage Roles permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<()>

Deletes this category.

Note: Requires the Manage Channels permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn edit<F>(&mut self, cache_http: impl CacheHttp, f: F) -> Result<()>where F: FnOnce(&mut EditChannel) -> &mut EditChannel,

Modifies the category’s settings, such as its position or name.

Refer to EditChannels documentation for a full list of methods.

Note: Requires the Manage Channels permission, also requires the Manage Roles permission if modifying permissions for the category.

Examples

Change a voice channels name and bitrate:

category.edit(&http, |c| c.name("test").bitrate(86400)).await;

Errors

Returns Error::Http if an invalid value is set, or if the current user lacks the necessary permissions.

pub fn is_nsfw(&self) -> bool

pub fn name(&self) -> &str

Returns the name of the category.