Struct serenity::model::channel::Group

pub struct Group {
    pub id: ChannelId,
    pub icon: Option<String>,
    pub last_message_id: Option<MessageId>,
    pub last_pin_timestamp: Option<DateTime<Utc>>,
    pub name: Option<String>,
    pub owner_id: UserId,
    pub recipients: HashMap<UserId, User>,
}

A group channel, potentially including other users, separate from a Guild.

Fields
id: ChannelId

The Id of the group channel.

icon: Option<String>

The optional icon of the group channel.

last_message_id: Option<MessageId>

The Id of the last message sent.

last_pin_timestamp: Option<DateTime<Utc>>

Timestamp of the latest pinned message.

name: Option<String>

The name of the group channel.

owner_id: UserId

The Id of the group channel creator.

recipients: HashMap<UserId, User>

Group channel’s members.

Implementations

impl Group

pub async fn add_user(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<()>

Available on crate feature model only.

Adds a user to the group

Errors

Returns Error::Http when trying to add an already added user or someone whose privacy settings do not allow being added to a group under current circumstances.

pub async fn remove_user(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<()>

Available on crate feature model only.

Removes a user from the group

Errors

Returns Error::Http when trying to remove a user from the group, which is not in the group or when not group owner.

pub async fn broadcast_typing(&self, http: impl AsRef<Http>) -> Result<()>

Available on crate feature model only.

Broadcasts that the current user is typing to the recipients.

For bots, this is a good indicator for long-running commands.

Errors

Returns Error::Http never

pub async fn create_invite<F>(
&self,
cache_http: impl CacheHttp,
f: F
) -> Result<RichInvite> where
F: FnOnce(&mut CreateInvite) -> &mut CreateInvite,

Available on crate features model and utils only.

Creates an invite leading to the given channel.

Examples

Create an invite that is valid for one day:

let invite = channel.create_invite(&context, |i| i.max_age(86400)).await;

pub async fn delete_reaction(
&self,
http: impl AsRef<Http>,
message_id: impl Into<MessageId>,
user_id: Option<UserId>,
reaction_type: impl Into<ReactionType>
) -> Result<()>

Available on crate feature model only.

Deletes the given Reaction from the group.

Errors

Returns Error::Http if attempting to delete another users reaction.

pub async fn edit<F>(&mut self, cache_http: impl CacheHttp, f: F) -> Result<()> where
F: FnOnce(&mut EditGuild) -> &mut EditGuild,

Available on crate features model and utils only.

Modifies a group’s settings (name).

Examples

Change a groups name:

group.edit(&context, |c| c.name("test")).await;

pub async fn edit_message<F>(
&self,
http: impl AsRef<Http>,
message_id: impl Into<MessageId>,
f: F
) -> Result<Message> where
F: FnOnce(&mut EditMessage) -> &mut EditMessage,

Available on crate feature model only.

Edits a Message in the channel given its Id.

Message editing preserves all unchanged message data.

Refer to the documentation for EditMessage for more information regarding message restrictions and requirements.

Note: Requires that the current user be the author of the message.

Errors

Returns a ModelError::MessageTooLong if the content of the message is over the the limit, containing the number of unicode code points over the limit.

pub async fn invites(&self, http: impl AsRef<Http>) -> Result<Vec<RichInvite>>

Available on crate feature model only.

Gets all of the group’s invites.

pub async fn message(
&self,
http: impl AsRef<Http>,
message_id: impl Into<MessageId>
) -> Result<Message>

Available on crate feature model only.

Gets a message from the channel.

pub async fn messages<F>(
&self,
http: impl AsRef<Http>,
builder: F
) -> Result<Vec<Message>> where
F: FnOnce(&mut GetMessages) -> &mut GetMessages,

Available on crate feature model only.

Gets messages from the channel.

Refer to the GetMessages-builder for more information on how to use builder.

pub fn name(&self) -> String

Available on crate feature model only.

Returns the name of the group.

pub async fn pin(
&self,
http: impl AsRef<Http>,
message_id: impl Into<MessageId>
) -> Result<()>

Available on crate feature model only.

Pins a Message to the channel.

pub async fn pins(&self, http: impl AsRef<Http>) -> Result<Vec<Message>>

Available on crate feature model only.

Gets all channel’s pins.

pub async fn reaction_users(
&self,
http: impl AsRef<Http>,
message_id: impl Into<MessageId>,
reaction_type: impl Into<ReactionType>,
limit: Option<u8>,
after: impl Into<Option<UserId>>
) -> Result<Vec<User>>

Available on crate feature model only.

Gets the list of Users who have reacted to a Message with a certain Emoji.

The default limit is 50 - specify otherwise to receive a different maximum number of users. The maximum that may be retrieve at a time is 100, if a greater number is provided then it is automatically reduced.

The optional after attribute is to retrieve the users after a certain user. This is useful for pagination.

pub async fn say(
&self,
http: impl AsRef<Http>,
content: impl Display
) -> Result<Message>

Available on crate feature model only.

Sends a message with just the given message content in the channel.

Errors

Returns a ModelError::MessageTooLong if the content of the message is over the above limit, containing the number of unicode code points over the limit.

pub async fn send_message<'a, F>(
&self,
cache_http: impl CacheHttp,
f: F
) -> Result<Message> where
for<'b> F: FnOnce(&'b mut CreateMessage<'a>) -> &'b mut CreateMessage<'a>,

Available on crate feature model only.

Sends a message to the group with the given content.

Errors

Returns a ModelError::MessageTooLong if the content of the message is over the above limit, containing the number of unicode code points over the limit.

pub fn start_typing(self, http: &Arc<Http>) -> Result<Typing>

Available on crate feature model only.

Starts typing in the group for an indefinite period of time.

Returns Typing that is used to trigger the typing. Typing::stop must be called on the returned struct to stop typing. Note that on some clients, typing may persist for a few seconds after Typing::stop is called. Typing is also stopped when the struct is dropped.

If a message is sent while typing is triggered, the user will stop typing for a brief period of time and then resume again until either Typing::stop is called or the struct is dropped.

This should rarely be used for bots, although it is a good indicator that a long-running command is still being processed.

Examples

// Initiate typing (assuming http is `Arc<Http>` and `group` is bound)
let typing = group.start_typing(&http)?;

// Run some long-running process
long_process();

// Stop typing
typing.stop();

pub async fn unpin(
&self,
http: impl AsRef<Http>,
message_id: impl Into<MessageId>
) -> Result<()>

Available on crate feature model only.

Unpins a Message in the group given by its Id.

pub async fn delete(&self, http: impl AsRef<Http>) -> Result<Channel>

Available on crate feature model only.

Deletes the group.

Note: This does not delete the contents of the group channel, and is equivalent to closing a private channel on the client, which can be re-opened.

pub async fn members(&self) -> Vec<User>

Available on crate features model and cache only.

Retrieves Users from the current group.

pub fn await_reply<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> CollectReply<'a>

Available on crate features model and collector only.

Returns a future that will await one message by this guild channel.

pub fn await_replies<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> MessageCollectorBuilder<'a>

Available on crate features model and collector only.

Returns a stream builder which can be awaited to obtain a stream of messages sent by this guild channel.

pub fn await_reaction<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> CollectReaction<'a>

Available on crate features model and collector only.

Await a single reaction by this guild channel.

pub fn await_reactions<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> ReactionCollectorBuilder<'a>

Available on crate features model and collector only.

Returns a stream builder which can be awaited to obtain a stream of reactions sent by this guild channel.