Struct serenity::model::channel::MessageGeneratedOriginal

#[non_exhaustive]
pub struct MessageGeneratedOriginal {

    pub id: MessageId,
    pub channel_id: ChannelId,
    pub author: User,
    pub content: FixedString<u16>,
    pub timestamp: Timestamp,
    pub edited_timestamp: Option<Timestamp>,
    pub tts: bool,
    pub mention_everyone: bool,
    pub mentions: FixedArray<User>,
    pub mention_roles: FixedArray<RoleId>,
    pub mention_channels: FixedArray<ChannelMention>,
    pub attachments: FixedArray<Attachment>,
    pub embeds: FixedArray<Embed>,
    pub reactions: FixedArray<MessageReaction>,
    pub nonce: Option<Nonce>,
    pub pinned: bool,
    pub webhook_id: Option<WebhookId>,
    pub kind: MessageType,
    pub activity: Option<MessageActivity>,
    pub application: Option<MessageApplication>,
    pub application_id: Option<ApplicationId>,
    pub message_reference: Option<MessageReference>,
    pub flags: Option<MessageFlags>,
    pub referenced_message: Option<Box<Message>>,
    pub interaction: Option<Box<MessageInteraction>>,
    pub thread: Option<Box<GuildChannel>>,
    pub components: FixedArray<ActionRow>,
    pub sticker_items: FixedArray<StickerItem>,
    pub position: Option<u64>,
    pub role_subscription_data: Option<RoleSubscriptionData>,
    pub guild_id: Option<GuildId>,
    pub member: Option<Box<PartialMember>>,

}

A representation of a message over a guild’s text channel, a group, or a private channel.

Discord docs with some extra fields.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: MessageId

The unique Id of the message. Can be used to calculate the creation date of the message.

channel_id: ChannelId

The Id of the Channel that the message was sent to.

author: User

The user that sent the message.

content: FixedString<u16>

The content of the message.

timestamp: Timestamp

Initial message creation timestamp, calculated from its Id.

edited_timestamp: Option<Timestamp>

The timestamp of the last time the message was updated, if it was.

tts: bool

Indicator of whether the command is to be played back via text-to-speech.

In the client, this is done via the /tts slash command.

mention_everyone: bool

Indicator of whether the message mentions everyone.

mentions: FixedArray<User>

Array of users mentioned in the message.

mention_roles: FixedArray<RoleId>

Array of Roles’ Ids mentioned in the message.

mention_channels: FixedArray<ChannelMention>

Channels specifically mentioned in this message.

Note: Not all channel mentions in a message will appear in Self::mention_channels. Only textual channels that are visible to everyone in a lurkable guild will ever be included.

A lurkable guild is one that allows users to read public channels in a server without actually joining the server. It also allows users to look at these channels without being logged in to Discord.

Only crossposted messages (via Channel Following) currently include Self::mention_channels at all. If no mentions in the message meet these requirements, this field will not be sent.

Refer to Discord’s documentation for more information.

attachments: FixedArray<Attachment>

An vector of the files attached to a message.

embeds: FixedArray<Embed>

Array of embeds sent with the message.

reactions: FixedArray<MessageReaction>

Array of reactions performed on the message.

nonce: Option<Nonce>

Non-repeating number used for ensuring message order.

pinned: bool

Indicator of whether the message is pinned.

webhook_id: Option<WebhookId>

The Id of the webhook that sent this message, if one did.

kind: MessageType

Indicator of the type of message this is, i.e. whether it is a regular message or a system message.

activity: Option<MessageActivity>

Sent with Rich Presence-related chat embeds.

application: Option<MessageApplication>

Sent with Rich Presence-related chat embeds.

application_id: Option<ApplicationId>

If the message is an Interaction or application-owned webhook, this is the id of the application.

message_reference: Option<MessageReference>

Reference data sent with crossposted messages.

flags: Option<MessageFlags>

Bit flags describing extra features of the message.

referenced_message: Option<Box<Message>>

The message that was replied to using this message.

interaction: Option<Box<MessageInteraction>>

Sent if the message is a response to an Interaction.

thread: Option<Box<GuildChannel>>

The thread that was started from this message, includes thread member object.

components: FixedArray<ActionRow>

The components of this message

sticker_items: FixedArray<StickerItem>

Array of message sticker item objects.

position: Option<u64>

A generally increasing integer (there may be gaps or duplicates) that represents the approximate position of the message in a thread, it can be used to estimate the relative position of the message in a thread in company with total_message_sent on parent thread.

role_subscription_data: Option<RoleSubscriptionData>

Data of the role subscription purchase or renewal that prompted this MessageType::RoleSubscriptionPurchase message.

guild_id: Option<GuildId>

The Id of the Guild that the message was sent in. This value will only be present if this message was received over the gateway, therefore do not use this to check if message is in DMs, it is not a reliable method.

member: Option<Box<PartialMember>>

A partial amount of data about the user’s member data

Only present in MessageCreateEvent.