Module serenity::model::colour::colours::css

Constants

DANGER

Creates a new Colour, setting its value to hsl(359, 82.6%, 59.4%).

POSITIVE

Creates a new Colour, setting its value to hsl(139, 47.3%, 43.9%).

WARNING

Creates a new Colour, setting its value to hsl(38, 95.7%, 54.1%).