serenity::model::colour::colours

Module roles

Constants

BLUE

Creates a new Colour, setting its value to rgb(52, 152, 219).

DARKER_GREY

Creates a new Colour, setting its value to rgb(84, 110, 122).

DARK_BLUE

Creates a new Colour, setting its value to rgb(32, 102, 148).

DARK_GOLD

Creates a new Colour, setting its value to rgb(194, 124, 14).

DARK_GREEN

Creates a new Colour, setting its value to rgb(31, 139, 76).

DARK_GREY

Creates a new Colour, setting its value to rgb(96, 125, 139).

DARK_MAGENTA

Creates a new Colour, setting its value to rgb(173, 20, 87).

DARK_ORANGE

Creates a new Colour, setting its value to rgb(168, 67, 0).

DARK_PURPLE

Creates a new Colour, setting its value to rgb(113, 54, 138).

DARK_RED

Creates a new Colour, setting its value to rgb(153, 45, 34).

DARK_TEAL

Creates a new Colour, setting its value to rgb(17, 128, 106).

DEFAULT

Creates a new Colour, setting its value to rgb(153, 170, 181).

GOLD

Creates a new Colour, setting its value to rgb(241, 196, 15).

GREEN

Creates a new Colour, setting its value to rgb(46, 204, 113).

LIGHTER_GREY

Creates a new Colour, setting its value to rgb(149, 165, 166).

LIGHT_GREY

Creates a new Colour, setting its value to rgb(151, 156, 159).

MAGENTA

Creates a new Colour, setting its value to rgb(233, 30, 99).

ORANGE

Creates a new Colour, setting its value to rgb(230, 126, 34).

PURPLE

Creates a new Colour, setting its value to rgb(155, 89, 182).

RED

Creates a new Colour, setting its value to rgb(231, 76, 60).

TEAL

Creates a new Colour, setting its value to rgb(26, 188, 156).