Enum serenity::model::event::Event

pub enum Event {
    ChannelCreate(ChannelCreateEvent),
    ChannelDelete(ChannelDeleteEvent),
    ChannelPinsUpdate(ChannelPinsUpdateEvent),
    ChannelRecipientAdd(ChannelRecipientAddEvent),
    ChannelRecipientRemove(ChannelRecipientRemoveEvent),
    ChannelUpdate(ChannelUpdateEvent),
    GuildBanAdd(GuildBanAddEvent),
    GuildBanRemove(GuildBanRemoveEvent),
    GuildCreate(GuildCreateEvent),
    GuildDelete(GuildDeleteEvent),
    GuildEmojisUpdate(GuildEmojisUpdateEvent),
    GuildIntegrationsUpdate(GuildIntegrationsUpdateEvent),
    GuildMemberAdd(GuildMemberAddEvent),
    GuildMemberRemove(GuildMemberRemoveEvent),
    GuildMemberUpdate(GuildMemberUpdateEvent),
    GuildMembersChunk(GuildMembersChunkEvent),
    GuildRoleCreate(GuildRoleCreateEvent),
    GuildRoleDelete(GuildRoleDeleteEvent),
    GuildRoleUpdate(GuildRoleUpdateEvent),
    GuildUnavailable(GuildUnavailableEvent),
    GuildUpdate(GuildUpdateEvent),
    MessageCreate(MessageCreateEvent),
    MessageDelete(MessageDeleteEvent),
    MessageDeleteBulk(MessageDeleteBulkEvent),
    MessageUpdate(MessageUpdateEvent),
    PresenceUpdate(PresenceUpdateEvent),
    PresencesReplace(PresencesReplaceEvent),
    ReactionAdd(ReactionAddEvent),
    ReactionRemove(ReactionRemoveEvent),
    ReactionRemoveAll(ReactionRemoveAllEvent),
    Ready(ReadyEvent),
    Resumed(ResumedEvent),
    TypingStart(TypingStartEvent),
    UserUpdate(UserUpdateEvent),
    VoiceStateUpdate(VoiceStateUpdateEvent),
    VoiceServerUpdate(VoiceServerUpdateEvent),
    WebhookUpdate(WebhookUpdateEvent),
    Unknown(UnknownEvent),
}

Event received over a websocket connection

Variants
ChannelCreate(ChannelCreateEvent)

A Channel was created.

Fires the Client::on_channel_create event.

ChannelDelete(ChannelDeleteEvent)

A Channel has been deleted.

Fires the [Client::on_channel_delete] event.

ChannelPinsUpdate(ChannelPinsUpdateEvent)

The pins for a Channel have been updated.

Fires the Client::on_channel_pins_update event.

ChannelRecipientAdd(ChannelRecipientAddEvent)

A User has been added to a [Group].

Fires the Client::on_recipient_add event.

ChannelRecipientRemove(ChannelRecipientRemoveEvent)

A User has been removed from a [Group].

Fires the Client::on_recipient_remove event.

ChannelUpdate(ChannelUpdateEvent)

A [Channel] has been updated.

Fires the Client::on_channel_update event.

GuildBanAdd(GuildBanAddEvent)GuildBanRemove(GuildBanRemoveEvent)GuildCreate(GuildCreateEvent)GuildDelete(GuildDeleteEvent)GuildEmojisUpdate(GuildEmojisUpdateEvent)GuildIntegrationsUpdate(GuildIntegrationsUpdateEvent)GuildMemberAdd(GuildMemberAddEvent)GuildMemberRemove(GuildMemberRemoveEvent)GuildMemberUpdate(GuildMemberUpdateEvent)

A member's roles have changed

GuildMembersChunk(GuildMembersChunkEvent)GuildRoleCreate(GuildRoleCreateEvent)GuildRoleDelete(GuildRoleDeleteEvent)GuildRoleUpdate(GuildRoleUpdateEvent)GuildUnavailable(GuildUnavailableEvent)

When a guild is unavailable, such as due to a Discord server outage.

GuildUpdate(GuildUpdateEvent)MessageCreate(MessageCreateEvent)MessageDelete(MessageDeleteEvent)MessageDeleteBulk(MessageDeleteBulkEvent)MessageUpdate(MessageUpdateEvent)

A message has been edited, either by the user or the system

PresenceUpdate(PresenceUpdateEvent)

A member's presence state (or username or avatar) has changed

PresencesReplace(PresencesReplaceEvent)

The precense list of the user's friends should be replaced entirely

ReactionAdd(ReactionAddEvent)

A reaction was added to a message.

Fires the on_message_reaction_add event handler.

ReactionRemove(ReactionRemoveEvent)

A reaction was removed to a message.

Fires the on_message_reaction_remove event handler.

ReactionRemoveAll(ReactionRemoveAllEvent)

A request was issued to remove all Reactions from a Message.

Fires the on_reaction_remove_all event handler.

Ready(ReadyEvent)

The first event in a connection, containing the initial ready cache.

May also be received at a later time in the event of a reconnect.

Resumed(ResumedEvent)

The connection has successfully resumed after a disconnect.

TypingStart(TypingStartEvent)

A user is typing; considered to last 5 seconds

UserUpdate(UserUpdateEvent)

Update to the logged-in user's information

VoiceStateUpdate(VoiceStateUpdateEvent)

A member's voice state has changed

VoiceServerUpdate(VoiceServerUpdateEvent)

Voice server information is available

WebhookUpdate(WebhookUpdateEvent)

A webhook for a channel was updated in a Guild.

Unknown(UnknownEvent)

An event type not covered by the above