serenity::model

Module gateway

Models pertaining to the gateway.

Structs

Activity

Representation of an activity that a User is performing.

ActivityAssets

The assets for an activity.

ActivityButton

Discord docs.

ActivityEmoji

Representation of an emoji used in a custom status

ActivityFlags

A set of flags defining what is in an activity’s payload.

ActivityParty

Information about an activity’s party.

ActivitySecrets

Secrets for an activity.

ActivityTimestamps

Timestamps of when a user started and/or is ending their activity.

ActivityType

Discord docs.

BotGateway

A representation of the data retrieved from the bot gateway endpoint.

ClientStatus

Information detailing the current active status of a User.

Gateway

A representation of the data retrieved from the gateway endpoint.

GatewayIntents

Gateway Intents will limit the events your bot will receive via the gateway. By default, all intents except Privileged Intents are selected.

Presence

Information detailing the current online status of a User.

PresenceUser

Information about the user of a Presence event.

PresenceUserGeneratedOriginal

Information about the user of a Presence event.

Ready

An initial set of information given after IDENTIFYing to the gateway.

SessionStartLimit

Information describing how many gateway sessions you can initiate within a ratelimit period.

ShardInfo

Enums

ConnectionStage

Indicates the current connection stage of a shard.