Enum serenity::model::gateway::ActivityType

#[non_exhaustive]
pub enum ActivityType {
    Playing,
    Streaming,
    Listening,
    Watching,
    Custom,
    Competing,
    Unknown,
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Playing

An indicator that the user is playing a game.

Streaming

An indicator that the user is streaming to a service.

Listening

An indicator that the user is listening to something.

Watching

An indicator that the user is watching something.

Custom

An indicator that the user uses custom statuses

Competing

An indicator that the user is competing somewhere.

Unknown

An indicator that the activity is of unknown type.

Implementations

impl ActivityType

pub fn num(&self) -> u64