Struct serenity::model::gateway::PresenceUser

#[non_exhaustive]
pub struct PresenceUser {
    pub id: UserId,
    pub avatar: Option<String>,
    pub bot: Option<bool>,
    pub discriminator: Option<u16>,
    pub email: Option<String>,
    pub mfa_enabled: Option<bool>,
    pub name: Option<String>,
    pub verified: Option<bool>,
    pub public_flags: Option<UserPublicFlags>,
}

Information about the user of a Presence event.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: UserIdavatar: Option<String>bot: Option<bool>discriminator: Option<u16>email: Option<String>mfa_enabled: Option<bool>name: Option<String>verified: Option<bool>public_flags: Option<UserPublicFlags>
Implementations

impl PresenceUser

pub fn into_user(self) -> Option<User>

Attempts to convert this PresenceUser instance into a User.

If one of User’s required fields is None in self, None is returned.

pub fn to_user(&self) -> Option<User>

Attempts to convert this PresenceUser instance into a User.

Will clone individual fields if needed.

If one of User’s required fields is None in self, None is returned.