serenity::model

Module guild

Models relating to guilds and types that it owns.

Modules

audit_log

Audit log types for administrative actions within guilds.

automod

Auto moderation types

Structs

AfkMetadata

AfkTimeout

The Guild AFK timeout length.

Ban

A representation of a banning of a user.

BulkBanResponse

The response from GuildId::bulk_ban.

DefaultMessageNotificationLevel

Default message notification level for a guild.

Emoji

Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

EmojiGeneratedOriginal

Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

ExplicitContentFilter

Setting used to filter explicit messages from members.

Guild

Information about a Discord guild, such as channels, emojis, etc.

GuildGeneratedOriginal

Information about a Discord guild, such as channels, emojis, etc.

GuildInfo

Variant of Guild returned from Http::get_guilds.

GuildMemberFlags

Flags for a guild member.

GuildPreview

Preview Guild information.

GuildPrune

Representation of the number of members that would be pruned by a guild prune operation.

GuildWelcomeChannel

A channel shown in the GuildWelcomeScreen.

GuildWelcomeScreen

Information relating to a guild’s welcome screen.

GuildWidget

A Guild widget.

IncidentsData

The Guild’s incident’s data.

Integration

Various information about integrations.

IntegrationAccount

Integration account object.

IntegrationApplication

Integration application object.

IntegrationExpireBehaviour

The behavior once the integration expires.

IntegrationGeneratedOriginal

Various information about integrations.

Member

Information about a member of a guild.

MemberGeneratedOriginal

Information about a member of a guild.

MembersItermodel

A helper class returned by GuildId::members_iter

MfaLevel

Multi-Factor Authentication level for guild moderators.

NsfwLevel

The Guild nsfw level.

PartialGuild

Partial information about a Guild. This does not include information like member data.

PartialGuildGeneratedOriginal

Partial information about a Guild. This does not include information like member data.

PartialMember

A partial amount of data for a member.

PartialMemberGeneratedOriginal

A partial amount of data for a member.

PremiumTier

The guild’s premium tier, depends on the amount of users boosting the guild currently

Role

Information about a role within a guild.

RoleColours

The colours of a Discord role, secondary_colour and tertiary_colour may only be set if the Guild has the ENHANCED_ROLE_COLORS feature.

RoleGeneratedOriginal

Information about a role within a guild.

RoleTags

The tags of a Role.

RoleTagsGeneratedOriginal

The tags of a Role.

ScheduledEvent

Information about a guild scheduled event.

ScheduledEventMetadata

Discord docs.

ScheduledEventPrivacyLevel

See ScheduledEvent::privacy_level.

ScheduledEventStatus

Discord docs.

ScheduledEventType

Discord docs.

ScheduledEventUser

Discord docs.

SystemChannelFlags

Describes a system channel flags.

UnavailableGuild

Data for an unavailable guild.

VerificationLevel

The level to set as criteria prior to a user being able to send messages in a Guild.

Enums

GuildWelcomeChannelEmoji

A GuildWelcomeScreen emoji.

GuildWidgetStyle