serenity::model::guild::audit_log

Enum Change

#[non_exhaustive]
pub enum Change {

    Actions {
        old: Option<FixedArray<Action>>,
        new: Option<FixedArray<Action>>,
    },
    AfkChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    AfkTimeout {
        old: Option<AfkTimeout>,
        new: Option<AfkTimeout>,
    },
    Allow {
        old: Option<Permissions>,
        new: Option<Permissions>,
    },
    ApplicationId {
        old: Option<ApplicationId>,
        new: Option<ApplicationId>,
    },
    AppliedTags {
        old: Option<FixedArray<ForumTagId>>,
        new: Option<FixedArray<ForumTagId>>,
    },
    Archived {
        old: Option<bool>,
        new: Option<bool>,
    },
    Asset {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    AutoArchiveDuration {
        old: Option<u16>,
        new: Option<u16>,
    },
    Available {
        old: Option<bool>,
        new: Option<bool>,
    },
    AvailableTags {
        old: Option<FixedArray<ForumTag>>,
        new: Option<FixedArray<ForumTag>>,
    },
    AvatarHash {
        old: Option<ImageHash>,
        new: Option<ImageHash>,
    },
    BannerHash {
        old: Option<ImageHash>,
        new: Option<ImageHash>,
    },
    Bitrate {
        old: Option<u32>,
        new: Option<u32>,
    },
    BrandColorPrimary {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    ChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    Code {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Color {
        old: Option<u32>,
        new: Option<u32>,
    },
    Colors {
        old: Option<RoleColours>,
        new: Option<RoleColours>,
    },
    CommunicationDisabledUntil {
        old: Option<Timestamp>,
        new: Option<Timestamp>,
    },
    Deaf {
        old: Option<bool>,
        new: Option<bool>,
    },
    DefaultAutoArchiveDuration {
        old: Option<u16>,
        new: Option<u16>,
    },
    DefaultChannelIds {
        old: Option<FixedArray<ChannelId>>,
        new: Option<FixedArray<ChannelId>>,
    },
    DefaultMessageNotifications {
        old: Option<DefaultMessageNotificationLevel>,
        new: Option<DefaultMessageNotificationLevel>,
    },
    DefaultReactionEmoji {
        old: Option<ForumEmoji>,
        new: Option<ForumEmoji>,
    },
    DefaultThreadRateLimitPerUser {
        old: Option<u16>,
        new: Option<u16>,
    },
    Deny {
        old: Option<Permissions>,
        new: Option<Permissions>,
    },
    Description {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    DiscoverySplashHash {
        old: Option<ImageHash>,
        new: Option<ImageHash>,
    },
    EmojiId {
        old: Option<EmojiId>,
        new: Option<EmojiId>,
    },
    EmojiName {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Enabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    EnableEmoticons {
        old: Option<bool>,
        new: Option<bool>,
    },
    EntityType {
        old: Option<u64>,
        new: Option<u64>,
    },
    EventType {
        old: Option<AutomodEventType>,
        new: Option<AutomodEventType>,
    },
    ExemptChannels {
        old: Option<FixedArray<ChannelId>>,
        new: Option<FixedArray<ChannelId>>,
    },
    ExemptRoles {
        old: Option<FixedArray<RoleId>>,
        new: Option<FixedArray<RoleId>>,
    },
    ExpireBehavior {
        old: Option<u64>,
        new: Option<u64>,
    },
    ExpireGracePeriod {
        old: Option<u64>,
        new: Option<u64>,
    },
    ExplicitContentFilter {
        old: Option<ExplicitContentFilter>,
        new: Option<ExplicitContentFilter>,
    },
    Flags {
        old: Option<u64>,
        new: Option<u64>,
    },
    FormatType {
        old: Option<StickerFormatType>,
        new: Option<StickerFormatType>,
    },
    GameApplicationIds {
        old: Option<FixedArray<ApplicationId>>,
        new: Option<FixedArray<ApplicationId>>,
    },
    GuildId {
        old: Option<GuildId>,
        new: Option<GuildId>,
    },
    Hoist {
        old: Option<bool>,
        new: Option<bool>,
    },
    IconHash {
        old: Option<ImageHash>,
        new: Option<ImageHash>,
    },
    Id {
        old: Option<GenericId>,
        new: Option<GenericId>,
    },
    ImageHash {
        old: Option<ImageHash>,
        new: Option<ImageHash>,
    },
    InOnboarding {
        old: Option<bool>,
        new: Option<bool>,
    },
    Invitable {
        old: Option<bool>,
        new: Option<bool>,
    },
    InviterId {
        old: Option<UserId>,
        new: Option<UserId>,
    },
    Location {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Locked {
        old: Option<bool>,
        new: Option<bool>,
    },
    ManualApprovalEnabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    MaxAge {
        old: Option<u32>,
        new: Option<u32>,
    },
    MaxUses {
        old: Option<u8>,
        new: Option<u8>,
    },
    Mentionable {
        old: Option<bool>,
        new: Option<bool>,
    },
    MfaLevel {
        old: Option<MfaLevel>,
        new: Option<MfaLevel>,
    },
    Mute {
        old: Option<bool>,
        new: Option<bool>,
    },
    Name {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Nick {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Nsfw {
        old: Option<bool>,
        new: Option<bool>,
    },
    OwnerId {
        old: Option<UserId>,
        new: Option<UserId>,
    },
    PermissionOverwrites {
        old: Option<FixedArray<PermissionOverwrite>>,
        new: Option<FixedArray<PermissionOverwrite>>,
    },
    Permissions {
        old: Option<Permissions>,
        new: Option<Permissions>,
    },
    Position {
        old: Option<u32>,
        new: Option<u32>,
    },
    PreferredLocale {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    PremiumProgressBarEnabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    PrivacyLevel {
        old: Option<u64>,
        new: Option<u64>,
    },
    PruneDeleteDays {
        old: Option<u64>,
        new: Option<u64>,
    },
    PublicUpdatesChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    RateLimitPerUser {
        old: Option<u16>,
        new: Option<u16>,
    },
    Region {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Required {
        old: Option<bool>,
        new: Option<bool>,
    },
    RoleIds {
        old: Option<FixedArray<RoleId>>,
        new: Option<FixedArray<RoleId>>,
    },
    RtcRegion {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    RulesChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    ScheduledEndTime {
        old: Option<Timestamp>,
        new: Option<Timestamp>,
    },
    ScheduledStartTime {
        old: Option<Timestamp>,
        new: Option<Timestamp>,
    },
    SingleSelect {
        old: Option<bool>,
        new: Option<bool>,
    },
    SoundId {
        old: Option<SoundId>,
        new: Option<SoundId>,
    },
    SplashHash {
        old: Option<ImageHash>,
        new: Option<ImageHash>,
    },
    Status {
        old: Option<u64>,
        new: Option<u64>,
    },
    SystemChannelFlags {
        old: Option<SystemChannelFlags>,
        new: Option<SystemChannelFlags>,
    },
    SystemChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    Tags {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Temporary {
        old: Option<bool>,
        new: Option<bool>,
    },
    Title {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    Topic {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    TriggerMetadata {
        old: Option<TriggerMetadata>,
        new: Option<TriggerMetadata>,
    },
    TriggerType {
        old: Option<TriggerType>,
        new: Option<TriggerType>,
    },
    Type {
        old: Option<EntityType>,
        new: Option<EntityType>,
    },
    UnicodeEmoji {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    UserId {
        old: Option<UserId>,
        new: Option<UserId>,
    },
    UserLimit {
        old: Option<NonMaxU16>,
        new: Option<NonMaxU16>,
    },
    Uses {
        old: Option<u64>,
        new: Option<u64>,
    },
    VanityUrlCode {
        old: Option<FixedString>,
        new: Option<FixedString>,
    },
    VerificationEnabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    VerificationLevel {
        old: Option<VerificationLevel>,
        new: Option<VerificationLevel>,
    },
    VideoQualityMode {
        old: Option<VideoQualityMode>,
        new: Option<VideoQualityMode>,
    },
    Volume {
        old: Option<f64>,
        new: Option<f64>,
    },
    WidgetChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    WidgetEnabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    RolesAdded {
        old: Option<FixedArray<AffectedRole>>,
        new: Option<FixedArray<AffectedRole>>,
    },
    RolesRemoved {
        old: Option<FixedArray<AffectedRole>>,
        new: Option<FixedArray<AffectedRole>>,
    },
    CommandPermissions {
        old_value: Option<CommandPermission>,
        new_value: Option<CommandPermission>,
    },
    Other {
        key: FixedString,
        old_value: Option<Value>,
        new_value: Option<Value>,
    },

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Actions

Actions that execute when an auto moderation rule is triggered were changed.

Fields

old: Option<FixedArray<Action>>

new: Option<FixedArray<Action>>

AfkChannelId

AFK channel was changed.

Fields

old: Option<ChannelId>

new: Option<ChannelId>

AfkTimeout

AFK timeout duration was changed.

Fields

old: Option<AfkTimeout>

new: Option<AfkTimeout>

Allow

Allow field of a permission overwrite was changed.

Fields

old: Option<Permissions>

new: Option<Permissions>

ApplicationId

Id of the application associated with an entity was changed.

Fields

old: Option<ApplicationId>

new: Option<ApplicationId>

AppliedTags

Ids of the set of tags applied to a thread in a forum channel was changed.

Fields

old: Option<FixedArray<ForumTagId>>

new: Option<FixedArray<ForumTagId>>

Archived

Whether a thread is archived was changed.

Fields

old: Option<bool>

new: Option<bool>

Asset

Entity asset was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

AutoArchiveDuration

Auto archive duration of a thread was changed.

Fields

old: Option<u16>

new: Option<u16>

Available

Availability status was changed.

Fields

old: Option<bool>

new: Option<bool>

AvailableTags

Set of tags that can be used in a forum channel was changed.

Fields

old: Option<FixedArray<ForumTag>>

new: Option<FixedArray<ForumTag>>

AvatarHash

User or webhook avatar was changed.

Fields

old: Option<ImageHash>

new: Option<ImageHash>

BannerHash

Banner image was changed.

Fields

old: Option<ImageHash>

new: Option<ImageHash>

Bitrate

Voice channel bitrate was changed.

Fields

old: Option<u32>

new: Option<u32>

BrandColorPrimary

Primary color of a server profile banner was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

ChannelId

Id of the channel associated with an entity was changed.

Fields

old: Option<ChannelId>

new: Option<ChannelId>

Code

Invite code was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Color

Role color was changed.

Fields

old: Option<u32>

new: Option<u32>

Colors

Role colors were changed.

Fields

old: Option<RoleColours>

new: Option<RoleColours>

CommunicationDisabledUntil

Member timeout state was changed.

Fields

old: Option<Timestamp>

new: Option<Timestamp>

Deaf

whether a user is deafened in voice channels was changed.

Fields

old: Option<bool>

new: Option<bool>

DefaultAutoArchiveDuration

Default auto archive duration for newly created threads was changed.

Fields

old: Option<u16>

new: Option<u16>

DefaultChannelIds

Default channels for onboarding were changed.

Fields

old: Option<FixedArray<ChannelId>>

new: Option<FixedArray<ChannelId>>

DefaultMessageNotifications

Default message notification level for a server was changed.

Fields

old: Option<DefaultMessageNotificationLevel>

new: Option<DefaultMessageNotificationLevel>

DefaultReactionEmoji

Emoji to show in the add reaction button on a thread in a forum channel was changed.

Fields

old: Option<ForumEmoji>

new: Option<ForumEmoji>

DefaultThreadRateLimitPerUser

Initial rate limit per user to set on newly created threads in a channel was changed.

Fields

old: Option<u16>

new: Option<u16>

Deny

Deny field of a permission overwrite was changed.

Fields

old: Option<Permissions>

new: Option<Permissions>

Description

Description of an entity was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

DiscoverySplashHash

Guild’s discovery splash was changed.

Fields

old: Option<ImageHash>

new: Option<ImageHash>

EmojiId

Id of the emoji for a soundboard sound was changed.

Fields

old: Option<EmojiId>

new: Option<EmojiId>

EmojiName

Unicode character of the emoji for a soundboard sound was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Enabled

Enabled status was changed.

Fields

old: Option<bool>

new: Option<bool>

EnableEmoticons

Whether emoticons should be synced for an integration was changed.

Fields

old: Option<bool>

new: Option<bool>

EntityType

Entity type of a scheduled event was changed.

Fields

old: Option<u64>

new: Option<u64>

EventType

Event type of an auto moderation rule was changed.

Fields

old: Option<AutomodEventType>

new: Option<AutomodEventType>

ExemptChannels

Channels not affected by an auto moderation rule were changed.

Fields

old: Option<FixedArray<ChannelId>>

new: Option<FixedArray<ChannelId>>

ExemptRoles

Roles not affected by an auto moderation rule were changed.

Fields

old: Option<FixedArray<RoleId>>

new: Option<FixedArray<RoleId>>

ExpireBehavior

Behavior of expiring subscribers for an integration was changed.

Fields

old: Option<u64>

new: Option<u64>

ExpireGracePeriod

Grace period before expiring subscribers for an integration was changed.

Fields

old: Option<u64>

new: Option<u64>

ExplicitContentFilter

Explicit content filter level of a guild was changed.

Fields

old: Option<ExplicitContentFilter>

new: Option<ExplicitContentFilter>

Flags

Flags of an entity were changed.

Fields

old: Option<u64>

new: Option<u64>

FormatType

Format type of a sticker was changed.

Fields

old: Option<StickerFormatType>

new: Option<StickerFormatType>

GameApplicationIds

Ids of games included in a server profile were changed.

Fields

old: Option<FixedArray<ApplicationId>>

new: Option<FixedArray<ApplicationId>>

GuildId

Id of the guild associated with an entity was changed.

Fields

old: Option<GuildId>

new: Option<GuildId>

Hoist

Whether a role is pinned in the user listing was changed.

Fields

old: Option<bool>

new: Option<bool>

IconHash

Guild or role icon was changed.

Fields

old: Option<ImageHash>

new: Option<ImageHash>

Id

Id of an entity was changed.

Fields

old: Option<GenericId>

new: Option<GenericId>

ImageHash

Cover image of a scheduled event was changed.

Fields

old: Option<ImageHash>

new: Option<ImageHash>

InOnboarding

Whether a prompt is present in an onboarding flow was changed.

Fields

old: Option<bool>

new: Option<bool>

Invitable

Private thread’s invitable state was changed.

Fields

old: Option<bool>

new: Option<bool>

InviterId

Id of the user who created an invite was changed.

Fields

old: Option<UserId>

new: Option<UserId>

Location

Location for a scheduled event was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Locked

Locked status of a thread was changed.

Fields

old: Option<bool>

new: Option<bool>

ManualApprovalEnabled

Whether users must apply to join a guild was changed.

Fields

old: Option<bool>

new: Option<bool>

MaxAge

How long an invite code lasts was changed.

Fields

old: Option<u32>

new: Option<u32>

MaxUses

Maximum uses of an invite was changed.

Fields

old: Option<u8>

new: Option<u8>

Mentionable

Whether a role can be mentioned in a message was changed.

Fields

old: Option<bool>

new: Option<bool>

MfaLevel

Multi-factor authentication requirement was changed.

Fields

old: Option<MfaLevel>

new: Option<MfaLevel>

Mute

Whether a user is server muted was changed.

Fields

old: Option<bool>

new: Option<bool>

Name

Name of an entity was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Nick

Nickname of a member was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Nsfw

Whether a channel is age-restricted was changed.

Fields

old: Option<bool>

new: Option<bool>

OwnerId

Owner of a guild was changed.

Fields

old: Option<UserId>

new: Option<UserId>

PermissionOverwrites

Permissions on a channel were changed.

Fields

old: Option<FixedArray<PermissionOverwrite>>

new: Option<FixedArray<PermissionOverwrite>>

Permissions

Permissions for an entity were changed.

Fields

old: Option<Permissions>

new: Option<Permissions>

Position

Channel or role position was changed.

Fields

old: Option<u32>

new: Option<u32>

PreferredLocale

Preferred locale of a guild was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

PremiumProgressBarEnabled

Whether a guild has the boost progress bar enabled was changed.

Fields

old: Option<bool>

new: Option<bool>

PrivacyLevel

Privacy level of a stage instance was changed.

Fields

old: Option<u64>

new: Option<u64>

PruneDeleteDays

Number of days after which inactive and role-unassigned members are kicked was changed.

Fields

old: Option<u64>

new: Option<u64>

PublicUpdatesChannelId

Id of a public updates channel was changed.

Fields

old: Option<ChannelId>

new: Option<ChannelId>

RateLimitPerUser

Rate limit per user in a text channel was changed.

Fields

old: Option<u16>

new: Option<u16>

Region

Region of a guild was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Required

Whether an onboarding prompt is required was changed.

Fields

old: Option<bool>

new: Option<bool>

RoleIds

Roles assigned to a user upon accepting an invite were changed.

Fields

old: Option<FixedArray<RoleId>>

new: Option<FixedArray<RoleId>>

RtcRegion

Voice region Id for a voice or stage channel was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

RulesChannelId

Id of a rules channel was changed.

Fields

old: Option<ChannelId>

new: Option<ChannelId>

ScheduledEndTime

End time of a scheduled event was changed.

Fields

old: Option<Timestamp>

new: Option<Timestamp>

ScheduledStartTime

Start time of a scheduled event was changed.

Fields

old: Option<Timestamp>

new: Option<Timestamp>

SingleSelect

Whether only one option can be selected for an onboarding prompt was changed.

Fields

old: Option<bool>

new: Option<bool>

SoundId

Id of a soundboard sound was changed.

Fields

old: Option<SoundId>

new: Option<SoundId>

SplashHash

Guild splash image was changed.

Fields

old: Option<ImageHash>

new: Option<ImageHash>

Status

Status of a scheduled event was changed.

Fields

old: Option<u64>

new: Option<u64>

SystemChannelFlags

System channel settings were changed.

Fields

old: Option<SystemChannelFlags>

new: Option<SystemChannelFlags>

SystemChannelId

Id of a system channel was changed.

Fields

old: Option<ChannelId>

new: Option<ChannelId>

Tags

Autocomplete/suggestion tags (related emoji) of a sticker were changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Temporary

Whether an invite only grants temporary membership was changed.

Fields

old: Option<bool>

new: Option<bool>

Title

Title of an entity was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

Topic

Topic of a text channel or stage instance was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

TriggerMetadata

Trigger metadata of an auto moderation rule was changed.

Fields

old: Option<TriggerMetadata>

new: Option<TriggerMetadata>

TriggerType

Trigger type of an auto moderation rule was changed.

Fields

old: Option<TriggerType>

new: Option<TriggerType>

Type

Type of an entity was changed.

Fields

old: Option<EntityType>

new: Option<EntityType>

UnicodeEmoji

Unicode emoji of a role icon was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

UserId

Id of the user associated with an entity was changed.

Fields

old: Option<UserId>

new: Option<UserId>

UserLimit

User limit of a voice channel was changed.

Fields

old: Option<NonMaxU16>

new: Option<NonMaxU16>

Uses

Number of times an invite has been used was changed.

Fields

old: Option<u64>

new: Option<u64>

VanityUrlCode

Guild invite vanity url was changed.

Fields

old: Option<FixedString>

new: Option<FixedString>

VerificationEnabled

Whether server rules are enabled was changed.

Fields

old: Option<bool>

new: Option<bool>

VerificationLevel

Required verification level for new members was changed.

Fields

old: Option<VerificationLevel>

new: Option<VerificationLevel>

VideoQualityMode

Video quality mode for a voice channel was changed.

Fields

old: Option<VideoQualityMode>

new: Option<VideoQualityMode>

Volume

Volume of a soundboard sound was changed.

Fields

old: Option<f64>

new: Option<f64>

WidgetChannelId

Channel of a server widget was changed.

Fields

old: Option<ChannelId>

new: Option<ChannelId>

WidgetEnabled

Whether a server widget is enabled was changed.

Fields

old: Option<bool>

new: Option<bool>

RolesAdded

Role was added to a member.

Fields

old: Option<FixedArray<AffectedRole>>

new: Option<FixedArray<AffectedRole>>

RolesRemoved

Role was removed from a member.

Fields

old: Option<FixedArray<AffectedRole>>

new: Option<FixedArray<AffectedRole>>

CommandPermissions

Permissions were updated for a command.

Fields

old_value: Option<CommandPermission>

new_value: Option<CommandPermission>

Other

Unknown key was changed.

Fields

key: FixedString

old_value: Option<Value>

new_value: Option<Value>

Implementations

impl Change

pub fn key(&self) -> FixedString