Struct serenity::model::guild::Emoji

#[non_exhaustive]
pub struct Emoji {
    pub animated: bool,
    pub available: bool,
    pub id: EmojiId,
    pub name: String,
    pub managed: bool,
    pub require_colons: bool,
    pub roles: Vec<RoleId>,
    pub user: Option<User>,
}

Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

animated: bool

Whether the emoji is animated.

available: bool

Whether the emoji can be used. This may be false when the guild loses boosts, reducing the emoji limit.

id: EmojiId

The Id of the emoji.

name: String

The name of the emoji. It must be at least 2 characters long and can only contain alphanumeric characters and underscores.

managed: bool

Whether the emoji is managed via an Integration service.

require_colons: bool

Whether the emoji name needs to be surrounded by colons in order to be used by the client.

roles: Vec<RoleId>

A list of Roles that are allowed to use the emoji. If there are no roles specified, then usage is unrestricted.

user: Option<User>

The user who created the emoji.

Implementations

impl Emoji

pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<()>

Available on crate features model and cache only.

Deletes the emoji. This method requires the cache to fetch the guild ID.

Note: The Manage Emojis and Stickers permission is required.

Examples

Delete a given emoji:

// assuming emoji has been set already
match emoji.delete(&ctx).await {
    Ok(()) => println!("Emoji deleted."),
    Err(_) => println!("Could not delete emoji."),
}

Errors

Returns Error::Http if the current user lacks permission, or may return ModelError::ItemMissing if the emoji is not in the cache.

pub async fn edit( &mut self, cache_http: impl CacheHttp, name: &str ) -> Result<()>

Available on crate features model and cache only.

Edits the emoji by updating it with a new name. This method requires the cache to fetch the guild ID.

Note: The Manage Emojis and Stickers permission is required.

Errors

Returns Error::Http if the current user lacks permission, or if an invalid name is given.

pub fn find_guild_id(&self, cache: impl AsRef<Cache>) -> Option<GuildId>

Available on crate features model and cache only.

Finds the Guild that owns the emoji by looking through the Cache.

Examples

Print the guild id that owns this emoji:

// assuming emoji has been set already
if let Some(guild_id) = emoji.find_guild_id(&cache) {
    println!("{} is owned by {}", emoji.name, guild_id);
}

pub fn url(&self) -> String

Available on crate feature model only.

Generates a URL to the emoji’s image.

Examples

Print the direct link to the given emoji:

// assuming emoji has been set already
println!("Direct link to emoji image: {}", emoji.url());