Struct serenity::model::id::UserId

pub struct UserId(pub u64);

An identifier for a User

Implementations

impl UserId

pub fn created_at(&self) -> DateTime<Utc>

Retrieves the time that the Id was created at.

pub fn as_u64(&self) -> &u64

Immutably borrow inner Id.

pub fn as_mut_u64(&mut self) -> &mut u64

Mutably borrow inner Id.

impl UserId

pub async fn create_dm_channel(
self,
cache_http: impl CacheHttp
) -> Result<PrivateChannel>

Creates a direct message channel between the current user and the user. This can also retrieve the channel if one already exists.

Errors

Returns Error::Http if a User with that UserId does not exist, or is otherwise unavailable.

May also return an Error::Json if there is an error in deserializing the channel data returned by the Discord API.

pub async fn to_user_cached(self, cache: impl AsRef<Cache>) -> Option<User>

Attempts to find a User by its Id in the cache.

pub async fn to_user(self, cache_http: impl CacheHttp) -> Result<User>

First attempts to find a User by its Id in the cache, upon failure requests it via the REST API.

Note: If the cache is not enabled, REST API will be used only.

Errors

May return an Error::Http if a User with that UserId does not exist, or otherwise cannot be fetched.

May also return an Error::Json if there is an error in deserializing the user.