Module serenity::model::interactions

This is supported on crate feature unstable_discord_api only.

Interactions information-related models.

Structs

ActionRow

An action row.

ApplicationCommand

The base command model that belongs to an application.

ApplicationCommandInteractionData

The command data payload.

ApplicationCommandInteractionDataOption

A set of a parameter and a value from the user.

ApplicationCommandInteractionDataResolved

The resolved data of a command data interaction payload. It contains the objects of ApplicationCommandInteractionDataOptions.

ApplicationCommandOption

The parameters for an ApplicationCommand.

ApplicationCommandOptionChoice

The only valid values a user can pick in an ApplicationCommandOption.

ApplicationCommandPermission

An ApplicationCommand permission.

ApplicationCommandPermissionData

The ApplicationCommandPermission data.

Button

A button component.

EphemeralMessage

An ephemeral message given in an interaction.

Interaction

Information about an interaction.

InteractionApplicationCommandCallbackDataFlags

The flags for an interaction response.

MessageComponent

A message component data, provided by Interaction::data

MessageInteraction

Sent when a Message is a response to an Interaction.

SelectMenu

A select menu component.

SelectMenuOption

A select menu component options.

Enums

ApplicationCommandInteractionDataOptionValue

The resolved value of an ApplicationCommandInteractionDataOption.

ApplicationCommandOptionType

The type of an ApplicationCommandOption.

ApplicationCommandPermissionType

The type of an ApplicationCommandPermissionData.

ButtonStyle

The style of a button.

Component

A component.

ComponentType

The type of a component

InteractionData

The Interaction::data field.

InteractionMessage

The Interaction::message field.

InteractionResponseType

The available responses types for an interaction response.

InteractionType

The type of an Interaction