Type Alias serenity::model::interactions::application_command::ApplicationCommand

pub type ApplicationCommand = Command;

๐Deprecated: use model::application::command::Command

The base command model that belongs to an application.

Aliased Type

struct ApplicationCommand {

    pub id: CommandId,
    pub kind: CommandType,
    pub application_id: ApplicationId,
    pub guild_id: Option<GuildId>,
    pub name: String,
    pub name_localized: Option<String>,
    pub name_localizations: Option<HashMap<String, String, RandomState>>,
    pub description: String,
    pub description_localized: Option<String>,
    pub description_localizations: Option<HashMap<String, String, RandomState>>,
    pub options: Vec<CommandOption, Global>,
    pub default_member_permissions: Option<Permissions>,
    pub dm_permission: Option<bool>,
    pub default_permission: bool,
    pub version: CommandVersionId,

}

Fields
id: CommandId

The command Id.

kind: CommandType

The application command kind.

application_id: ApplicationId

The parent application Id.

guild_id: Option<GuildId>

The command guild Id, if it is a guild command.

Note: It may only be present if it is received through the gateway.

name: String

The command name.

name_localized: Option<String>

The localized command name of the selected locale.

If the name is localized, either this field or Self::name_localizations is set, depending on which endpoint this data was retrieved from (source).

name_localizations: Option<HashMap<String, String, RandomState>>

All localized command names.

If the name is localized, either this field or Self::name_localized is set, depending on which endpoint this data was retrieved from (source).

description: String

The command description.

description_localized: Option<String>

The localized command description of the selected locale.

If the description is localized, either this field or Self::description_localizations is set, depending on which endpoint this data was retrieved from (source).

description_localizations: Option<HashMap<String, String, RandomState>>

All localized command descriptions.

If the description is localized, either this field or Self::description_localized is set, depending on which endpoint this data was retrieved from (source).

options: Vec<CommandOption, Global>

The parameters for the command.

default_member_permissions: Option<Permissions>

The default permissions required to execute the command.

dm_permission: Option<bool>

Indicates whether the command is available in DMs with the app, only for globally-scoped commands. By default, commands are visible.

default_permission: bool
๐Deprecated: replaced by default_member_permissions

Whether the command is enabled by default when the application is added to a guild.

version: CommandVersionId

An autoincremented version identifier updated during substantial record changes.

Implementations

impl Command

pub async fn create_global_application_command<F>( http: impl AsRef<Http>, f: F ) -> Result<Command>where F: FnOnce(&mut CreateApplicationCommand) -> &mut CreateApplicationCommand,

Creates a global Command, overriding an existing one with the same name if it exists.

When a created Command is used, the InteractionCreate event will be emitted.

Note: Global commands may take up to an hour to be updated in the user slash commands list. If an outdated command data is sent by a user, discord will consider it as an error and then will instantly update that command.

As such, it is recommended that guild application commands be used for testing purposes.

Examples

Create a simple ping command:

use serenity::model::application::command::Command;
use serenity::model::id::ApplicationId;

let _ = Command::create_global_application_command(&http, |command| {
    command.name("ping").description("A simple ping command")
})
.await;

Create a command that echoes what is inserted:

use serenity::model::application::command::{Command, CommandOptionType};
use serenity::model::id::ApplicationId;

let _ = Command::create_global_application_command(&http, |command| {
    command.name("echo").description("Makes the bot send a message").create_option(|option| {
        option
            .name("message")
            .description("The message to send")
            .kind(CommandOptionType::String)
            .required(true)
    })
})
.await;

Errors

May return an Error::Http if the Command is illformed, such as if more than 10 choices are set. See the API Docs for further details.

Can also return an Error::Json if there is an error in deserializing the response.

pub async fn set_global_application_commands<F>( http: impl AsRef<Http>, f: F ) -> Result<Vec<Command>>where F: FnOnce(&mut CreateApplicationCommands) -> &mut CreateApplicationCommands,

Overrides all global application commands.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn edit_global_application_command<F>( http: impl AsRef<Http>, command_id: CommandId, f: F ) -> Result<Command>where F: FnOnce(&mut CreateApplicationCommand) -> &mut CreateApplicationCommand,

Edits a global command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_global_application_commands( http: impl AsRef<Http> ) -> Result<Vec<Command>>

Gets all global commands.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_global_application_commands_with_localizations( http: impl AsRef<Http> ) -> Result<Vec<Command>>

Gets all global commands with localizations.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_global_application_command( http: impl AsRef<Http>, command_id: CommandId ) -> Result<Command>

Gets a global command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn delete_global_application_command( http: impl AsRef<Http>, command_id: CommandId ) -> Result<()>

Deletes a global command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.