Type Alias serenity::model::interactions::application_command::ApplicationCommandInteractionDataOption

pub type ApplicationCommandInteractionDataOption = CommandDataOption;

👎Deprecated: use model::application::interaction::application_command::CommandDataOption

A set of a parameter and a value from the user.

All options have names and an option can either be a parameter and input value or it can denote a sub-command or group, in which case it will contain a top-level key and another vector of options.

Their resolved objects can be found on ApplicationCommandInteractionData::resolved.

Aliased Type

struct ApplicationCommandInteractionDataOption {
    pub name: String,
    pub value: Option<Value>,
    pub kind: CommandOptionType,
    pub options: Vec<CommandDataOption, Global>,
    pub resolved: Option<CommandDataOptionValue>,
    pub focused: bool,
}

Fields
name: String

The name of the parameter.

value: Option<Value>

The given value.

kind: CommandOptionType

The value type.

options: Vec<CommandDataOption, Global>

The nested options.

Note: It is only present if the option is a group or a subcommand.

resolved: Option<CommandDataOptionValue>

The resolved object of the given value, if there is one.

focused: bool

For Autocomplete Interactions this will be true if this option is currently focused by the user.