Type Alias serenity::model::interactions::application_command::ApplicationCommandOption

pub type ApplicationCommandOption = CommandOption;

👎Deprecated: use model::application::command::CommandOption

The parameters for an ApplicationCommand.

Aliased Type

struct ApplicationCommandOption {

    pub kind: CommandOptionType,
    pub name: String,
    pub name_localizations: Option<HashMap<String, String>>,
    pub description: String,
    pub description_localizations: Option<HashMap<String, String>>,
    pub required: bool,
    pub choices: Vec<CommandOptionChoice>,
    pub options: Vec<CommandOption>,
    pub channel_types: Vec<ChannelType>,
    pub min_value: Option<Number>,
    pub max_value: Option<Number>,
    pub min_length: Option<u16>,
    pub max_length: Option<u16>,
    pub autocomplete: bool,

}

Fields
kind: CommandOptionType

The option type.

name: String

The option name.

name_localizations: Option<HashMap<String, String>>

Localizations of the option name with locale as the key

description: String

The option description.

description_localizations: Option<HashMap<String, String>>

Localizations of the option description with locale as the key

required: bool

Whether the parameter is optional or required.

choices: Vec<CommandOptionChoice>

Choices the user can pick from.

Note: Only available for String and Integer CommandOptionType.

options: Vec<CommandOption>

The nested options.

Note: Only available for SubCommand or SubCommandGroup.

channel_types: Vec<ChannelType>

If the option is a Channel, it will only be able to show these types.

min_value: Option<Number>

Minimum permitted value for Integer or Number options

max_value: Option<Number>

Maximum permitted value for Integer or Number options

min_length: Option<u16>

Minimum permitted length for String options

max_length: Option<u16>

Maximum permitted length for String options

autocomplete: bool