Type Alias serenity::model::interactions::autocomplete::AutocompleteInteraction

pub type AutocompleteInteraction = AutocompleteInteraction;

👎Deprecated: use model::application::interaction::autocomplete::AutocompleteInteraction

An interaction received when the user fills in an autocomplete option

Aliased Type

struct AutocompleteInteraction {

    pub id: InteractionId,
    pub application_id: ApplicationId,
    pub kind: InteractionType,
    pub data: CommandData,
    pub guild_id: Option<GuildId>,
    pub channel_id: ChannelId,
    pub member: Option<Member>,
    pub user: User,
    pub token: String,
    pub version: u8,
    pub app_permissions: Option<Permissions>,
    pub locale: String,
    pub guild_locale: Option<String>,

}

Fields
id: InteractionId

Id of the interaction.

application_id: ApplicationId

Id of the application this interaction is for.

kind: InteractionType

The type of interaction.

data: CommandData

The data of the interaction which was triggered.

guild_id: Option<GuildId>

The guild Id this interaction was sent from, if there is one.

channel_id: ChannelId

The channel Id this interaction was sent from.

member: Option<Member>

The member data for the invoking user.

Note: It is only present if the interaction is triggered in a guild.

user: User

The user object for the invoking user.

token: String

A continuation token for responding to the interaction.

version: u8

Always 1.

app_permissions: Option<Permissions>

Permissions the app or bot has within the channel the interaction was sent from.

locale: String

The selected language of the invoking user.

guild_locale: Option<String>

The guild’s preferred locale.

Implementations

impl AutocompleteInteraction

pub async fn create_autocomplete_response<F>( &self, http: impl AsRef<Http>, f: F ) -> Result<()>where F: FnOnce(&mut CreateAutocompleteResponse) -> &mut CreateAutocompleteResponse,

Creates a response to an autocomplete interaction.

Errors

Returns an Error::Http if the API returns an error.