serenity::model
Module monetization

Structs

Entitlement

Represents that a user or guild has access to a premium offering in the application.

EntitlementKind

Differentiates between Entitlement types.

Sku

A premium offering that can be made available to an application’s users and guilds.

SkuFlags

Differentates between user and server subscriptions.

SkuKind

Differentiates between SKU classes.

Enums

EntitlementOwner