Enum serenity::model::oauth2::OAuth2Scope

#[non_exhaustive]
pub enum OAuth2Scope {

    Bot,
    ApplicationsCommands,
    ApplicationsCommandsUpdate,
    Identify,
    Email,
    Connections,
    Guilds,
    GuildsJoin,
    GdmJoin,
    Rpc,
    RpcNotificationsRead,
    RpcVoiceRead,
    RpcVoiceWrite,
    RpcActivitiesWrite,
    WebhookIncomming,
    MessagesRead,
    ApplicationsBuildsUpload,
    ApplicationsBuildsRead,
    ApplicationsStoreUpdate,
    ApplicationsEntitlements,
    ActivitiesRead,
    ActivitiesWrite,
    RelactionshipsRead,

}

The available OAuth2 Scopes.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Bot

For oauth2 bots, this puts the bot in the user’s selected guild by default.

ApplicationsCommands

Allows your app to use Slash Commands in a guild.

ApplicationsCommandsUpdate

Allows your app to update its Slash Commands via this bearer token - client credentials grant only.

Identify

Allows /users/@me without Self::Email.

Email

Enables /users/@me to return an email field.

Connections

Allows /users/@me/connections to return linked third-party accounts.

Guilds

Allows /users/@me/guilds to return basic information about all of a user’s guilds.

GuildsJoin

Allows /guilds/{guild.id}/members/{user.id} to be used for joining users to a guild.

GdmJoin

Allows your app to join users to a group dm.

Rpc

For local rpc server access, this allows you to control a user’s local Discord client - requires Discord approval.

RpcNotificationsRead

For local rpc server api access, this allows you to receive notifications pushed out to the user - requires Discord approval.

RpcVoiceRead

RpcVoiceWrite

RpcActivitiesWrite

WebhookIncomming

This generates a webhook that is returned in the oauth token response for authorization code grants.

MessagesRead

For local rpc server api access, this allows you to read messages from all client channels (otherwise restricted to channels/guilds your app creates).

ApplicationsBuildsUpload

Allows your app to upload/update builds for a user’s applications - requires Discord approval.

ApplicationsBuildsRead

Allows your app to read build data for a user’s applications.

ApplicationsStoreUpdate

Allows your app to read and update store data (SKUs, store listings, achievements, etc.) for a user’s applications.

ApplicationsEntitlements

Allows your app to read entitlements for a user’s applications.

ActivitiesRead

Allows your app to fetch data from a user’s “Now Playing/Recently Played” list - requires Discord approval.

ActivitiesWrite

allows your app to update a user’s activity - requires Discord approval (Not required for gamesdk activity manager!).

RelactionshipsRead

Allows your app to know a user’s friends and implicit relationships - requires Discord approval.