Struct serenity::model::permissions::Permissions

pub struct Permissions { /* fields omitted */ }

A set of permissions that can be assigned to Users and Roles via PermissionOverwrites, roles globally in a Guild, and to GuildChannels.

Methods

impl Permissions

const CREATE_INVITE: Permissions

CREATE_INVITE: Permissions = Permissions{bits: 1,}

Allows for the creation of RichInvites.

const KICK_MEMBERS: Permissions

KICK_MEMBERS: Permissions = Permissions{bits: 2,}

Allows for the kicking of guild members.

const BAN_MEMBERS: Permissions

BAN_MEMBERS: Permissions = Permissions{bits: 4,}

Allows the banning of guild members.

const ADMINISTRATOR: Permissions

ADMINISTRATOR: Permissions = Permissions{bits: 8,}

Allows all permissions, bypassing channel permission overwrites.

const MANAGE_CHANNELS: Permissions

MANAGE_CHANNELS: Permissions = Permissions{bits: 16,}

Allows management and editing of guild channels.

const MANAGE_GUILD: Permissions

MANAGE_GUILD: Permissions = Permissions{bits: 32,}

Allows management and editing of the guild.

const ADD_REACTIONS: Permissions

ADD_REACTIONS: Permissions = Permissions{bits: 64,}

Members with this permission can add new Reactions to a Message. Members can still react using reactions already added to messages without this permission.

const READ_MESSAGES: Permissions

READ_MESSAGES: Permissions = Permissions{bits: 1024,}

Allows reading messages in a guild channel. If a user does not have this permission, then they will not be able to see the channel.

const SEND_MESSAGES: Permissions

SEND_MESSAGES: Permissions = Permissions{bits: 2048,}

Allows sending messages in a guild channel.

const SEND_TTS_MESSAGES: Permissions

SEND_TTS_MESSAGES: Permissions = Permissions{bits: 4096,}

Allows the sending of text-to-speech messages in a channel.

const MANAGE_MESSAGES: Permissions

MANAGE_MESSAGES: Permissions = Permissions{bits: 8192,}

Allows the deleting of other messages in a guild channel.

Note: This does not allow the editing of other messages.

const EMBED_LINKS: Permissions

EMBED_LINKS: Permissions = Permissions{bits: 16384,}

Allows links from this user - or users of this role - to be embedded, with potential data such as a thumbnail, description, and page name.

const ATTACH_FILES: Permissions

ATTACH_FILES: Permissions = Permissions{bits: 32768,}

Allows uploading of files.

const READ_MESSAGE_HISTORY: Permissions

READ_MESSAGE_HISTORY: Permissions = Permissions{bits: 65536,}

Allows the reading of a channel's message history.

const MENTION_EVERYONE: Permissions

MENTION_EVERYONE: Permissions = Permissions{bits: 131072,}

Allows the usage of the @everyone mention, which will notify all users in a channel. The @here mention will also be available, and can be used to mention all non-offline users.

Note: You probably want this to be disabled for most roles and users.

const USE_EXTERNAL_EMOJIS: Permissions

USE_EXTERNAL_EMOJIS: Permissions = Permissions{bits: 262144,}

Allows the usage of custom emojis from other guilds.

This does not dictate whether custom emojis in this guild can be used in other guilds.

const CONNECT: Permissions

CONNECT: Permissions = Permissions{bits: 1048576,}

Allows the joining of a voice channel.

const SPEAK: Permissions

SPEAK: Permissions = Permissions{bits: 2097152,}

Allows the user to speak in a voice channel.

const MUTE_MEMBERS: Permissions

MUTE_MEMBERS: Permissions = Permissions{bits: 4194304,}

Allows the muting of members in a voice channel.

const DEAFEN_MEMBERS: Permissions

DEAFEN_MEMBERS: Permissions = Permissions{bits: 8388608,}

Allows the deafening of members in a voice channel.

const MOVE_MEMBERS: Permissions

MOVE_MEMBERS: Permissions = Permissions{bits: 16777216,}

Allows the moving of members from one voice channel to another.

const USE_VAD: Permissions

USE_VAD: Permissions = Permissions{bits: 33554432,}

Allows the usage of voice-activity-detection in a voice channel.

If this is disabled, then Members must use push-to-talk.

const CHANGE_NICKNAME: Permissions

CHANGE_NICKNAME: Permissions = Permissions{bits: 67108864,}

Allows members to change their own nickname in the guild.

const MANAGE_NICKNAMES: Permissions

MANAGE_NICKNAMES: Permissions = Permissions{bits: 134217728,}

Allows members to change other members' nicknames.

const MANAGE_ROLES: Permissions

MANAGE_ROLES: Permissions = Permissions{bits: 268435456,}

Allows management and editing of roles below their own.

const MANAGE_WEBHOOKS: Permissions

MANAGE_WEBHOOKS: Permissions = Permissions{bits: 536870912,}

Allows management of webhooks.

const MANAGE_EMOJIS: Permissions

MANAGE_EMOJIS: Permissions = Permissions{bits: 1073741824,}

Allows management of emojis created without the use of an Integration.

fn empty() -> Permissions

Returns an empty set of flags.

fn all() -> Permissions

Returns the set containing all flags.

fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

fn from_bits(bits: u64) -> Option<Permissions>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: u64) -> Permissions

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Permissions) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: Permissions) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Permissions)

Inserts the specified flags in-place.

fn remove(&mut self, other: Permissions)

Removes the specified flags in-place.

fn toggle(&mut self, other: Permissions)

Toggles the specified flags in-place.

fn set(&mut self, other: Permissions, value: bool)

Inserts or removes the specified flags depending on the passed value.

impl Permissions

fn add_reactions(&self) -> bool

Shorthand for checking that the set of permissions contains the Add Reactions permission.

fn administrator(&self) -> bool

Shorthand for checking that the set of permissions contains the Administrator permission.

fn attach_files(&self) -> bool

Shorthand for checking that the set of permissions contains the Attach Files permission.

fn ban_members(&self) -> bool

Shorthand for checking that the set of permissions contains the Ban Members permission.

fn change_nickname(&self) -> bool

Shorthand for checking that the set of permissions contains the Change Nickname permission.

fn connect(&self) -> bool

Shorthand for checking that the set of permissions contains the Connect permission.

fn create_invite(&self) -> bool

Shorthand for checking that the set of permissions contains the Create Invite permission.

fn deafen_members(&self) -> bool

Shorthand for checking that the set of permissions contains the Deafen Members permission.

fn embed_links(&self) -> bool

Shorthand for checking that the set of permissions contains the Embed Links permission.

fn external_emojis(&self) -> bool

Shorthand for checking that the set of permissions contains the Use External Emojis permission.

fn kick_members(&self) -> bool

Shorthand for checking that the set of permissions contains the Kick Members permission.

fn manage_channels(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Channels permission.

fn manage_emojis(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Emojis permission.

fn manage_guild(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Guild permission.

fn manage_messages(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Messages permission.

fn manage_nicknames(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Nicknames permission.

fn manage_roles(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Roles permission.

fn manage_webhooks(&self) -> bool

Shorthand for checking that the set of permissions contains the Manage Webhooks permission.

fn mention_everyone(&self) -> bool

Shorthand for checking that the set of permissions contains the Mention Everyone permission.

fn move_members(&self) -> bool

Shorthand for checking that the set of permissions contains the Move Members permission.

fn mute_members(&self) -> bool

Shorthand for checking that the set of permissions contains the Mute Members permission.

fn read_message_history(&self) -> bool

Shorthand for checking that the set of permissions contains the Read Message History permission.

fn read_messages(&self) -> bool

Shorthand for checking that the set of permissions contains the Read Messages permission.

fn send_messages(&self) -> bool

Shorthand for checking that the set of permissions contains the Send Messages permission.

fn send_tts_messages(&self) -> bool

Shorthand for checking that the set of permissions contains the Send TTS Messages permission.

fn speak(&self) -> bool

Shorthand for checking that the set of permissions contains the Speak permission.

fn use_external_emojis(&self) -> bool

Shorthand for checking that the set of permissions contains the Use External Emojis permission.

fn use_vad(&self) -> bool

Shorthand for checking that the set of permissions contains the Use VAD permission.