Enum serenity::model::prelude::application::CommandDataOptionValue

#[non_exhaustive]
pub enum CommandDataOptionValue {

    Autocomplete {
        kind: CommandOptionType,
        value: String,
    },
    Boolean(bool),
    Integer(i64),
    Number(f64),
    String(String),
    SubCommand(Vec<CommandDataOption>),
    SubCommandGroup(Vec<CommandDataOption>),
    Attachment(AttachmentId),
    Channel(ChannelId),
    Mentionable(GenericId),
    Role(RoleId),
    User(UserId),
    Unknown(u8),

}

The value of an CommandDataOption.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Autocomplete

Fields

kind: CommandOptionType

value: String

Boolean(bool)

Integer(i64)

Number(f64)

String(String)

SubCommand(Vec<CommandDataOption>)

SubCommandGroup(Vec<CommandDataOption>)

Attachment(AttachmentId)

Channel(ChannelId)

Mentionable(GenericId)

Role(RoleId)

User(UserId)

Unknown(u8)

Implementations

impl CommandDataOptionValue

pub fn kind(&self) -> CommandOptionType

pub fn as_bool(&self) -> Option<bool>

If the value is a boolean, returns the associated f64. Returns None otherwise.

pub fn as_i64(&self) -> Option<i64>

If the value is an integer, returns the associated f64. Returns None otherwise.

pub fn as_f64(&self) -> Option<f64>

If the value is a number, returns the associated f64. Returns None otherwise.

pub fn as_str(&self) -> Option<&str>

If the value is a string, returns the associated str. Returns None otherwise.

pub fn as_attachment_id(&self) -> Option<AttachmentId>

If the value is an AttachmentId, returns the associated ID. Returns None otherwise.

pub fn as_channel_id(&self) -> Option<ChannelId>

If the value is an ChannelId, returns the associated ID. Returns None otherwise.

pub fn as_mentionable(&self) -> Option<GenericId>

If the value is an GenericId, returns the associated ID. Returns None otherwise.

pub fn as_user_id(&self) -> Option<UserId>

If the value is an UserId, returns the associated ID. Returns None otherwise.

pub fn as_role_id(&self) -> Option<RoleId>

If the value is an RoleId, returns the associated ID. Returns None otherwise.