Enum serenity::model::prelude::audit_log::Change

#[non_exhaustive]
pub enum Change {

    Actions {
        old: Option<Vec<Action>>,
        new: Option<Vec<Action>>,
    },
    AfkChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    AfkTimeout {
        old: Option<u64>,
        new: Option<u64>,
    },
    Allow {
        old: Option<Permissions>,
        new: Option<Permissions>,
    },
    ApplicationId {
        old: Option<ApplicationId>,
        new: Option<ApplicationId>,
    },
    Archived {
        old: Option<bool>,
        new: Option<bool>,
    },
    Asset {
        old: Option<String>,
        new: Option<String>,
    },
    AutoArchiveDuration {
        old: Option<u64>,
        new: Option<u64>,
    },
    Available {
        old: Option<bool>,
        new: Option<bool>,
    },
    AvatarHash {
        old: Option<String>,
        new: Option<String>,
    },
    BannerHash {
        old: Option<String>,
        new: Option<String>,
    },
    Bitrate {
        old: Option<u32>,
        new: Option<u32>,
    },
    ChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    Code {
        old: Option<String>,
        new: Option<String>,
    },
    Color {
        old: Option<u64>,
        new: Option<u64>,
    },
    CommunicationDisabledUntil {
        old: Option<Timestamp>,
        new: Option<Timestamp>,
    },
    Deaf {
        old: Option<bool>,
        new: Option<bool>,
    },
    DefaultAutoArchiveDuration {
        old: Option<u64>,
        new: Option<u64>,
    },
    DefaultMessageNotifications {
        old: Option<DefaultMessageNotificationLevel>,
        new: Option<DefaultMessageNotificationLevel>,
    },
    Deny {
        old: Option<Permissions>,
        new: Option<Permissions>,
    },
    Description {
        old: Option<String>,
        new: Option<String>,
    },
    DiscoverySplashHash {
        old: Option<String>,
        new: Option<String>,
    },
    Enabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    EnableEmoticons {
        old: Option<bool>,
        new: Option<bool>,
    },
    EntityType {
        old: Option<u64>,
        new: Option<u64>,
    },
    EventType {
        old: Option<EventType>,
        new: Option<EventType>,
    },
    ExemptChannels {
        old: Option<Vec<ChannelId>>,
        new: Option<Vec<ChannelId>>,
    },
    ExemptRoles {
        old: Option<Vec<RoleId>>,
        new: Option<Vec<RoleId>>,
    },
    ExpireBehavior {
        old: Option<u64>,
        new: Option<u64>,
    },
    ExpireGracePeriod {
        old: Option<u64>,
        new: Option<u64>,
    },
    ExplicitContentFilter {
        old: Option<ExplicitContentFilter>,
        new: Option<ExplicitContentFilter>,
    },
    Flags {
        old: Option<u64>,
        new: Option<u64>,
    },
    FormatType {
        old: Option<StickerFormatType>,
        new: Option<StickerFormatType>,
    },
    GuildId {
        old: Option<GuildId>,
        new: Option<GuildId>,
    },
    Hoist {
        old: Option<bool>,
        new: Option<bool>,
    },
    IconHash {
        old: Option<String>,
        new: Option<String>,
    },
    Id {
        old: Option<GenericId>,
        new: Option<GenericId>,
    },
    ImageHash {
        old: Option<String>,
        new: Option<String>,
    },
    Invitable {
        old: Option<bool>,
        new: Option<bool>,
    },
    InviterId {
        old: Option<UserId>,
        new: Option<UserId>,
    },
    Location {
        old: Option<String>,
        new: Option<String>,
    },
    Locked {
        old: Option<bool>,
        new: Option<bool>,
    },
    MaxAge {
        old: Option<u64>,
        new: Option<u64>,
    },
    MaxUses {
        old: Option<u64>,
        new: Option<u64>,
    },
    Mentionable {
        old: Option<bool>,
        new: Option<bool>,
    },
    MfaLevel {
        old: Option<MfaLevel>,
        new: Option<MfaLevel>,
    },
    Mute {
        old: Option<bool>,
        new: Option<bool>,
    },
    Name {
        old: Option<String>,
        new: Option<String>,
    },
    Nick {
        old: Option<String>,
        new: Option<String>,
    },
    Nsfw {
        old: Option<bool>,
        new: Option<bool>,
    },
    OwnerId {
        old: Option<UserId>,
        new: Option<UserId>,
    },
    PermissionOverwrites {
        old: Option<Vec<PermissionOverwrite>>,
        new: Option<Vec<PermissionOverwrite>>,
    },
    Permissions {
        old: Option<Permissions>,
        new: Option<Permissions>,
    },
    Position {
        old: Option<u32>,
        new: Option<u32>,
    },
    PreferredLocale {
        old: Option<String>,
        new: Option<String>,
    },
    PrivacyLevel {
        old: Option<u64>,
        new: Option<u64>,
    },
    PruneDeleteDays {
        old: Option<u64>,
        new: Option<u64>,
    },
    PublicUpdatesChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    RateLimitPerUser {
        old: Option<u64>,
        new: Option<u64>,
    },
    Region {
        old: Option<String>,
        new: Option<String>,
    },
    RolesAdded {
        old: Option<Vec<AffectedRole>>,
        new: Option<Vec<AffectedRole>>,
    },
    RolesRemove {
        old: Option<Vec<AffectedRole>>,
        new: Option<Vec<AffectedRole>>,
    },
    RulesChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    SplashHash {
        old: Option<String>,
        new: Option<String>,
    },
    Status {
        old: Option<u64>,
        new: Option<u64>,
    },
    SystemChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    Tags {
        old: Option<String>,
        new: Option<String>,
    },
    Temporary {
        old: Option<bool>,
        new: Option<bool>,
    },
    Topic {
        old: Option<String>,
        new: Option<String>,
    },
    TriggerMetadata {
        old: Option<TriggerMetadata>,
        new: Option<TriggerMetadata>,
    },
    TriggerType {
        old: Option<TriggerType>,
        new: Option<TriggerType>,
    },
    Type {
        old: Option<EntityType>,
        new: Option<EntityType>,
    },
    UnicodeEmoji {
        old: Option<String>,
        new: Option<String>,
    },
    UserLimit {
        old: Option<u64>,
        new: Option<u64>,
    },
    Uses {
        old: Option<u64>,
        new: Option<u64>,
    },
    VanityUrlCode {
        old: Option<String>,
        new: Option<String>,
    },
    VerificationLevel {
        old: Option<VerificationLevel>,
        new: Option<VerificationLevel>,
    },
    WidgetChannelId {
        old: Option<ChannelId>,
        new: Option<ChannelId>,
    },
    WidgetEnabled {
        old: Option<bool>,
        new: Option<bool>,
    },
    Other {
        name: String,
        old_value: Option<Value>,
        new_value: Option<Value>,
    },
    Unknown,

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Actions

Fields

old: Option<Vec<Action>>

new: Option<Vec<Action>>

AfkChannelId

Fields

old: Option<ChannelId>

new: Option<ChannelId>

AFK channel was changed.

AfkTimeout

Fields

old: Option<u64>

new: Option<u64>

AFK timeout duration was changed.

Allow

Fields

old: Option<Permissions>

new: Option<Permissions>

Permission on a text or voice channel was allowed for a role.

ApplicationId

Fields

old: Option<ApplicationId>

new: Option<ApplicationId>

Application ID of the added or removed webhook or bot.

Archived

Fields

old: Option<bool>

new: Option<bool>

Thread is now archived/unarchived.

Asset

Fields

old: Option<String>

new: Option<String>

AutoArchiveDuration

Fields

old: Option<u64>

new: Option<u64>

Auto archive duration of a thread was changed.

Available

Fields

old: Option<bool>

new: Option<bool>

Availability of a sticker was changed.

AvatarHash

Fields

old: Option<String>

new: Option<String>

User avatar was changed.

BannerHash

Fields

old: Option<String>

new: Option<String>

Guild banner was changed.

Bitrate

Fields

old: Option<u32>

new: Option<u32>

Voice channel bitrate was changed.

ChannelId

Fields

old: Option<ChannelId>

new: Option<ChannelId>

Channel for invite code or guild scheduled event was changed.

Code

Fields

old: Option<String>

new: Option<String>

Invite code was changed.

Color

Fields

old: Option<u64>

new: Option<u64>

Role color was changed.

CommunicationDisabledUntil

Fields

old: Option<Timestamp>

new: Option<Timestamp>

Member timeout state was changed.

Deaf

Fields

old: Option<bool>

new: Option<bool>

User was server deafened/undeafened.

DefaultAutoArchiveDuration

Fields

old: Option<u64>

new: Option<u64>

Default auto archive duration for newly created threads was changed.

DefaultMessageNotifications

Fields

old: Option<DefaultMessageNotificationLevel>

new: Option<DefaultMessageNotificationLevel>

Default message notification level for a server was changed.

Deny

Fields

old: Option<Permissions>

new: Option<Permissions>

Permission on a text or voice channel was denied for a role.

Description

Fields

old: Option<String>

new: Option<String>

Description for guild, sticker, or guild scheduled event was changed.

DiscoverySplashHash

Fields

old: Option<String>

new: Option<String>

Guild’s discovery splash was changed.

Enabled

Fields

old: Option<bool>

new: Option<bool>

EnableEmoticons

Fields

old: Option<bool>

new: Option<bool>

Integration emoticons was enabled/disabled.

EntityType

Fields

old: Option<u64>

new: Option<u64>

Entity type of guild scheduled event was changed.

EventType

Fields

old: Option<EventType>

new: Option<EventType>

ExemptChannels

Fields

old: Option<Vec<ChannelId>>

new: Option<Vec<ChannelId>>

ExemptRoles

Fields

old: Option<Vec<RoleId>>

new: Option<Vec<RoleId>>

ExpireBehavior

Fields

old: Option<u64>

new: Option<u64>

Behavior of the expiration of an integration was changed.

ExpireGracePeriod

Fields

old: Option<u64>

new: Option<u64>

Grace period of the expiration of an integration was changed.

ExplicitContentFilter

Fields

old: Option<ExplicitContentFilter>

new: Option<ExplicitContentFilter>

Explicit content filter level of a guild was changed.

Flags

Fields

old: Option<u64>

new: Option<u64>

Unknown but sent by discord

FormatType

Fields

old: Option<StickerFormatType>

new: Option<StickerFormatType>

Format type of a sticker was changed.

GuildId

Fields

old: Option<GuildId>

new: Option<GuildId>

Guild a sticker is in was changed.

Hoist

Fields

old: Option<bool>

new: Option<bool>

Role is now displayed/no longer displayed separate from online users.

IconHash

Fields

old: Option<String>

new: Option<String>

Guild icon was changed.

Id

Fields

old: Option<GenericId>

new: Option<GenericId>

Guild scheduled event cover image was changed.

ImageHash

Fields

old: Option<String>

new: Option<String>

ID of the changed entity.

Invitable

Fields

old: Option<bool>

new: Option<bool>

Private thread’s invitable state was changed.

InviterId

Fields

old: Option<UserId>

new: Option<UserId>

ID of the user who created the invite.

Location

Fields

old: Option<String>

new: Option<String>

Location for a guild scheduled event was changed.

Locked

Fields

old: Option<bool>

new: Option<bool>

Thread was locked/unlocked.

MaxAge

Fields

old: Option<u64>

new: Option<u64>

How long invite code lasts was changed.

MaxUses

Fields

old: Option<u64>

new: Option<u64>

Maximum uses of an invite was changed.

Mentionable

Fields

old: Option<bool>

new: Option<bool>

Whether a role can be mentioned in a message was changed.

MfaLevel

Fields

old: Option<MfaLevel>

new: Option<MfaLevel>

Multi-factor authentication requirement was changed.

Mute

Fields

old: Option<bool>

new: Option<bool>

User was server muted/unmuted.

Name

Fields

old: Option<String>

new: Option<String>

Name of an entity was changed.

Nick

Fields

old: Option<String>

new: Option<String>

Nickname of a member was changed.

Nsfw

Fields

old: Option<bool>

new: Option<bool>

Channel NSFW restriction was changed.

OwnerId

Fields

old: Option<UserId>

new: Option<UserId>

Owner of a guild was changed.

PermissionOverwrites

Fields

old: Option<Vec<PermissionOverwrite>>

new: Option<Vec<PermissionOverwrite>>

Permissions on a channel were changed.

Permissions

Fields

old: Option<Permissions>

new: Option<Permissions>

Permissions for a role were changed.

Position

Fields

old: Option<u32>

new: Option<u32>

Channel or role position was changed.

PreferredLocale

Fields

old: Option<String>

new: Option<String>

Preferred locale of a guild was changed.

PrivacyLevel

Fields

old: Option<u64>

new: Option<u64>

Privacy level of the stage instance was changed.

PruneDeleteDays

Fields

old: Option<u64>

new: Option<u64>

Number of days after which inactive and role-unassigned members are kicked was changed.

PublicUpdatesChannelId

Fields

old: Option<ChannelId>

new: Option<ChannelId>

ID of the public updates channel was changed.

RateLimitPerUser

Fields

old: Option<u64>

new: Option<u64>

Ratelimit per user in a text channel was changed.

Region

Fields

old: Option<String>

new: Option<String>

Region of a guild was changed.

RolesAdded

Fields

old: Option<Vec<AffectedRole>>

new: Option<Vec<AffectedRole>>

Role was added to a member.

RolesRemove

Fields

old: Option<Vec<AffectedRole>>

new: Option<Vec<AffectedRole>>

Role was removed to a member.

RulesChannelId

Fields

old: Option<ChannelId>

new: Option<ChannelId>

ID of the rules channel was changed.

SplashHash

Fields

old: Option<String>

new: Option<String>

Invite splash page artwork was changed.

Status

Fields

old: Option<u64>

new: Option<u64>

Status of guild scheduled event was changed.

SystemChannelId

Fields

old: Option<ChannelId>

new: Option<ChannelId>

ID of the system channel was changed.

Tags

Fields

old: Option<String>

new: Option<String>

Related emoji of a sticker was changed.

Temporary

Fields

old: Option<bool>

new: Option<bool>

Whether an invite is temporary or never expires was changed.

Topic

Fields

old: Option<String>

new: Option<String>

Topic of a text channel or stage instance was changed.

TriggerMetadata

Fields

old: Option<TriggerMetadata>

new: Option<TriggerMetadata>

TriggerType

Fields

old: Option<TriggerType>

new: Option<TriggerType>

Type

Fields

old: Option<EntityType>

new: Option<EntityType>

Type of a created entity.

UnicodeEmoji

Fields

old: Option<String>

new: Option<String>

Unicode emoji of a role icon was changed.

UserLimit

Fields

old: Option<u64>

new: Option<u64>

Maximum number of users in a voice channel was changed.

Uses

Fields

old: Option<u64>

new: Option<u64>

Number of uses of an invite was changed.

VanityUrlCode

Fields

old: Option<String>

new: Option<String>

Guild invite vanity url was changed.

VerificationLevel

Fields

old: Option<VerificationLevel>

new: Option<VerificationLevel>

Required verification level for new members was changed.

WidgetChannelId

Fields

old: Option<ChannelId>

new: Option<ChannelId>

Channel of the server widget was changed.

WidgetEnabled

Fields

old: Option<bool>

new: Option<bool>

Whether a widget is enabled or not was changed.

Other

Fields

name: String

old_value: Option<Value>

new_value: Option<Value>

Unknown key was changed.

Unknown

Unknown key was changed and was invalid

Implementations

impl Change

pub fn key(&self) -> &str