Module serenity::model::prelude::colour::colours::branding

Constants

BLACK

Creates a new Colour, setting its value to rgb(35, 39, 42).

BLURPLE

Creates a new Colour, setting its value to rgb(88, 101, 242).

FUCHSIA

Creates a new Colour, setting its value to rgb(235, 69, 158).

GREEN

Creates a new Colour, setting its value to rgb(87, 242, 135).

RED

Creates a new Colour, setting its value to rgb(237, 66, 69).

WHITE

Creates a new Colour, setting its value to rgb(255, 255, 255).

YELLOW

Creates a new Colour, setting its value to rgb(254, 231, 92).