Module serenity::model::prelude::command

Structs

Command

The base command model that belongs to an application.

CommandOption

The parameters for an Command.

CommandOptionChoice

The only valid values a user can pick in an CommandOption.

CommandPermission

An Command permission.

CommandPermissionData

The CommandPermission data.

Enums

CommandOptionType

The type of an CommandOption.

CommandPermissionType

The type of an CommandPermissionData.

CommandType

The type of an application command.