Module serenity::model::prelude::interactions::application_command

👎Deprecated: use `model::application::interaction::application_command

Type Aliases

ApplicationCommandDeprecated

The base command model that belongs to an application.

ApplicationCommandInteractionDeprecated

An interaction when a user invokes a slash command.

ApplicationCommandInteractionDataDeprecated

The command data payload.

ApplicationCommandInteractionDataOptionDeprecated

A set of a parameter and a value from the user.

ApplicationCommandInteractionDataOptionValueDeprecated

The resolved value of an ApplicationCommandInteractionDataOption.

ApplicationCommandInteractionDataResolvedDeprecated

The resolved data of a command data interaction payload. It contains the objects of ApplicationCommandInteractionDataOptions.

ApplicationCommandOptionDeprecated

The parameters for an ApplicationCommand.

ApplicationCommandOptionChoiceDeprecated

The only valid values a user can pick in an ApplicationCommandOption.

ApplicationCommandOptionTypeDeprecated

The type of an ApplicationCommandOption.

ApplicationCommandPermissionDeprecated

An ApplicationCommand permission.

ApplicationCommandPermissionDataDeprecated

The ApplicationCommandPermission data.

ApplicationCommandPermissionTypeDeprecated

The type of an ApplicationCommandPermissionData.

ApplicationCommandTypeDeprecated

The type of an application command.

ResolvedTargetDeprecated

The resolved value of a ApplicationCommandInteractionData::target_id.