Type Alias serenity::model::prelude::interactions::application_command::ApplicationCommand

pub type ApplicationCommand = Command;

👎Deprecated: use model::application::command::Command

The base command model that belongs to an application.

Aliased Type

struct ApplicationCommand {

    pub id: CommandId,
    pub kind: CommandType,
    pub application_id: ApplicationId,
    pub guild_id: Option<GuildId>,
    pub name: String,
    pub name_localized: Option<String>,
    pub name_localizations: Option<HashMap<String, String>>,
    pub description: String,
    pub description_localized: Option<String>,
    pub description_localizations: Option<HashMap<String, String>>,
    pub options: Vec<CommandOption>,
    pub default_member_permissions: Option<Permissions>,
    pub dm_permission: Option<bool>,
    pub default_permission: bool,
    pub version: CommandVersionId,

}

Fields
id: CommandId

The command Id.

kind: CommandType

The application command kind.

application_id: ApplicationId

The parent application Id.

guild_id: Option<GuildId>

The command guild Id, if it is a guild command.

Note: It may only be present if it is received through the gateway.

name: String

The command name.

name_localized: Option<String>

The localized command name of the selected locale.

If the name is localized, either this field or Self::name_localizations is set, depending on which endpoint this data was retrieved from (source).

name_localizations: Option<HashMap<String, String>>

All localized command names.

If the name is localized, either this field or Self::name_localized is set, depending on which endpoint this data was retrieved from (source).

description: String

The command description.

description_localized: Option<String>

The localized command description of the selected locale.

If the description is localized, either this field or Self::description_localizations is set, depending on which endpoint this data was retrieved from (source).

description_localizations: Option<HashMap<String, String>>

All localized command descriptions.

If the description is localized, either this field or Self::description_localized is set, depending on which endpoint this data was retrieved from (source).

options: Vec<CommandOption>

The parameters for the command.

default_member_permissions: Option<Permissions>

The default permissions required to execute the command.

dm_permission: Option<bool>

Indicates whether the command is available in DMs with the app, only for globally-scoped commands. By default, commands are visible.

default_permission: bool
👎Deprecated: replaced by default_member_permissions

Whether the command is enabled by default when the application is added to a guild.

version: CommandVersionId

An autoincremented version identifier updated during substantial record changes.