Type Alias serenity::model::prelude::interactions::application_command::ApplicationCommandInteraction

pub type ApplicationCommandInteraction = ApplicationCommandInteraction;

👎Deprecated: use model::application::interaction::application_command::ApplicationCommandInteraction

An interaction when a user invokes a slash command.

Aliased Type

struct ApplicationCommandInteraction {

    pub id: InteractionId,
    pub application_id: ApplicationId,
    pub kind: InteractionType,
    pub data: CommandData,
    pub guild_id: Option<GuildId>,
    pub channel_id: ChannelId,
    pub member: Option<Member>,
    pub user: User,
    pub token: String,
    pub version: u8,
    pub app_permissions: Option<Permissions>,
    pub locale: String,
    pub guild_locale: Option<String>,

}

Fields
id: InteractionId

Id of the interaction.

application_id: ApplicationId

Id of the application this interaction is for.

kind: InteractionType

The type of interaction.

data: CommandData

The data of the interaction which was triggered.

guild_id: Option<GuildId>

The guild Id this interaction was sent from, if there is one.

channel_id: ChannelId

The channel Id this interaction was sent from.

member: Option<Member>

The member data for the invoking user.

Note: It is only present if the interaction is triggered in a guild.

user: User

The user object for the invoking user.

token: String

A continuation token for responding to the interaction.

version: u8

Always 1.

app_permissions: Option<Permissions>

Permissions the app or bot has within the channel the interaction was sent from.

locale: String

The selected language of the invoking user.

guild_locale: Option<String>

The guild’s preferred locale.

Implementations

impl ApplicationCommandInteraction

pub async fn get_interaction_response( &self, http: impl AsRef<Http> ) -> Result<Message>

Gets the interaction response.

Errors

Returns an Error::Http if there is no interaction response.

pub async fn create_interaction_response<'a, F>( &self, http: impl AsRef<Http>, f: F ) -> Result<()>where for<'b> F: FnOnce(&'b mut CreateInteractionResponse<'a>) -> &'b mut CreateInteractionResponse<'a>,

Creates a response to the interaction received.

Note: Message contents must be under 2000 unicode code points.

Errors

Returns an Error::Model if the message content is too long. May also return an Error::Http if the API returns an error, or an Error::Json if there is an error in deserializing the API response.

pub async fn edit_original_interaction_response<F>( &self, http: impl AsRef<Http>, f: F ) -> Result<Message>where F: FnOnce(&mut EditInteractionResponse) -> &mut EditInteractionResponse,

Edits the initial interaction response.

application_id will usually be the bot’s UserId, except in cases of bots being very old.

Refer to Discord’s docs for Edit Webhook Message for field information.

Note: Message contents must be under 2000 unicode code points.

Errors

Returns Error::Model if the edited content is too long. May also return Error::Http if the API returns an error, or an Error::Json if there is an error deserializing the response.

pub async fn delete_original_interaction_response( &self, http: impl AsRef<Http> ) -> Result<()>

Deletes the initial interaction response.

Does not work on ephemeral messages.

Errors

May return Error::Http if the API returns an error. Such as if the response was already deleted.

pub async fn create_followup_message<'a, F>( &self, http: impl AsRef<Http>, f: F ) -> Result<Message>where for<'b> F: FnOnce(&'b mut CreateInteractionResponseFollowup<'a>) -> &'b mut CreateInteractionResponseFollowup<'a>,

Creates a followup response to the response sent.

Note: Message contents must be under 2000 unicode code points.

Errors

Will return Error::Model if the content is too long. May also return Error::Http if the API returns an error, or a Error::Json if there is an error in deserializing the response.

pub async fn edit_followup_message<'a, F, M: Into<MessageId>>( &self, http: impl AsRef<Http>, message_id: M, f: F ) -> Result<Message>where for<'b> F: FnOnce(&'b mut CreateInteractionResponseFollowup<'a>) -> &'b mut CreateInteractionResponseFollowup<'a>,

Edits a followup response to the response sent.

Note: Message contents must be under 2000 unicode code points.

Errors

Will return Error::Model if the content is too long. May also return Error::Http if the API returns an error, or a Error::Json if there is an error in deserializing the response.

pub async fn delete_followup_message<M: Into<MessageId>>( &self, http: impl AsRef<Http>, message_id: M ) -> Result<()>

Deletes a followup message.

Errors

May return Error::Http if the API returns an error. Such as if the response was already deleted.

pub async fn get_followup_message<M: Into<MessageId>>( &self, http: impl AsRef<Http>, message_id: M ) -> Result<Message>

Gets a followup message.

Errors

May return Error::Http if the API returns an error. Such as if the response was deleted.

pub async fn defer(&self, http: impl AsRef<Http>) -> Result<()>

Helper function to defer an interaction

Errors

May also return an Error::Http if the API returns an error, or an Error::Json if there is an error in deserializing the API response.

pub async fn defer_ephemeral(&self, http: impl AsRef<Http>) -> Result<()>

Helper function to defer an interaction ephemerally

Errors

May also return an Error::Http if the API returns an error, or an Error::Json if there is an error in deserializing the API response.