Type Alias serenity::model::prelude::interactions::application_command::ApplicationCommandInteractionData

pub type ApplicationCommandInteractionData = CommandData;

👎Deprecated: use model::application::interaction::application_command::CommandData

The command data payload.

Aliased Type

struct ApplicationCommandInteractionData {
    pub id: CommandId,
    pub name: String,
    pub kind: CommandType,
    pub resolved: CommandDataResolved,
    pub options: Vec<CommandDataOption>,
    pub guild_id: Option<GuildId>,
    pub target_id: Option<TargetId>,
}

Fields
id: CommandId

The Id of the invoked command.

name: String

The name of the invoked command.

kind: CommandType

The application command type of the triggered application command.

resolved: CommandDataResolved

The parameters and the given values. The converted objects from the given options.

options: Vec<CommandDataOption>guild_id: Option<GuildId>

The Id of the guild the command is registered to.

target_id: Option<TargetId>

The targeted user or message, if the triggered application command type is User or Message.

Its object data can be found in the resolved field.