Type Alias serenity::model::prelude::interactions::application_command::ApplicationCommandOptionChoice

pub type ApplicationCommandOptionChoice = CommandOptionChoice;

👎Deprecated: use model::application::command::CommandOptionChoice

The only valid values a user can pick in an ApplicationCommandOption.

Aliased Type

struct ApplicationCommandOptionChoice {
    pub name: String,
    pub name_localizations: Option<HashMap<String, String>>,
    pub value: Value,
}

Fields
name: String

The choice name.

name_localizations: Option<HashMap<String, String>>

Localizations of the choice name, with locale as key

value: Value

The choice value.