Type Alias serenity::model::prelude::interactions::message_component::MessageComponentInteraction

pub type MessageComponentInteraction = MessageComponentInteraction;

👎Deprecated: use model::application::interaction::message_component::MessageComponentInteraction

An interaction triggered by a message component.

Aliased Type

struct MessageComponentInteraction {

    pub id: InteractionId,
    pub application_id: ApplicationId,
    pub kind: InteractionType,
    pub data: MessageComponentInteractionData,
    pub guild_id: Option<GuildId>,
    pub channel_id: ChannelId,
    pub member: Option<Member>,
    pub user: User,
    pub token: String,
    pub version: u8,
    pub message: Message,
    pub app_permissions: Option<Permissions>,
    pub locale: String,
    pub guild_locale: Option<String>,

}

Fields
id: InteractionId

Id of the interaction.

application_id: ApplicationId

Id of the application this interaction is for.

kind: InteractionType

The type of interaction.

data: MessageComponentInteractionData

The data of the interaction which was triggered.

guild_id: Option<GuildId>

The guild Id this interaction was sent from, if there is one.

channel_id: ChannelId

The channel Id this interaction was sent from.

member: Option<Member>

The member data for the invoking user.

Note: It is only present if the interaction is triggered in a guild.

user: User

The user object for the invoking user.

token: String

A continuation token for responding to the interaction.

version: u8

Always 1.

message: Message

The message this interaction was triggered by, if it is a component.

app_permissions: Option<Permissions>

Permissions the app or bot has within the channel the interaction was sent from.

locale: String

The selected language of the invoking user.

guild_locale: Option<String>

The guild’s preferred locale.