Module serenity::model::prelude::misc

Miscellaneous helper traits, enums, and structs for models.

Structs

AffectedComponent

A component that was affected during a service incident.

EmojiIdentifier

A version of an emoji used only when solely the animated state, Id, and name are known.

EmojiIdentifierParseErrormodel and utils

Incident

An incident retrieved from the Discord status page.

IncidentUpdate

An update to an incident from the Discord status page.

Maintenance

A Discord status maintenance message. This can be either for active maintenances or for scheduled maintenances.

Enums

ChannelIdParseErrormodel and utils

IncidentStatus

The type of status update during a service incident.

RoleIdParseErrormodel and utils

UserIdParseErrormodel and utils