Module serenity::model::prelude::prelude::application_command

Structs

ApplicationCommandInteraction

An interaction when a user invokes a slash command.

AutocompleteOption

The focused option for autocomplete interactions return by CommandData::autocomplete.

CommandData

The command data payload.

CommandDataOption

A set of a parameter and a value from the user.

CommandDataResolved

The resolved data of a command data interaction payload. It contains the objects of CommandDataOptions.

ResolvedOption

Enums

CommandDataOptionValue

The value of an CommandDataOption.

ResolvedTarget

The resolved value of a CommandData::target_id.

ResolvedValue

The resolved value of a CommandDataOption.

Unresolved

Option value variants that couldn’t be resolved by CommandData::options().