Struct serenity::model::prelude::prelude::Colour

pub struct Colour(pub u32);

A utility struct to help with working with the basic representation of a colour. This is particularly useful when working with a Role’s colour, as the API works with an integer value instead of an RGB value.

Instances can be created by using the struct’s associated functions. These produce presets equivalent to those found in the official client’s colour picker.

Examples

Passing in a role’s colour, and then retrieving its green component via Self::g:

use serenity::model::colour::Colour;

// assuming a `role` has already been bound

let green = role.colour.g();

println!("The green component is: {}", green);

Creating an instance with the Self::DARK_TEAL preset:

use serenity::model::colour::Colour;

let colour = Colour::DARK_TEAL;

assert_eq!(colour.tuple(), (17, 128, 106));

Colours can also be directly compared for equivalence:

use serenity::model::colour::Colour;

let blitz_blue = Colour::BLITZ_BLUE;
let fooyoo = Colour::FOOYOO;
let fooyoo2 = Colour::FOOYOO;
assert!(blitz_blue != fooyoo);
assert_eq!(fooyoo, fooyoo2);
assert!(blitz_blue > fooyoo);

Tuple Fields
0: u32
Implementations

impl Colour

pub const fn new(value: u32) -> Colour

Generates a new Colour with the given integer value set.

Examples

Create a new Colour, and then ensure that its inner value is equivalent to a specific RGB value, retrieved via Self::tuple:

use serenity::model::colour::Colour;

let colour = Colour::new(6573123);

assert_eq!(colour.tuple(), (100, 76, 67));

pub const fn from_rgb(red: u8, green: u8, blue: u8) -> Colour

Generates a new Colour from an RGB value, creating an inner u32 representation.

Examples

Creating a Colour via its RGB values will set its inner u32 correctly:

use serenity::model::colour::Colour;

assert!(Colour::from_rgb(255, 0, 0).0 == 0xFF0000);
assert!(Colour::from_rgb(217, 23, 211).0 == 0xD917D3);

And you can then retrieve those same RGB values via its methods:

use serenity::model::colour::Colour;

let colour = Colour::from_rgb(217, 45, 215);

assert_eq!(colour.r(), 217);
assert_eq!(colour.g(), 45);
assert_eq!(colour.b(), 215);
assert_eq!(colour.tuple(), (217, 45, 215));

pub const fn r(self) -> u8

Returns the red RGB component of this Colour.

Examples

use serenity::model::colour::Colour;

assert_eq!(Colour::new(6573123).r(), 100);

pub const fn g(self) -> u8

Returns the green RGB component of this Colour.

Examples

use serenity::model::colour::Colour;

assert_eq!(Colour::new(6573123).g(), 76);

pub const fn b(self) -> u8

Returns the blue RGB component of this Colour.

Examples

use serenity::model::colour::Colour;

assert_eq!(Colour::new(6573123).b(), 67);

pub const fn tuple(self) -> (u8, u8, u8)

Returns a tuple of the red, green, and blue components of this Colour.

This is equivalent to creating a tuple with the return values of Self::r, Self::g, and Self::b.

Examples

use serenity::model::colour::Colour;

assert_eq!(Colour::new(6573123).tuple(), (100, 76, 67));

pub fn hex(self) -> String

Returns a hexadecimal string of this Colour.

This is equivalent to passing the integer value through std::fmt::UpperHex with 0 padding and 6 width.

Examples

use serenity::model::colour::Colour;

assert_eq!(Colour::new(6573123).hex(), "644C43");

impl Colour

pub const BLITZ_BLUE: Colour = _

Creates a new Colour, setting its RGB value to (111, 198, 226).

pub const BLUE: Colour = _

Creates a new Colour, setting its RGB value to (52, 152, 219).

pub const BLURPLE: Colour = _

Creates a new Colour, setting its RGB value to (114, 137, 218).

pub const DARK_BLUE: Colour = _

Creates a new Colour, setting its RGB value to (32, 102, 148).

pub const DARK_GOLD: Colour = _

Creates a new Colour, setting its RGB value to (194, 124, 14).

pub const DARK_GREEN: Colour = _

Creates a new Colour, setting its RGB value to (31, 139, 76).

pub const DARK_GREY: Colour = _

Creates a new Colour, setting its RGB value to (96, 125, 139).

pub const DARK_MAGENTA: Colour = _

Creates a new Colour, setting its RGB value to (173, 20, 87).

pub const DARK_ORANGE: Colour = _

Creates a new Colour, setting its RGB value to (168, 67, 0).

pub const DARK_PURPLE: Colour = _

Creates a new Colour, setting its RGB value to (113, 54, 138).

pub const DARK_RED: Colour = _

Creates a new Colour, setting its RGB value to (153, 45, 34).

pub const DARK_TEAL: Colour = _

Creates a new Colour, setting its RGB value to (17, 128, 106).

pub const DARKER_GREY: Colour = _

Creates a new Colour, setting its RGB value to (84, 110, 122).

pub const FABLED_PINK: Colour = _

Creates a new Colour, setting its RGB value to (250, 177, 237).

pub const FADED_PURPLE: Colour = _

Creates a new Colour, setting its RGB value to (136, 130, 196).

pub const FOOYOO: Colour = _

Creates a new Colour, setting its RGB value to (17, 202, 128).

pub const GOLD: Colour = _

Creates a new Colour, setting its RGB value to (241, 196, 15).

pub const KERBAL: Colour = _

Creates a new Colour, setting its RGB value to (186, 218, 85).

pub const LIGHT_GREY: Colour = _

Creates a new Colour, setting its RGB value to (151, 156, 159).

pub const LIGHTER_GREY: Colour = _

Creates a new Colour, setting its RGB value to (149, 165, 166).

pub const MAGENTA: Colour = _

Creates a new Colour, setting its RGB value to (233, 30, 99).

pub const MEIBE_PINK: Colour = _

Creates a new Colour, setting its RGB value to (230, 131, 151).

pub const ORANGE: Colour = _

Creates a new Colour, setting its RGB value to (230, 126, 34).

pub const PURPLE: Colour = _

Creates a new Colour, setting its RGB value to (155, 89, 182).

pub const RED: Colour = _

Creates a new Colour, setting its RGB value to (231, 76, 60).

pub const ROHRKATZE_BLUE: Colour = _

Creates a new Colour, setting its RGB value to (117, 150, 255).

pub const ROSEWATER: Colour = _

Creates a new Colour, setting its RGB value to (246, 219, 216).

pub const TEAL: Colour = _

Creates a new Colour, setting its RGB value to (26, 188, 156).