Struct serenity::model::prelude::prelude::Command

#[non_exhaustive]
pub struct Command {

    pub id: CommandId,
    pub kind: CommandType,
    pub application_id: ApplicationId,
    pub guild_id: Option<GuildId>,
    pub name: String,
    pub name_localized: Option<String>,
    pub name_localizations: Option<HashMap<String, String>>,
    pub description: String,
    pub description_localized: Option<String>,
    pub description_localizations: Option<HashMap<String, String>>,
    pub options: Vec<CommandOption>,
    pub default_member_permissions: Option<Permissions>,
    pub dm_permission: Option<bool>,
    pub version: CommandVersionId,

}

The base command model that belongs to an application.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: CommandId

The command Id.

kind: CommandType

The application command kind.

application_id: ApplicationId

The parent application Id.

guild_id: Option<GuildId>

The command guild Id, if it is a guild command.

Note: It may only be present if it is received through the gateway.

name: String

The command name.

name_localized: Option<String>

The localized command name of the selected locale.

If the name is localized, either this field or Self::name_localizations is set, depending on which endpoint this data was retrieved from (source).

name_localizations: Option<HashMap<String, String>>

All localized command names.

If the name is localized, either this field or Self::name_localized is set, depending on which endpoint this data was retrieved from (source).

description: String

The command description.

description_localized: Option<String>

The localized command description of the selected locale.

If the description is localized, either this field or Self::description_localizations is set, depending on which endpoint this data was retrieved from (source).

description_localizations: Option<HashMap<String, String>>

All localized command descriptions.

If the description is localized, either this field or Self::description_localized is set, depending on which endpoint this data was retrieved from (source).

options: Vec<CommandOption>

The parameters for the command.

default_member_permissions: Option<Permissions>

The default permissions required to execute the command.

dm_permission: Option<bool>

Indicates whether the command is available in DMs with the app, only for globally-scoped commands. By default, commands are visible.

version: CommandVersionId

An autoincremented version identifier updated during substantial record changes.

Implementations

impl Command

pub async fn create_global_application_command( cache_http: impl CacheHttp, builder: CreateCommand ) -> Result<Command>

Available on crate feature model only.

Create a global Command, overriding an existing one with the same name if it exists.

When a created Command is used, the InteractionCreate event will be emitted.

Note: Global commands may take up to an hour to be updated in the user slash commands list. If an outdated command data is sent by a user, discord will consider it as an error and then will instantly update that command.

As such, it is recommended that guild application commands be used for testing purposes.

Examples

Create a simple ping command:

use serenity::builder::CreateCommand;
use serenity::model::application::Command;
use serenity::model::id::ApplicationId;

let builder = CreateCommand::new("ping").description("A simple ping command");
let _ = Command::create_global_application_command(&http, builder).await;

Create a command that echoes what is inserted:

use serenity::builder::{CreateCommand, CreateCommandOption as CreateOption};
use serenity::model::application::{Command, CommandOptionType};
use serenity::model::id::ApplicationId;

let builder =
    CreateCommand::new("echo").description("Makes the bot send a message").add_option(
        CreateOption::new(CommandOptionType::String, "message", "The message to send")
            .required(true),
    );
let _ = Command::create_global_application_command(&http, builder).await;

Errors

See CreateCommand::execute for a list of possible errors.

pub async fn set_global_application_commands( http: impl AsRef<Http>, commands: Vec<CreateCommand> ) -> Result<Vec<Command>>

Available on crate feature model only.

Override all global application commands.

Errors

Returns the same errors as Self::create_global_application_command.

pub async fn edit_global_application_command( cache_http: impl CacheHttp, command_id: CommandId, builder: CreateCommand ) -> Result<Command>

Available on crate feature model only.

Edit a global command, given its Id.

Errors

See CreateCommand::execute for a list of possible errors.

pub async fn get_global_application_commands( http: impl AsRef<Http> ) -> Result<Vec<Command>>

Available on crate feature model only.

Gets all global commands.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_global_application_command( http: impl AsRef<Http>, command_id: CommandId ) -> Result<Command>

Available on crate feature model only.

Gets a global command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn delete_global_application_command( http: impl AsRef<Http>, command_id: CommandId ) -> Result<()>

Available on crate feature model only.

Deletes a global command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.