Struct serenity::model::prelude::prelude::RichInvite

pub struct RichInvite {
    pub channel: InviteChannel,
    pub code: String,
    pub created_at: Timestamp,
    pub guild: Option<InviteGuild>,
    pub inviter: Option<User>,
    pub max_age: u64,
    pub max_uses: u64,
    pub temporary: bool,
    pub uses: u64,
}

Detailed information about an invite. This information can only be retrieved by anyone with the Manage Guild permission. Otherwise, a minimal amount of information can be retrieved via the Invite struct.

Discord docs (extends Invite fields).

Fields
channel: InviteChannel

A representation of the minimal amount of information needed about the channel being invited to.

code: String

The unique code for the invite.

created_at: Timestamp

When the invite was created.

guild: Option<InviteGuild>

A representation of the minimal amount of information needed about the Guild being invited to.

inviter: Option<User>

The user that created the invite.

max_age: u64

The maximum age of the invite in seconds, from when it was created.

max_uses: u64

The maximum number of times that an invite may be used before it expires.

Note that this does not supersede the Self::max_age value, if the value of Self::temporary is true. If the value of temporary is false, then the invite will self-expire after the given number of max uses.

If the value is 0, then the invite is permanent.

temporary: bool

Indicator of whether the invite self-expires after a certain amount of time or uses.

uses: u64

The amount of times that an invite has been used.

Implementations

impl RichInvite

pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<Invite>

Available on crate feature model only.

Deletes the invite.

Refer to Http::delete_invite for more information.

Note: Requires the Manage Guild permission.

Errors

If the cache feature is enabled, then this returns a ModelError::InvalidPermissions if the current user does not have the required permission.

pub fn url(&self) -> String

Available on crate feature model only.

Returns a URL to use for the invite.

Examples

Retrieve the URL for an invite with the code WxZumR:

assert_eq!(invite.url(), "https://discord.gg/WxZumR");