Struct serenity::model::prelude::GatewayIntents

pub struct GatewayIntents { /* private fields */ }

[Gateway Intents] will limit the events your bot will receive via the gateway. By default, all intents except Privileged Intents are selected.

What are Intents

A gateway intent sets the types of gateway events (e.g. member joins, guild integrations, guild emoji updates, …) the bot shall receive. Carefully picking the needed intents greatly helps the bot to scale, as less intents will result in less events to be received via the network from Discord and less processing needed for handling the data.

Privileged Intents

The intents GatewayIntents::GUILD_PRESENCES, GatewayIntents::GUILD_MEMBERS and GatewayIntents::MESSAGE_CONTENT are Privileged Intents. They need to be enabled in the developer portal.

Note: Once the bot is in 100 guilds or more, the bot must be verified in order to use privileged intents.

Discord docs.

Implementations

impl GatewayIntents

pub const GUILDS: Self = _

Enables following gateway events:

GUILD_CREATE

GUILD_DELETE

GUILD_ROLE_CREATE

GUILD_ROLE_UPDATE

GUILD_ROLE_DELETE

CHANNEL_CREATE

CHANNEL_UPDATE

CHANNEL_DELETE

CHANNEL_PINS_UPDATE

THREAD_CREATE

THREAD_UPDATE

THREAD_DELETE

THREAD_LIST_SYNC

THREAD_MEMBER_UPDATE

THREAD_MEMBERS_UPDATE

STAGE_INSTANCE_CREATE

STAGE_INSTANCE_UPDATE

STAGE_INSTANCE_DELETE

pub const GUILD_MEMBERS: Self = _

Enables following gateway events:

GUILD_MEMBER_ADD

GUILD_MEMBER_UPDATE

GUILD_MEMBER_REMOVE

THREAD_MEMBERS_UPDATE

Info: This intent is privileged. In order to use it, you must head to your application in the Developer Portal and enable the toggle for Privileged Intents.

This intent is also necessary to even receive the events in contains.

pub const GUILD_MODERATION: Self = _

Enables following gateway events:

GUILD_AUDIT_LOG_ENTRY_CREATE

GUILD_BAN_ADD

GUILD_BAN_REMOVE

pub const GUILD_BANS: Self = _

Backwards compatibility with old gateway event name. Same as GUILD_MODERATION

pub const GUILD_EMOJIS_AND_STICKERS: Self = _

Enables following gateway event:

GUILD_EMOJIS_UPDATE

GUILD_STICKERS_UPDATE

pub const GUILD_INTEGRATIONS: Self = _

Enables following gateway event:

GUILD_INTEGRATIONS_UPDATE

INTEGRATION_CREATE

INTEGRATION_UPDATE

INTEGRATION_DELETE

pub const GUILD_WEBHOOKS: Self = _

Enables following gateway event:

WEBHOOKS_UPDATE

pub const GUILD_INVITES: Self = _

Enables following gateway events:

INVITE_CREATE

INVITE_DELETE

pub const GUILD_VOICE_STATES: Self = _

Enables following gateway event:

VOICE_STATE_UPDATE

pub const GUILD_PRESENCES: Self = _

Enables following gateway event:

PRESENCE_UPDATE

Info: This intent is privileged. In order to use it, you must head to your application in the Developer Portal and enable the toggle for Privileged Intents.

This intent is also necessary to even receive the events in contains.

pub const GUILD_MESSAGES: Self = _

Enables following gateway events:

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

MESSAGE_DELETE_BULK

pub const GUILD_MESSAGE_REACTIONS: Self = _

Enables following gateway events:

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOJI

pub const GUILD_MESSAGE_TYPING: Self = _

Enable following gateway event:

TYPING_START

pub const DIRECT_MESSAGES: Self = _

Enable following gateway events:

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

CHANNEL_PINS_UPDATE

pub const DIRECT_MESSAGE_REACTIONS: Self = _

Enable following gateway events:

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOJI

pub const DIRECT_MESSAGE_TYPING: Self = _

Enable following gateway event:

TYPING_START

pub const MESSAGE_CONTENT: Self = _

Enable message content in following gateway events:

MESSAGE_CREATE

Info: This intent is privileged. In order to use it, you must head to your application in the Developer Portal and enable the toggle for Privileged Intents.

pub const GUILD_SCHEDULED_EVENTS: Self = _

Enable following gateway events:

GUILD_SCHEDULED_EVENT_CREATE

GUILD_SCHEDULED_EVENT_UPDATE

GUILD_SCHEDULED_EVENT_DELETE

GUILD_SCHEDULED_EVENT_USER_ADD

GUILD_SCHEDULED_EVENT_USER_REMOVE

pub const AUTO_MODERATION_CONFIGURATION: Self = _

Enable following gateway events:

AUTO_MODERATION_RULE_CREATE

AUTO_MODERATION_RULE_UPDATE

AUTO_MODERATION_RULE_DELETE

pub const AUTO_MODERATION_EXECUTION: Self = _

Enable following gateway events:

AUTO_MODERATION_ACTION_EXECUTION

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl GatewayIntents

pub const fn non_privileged() -> GatewayIntents

Gets all of the intents that aren’t considered privileged by Discord.

pub const fn privileged() -> GatewayIntents

Gets all of the intents that are considered privileged by Discord. Use of these intents will require explicitly whitelisting the bot.

impl GatewayIntents

pub const fn is_privileged(self) -> bool

Available on crate feature model only.

Checks if any of the included intents are privileged.

pub const fn guilds(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILDS intent.

pub const fn guild_members(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_MEMBERS intent.

pub const fn guild_bans(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_BANS intent.

This is the same as calling guild_moderation since Discord changed the name

pub const fn guild_moderation(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_MODERATION intent.

pub const fn guild_emojis_and_stickers(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_EMOJIS_AND_STICKERS intent.

pub const fn guild_integrations(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_INTEGRATIONS intent.

pub const fn guild_webhooks(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_WEBHOOKS intent.

pub const fn guild_invites(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_INVITES intent.

pub const fn guild_voice_states(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_VOICE_STATES intent.

pub const fn guild_presences(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_PRESENCES intent.

pub const fn guild_message_reactions(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_MESSAGE_REACTIONS intent.

pub const fn guild_message_typing(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_MESSAGE_TYPING intent.

pub const fn direct_messages(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the DIRECT_MESSAGES intent.

pub const fn direct_message_reactions(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the DIRECT_MESSAGE_REACTIONS intent.

pub const fn direct_message_typing(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the DIRECT_MESSAGE_TYPING intent.

pub const fn message_content(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the MESSAGE_CONTENT intent.

pub const fn guild_scheduled_events(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the GUILD_SCHEDULED_EVENTS intent.

pub fn auto_moderation_configuration(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the AUTO_MODERATION_CONFIGURATION intent.

pub fn auto_moderation_execution(self) -> bool

Available on crate feature model only.

Shorthand for checking that the set of intents contains the AUTO_MODERATION_EXECUTION intent.