Struct serenity::model::prelude::GuildId

pub struct GuildId(pub NonZeroU64);

An identifier for a Guild

Tuple Fields
0: NonZeroU64
Implementations

impl GuildId

pub async fn automod_rules(self, http: impl AsRef<Http>) -> Result<Vec<Rule>>

Available on crate feature model only.

Gets all auto moderation Rules of this guild via HTTP.

Note: Requires the Manage Guild permission.

Errors

Returns an Error::Http if the guild is unavailable.

pub async fn automod_rule(
self,
http: impl AsRef<Http>,
rule_id: impl Into<RuleId>
) -> Result<Rule>

Available on crate feature model only.

Gets an auto moderation Rule of this guild by its ID via HTTP.

Note: Requires the Manage Guild permission.

Errors

Returns an Error::Http if a rule with the given ID does not exist.

pub async fn create_automod_rule(
self,
http: impl AsRef<Http>,
builder: EditAutoModRule<'_>
) -> Result<Rule>

Available on crate feature model only.

Creates an auto moderation Rule in the guild.

Note: Requires the Manage Guild permission.

Examples

Create a custom keyword filter to block the message and timeout the author.

use std::time::Duration;

use serenity::builder::EditAutoModRule;
use serenity::model::guild::automod::{Action, Trigger};
use serenity::model::id::GuildId;

let builder = EditAutoModRule::new()
    .name("foobar filter")
    .trigger(Trigger::Keyword(vec!["foo*".to_string(), "*bar".to_string()]))
    .actions(vec![Action::BlockMessage, Action::Timeout(Duration::from_secs(60))]);
let _rule = GuildId::new(7).create_automod_rule(&http, builder).await;

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub async fn edit_automod_rule(
self,
http: impl AsRef<Http>,
rule_id: impl Into<RuleId>,
builder: EditAutoModRule<'_>
) -> Result<Rule>

Available on crate feature model only.

Edit an auto moderation Rule, given its Id.

Note: Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub async fn delete_automod_rule(
self,
http: impl AsRef<Http>,
rule_id: impl Into<RuleId>
) -> Result<()>

Available on crate feature model only.

Deletes an auto moderation Rule from the guild.

Note: Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission, or if a rule with that Id does not exist.

pub async fn add_member(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
builder: AddMember
) -> Result<Option<Member>>

Available on crate feature model only.

Adds a User to this guild with a valid OAuth2 access token.

Returns the created Member object, or nothing if the user is already a member of the guild.

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub async fn ban(
self,
http: impl AsRef<Http>,
user: impl Into<UserId>,
dmd: u8
) -> Result<()>

Available on crate feature model only.

Ban a User from the guild, deleting a number of days’ worth of messages (dmd) between the range 0 and 7.

Refer to the documentation for Guild::ban for more information.

Note: Requires the Ban Members permission.

Examples

Ban a member and remove all messages they’ve sent in the last 4 days:

use serenity::model::id::{GuildId, UserId};

// assuming a `user` has already been bound
let _ = GuildId::new(81384788765712384).ban(&http, user, 4).await;

Errors

Returns a ModelError::DeleteMessageDaysAmount if the number of days’ worth of messages to delete is over the maximum.

Also can return Error::Http if the current user lacks permission.

pub async fn ban_with_reason(
self,
http: impl AsRef<Http>,
user: impl Into<UserId>,
dmd: u8,
reason: impl AsRef<str>
) -> Result<()>

Available on crate feature model only.

Ban a User from the guild with a reason. Refer to Self::ban to further documentation.

Errors

In addition to the reasons Self::ban may return an error, may also return Error::ExceededLimit if reason is too long.

pub async fn bans(self, http: impl AsRef<Http>) -> Result<Vec<Ban>>

Available on crate feature model only.

Gets a list of the guild’s bans.

Note: Requires the Ban Members permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn audit_logs(
self,
http: impl AsRef<Http>,
action_type: Option<u8>,
user_id: Option<UserId>,
before: Option<AuditLogEntryId>,
limit: Option<u8>
) -> Result<AuditLogs>

Available on crate feature model only.

Gets a list of the guild’s audit log entries

Note: Requires the View Audit Log permission.

Errors

Returns Error::Http if the current user lacks permission, or if an invalid value is given.

pub async fn channels(
self,
http: impl AsRef<Http>
) -> Result<HashMap<ChannelId, GuildChannel>>

Available on crate feature model only.

Gets all of the guild’s channels over the REST API.

Errors

Returns Error::Http if the current user is not in the guild.

pub async fn create_channel(
self,
cache_http: impl CacheHttp,
builder: CreateChannel<'_>
) -> Result<GuildChannel>

Available on crate feature model only.

Creates a GuildChannel in the the guild.

Refer to Http::create_channel for more information.

Note: Requires the Manage Channels permission.

Examples

Create a voice channel in a guild with the name test:

use serenity::builder::CreateChannel;
use serenity::model::channel::ChannelType;
use serenity::model::id::GuildId;

let builder = CreateChannel::new("test").kind(ChannelType::Voice);
let _channel = GuildId::new(7).create_channel(&http, builder).await?;

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn create_emoji(
self,
http: impl AsRef<Http>,
name: &str,
image: &str
) -> Result<Emoji>

Available on crate feature model only.

Creates an emoji in the guild with a name and base64-encoded image.

Refer to the documentation for Guild::create_emoji for more information.

Requires the Manage Emojis and Stickers permission.

Examples

See the EditProfile::avatar example for an in-depth example as to how to read an image from the filesystem and encode it as base64. Most of the example can be applied similarly for this method.

Errors

Returns Error::Http if the current user lacks permission, if the name is too long, or if the image is too big.

pub async fn create_integration(
self,
http: impl AsRef<Http>,
integration_id: impl Into<IntegrationId>,
kind: &str
) -> Result<()>

Available on crate feature model only.

Creates an integration for the guild.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn create_role(
self,
cache_http: impl CacheHttp,
builder: EditRole<'_>
) -> Result<Role>

Available on crate feature model only.

Creates a new role in the guild with the data set, if any.

See the documentation for Guild::create_role on how to use this.

Note: Requires the Manage Roles permission.

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn create_scheduled_event(
self,
cache_http: impl CacheHttp,
builder: CreateScheduledEvent<'_>
) -> Result<ScheduledEvent>

Available on crate feature model only.

Creates a new scheduled event in the guild with the data set, if any.

Note: Requires the Manage Events permission.

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn create_sticker(
self,
cache_http: impl CacheHttp,
builder: CreateSticker<'_>
) -> Result<Sticker>

Available on crate feature model only.

Creates a new sticker in the guild with the data set, if any.

Note: Requires the Manage Emojis and Stickers permission.

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn delete(self, http: impl AsRef<Http>) -> Result<()>

Available on crate feature model only.

Deletes the current guild if the current account is the owner of the guild.

Refer to Guild::delete for more information.

Note: Requires the current user to be the owner of the guild.

Errors

Returns Error::Http if the current user is not the owner of the guild.

pub async fn delete_emoji(
self,
http: impl AsRef<Http>,
emoji_id: impl Into<EmojiId>
) -> Result<()>

Available on crate feature model only.

Deletes an Emoji from the guild.

Note: Requires the Manage Emojis and Stickers permission.

Errors

Returns Error::Http if the current user lacks permission, or if an Emoji with that Id does not exist.

pub async fn delete_integration(
self,
http: impl AsRef<Http>,
integration_id: impl Into<IntegrationId>
) -> Result<()>

Available on crate feature model only.

Deletes an integration by Id from the guild.

Note: Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission, or if an integration with that Id does not exist.

pub async fn delete_role(
self,
http: impl AsRef<Http>,
role_id: impl Into<RoleId>
) -> Result<()>

Available on crate feature model only.

Deletes a Role by Id from the guild.

Also see Role::delete if you have the cache and model features enabled.

Note: Requires the Manage Roles permission.

Errors

Returns Error::Http if the current user lacks permission, or if a role with that Id does not exist.

pub async fn delete_scheduled_event(
self,
http: impl AsRef<Http>,
event_id: impl Into<ScheduledEventId>
) -> Result<()>

Available on crate feature model only.

Deletes a specified scheduled event in the guild.

Note: Requires the Manage Events permission.

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub async fn delete_sticker(
self,
http: impl AsRef<Http>,
sticker_id: impl Into<StickerId>
) -> Result<()>

Available on crate feature model only.

Deletes a Sticker by Id from the guild.

Note: Requires the Manage Emojis and Stickers permission.

Errors

Returns Error::Http if the current user lacks permission, or if a sticker with that Id does not exist.

pub async fn edit(
self,
cache_http: impl CacheHttp,
builder: EditGuild<'_>
) -> Result<PartialGuild>

Available on crate feature model only.

Edits the current guild with new data where specified.

Note: Requires the Manage Guild permission.

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn edit_emoji(
self,
http: impl AsRef<Http>,
emoji_id: impl Into<EmojiId>,
name: &str
) -> Result<Emoji>

Available on crate feature model only.

Edits an Emoji’s name in the guild.

Also see Emoji::edit if you have the cache and methods features enabled.

Requires the Manage Emojis and Stickers permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn edit_member(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
builder: EditMember<'_>
) -> Result<Member>

Available on crate feature model only.

Edits the properties a guild member, such as muting or nicknaming them. Returns the new member.

Refer to the documentation of EditMember for a full list of methods and permission restrictions.

Examples

Mute a member and set their roles to just one role with a predefined Id:

let builder = EditMember::new().mute(true).roles(vec![role_id]);
let _ = GuildId::new(7).edit_member(&http, user_id, builder).await?;

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub async fn edit_nickname(
self,
http: impl AsRef<Http>,
new_nickname: Option<&str>
) -> Result<()>

Available on crate feature model only.

Edits the current user’s nickname for the guild.

Pass None to reset the nickname.

Requires the Change Nickname permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn edit_role(
self,
cache_http: impl CacheHttp,
role_id: impl Into<RoleId>,
builder: EditRole<'_>
) -> Result<Role>

Available on crate feature model only.

Edits a Role, optionally setting its new fields.

Note: Requires the Manage Roles permission.

Examples

Make a role hoisted, and change its name:

// assuming a `role_id` and `guild_id` has been bound
let builder = EditRole::new().name("a test role").hoist(true);
let role = guild_id.edit_role(&http, role_id, builder).await?;

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn edit_scheduled_event(
self,
cache_http: impl CacheHttp,
event_id: impl Into<ScheduledEventId>,
builder: EditScheduledEvent<'_>
) -> Result<ScheduledEvent>

Available on crate feature model only.

Modifies a scheduled event in the guild with the data set, if any.

Note: Requires the Manage Events permission.

Errors

If the cache is enabled, returns a ModelError::InvalidPermissions if the current user lacks permission. Otherwise returns Error::Http, as well as if invalid data is given.

pub async fn edit_sticker(
self,
http: impl AsRef<Http>,
sticker_id: impl Into<StickerId>,
builder: EditSticker<'_>
) -> Result<Sticker>

Available on crate feature model only.

Edits a sticker.

Note: Requires the Manage Emojis and Stickers permission.

Examples

Rename a sticker:

use serenity::builder::EditSticker;
use serenity::model::id::{GuildId, StickerId};

let builder = EditSticker::new().name("Bun bun meow");
let _ = GuildId::new(7).edit_sticker(&http, StickerId::new(7), builder).await?;

Errors

Returns Error::Http if the current user lacks permission, or if invalid data is given.

pub async fn edit_role_position(
self,
http: impl AsRef<Http>,
role_id: impl Into<RoleId>,
position: u32
) -> Result<Vec<Role>>

Available on crate feature model only.

Edit the position of a Role relative to all others in the Guild.

Note: Requires the Manage Roles permission.

Examples

use serenity::model::{GuildId, RoleId};
GuildId::new(7).edit_role_position(&context, RoleId::new(8), 2);

Errors

Returns an Error::Http if the current user lacks permission.

pub async fn edit_welcome_screen(
self,
http: impl AsRef<Http>,
builder: EditGuildWelcomeScreen<'_>
) -> Result<GuildWelcomeScreen>

Available on crate feature model only.

Edits the guild’s welcome screen.

Note: Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn edit_widget(
self,
http: impl AsRef<Http>,
builder: EditGuildWidget<'_>
) -> Result<GuildWidget>

Available on crate feature model only.

Edits the guild’s widget.

Note: Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn roles(self, http: impl AsRef<Http>) -> Result<HashMap<RoleId, Role>>

Available on crate feature model only.

Gets all of the guild’s roles over the REST API.

Errors

Returns Error::Http if the current user is not in the guild.

pub fn to_guild_cached(self, cache: &impl AsRef<Cache>) -> Option<GuildRef<'_>>

Available on crate features model and cache only.

Tries to find the Guild by its Id in the cache.

pub async fn to_partial_guild(
self,
cache_http: impl CacheHttp
) -> Result<PartialGuild>

Available on crate feature model only.

Requests PartialGuild over REST API.

Note: This will not be a Guild, as the REST API does not send all data with a guild retrieval.

Errors

Returns an Error::Http if the current user is not in the guild.

pub async fn to_partial_guild_with_counts(
self,
http: impl AsRef<Http>
) -> Result<PartialGuild>

Available on crate feature model only.

Requests PartialGuild over REST API with counts.

Note: This will not be a Guild, as the REST API does not send all data with a guild retrieval.

Errors

Returns an Error::Http if the current user is not in the guild.

pub async fn emojis(self, http: impl AsRef<Http>) -> Result<Vec<Emoji>>

Available on crate feature model only.

Gets all Emojis of this guild via HTTP.

Errors

Returns an Error::Http if the guild is unavailable.

pub async fn emoji(
self,
http: impl AsRef<Http>,
emoji_id: EmojiId
) -> Result<Emoji>

Available on crate feature model only.

Gets an Emoji of this guild by its ID via HTTP.

Errors

Returns an Error::Http if an emoji with that Id does not exist.

pub async fn stickers(self, http: impl AsRef<Http>) -> Result<Vec<Sticker>>

Available on crate feature model only.

Gets all Stickers of this guild via HTTP.

Errors

Returns an Error::Http if the guild is unavailable.

pub async fn sticker(
self,
http: impl AsRef<Http>,
sticker_id: StickerId
) -> Result<Sticker>

Available on crate feature model only.

Gets an Sticker of this guild by its ID via HTTP.

Errors

Returns an Error::Http if an sticker with that Id does not exist.

pub async fn integrations(
self,
http: impl AsRef<Http>
) -> Result<Vec<Integration>>

Available on crate feature model only.

Gets all integration of the guild.

Requires the Manage Guild permission.

Errors

Returns an Error::Http if the current user lacks permission, also may return Error::Json if there is an error in deserializing the API response.

pub async fn invites(self, http: impl AsRef<Http>) -> Result<Vec<RichInvite>>

Available on crate feature model only.

Gets all of the guild’s invites.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission, also may return Error::Json if there is an error in deserializing the API response.

pub async fn kick(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<()>

Available on crate feature model only.

Kicks a Member from the guild.

Requires the Kick Members permission.

Errors

Returns Error::Http if the member cannot be kicked by the current user.

pub async fn kick_with_reason(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
reason: &str
) -> Result<()>

Available on crate feature model only.

Errors

In addition to the reasons Self::kick may return an error, may also return an error if the reason is too long.

pub async fn leave(self, http: impl AsRef<Http>) -> Result<()>

Available on crate feature model only.

Leaves the guild.

Errors

May return an Error::Http if the current user cannot leave the guild, or currently is not in the guild.

pub async fn member(
self,
cache_http: impl CacheHttp,
user_id: impl Into<UserId>
) -> Result<Member>

Available on crate feature model only.

Gets a user’s Member for the guild by Id.

If the cache feature is enabled the cache will be checked first. If not found it will resort to an http request.

Errors

Returns an Error::Http if the user is not in the guild, or if the guild is otherwise unavailable

pub async fn members(
self,
http: impl AsRef<Http>,
limit: Option<u64>,
after: impl Into<Option<UserId>>
) -> Result<Vec<Member>>

Available on crate feature model only.

Gets a list of the guild’s members.

Optionally pass in the limit to limit the number of results. Minimum value is 1, maximum and default value is 1000.

Optionally pass in after to offset the results by a User’s Id.

Errors

Returns an Error::Http if the API returns an error, may also return Error::NotInRange if the input is not within range.

pub fn members_iter<H: AsRef<Http>>(
self,
http: H
) -> impl Stream<Item = Result<Member>>

Available on crate feature model only.

Streams over all the members in a guild.

This is accomplished and equivalent to repeated calls to Self::members. A buffer of at most 1,000 members is used to reduce the number of calls necessary.

Examples

use serenity::futures::StreamExt;
use serenity::model::guild::MembersIter;

let mut members = guild_id.members_iter(&ctx).boxed();
while let Some(member_result) = members.next().await {
    match member_result {
        Ok(member) => println!("{} is {}", member, member.display_name(),),
        Err(error) => eprintln!("Uh oh!  Error: {}", error),
    }
}

pub async fn move_member(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
channel_id: impl Into<ChannelId>
) -> Result<Member>

Available on crate feature model only.

Moves a member to a specific voice channel.

Note: Requires the Move Members permission.

Errors

Returns Error::Http if the current user lacks permission, or if the member is not currently in a voice channel for this Guild.

pub fn name(self, cache: impl AsRef<Cache>) -> Option<String>

Available on crate features model and cache only.

Returns the name of whatever guild this id holds.

pub async fn disconnect_member(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<Member>

Available on crate feature model only.

Disconnects a member from a voice channel in the guild.

Note: Requires the Move Members permission.

Errors

Returns Error::Http if the current user lacks permission, or if the member is not currently in a voice channel for this Guild.

pub async fn prune_count(
self,
http: impl AsRef<Http>,
days: u8
) -> Result<GuildPrune>

Available on crate feature model only.

Gets the number of Members that would be pruned with the given number of days.

Requires the Kick Members permission.

Errors

Returns Error::Http if the current user does not have permission.

pub async fn reorder_channels<It>(
self,
http: impl AsRef<Http>,
channels: It
) -> Result<()>where
It: IntoIterator<Item = (ChannelId, u64)>,

Available on crate feature model only.

Re-orders the channels of the guild.

Accepts an iterator of a tuple of the channel ID to modify and its new position.

Although not required, you should specify all channels’ positions, regardless of whether they were updated. Otherwise, positioning can sometimes get weird.

Note: Requires the Manage Channels permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn search_members(
self,
http: impl AsRef<Http>,
query: &str,
limit: Option<u64>
) -> Result<Vec<Member>>

Available on crate feature model only.

Returns a list of Members in a Guild whose username or nickname starts with a provided string.

Optionally pass in the limit to limit the number of results. Minimum value is 1, maximum and default value is 1000.

Errors

Returns an Error::Http if the API returns an error.

pub async fn scheduled_event(
self,
http: impl AsRef<Http>,
event_id: impl Into<ScheduledEventId>,
with_user_count: bool
) -> Result<ScheduledEvent>

Available on crate feature model only.

Fetches a specified scheduled event in the guild, by Id. If with_user_count is set to true, then the user_count field will be populated, indicating the number of users interested in the event.

Note: Requires the Manage Events permission.

Errors

Returns Error::Http if the current user lacks permission, or if the provided Id is invalid.

pub async fn scheduled_events(
self,
http: impl AsRef<Http>,
with_user_count: bool
) -> Result<Vec<ScheduledEvent>>

Available on crate feature model only.

Fetches a list of all scheduled events in the guild. If with_user_count is set to true, then each event returned will have its user_count field populated.

Note: Requires the Manage Events permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn scheduled_event_users(
self,
http: impl AsRef<Http>,
event_id: impl Into<ScheduledEventId>,
limit: Option<u64>
) -> Result<Vec<ScheduledEventUser>>

Available on crate feature model only.

Fetches a list of interested users for the specified event.

If limit is left unset, by default at most 100 users are returned.

Note: Requires the Manage Events permission.

Errors

Returns Error::Http if the current user lacks permission, or if the provided Id is invalid.

pub async fn scheduled_event_users_optioned(
self,
http: impl AsRef<Http>,
event_id: impl Into<ScheduledEventId>,
limit: Option<u64>,
target: Option<UserPagination>,
with_member: Option<bool>
) -> Result<Vec<ScheduledEventUser>>

Available on crate feature model only.

Fetches a list of interested users for the specified event, with additional options and filtering. See Http::get_scheduled_event_users for details.

Note: Requires the Manage Events permission.

Errors

Returns Error::Http if the current user lacks permission, or if the provided Id is invalid.

pub fn shard_id(self, cache: impl AsRef<Cache>) -> u32

Available on crate features cache and utils and model only.

Returns the Id of the shard associated with the guild.

When the cache is enabled this will automatically retrieve the total number of shards.

Note: When the cache is enabled, this function unlocks the cache to retrieve the total number of shards in use. If you already have the total, consider using utils::shard_id.

pub async fn start_integration_sync(
self,
http: impl AsRef<Http>,
integration_id: impl Into<IntegrationId>
) -> Result<()>

Available on crate feature model only.

Starts an integration sync for the given integration Id.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission, or if an Integration with that Id does not exist.

pub async fn start_prune(
self,
http: impl AsRef<Http>,
days: u8
) -> Result<GuildPrune>

Available on crate feature model only.

Starts a prune of Members.

See the documentation on GuildPrune for more information.

Note: Requires the Kick Members permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn unban(
self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<()>

Available on crate feature model only.

Unbans a User from the guild.

Note: Requires the Ban Members permission.

Errors

Returns Error::Http if the current user does not have permission.

pub async fn vanity_url(self, http: impl AsRef<Http>) -> Result<String>

Available on crate feature model only.

Retrieve’s the guild’s vanity URL.

Note: Requires the Manage Guild permission.

Errors

Will return Error::Http if the current user lacks permission. Can also return Error::Json if there is an error deserializing the API response.

pub async fn webhooks(self, http: impl AsRef<Http>) -> Result<Vec<Webhook>>

Available on crate feature model only.

Retrieves the guild’s webhooks.

Note: Requires the Manage Webhooks permission.

Errors

Will return an Error::Http if the bot is lacking permissions. Can also return an Error::Json if there is an error deserializing the API response.

pub fn reply_collector(
self,
shard_messenger: &ShardMessenger
) -> MessageCollectorBuilder<'_>

Available on crate features model and collector only.

Returns a builder which can be awaited to obtain a message or stream of messages in this guild.

pub fn reaction_collector(
self,
shard_messenger: &ShardMessenger
) -> ReactionCollectorBuilder<'_>

Available on crate features model and collector only.

Returns a builder which can be awaited to obtain a message or stream of reactions sent in this guild.

pub async fn create_application_command(
self,
http: impl AsRef<Http>,
builder: CreateCommand
) -> Result<Command>

Available on crate feature model only.

Create a guild specific application Command.

Note: Unlike global commands, guild commands will update instantly.

Errors

See CreateCommand::execute for a list of possible errors.

pub async fn set_application_commands(
self,
http: impl AsRef<Http>,
commands: Vec<CreateCommand>
) -> Result<Vec<Command>>

Available on crate feature model only.

Override all guild application commands.

Errors

Returns the same errors as Self::create_application_command.

pub async fn create_application_command_permission(
self,
http: impl AsRef<Http>,
command_id: CommandId,
builder: CreateCommandPermissionsData
) -> Result<CommandPermission>

Available on crate feature model only.

Create a guild specific CommandPermission.

Note: It will update instantly.

Errors

See CreateCommandPermissionsData::execute for a list of possible errors.

pub async fn get_application_commands(
self,
http: impl AsRef<Http>
) -> Result<Vec<Command>>

Available on crate feature model only.

Get all guild application commands.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_application_command(
self,
http: impl AsRef<Http>,
command_id: CommandId
) -> Result<Command>

Available on crate feature model only.

Get a specific guild application command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn edit_application_command(
self,
http: impl AsRef<Http>,
command_id: CommandId,
builder: CreateCommand
) -> Result<Command>

Available on crate feature model only.

Edit a guild application command, given its Id.

Errors

See CreateCommand::execute for a list of possible errors.

pub async fn delete_application_command(
self,
http: impl AsRef<Http>,
command_id: CommandId
) -> Result<()>

Available on crate feature model only.

Delete guild application command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_application_commands_permissions(
self,
http: impl AsRef<Http>
) -> Result<Vec<CommandPermission>>

Available on crate feature model only.

Get all guild application commands permissions only.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_application_command_permissions(
self,
http: impl AsRef<Http>,
command_id: CommandId
) -> Result<CommandPermission>

Available on crate feature model only.

Get permissions for specific guild application command by its Id.

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn get_welcome_screen(
self,
http: impl AsRef<Http>
) -> Result<GuildWelcomeScreen>

Available on crate feature model only.

Get the guild welcome screen.

Errors

Returns Error::Http if the guild does not have a welcome screen.

pub async fn get_preview(self, http: impl AsRef<Http>) -> Result<GuildPreview>

Available on crate feature model only.

Get the guild preview.

Note: The bot need either to be part of the guild or the guild needs to have the DISCOVERABLE feature.

Errors

Returns Error::Http if the bot cannot see the guild preview, see the note.

pub async fn get_widget(self, http: impl AsRef<Http>) -> Result<GuildWidget>

Available on crate feature model only.

Get the guild widget.

Errors

Returns Error::Http if the bot does not have MANAGE_MESSAGES permission.

pub fn widget_image_url(self, style: GuildWidgetStyle) -> String

Available on crate feature model only.

Get the widget image URL.

pub async fn get_active_threads(
self,
http: impl AsRef<Http>
) -> Result<ThreadsData>

Available on crate feature model only.

Gets the guild active threads.

Errors

Returns Error::Http if there is an error in the deserialization, or if the bot issuing the request is not in the guild.

impl GuildId

pub const fn new(id: u64) -> Self

Creates a new Id from a u64

Panics

Panics if the id is zero.

pub const fn get(self) -> u64

Retrieves the inner ID as u64

pub fn created_at(&self) -> Timestamp

Retrieves the time that the Id was created at.

pub const fn as_inner(&self) -> &NonZeroU64

Immutably borrow inner Id.