Struct serenity::model::prelude::InviteCreateEvent

pub struct InviteCreateEvent {
    pub channel_id: ChannelId,
    pub code: String,
    pub created_at: Timestamp,
    pub guild_id: Option<GuildId>,
    pub inviter: Option<User>,
    pub max_age: u32,
    pub max_uses: u8,
    pub target_type: Option<InviteTargetType>,
    pub target_user: Option<User>,
    pub target_application: Option<Value>,
    pub temporary: bool,
    pub uses: u64,
}

Requires GatewayIntents::GUILD_INVITES and [`Permissions::MANAGE_CHANNELS´] permission.

Discord docs.

Fields
channel_id: ChannelId

Whether or not the invite is temporary (invited users will be kicked on disconnect unless Channel the invite is for.

code: String

Unique invite code.

created_at: Timestamp

Time at which the invite was created.

guild_id: Option<GuildId>

Guild of the invite.

inviter: Option<User>

User that created the invite.

max_age: u32

How long the invite is valid for (in seconds).

max_uses: u8

Maximum number of times the invite can be used.

target_type: Option<InviteTargetType>

Type of target for this voice channel invite.

target_user: Option<User>

User whose stream to display for this voice channel stream invite.

target_application: Option<Value>

Embedded application to open for this voice channel embedded application invite.

temporary: bool

they’re assigned a role).

uses: u64

How many times the invite has been used (always will be 0).