Struct serenity::model::prelude::MessageUpdateEvent

pub struct MessageUpdateEvent {

    pub id: MessageId,
    pub channel_id: ChannelId,
    pub author: Option<User>,
    pub content: Option<String>,
    pub timestamp: Option<Timestamp>,
    pub edited_timestamp: Option<Timestamp>,
    pub tts: Option<bool>,
    pub mention_everyone: Option<bool>,
    pub mentions: Option<Vec<User>>,
    pub mention_roles: Option<Vec<RoleId>>,
    pub mention_channels: Option<Vec<ChannelMention>>,
    pub attachments: Option<Vec<Attachment>>,
    pub embeds: Option<Vec<Embed>>,
    pub reactions: Option<Vec<MessageReaction>>,
    pub pinned: Option<bool>,
    pub webhook_id: Option<Option<WebhookId>>,
    pub kind: Option<MessageType>,
    pub activity: Option<Option<MessageActivity>>,
    pub application: Option<Option<MessageApplication>>,
    pub application_id: Option<Option<ApplicationId>>,
    pub message_reference: Option<Option<MessageReference>>,
    pub flags: Option<Option<MessageFlags>>,
    pub referenced_message: Option<Option<Box<Message>>>,
    pub interaction: Option<Option<Box<MessageInteraction>>>,
    pub thread: Option<Option<GuildChannel>>,
    pub components: Option<Vec<ActionRow>>,
    pub sticker_items: Option<Vec<StickerItem>>,
    pub position: Option<Option<u64>>,
    pub role_subscription_data: Option<Option<RoleSubscriptionData>>,
    pub guild_id: Option<GuildId>,
    pub member: Option<Option<Box<PartialMember>>>,

}

Requires GatewayIntents::GUILD_MESSAGES.

Contains identical fields to Message, except everything but id and channel_id are optional. Even fields that cannot change in a message update event are included, because Discord may include them anyways, independent from whether they have actually changed (like Self::guild_id)

Discord docs.

Fields
id: MessageIdchannel_id: ChannelIdauthor: Option<User>content: Option<String>timestamp: Option<Timestamp>edited_timestamp: Option<Timestamp>tts: Option<bool>mention_everyone: Option<bool>mentions: Option<Vec<User>>mention_roles: Option<Vec<RoleId>>mention_channels: Option<Vec<ChannelMention>>attachments: Option<Vec<Attachment>>embeds: Option<Vec<Embed>>reactions: Option<Vec<MessageReaction>>pinned: Option<bool>webhook_id: Option<Option<WebhookId>>kind: Option<MessageType>activity: Option<Option<MessageActivity>>application: Option<Option<MessageApplication>>application_id: Option<Option<ApplicationId>>message_reference: Option<Option<MessageReference>>flags: Option<Option<MessageFlags>>referenced_message: Option<Option<Box<Message>>>interaction: Option<Option<Box<MessageInteraction>>>thread: Option<Option<GuildChannel>>components: Option<Vec<ActionRow>>sticker_items: Option<Vec<StickerItem>>position: Option<Option<u64>>role_subscription_data: Option<Option<RoleSubscriptionData>>guild_id: Option<GuildId>member: Option<Option<Box<PartialMember>>>
Implementations

impl MessageUpdateEvent

pub fn apply_to_message(&self, message: &mut Message)

Writes the updated data in this message update event into the given Message.