Struct serenity::model::prelude::PartialMember

#[non_exhaustive]
pub struct PartialMember {
    pub deaf: bool,
    pub joined_at: Option<Timestamp>,
    pub mute: bool,
    pub nick: Option<String>,
    pub roles: Vec<RoleId>,
    pub pending: bool,
    pub premium_since: Option<Timestamp>,
    pub guild_id: Option<GuildId>,
    pub user: Option<User>,
    pub permissions: Option<Permissions>,
}

A partial amount of data for a member.

This is used in Messages from Guilds.

Discord docs, subset specification unknown

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

deaf: bool

Indicator of whether the member can hear in voice channels.

joined_at: Option<Timestamp>

Timestamp representing the date when the member joined.

mute: bool

Indicator of whether the member can speak in voice channels

nick: Option<String>

The member’s nickname, if present.

Can’t be longer than 32 characters.

roles: Vec<RoleId>

Vector of Ids of Roles given to the member.

pending: bool

Indicator that the member hasn’t accepted the rules of the guild yet.

premium_since: Option<Timestamp>

Timestamp representing the date since the member is boosting the guild.

guild_id: Option<GuildId>

The unique Id of the guild that the member is a part of.

user: Option<User>

Attached User struct.

permissions: Option<Permissions>

The total permissions of the member in a channel, including overrides.

This is only Some when returned in an Interaction object.