Struct serenity::model::prelude::Reaction

#[non_exhaustive]
pub struct Reaction {
    pub channel_id: ChannelId,
    pub emoji: ReactionType,
    pub message_id: MessageId,
    pub user_id: Option<UserId>,
    pub guild_id: Option<GuildId>,
    pub member: Option<PartialMember>,
}

An emoji reaction to a message.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

channel_id: ChannelId

The Channel of the associated Message.

emoji: ReactionType

The reactive emoji used.

message_id: MessageId

The Id of the Message that was reacted to.

user_id: Option<UserId>

The Id of the User that sent the reaction.

Set to None by Message::react when cache is not available.

guild_id: Option<GuildId>

The optional Id of the Guild where the reaction was sent.

member: Option<PartialMember>

The optional object of the member which added the reaction.

Implementations

impl<'de> Reaction

pub fn deserialize<__D>(__deserializer: __D) -> Result<Reaction, __D::Error>where __D: Deserializer<'de>,

impl Reaction

pub fn serialize<__S>( __self: &Reaction, __serializer: __S ) -> Result<__S::Ok, __S::Error>where __S: Serializer,

impl Reaction

pub async fn channel(&self, cache_http: impl CacheHttp) -> Result<Channel>

Available on crate feature model only.

Retrieves the associated the reaction was made in.

If the cache is enabled, this will search for the already-cached channel. If not - or the channel was not found - this will perform a request over the REST API for the channel.

Requires the Read Message History permission.

Errors

Returns Error::Http if the current user lacks permission, or if the channel no longer exists.

pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<()>

Available on crate feature model only.

Deletes the reaction, but only if the current user is the user who made the reaction or has permission to.

Requires the Manage Messages permission, if the current user did not perform the reaction.

Errors

If the cache is enabled, then returns a ModelError::InvalidPermissions if the current user does not have the required permissions.

Otherwise returns Error::Http if the current user lacks permission.

pub async fn delete_all(&self, cache_http: impl CacheHttp) -> Result<()>

Available on crate feature model only.

Deletes all reactions from the message with this emoji.

Requires the Manage Messages permission

Errors

If the cache is enabled, then returns a ModelError::InvalidPermissions if the current user does not have the required permissions.

Otherwise returns Error::Http if the current user lacks permission.

pub async fn message(&self, cache_http: impl CacheHttp) -> Result<Message>

Available on crate feature model only.

Retrieves the Message associated with this reaction.

Requires the Read Message History permission.

Errors

Returns Error::Http if the current user lacks permission to read message history, or if the message was deleted.

pub async fn user(&self, cache_http: impl CacheHttp) -> Result<User>

Available on crate feature model only.

Retrieves the user that made the reaction.

If the cache is enabled, this will search for the already-cached user. If not - or the user was not found - this will perform a request over the REST API for the user.

Errors

Returns Error::Http if the user that made the reaction is unable to be retrieved from the API.

pub async fn users<R, U>( &self, http: impl AsRef<Http>, reaction_type: R, limit: Option<u8>, after: Option<U> ) -> Result<Vec<User>>where R: Into<ReactionType>, U: Into<UserId>,

Available on crate feature model only.

Retrieves the list of Users who have reacted to a Message with a certain Emoji.

The default limit is 50 - specify otherwise to receive a different maximum number of users. The maximum that may be retrieve at a time is 100, if a greater number is provided then it is automatically reduced.

The optional after attribute is to retrieve the users after a certain user. This is useful for pagination.

Requires the Read Message History permission.

Note: This will send a request to the REST API.

Errors

Returns a ModelError::InvalidPermissions if the current user does not have the required permissions.