Struct serenity::model::prelude::Role

#[non_exhaustive]
pub struct Role {
    pub id: RoleId,
    pub guild_id: GuildId,
    pub colour: Colour,
    pub hoist: bool,
    pub managed: bool,
    pub mentionable: bool,
    pub name: String,
    pub permissions: Permissions,
    pub position: u32,
    pub tags: RoleTags,
    pub icon: Option<String>,
    pub unicode_emoji: Option<String>,
}

Information about a role within a guild. A role represents a set of permissions, and can be attached to one or multiple users. A role has various miscellaneous configurations, such as being assigned a colour. Roles are unique per guild and do not cross over to other guilds in any way, and can have channel-specific permission overrides in addition to guild-level permissions.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: RoleId

The Id of the role. Can be used to calculate the role’s creation date.

guild_id: GuildId

The Id of the Guild the Role is in.

colour: Colour

The colour of the role.

hoist: bool

Indicator of whether the role is pinned above lesser roles.

In the client, this causes Members in the role to be seen above those in roles with a lower Self::position.

managed: bool

Indicator of whether the role is managed by an integration service.

mentionable: bool

Indicator of whether the role can be mentioned, similar to mentioning a specific member or @everyone.

Only members of the role will be notified if a role is mentioned with this set to true.

name: String

The name of the role.

permissions: Permissions

A set of permissions that the role has been assigned.

See the permissions module for more information.

position: u32

The role’s position in the position list. Roles are considered higher in hierarchy if their position is higher.

The @everyone role is usually either -1 or 0.

tags: RoleTags

The tags this role has. It can be used to determine if this role is a special role in this guild such as guild subscriber role, or if the role is linked to an Integration or a bot.

icon: Option<String>

Role icon image hash.

role-icons/<role_id>/<hash>.png - PNG, JPEG, WEBP role-icons/<role_id>/a_<hash>.gif - GIF, Animated WEBP

unicode_emoji: Option<String>

Role unicoded image.

Implementations

impl Role

pub async fn delete(&mut self, http: impl AsRef<Http>) -> Result<()>

Available on crate feature model only.

Deletes the role.

Note Requires the Manage Roles permission.

Errors

Returns Error::Http if the current user lacks permission to delete this role.

pub async fn edit( &mut self, http: impl AsRef<Http>, builder: EditRole<'_> ) -> Result<()>

Available on crate feature model only.

Edits a Role, optionally setting its new fields.

Requires the Manage Roles permission.

Examples

See the documentation of EditRole for details.

Errors

Returns Error::Http if the current user does not have permission to Manage Roles.

pub fn has_permission(&self, permission: Permissions) -> bool

Available on crate feature model only.

Check that the role has the given permission.

pub fn has_permissions(&self, permissions: Permissions, precise: bool) -> bool

Available on crate feature model only.

Checks whether the role has all of the given permissions.

The ‘precise’ argument is used to check if the role’s permissions are precisely equivalent to the given permissions. If you need only check that the role has at least the given permissions, pass false.