Struct serenity::model::ApplicationInfo

pub struct ApplicationInfo {
    pub bot: Option<BotApplication>,
    pub bot_public: bool,
    pub bot_require_code_grant: bool,
    pub description: String,
    pub flags: Option<u64>,
    pub icon: Option<String>,
    pub id: UserId,
    pub name: String,
    pub redirect_uris: Vec<String>,
    pub rpc_origins: Vec<String>,
    pub secret: String,
}

Information about a user's application. An application does not necessarily have an associated bot user.

Fields
bot: Option<BotApplication>

The bot user associated with the application. See BotApplication for more information.

bot_public: bool

Indicator of whether the bot is public.

If a bot is public, anyone may invite it to their Guild. While a bot is private, only the owner may add it to a guild.

bot_require_code_grant: bool

Indicator of whether the bot requires an OAuth2 code grant.

description: String

A description of the application, assigned by the application owner.

flags: Option<u64>

A set of bitflags assigned to the application, which represent gated feature flags that have been enabled for the application.

icon: Option<String>

A hash pointing to the application's icon.

This is not necessarily equivalent to the bot user's avatar.

id: UserId

The unique numeric Id of the application.

name: String

The name assigned to the application by the application owner.

redirect_uris: Vec<String>

A list of redirect URIs assigned to the application.

rpc_origins: Vec<String>

A list of RPC Origins assigned to the application.

secret: String

The given secret to the application.

This is not equivalent to the application's bot user's token.