Struct serenity::model::ChannelId

pub struct ChannelId(pub u64);

An identifier for a Channel

Methods

impl ChannelId

fn broadcast_typing(&self) -> Result<()>

Broadcasts that the current user is typing to a channel for the next 5 seconds.

After 5 seconds, another request must be made to continue broadcasting that the current user is typing.

This should rarely be used for bots, and should likely only be used for signifying that a long-running command is still being executed.

Note: Requires the Send Messages permission.

Examples

use serenity::model::ChannelId;

let _successful = ChannelId(7).broadcast_typing();

fn create_permission(&self, target: &PermissionOverwrite) -> Result<()>

Creates a permission overwrite for either a single Member or Role within the channel.

Refer to the documentation for GuildChannel::create_permission for more information.

Requires the Manage Channels permission.

fn create_reaction<M, R>(&self, message_id: M, reaction_type: R) -> Result<()> where
M: Into<MessageId>,
R: Into<ReactionType>,

React to a Message with a custom Emoji or unicode character.

Message::react may be a more suited method of reacting in most cases.

Requires the Add Reactions permission, if the current user is the first user to perform a react with a certain emoji.

fn delete(&self) -> Result<Channel>

Deletes this channel, returning the channel on a successful deletion.

fn delete_message<M: Into<MessageId>>(&self, message_id: M) -> Result<()>

Deletes a Message given its Id.

Refer to Message::delete for more information.

Requires the Manage Messages permission, if the current user is not the author of the message.

fn delete_messages(&self, message_ids: &[MessageId]) -> Result<()>

Deletes all messages by Ids from the given vector in the given channel.

Refer to the documentation for Channel::delete_messages for more information.

Requires the Manage Messages permission.

Note: This uses bulk delete endpoint which is not available for user accounts.

Note: Messages that are older than 2 weeks can't be deleted using this method.

fn delete_permission(
&self,
permission_type: PermissionOverwriteType
) -> Result<()>

Deletes all permission overrides in the channel from a member or role.

Note: Requires the Manage Channel permission.

fn delete_reaction<M, R>(
&self,
message_id: M,
user_id: Option<UserId>,
reaction_type: R
) -> Result<()> where
M: Into<MessageId>,
R: Into<ReactionType>,

Deletes the given Reaction from the channel.

Note: Requires the Manage Messages permission, if the current user did not perform the reaction.

fn edit<F: FnOnce(EditChannel) -> EditChannel>(
&self,
f: F
) -> Result<GuildChannel>

Edits the settings of a Channel, optionally setting new values.

Refer to EditChannel's documentation for its methods.

Requires the Manage Channel permission.

Examples

Change a voice channel's name and bitrate:

// assuming a `channel_id` has been bound

channel_id.edit(|c| c.name("test").bitrate(64000));

fn edit_message<F, M>(&self, message_id: M, f: F) -> Result<Message> where
F: FnOnce(CreateMessage) -> CreateMessage,
M: Into<MessageId>,

Edits a Message in the channel given its Id.

Message editing preserves all unchanged message data.

Refer to the documentation for CreateMessage for more information regarding message restrictions and requirements.

Note: Requires that the current user be the author of the message.

Errors

Returns a ModelError::MessageTooLong if the content of the message is over the the limit, containing the number of unicode code points over the limit.

fn find(&self) -> Option<Channel>

Search the cache for the channel with the Id.

fn get(&self) -> Result<Channel>

Search the cache for the channel. If it can't be found, the channel is requested over REST.

fn invites(&self) -> Result<Vec<RichInvite>>

Gets all of the channel's invites.

Requires the Manage Channels permission.

fn message<M: Into<MessageId>>(&self, message_id: M) -> Result<Message>

Gets a message from the channel.

Requires the Read Message History permission.

fn messages<F>(&self, f: F) -> Result<Vec<Message>> where
F: FnOnce(GetMessages) -> GetMessages,

Gets messages from the channel.

Refer to Channel::messages for more information.

Requires the Read Message History permission.

fn name(&self) -> Option<String>

Returns the name of whatever channel this id holds.

fn pin<M: Into<MessageId>>(&self, message_id: M) -> Result<()>

Pins a Message to the channel.

fn pins(&self) -> Result<Vec<Message>>

Gets the list of Messages which are pinned to the channel.

fn reaction_users<M, R, U>(
&self,
message_id: M,
reaction_type: R,
limit: Option<u8>,
after: Option<U>
) -> Result<Vec<User>> where
M: Into<MessageId>,
R: Into<ReactionType>,
U: Into<UserId>,

Gets the list of Users who have reacted to a Message with a certain Emoji.

Refer to Channel::reaction_users for more information.

Note: Requires the Read Message History permission.

fn say<D: Display>(&self, content: D) -> Result<Message>

Sends a message with just the given message content in the channel.

Errors

Returns a ModelError::MessageTooLong if the content of the message is over the above limit, containing the number of unicode code points over the limit.

fn send_files<'a, F, T>(&self, files: Vec<T>, f: F) -> Result<Message> where
F: FnOnce(CreateMessage) -> CreateMessage,
T: Into<AttachmentType<'a>>,

Sends a file along with optional message contents. The filename must be specified.

Message contents may be passed by using the CreateMessage::content method.

An embed can not be sent when sending a file. If you set one, it will be automatically removed.

The Attach Files and Send Messages permissions are required.

Note: Message contents must be under 2000 unicode code points.

Examples

Send files with the paths /path/to/file.jpg and /path/to/file2.jpg:

use serenity::model::ChannelId;

let channel_id = ChannelId(7);

let paths = vec!["/path/to/file.jpg", "path/to/file2.jpg"];

let _ = channel_id.send_files(paths, |m| m.content("a file"));

Send files using File:

use serenity::model::ChannelId;
use std::fs::File;

let channel_id = ChannelId(7);

let f1 = File::open("my_file.jpg").unwrap();
let f2 = File::open("my_file2.jpg").unwrap();

let files = vec![(&f1, "my_file.jpg"), (&f2, "my_file2.jpg")];

let _ = channel_id.send_files(files, |m| m.content("a file"));

Errors

If the content of the message is over the above limit, then a ClientError::MessageTooLong will be returned, containing the number of unicode code points over the limit.

Returns an HttpError::InvalidRequest(PayloadTooLarge) if the file is too large to send.

fn send_message<F>(&self, f: F) -> Result<Message> where
F: FnOnce(CreateMessage) -> CreateMessage,

Sends a message to the channel.

Refer to the documentation for CreateMessage for more information regarding message restrictions and requirements.

Requires the Send Messages permission.

Note: Message contents must be under 2000 unicode code points.

Errors

Returns a ModelError::MessageTooLong if the content of the message is over the above limit, containing the number of unicode code points over the limit.

fn unpin<M: Into<MessageId>>(&self, message_id: M) -> Result<()>

Unpins a Message in the channel given by its Id.

Requires the Manage Messages permission.

fn webhooks(&self) -> Result<Vec<Webhook>>

Retrieves the channel's webhooks.

Note: Requires the Manage Webhooks permission.

impl ChannelId

fn created_at(&self) -> NaiveDateTime

Retrieves the time that the Id was created at.