Struct serenity::model::CurrentUser

pub struct CurrentUser {
    pub id: UserId,
    pub avatar: Option<String>,
    pub bot: bool,
    pub discriminator: u16,
    pub email: Option<String>,
    pub mfa_enabled: bool,
    pub name: String,
    pub verified: bool,
}

Information about the current user.

Fields
id: UserId avatar: Option<String> bot: bool discriminator: u16 email: Option<String> mfa_enabled: bool name: String verified: bool
Methods

impl CurrentUser

fn avatar_url(&self) -> Option<String>

Returns the formatted URL of the user's icon, if one exists.

This will produce a WEBP image URL, or GIF if the user has a GIF avatar.

Examples

Print out the current user's avatar url if one is set:

// assuming the cache has been unlocked
let user = &cache.user;

match user.avatar_url() {
    Some(url) => println!("{}'s avatar can be found at {}", user.name, url),
    None => println!("{} does not have an avatar set.", user.name)
}

fn default_avatar_url(&self) -> String

Returns the formatted URL to the user's default avatar URL.

This will produce a PNG URL.

fn edit<F>(&mut self, f: F) -> Result<()> where
F: FnOnce(EditProfile) -> EditProfile,

Edits the current user's profile settings.

This mutates the current user in-place.

Refer to EditProfile's documentation for its methods.

Examples

Change the avatar:

use serenity::CACHE;

let avatar = serenity::utils::read_image("./avatar.png").unwrap();

CACHE.write().unwrap().user.edit(|p| p.avatar(Some(&avatar)));

fn face(&self) -> String

Retrieves the URL to the current user's avatar, falling back to the default avatar if needed.

This will call avatar_url first, and if that returns None, it then falls back to default_avatar_url.

fn guilds(&self) -> Result<Vec<GuildInfo>>

Gets a list of guilds that the current user is in.

Examples

Print out the names of all guilds the current user is in:

// assuming the cache has been unlocked
let user = &cache.user;

if let Ok(guilds) = user.guilds() {
    for (index, guild) in guilds.into_iter().enumerate() {
        println!("{}: {}", index, guild.name);
    }
}

fn invite_url(&self, permissions: Permissions) -> Result<String>

Returns the invite url for the bot with the given permissions.

This queries the REST API for the client id.

If the permissions passed are empty, the permissions part will be dropped.

Examples

Get the invite url with no permissions set:

use serenity::model::permissions::Permissions;

// assuming the cache has been unlocked
let url = match cache.user.invite_url(Permissions::empty()) {
    Ok(v) => v,
    Err(why) => {
        println!("Error getting invite url: {:?}", why);

        return;
    },
};

assert_eq!(url, "https://discordapp.com/api/oauth2/authorize? \
                 client_id=249608697955745802&scope=bot");

Get the invite url with some basic permissions set:

use serenity::model::Permissions;

// assuming the cache has been unlocked
let url = match cache.user.invite_url(Permissions::READ_MESSAGES | Permissions::SEND_MESSAGES | Permissions::EMBED_LINKS) {
    Ok(v) => v,
    Err(why) => {
        println!("Error getting invite url: {:?}", why);

        return;
    },
};

assert_eq!(url,
"https://discordapp.
com/api/oauth2/authorize?client_id=249608697955745802&scope=bot&permissions=19456");

Errors

Returns an HttpError::InvalidRequest(Unauthorized) If the user is not authorized for this end point.

May return Error::Format while writing url to the buffer.

fn static_avatar_url(&self) -> Option<String>

Returns a static formatted URL of the user's icon, if one exists.

This will always produce a WEBP image URL.

Examples

Print out the current user's static avatar url if one is set:

// assuming the cache has been unlocked
let user = &cache.user;

match user.static_avatar_url() {
    Some(url) => println!("{}'s static avatar can be found at {}", user.name, url),
    None => println!("Could not get static avatar for {}.", user.name)
}

fn tag(&self) -> String

Returns the tag of the current user.

Examples

Print out the current user's distinct identifier (e.g., Username#1234):

// assuming the cache has been unlocked
println!("The current user's distinct identifier is {}", cache.user.tag());