Struct serenity::model::Emoji

pub struct Emoji {
    pub id: EmojiId,
    pub name: String,
    pub managed: bool,
    pub require_colons: bool,
    pub roles: Vec<RoleId>,
}

Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

Fields
id: EmojiId

The Id of the emoji.

name: String

The name of the emoji. It must be at least 2 characters long and can only contain alphanumeric characters and underscores.

managed: bool

Whether the emoji is managed via an Integration service.

require_colons: bool

Whether the emoji name needs to be surrounded by colons in order to be used by the client.

roles: Vec<RoleId>

A list of Roles that are allowed to use the emoji. If there are no roles specified, then usage is unrestricted.

Methods

impl Emoji

fn delete(&self) -> Result<()>

Deletes the emoji.

Note: The Manage Emojis permission is required.

Note: Only user accounts may use this method.

Examples

Delete a given emoji:

// assuming emoji has been set already
match emoji.delete() {
    Ok(()) => println!("Emoji deleted."),
    Err(_) => println!("Could not delete emoji.")
}

fn edit(&mut self, name: &str) -> Result<()>

Edits the emoji by updating it with a new name.

Note: The Manage Emojis permission is required.

Note: Only user accounts may use this method.

Examples

Change the name of an emoji:

// assuming emoji has been set already
let _ = emoji.edit("blobuwu");
assert_eq!(emoji.name, "blobuwu");

fn find_guild_id(&self) -> Option<GuildId>

Finds the Guild that owns the emoji by looking through the Cache.

Examples

Print the guild id that owns this emoji:

// assuming emoji has been set already
if let Some(guild_id) = emoji.find_guild_id() {
    println!("{} is owned by {}", emoji.name, guild_id);
}

fn url(&self) -> String

Generates a URL to the emoji's image.

Examples

Print the direct link to the given emoji:

// assuming emoji has been set already
println!("Direct link to emoji image: {}", emoji.url());