Struct serenity::model::GuildId

pub struct GuildId(pub u64);

An identifier for a Guild

Methods

impl GuildId

fn as_channel_id(&self) -> ChannelId

Deprecated
: The concept of default channels is no more, use Guild::default_channel{_guaranteed} to simulate the concept.

Converts the guild Id into the default channel's Id.

fn ban<U: Into<UserId>, BO: BanOptions>(
&self,
user: U,
ban_options: BO
) -> Result<()>

Ban a User from the guild. All messages by the user within the last given number of days given will be deleted.

Refer to the documentation for Guild::ban for more information.

Note: Requires the Ban Members permission.

Examples

Ban a member and remove all messages they've sent in the last 4 days:

use serenity::model::GuildId;

// assuming a `user` has already been bound
let _ = GuildId(81384788765712384).ban(user, 4);

Errors

Returns a ModelError::DeleteMessageDaysAmount if the number of days' worth of messages to delete is over the maximum.

fn bans(&self) -> Result<Vec<Ban>>

Gets a list of the guild's bans.

Requires the Ban Members permission.

fn audit_logs(&self) -> Result<AuditLogs>

Gets a list of the guild's audit log entries

fn channels(&self) -> Result<HashMap<ChannelId, GuildChannel>>

Gets all of the guild's channels over the REST API.

fn create_channel(&self, name: &str, kind: ChannelType) -> Result<GuildChannel>

Creates a GuildChannel in the the guild.

Refer to http::create_channel for more information.

Requires the Manage Channels permission.

Examples

Create a voice channel in a guild with the name test:

use serenity::model::{ChannelType, GuildId};

let _channel = GuildId(7).create_channel("test", ChannelType::Voice);

fn create_emoji(&self, name: &str, image: &str) -> Result<Emoji>

Creates an emoji in the guild with a name and base64-encoded image.

Refer to the documentation for Guild::create_emoji for more information.

Requires the Manage Emojis permission.

Examples

See the EditProfile::avatar example for an in-depth example as to how to read an image from the filesystem and encode it as base64. Most of the example can be applied similarly for this method.

fn create_integration<I>(&self, integration_id: I, kind: &str) -> Result<()> where
I: Into<IntegrationId>,

Creates an integration for the guild.

Requires the Manage Guild permission.

fn create_role<F: FnOnce(EditRole) -> EditRole>(&self, f: F) -> Result<Role>

Creates a new role in the guild with the data set, if any.

See the documentation for Guild::create_role on how to use this.

Note: Requires the Manage Roles permission.

fn delete(&self) -> Result<PartialGuild>

Deletes the current guild if the current account is the owner of the guild.

Refer to Guild::delete for more information.

Note: Requires the current user to be the owner of the guild.

fn delete_emoji<E: Into<EmojiId>>(&self, emoji_id: E) -> Result<()>

Deletes an Emoji from the guild.

Requires the Manage Emojis permission.

fn delete_integration<I: Into<IntegrationId>>(
&self,
integration_id: I
) -> Result<()>

Deletes an integration by Id from the guild.

Requires the Manage Guild permission.

fn delete_role<R: Into<RoleId>>(&self, role_id: R) -> Result<()>

Deletes a Role by Id from the guild.

Also see Role::delete if you have the cache and methods features enabled.

Requires the Manage Roles permission.

fn edit<F: FnOnce(EditGuild) -> EditGuild>(
&mut self,
f: F
) -> Result<PartialGuild>

Edits the current guild with new data where specified.

Refer to Guild::edit for more information.

Note: Requires the current user to have the Manage Guild permission.

fn edit_emoji<E: Into<EmojiId>>(&self, emoji_id: E, name: &str) -> Result<Emoji>

Edits an Emoji's name in the guild.

Also see Emoji::edit if you have the cache and methods features enabled.

Requires the Manage Emojis permission.

fn edit_member<F, U>(&self, user_id: U, f: F) -> Result<()> where
F: FnOnce(EditMember) -> EditMember,
U: Into<UserId>,

Edits the properties of member of the guild, such as muting or nicknaming them.

Refer to EditMember's documentation for a full list of methods and permission restrictions.

Examples

Mute a member and set their roles to just one role with a predefined Id:

guild.edit_member(user_id, |m| m.mute(true).roles(&vec![role_id]));

fn edit_nickname(&self, new_nickname: Option<&str>) -> Result<()>

Edits the current user's nickname for the guild.

Pass None to reset the nickname.

Requires the Change Nickname permission.

fn edit_role<F, R>(&self, role_id: R, f: F) -> Result<Role> where
F: FnOnce(EditRole) -> EditRole,
R: Into<RoleId>,

Edits a Role, optionally setting its new fields.

Requires the Manage Roles permission.

Examples

Make a role hoisted:

use serenity::model::{GuildId, RoleId};

GuildId(7).edit_role(RoleId(8), |r| r.hoist(true));

fn find(&self) -> Option<Arc<RwLock<Guild>>>

Search the cache for the guild.

fn get(&self) -> Result<PartialGuild>

Requests the guild over REST.

Note that this will not be a complete guild, as REST does not send all data with a guild retrieval.

fn integrations(&self) -> Result<Vec<Integration>>

Gets all integration of the guild.

This performs a request over the REST API.

fn invites(&self) -> Result<Vec<RichInvite>>

Gets all of the guild's invites.

Requires the Manage Guild permission.

fn kick<U: Into<UserId>>(&self, user_id: U) -> Result<()>

Kicks a Member from the guild.

Requires the Kick Members permission.

fn leave(&self) -> Result<()>

Leaves the guild.

fn member<U: Into<UserId>>(&self, user_id: U) -> Result<Member>

Gets a user's Member for the guild by Id.

fn members<U>(
&self,
limit: Option<u64>,
after: Option<U>
) -> Result<Vec<Member>> where
U: Into<UserId>,

Gets a list of the guild's members.

Optionally pass in the limit to limit the number of results. Maximum value is 1000. Optionally pass in after to offset the results by a User's Id.

fn move_member<C, U>(&self, user_id: U, channel_id: C) -> Result<()> where
C: Into<ChannelId>,
U: Into<UserId>,

Moves a member to a specific voice channel.

Requires the Move Members permission.

fn prune_count(&self, days: u16) -> Result<GuildPrune>

Gets the number of Members that would be pruned with the given number of days.

Requires the Kick Members permission.

fn shard_id(&self) -> u64

Returns the Id of the shard associated with the guild.

When the cache is enabled this will automatically retrieve the total number of shards.

Note: When the cache is enabled, this function unlocks the cache to retrieve the total number of shards in use. If you already have the total, consider using utils::shard_id.

fn start_integration_sync<I: Into<IntegrationId>>(
&self,
integration_id: I
) -> Result<()>

Starts an integration sync for the given integration Id.

Requires the Manage Guild permission.

fn start_prune(&self, days: u16) -> Result<GuildPrune>

Starts a prune of Members.

See the documentation on GuildPrune for more information.

Note: Requires the Kick Members permission.

fn unban<U: Into<UserId>>(&self, user_id: U) -> Result<()>

Unbans a User from the guild.

Requires the Ban Members permission.

fn webhooks(&self) -> Result<Vec<Webhook>>

Retrieves the guild's webhooks.

Note: Requires the Manage Webhooks permission.

impl GuildId

fn created_at(&self) -> NaiveDateTime

Retrieves the time that the Id was created at.