Struct serenity::model::RichInvite

pub struct RichInvite {
    pub channel: InviteChannel,
    pub code: String,
    pub created_at: DateTime<FixedOffset>,
    pub guild: InviteGuild,
    pub inviter: User,
    pub max_age: u64,
    pub max_uses: u64,
    pub temporary: bool,
    pub uses: u64,
}

Detailed information about an invite. This information can only be retrieved by anyone with the Manage Guild permission. Otherwise, a minimal amount of information can be retrieved via the Invite struct.

Fields
channel: InviteChannel

A representation of the minimal amount of information needed about the channel being invited to.

code: String

The unique code for the invite.

created_at: DateTime<FixedOffset>

When the invite was created.

guild: InviteGuild

A representation of the minimal amount of information needed about the guild being invited to.

inviter: User

The user that created the invite.

max_age: u64

The maximum age of the invite in seconds, from when it was created.

max_uses: u64

The maximum number of times that an invite may be used before it expires. Note that this does not supercede the max_age value, if the value of temporary is true. If the value of temporary is false, then the invite will self-expire after the given number of max uses. If the value is 0, then the invite is permanent.

temporary: bool

Indicator of whether the invite self-expires after a certain amount of time or uses.

uses: u64

The amount of times that an invite has been used.

Methods

impl RichInvite

fn delete(&self) -> Result<Invite>

Deletes the invite.

Refer to http::delete_invite for more information.

Note: Requires the Manage Guild permission.

Errors

If the cache feature is enabled, then this returns a ModelError::InvalidPermissions if the current user does not have the required permission.

fn url(&self) -> String

Returns a URL to use for the invite.

Examples

Retrieve the URL for an invite with the code WxZumR:

assert_eq!(invite.url(), "https://discord.gg/WxZumR");