Struct serenity::model::UserId

pub struct UserId(pub u64);

An identifier for a User

Methods

impl UserId

fn create_dm_channel(&self) -> Result<PrivateChannel>

Creates a direct message channel between the current user and the user. This can also retrieve the channel if one already exists.

fn find(&self) -> Option<Arc<RwLock<User>>>

Search the cache for the user with the Id.

fn get(&self) -> Result<User>

Gets a user by its Id over the REST API.

Note: The current user must be a bot user.

impl UserId

fn created_at(&self) -> NaiveDateTime

Retrieves the time that the Id was created at.