Struct serenity::model::Webhook

pub struct Webhook {
    pub id: WebhookId,
    pub avatar: Option<String>,
    pub channel_id: ChannelId,
    pub guild_id: Option<GuildId>,
    pub name: Option<String>,
    pub token: String,
    pub user: Option<User>,
}

A representation of a webhook, which is a low-effort way to post messages to channels. They do not necessarily require a bot user or authentication to use.

Fields
id: WebhookId

The unique Id.

Can be used to calculate the creation date of the webhook.

avatar: Option<String>

The default avatar.

This can be modified via ExecuteWebhook::avatar.

channel_id: ChannelId

The Id of the channel that owns the webhook.

guild_id: Option<GuildId>

The Id of the guild that owns the webhook.

name: Option<String>

The default name of the webhook.

This can be modified via ExecuteWebhook::username.

token: String

The webhook's secure token.

user: Option<User>

The user that created the webhook.

Note: This is not received when getting a webhook by its token.

Methods

impl Webhook

fn delete(&self) -> Result<()>

Deletes the webhook.

As this calls the http::delete_webhook_with_token function, authentication is not required.

fn edit(&mut self, name: Option<&str>, avatar: Option<&str>) -> Result<()>

Edits the webhook in-place. All fields are optional.

To nullify the avatar, pass Some(""). Otherwise, passing None will not modify the avatar.

Refer to http::edit_webhook for httprictions on editing webhooks.

As this calls the http::edit_webhook_with_token function, authentication is not required.

Examples

Editing a webhook's name:

use serenity::http;

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";

let mut webhook = http::get_webhook_with_token(id, token)
    .expect("valid webhook");

let _ = webhook.edit(Some("new name"), None).expect("Error editing");

Setting a webhook's avatar:

use serenity::http;

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";

let mut webhook = http::get_webhook_with_token(id, token)
    .expect("valid webhook");

let image = serenity::utils::read_image("./webhook_img.png")
    .expect("Error reading image");

let _ = webhook.edit(None, Some(&image)).expect("Error editing");

fn execute<F: FnOnce(ExecuteWebhook) -> ExecuteWebhook>(
&self,
wait: bool,
f: F
) -> Result<Option<Message>>

Executes a webhook with the fields set via the given builder.

The builder provides a method of setting only the fields you need, without needing to pass a long set of arguments.

Examples

Execute a webhook with message content of test:

use serenity::http;

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";

let mut webhook = http::get_webhook_with_token(id, token)
    .expect("valid webhook");

let _ = webhook.execute(false, |w| w.content("test")).expect("Error executing");

Execute a webhook with message content of test, overriding the username to serenity, and sending an embed:

use serenity::http;
use serenity::model::Embed;

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";

let mut webhook = http::get_webhook_with_token(id, token)
    .expect("valid webhook");

let embed = Embed::fake(|e| e
    .title("Rust's website")
    .description("Rust is a systems programming language that runs
                  blazingly fast, prevents segfaults, and guarantees
                  thread safety.")
    .url("https://rust-lang.org"));

let _ = webhook.execute(false, |w| w
    .content("test")
    .username("serenity")
    .embeds(vec![embed]))
    .expect("Error executing");

fn refresh(&mut self) -> Result<()>

Retrieves the latest information about the webhook, editing the webhook in-place.

As this calls the http::get_webhook_with_token function, authentication is not required.