Struct serenity::model::timestamp::Timestamp

pub struct Timestamp(_);

Representation of a Unix timestamp.

The struct implements the std::fmt::Display trait to format the underlying type as an RFC 3339 date and string such as 2016-04-30T11:18:25.796Z.

let timestamp: Timestamp = GuildId(175928847299117063).created_at();
assert_eq!(timestamp.unix_timestamp(), 1462015105);
assert_eq!(timestamp.to_string(), "2016-04-30T11:18:25.796Z");

Implementations

impl Timestamp

pub fn now() -> Self

Create a new Timestamp with the current date and time in UTC.

pub fn from_unix_timestamp(secs: i64) -> Result<Self, InvalidTimestamp>

Create a new Timestamp from a UNIX timestamp.

Errors

Returns Err if the value is invalid.

pub fn unix_timestamp(&self) -> i64

Returns the number of non-leap seconds since January 1, 1970 0:00:00 UTC

pub fn parse(input: &str) -> Result<Timestamp, ParseError>

Parse a timestamp from an RFC 3339 date and time string.

Examples

let timestamp = Timestamp::parse("2016-04-30T11:18:25Z").unwrap();
let timestamp = Timestamp::parse("2016-04-30T11:18:25+00:00").unwrap();
let timestamp = Timestamp::parse("2016-04-30T11:18:25.796Z").unwrap();

assert!(Timestamp::parse("2016-04-30T11:18:25").is_err());
assert!(Timestamp::parse("2016-04-30T11:18").is_err());

Errors

Returns Err if the string is not a valid RFC 3339 date and time string.