serenity::model

Module voice

Representations of voice information.

Structs

AnimationType

The type of emoji animation, for emoji reaction and soundboard effects.

VoiceChannelEffect

Information about an effect, such as an emoji reaction or a soundboard sound, sent in a voice channel.

VoiceRegion

Information about an available voice region.

VoiceRegionGeneratedOriginal

Information about an available voice region.

VoiceState

A user’s state within a voice channel.

VoiceStateGeneratedOriginal

A user’s state within a voice channel.