serenity::model

Module webhook

Webhook model and implementations.

Structs

Webhook

A representation of a webhook, which is a low-effort way to post messages to channels. They do not necessarily require a bot user or authentication to use.

WebhookChannel

WebhookGuild

The guild object returned by a Webhook, of type WebhookType::ChannelFollower.

Enums

WebhookType

A representation of a type of webhook.