serenity

Module prelude

A set of exports which can be helpful to use.

Note that the SerenityError re-export is equivalent to serenity::Error, although is re-exported as a separate name to remove likely ambiguity with other crate error enums.

Examples

Import all of the exports:

use serenity::prelude::*;

Re-exports

pub use crate::gateway::GatewayError;gateway

pub use crate::gateway::client::Client;gateway

pub use crate::gateway::client::Context;gateway

pub use crate::gateway::client::EventHandler;gateway

pub use crate::gateway::client::RawEventHandler;gateway

pub use crate::http::CacheHttp;http

pub use crate::http::HttpError;http

pub use crate::model::mention::Mentionable;

pub use crate::model::ModelError;model

pub use crate::model::gateway::GatewayIntents;model

pub use crate::secrets::Token;

Structs

Mutex

An asynchronous Mutex-like type.

RwLock

An asynchronous reader-writer lock.

Enums

SerenityError

A common error enum returned by most of the library’s functionality within a custom Result.