Struct serenity::CACHE

pub struct CACHE { /* fields omitted */ }

A mutable and lazily-initialized static binding. It can be accessed across any function and in any context.

This Cache instance is updated for every event received, so you do not need to maintain your own cache.

See the cache module documentation for more details.

The Cache itself is wrapped within an RwLock, which allows for multiple readers or at most one writer at a time across threads. This means that you may have multiple commands reading from the Cache concurrently.

Examples

Retrieve the current user's Id, by opening a Read guard:

use serenity::CACHE;

println!("{}", CACHE.read().unwrap().user.id);

By unwrap()ing, the thread managing an event dispatch will be blocked until the guard can be opened.

If you do not want to block the current thread, you may instead use RwLock::try_read. Refer to RwLock's documentation in the stdlib for more information.