Enum serenity::utils::ContentModifier

pub enum ContentModifier {
    Italic,
    Bold,
    Strikethrough,
    Code,
    Underline,
}

Formatting modifiers for MessageBuilder content pushes

Provides an enum of formatting modifiers for a string, for combination with string types and Content types.

Examples

Create a new Content type which describes a bold-italic "text":

use serenity::utils::ContentModifier::{Bold, Italic};
use serenity::utils::Content;
let content: Content = Bold + Italic + "text";

Variants
ItalicBoldStrikethroughCodeUnderline