serenity::utils

Function content_safe

pub fn content_safe(
    guild: &Guild,
    s: &str,
    options: ContentSafeOptions,
    users: &[User],
) -> String

Available on crate features cache and utils only.

Transforms role, channel, user, @everyone and @here mentions into raw text by using the Guild and the users passed in with users.

ContentSafeOptions decides what kind of mentions should be filtered and how the raw-text will be displayed.

Examples

Sanitise an @everyone mention.

use serenity::utils::{ContentSafeOptions, content_safe};

let with_mention = "@everyone";
let without_mention = content_safe(&guild, &with_mention, ContentSafeOptions::default(), &[]);

assert_eq!("@\u{200B}everyone", without_mention);

Filtering out mentions from a message.

use serenity::cache::Cache;
use serenity::model::channel::Message;
use serenity::utils::{ContentSafeOptions, content_safe};

fn filter_message(cache: &Cache, message: &Message) -> String {
    if let Some(guild) = message.guild(cache) {
        content_safe(&guild, &message.content, ContentSafeOptions::default(), &message.mentions)
    } else {
        // We don't need to clean messages in DMs
        message.content.to_string()
    }
}