serenity::utils

Function parse_emoji

pub fn parse_emoji(mention: &str) -> Option<EmojiIdentifier>

Available on crate feature utils only.

Retrieves the animated state, name and Id from an emoji mention, in the form of an EmojiIdentifier.

If the emoji usage is invalid, then None is returned.

Examples

Ensure that a valid Emoji usage is correctly parsed:

use serenity::model::id::{EmojiId, GuildId};
use serenity::model::misc::EmojiIdentifier;
use serenity::utils::parse_emoji;

let emoji = parse_emoji("<:smugAnimeFace:302516740095606785>").unwrap();
assert_eq!(emoji.animated, false);
assert_eq!(emoji.id, EmojiId::new(302516740095606785));
assert_eq!(&*emoji.name, "smugAnimeFace");

Asserting that an invalid emoji usage returns None:

use serenity::utils::parse_emoji;

assert!(parse_emoji("<:smugAnimeFace:302516740095606785").is_none());