serenity::utils

Function parse_user_mention

pub fn parse_user_mention(mention: &str) -> Option<UserId>

Available on crate feature utils only.

Retrieves an Id from a user mention.

If the mention is invalid, then None is returned.

Examples

Retrieving an Id from a valid User mention:

use serenity::model::id::UserId;
use serenity::utils::parse_user_mention;

// regular username mention
assert_eq!(parse_user_mention("<@114941315417899012>"), Some(UserId::new(114941315417899012)));

// nickname mention
assert_eq!(parse_user_mention("<@!114941315417899012>"), Some(UserId::new(114941315417899012)));

Asserting that an invalid username or nickname mention returns None:

use serenity::utils::parse_user_mention;

assert!(parse_user_mention("<@1149413154aa17899012").is_none());
assert!(parse_user_mention("<@!11494131541789a90b1c2").is_none());