serenity::utils

Function shard_id

pub fn shard_id(guild_id: GuildId, shard_count: NonZeroU16) -> u16

Available on crate feature utils only.

Calculates the Id of the shard responsible for a guild, given its Id and total number of shards used.

Examples

Retrieve the Id of the shard for a guild with Id 81384788765712384, using 17 shards:

use serenity::model::id::GuildId;
use serenity::utils;

let guild_id = GuildId::new(81384788765712384);
let shard_total = std::num::NonZeroU16::new(17).unwrap();

assert_eq!(utils::shard_id(guild_id, shard_total), 7);