Struct serenity::utils::Colour

pub struct Colour(pub u32);

A utility struct to help with working with the basic representation of a colour. This is particularly useful when working with a Role's colour, as the API works with an integer value instead of an RGB value.

Instances can be created by using the struct's associated functions. These produce presets equivalent to those found in the official client's colour picker.

Examples

Passing in a role's colour, and then retrieving its green component via g:

use serenity::utils::Colour;

// assuming a `role` has already been bound

let green = role.colour.g();

println!("The green component is: {}", green);

Creating an instance with the dark_teal presets:

use serenity::utils::Colour;

let colour = Colour::dark_teal();

assert_eq!(colour.tuple(), (17, 128, 106));

Colours can also be directly compared for equivalence:

use serenity::utils::Colour;

let blitz_blue = Colour::blitz_blue();
let fooyoo = Colour::fooyoo();
let fooyoo2 = Colour::fooyoo();
assert!(blitz_blue != fooyoo);
assert_eq!(fooyoo, fooyoo2);
assert!(blitz_blue > fooyoo);

Methods

impl Colour

fn new(value: u32) -> Colour

Generates a new Colour with the given integer value set.

Examples

Create a new Colour, and then ensure that its inner value is equivalent to a specific RGB value, retrieved via tuple:

use serenity::utils::Colour;

let colour = Colour::new(6573123);

assert_eq!(colour.tuple(), (100, 76, 67));

fn from_rgb(r: u8, g: u8, b: u8) -> Colour

Generates a new Colour from an RGB value, creating an inner u32 representation.

Examples

Creating a Colour via its RGB values will set its inner u32 correctly:

use serenity::utils::Colour;

assert!(Colour::from_rgb(255, 0, 0).0 == 0xFF0000);
assert!(Colour::from_rgb(217, 23, 211).0 == 0xD917D3);

And you can then retrieve those same RGB values via its methods:

use serenity::utils::Colour;

let colour = Colour::from_rgb(217, 45, 215);

assert_eq!(colour.r(), 217);
assert_eq!(colour.g(), 45);
assert_eq!(colour.b(), 215);
assert_eq!(colour.tuple(), (217, 45, 215));

fn r(&self) -> u8

Returns the red RGB component of this Colour.

Examples

use serenity::utils::Colour;

assert_eq!(Colour::new(6573123).r(), 100);

fn g(&self) -> u8

Returns the green RGB component of this Colour.

Examples

use serenity::utils::Colour;

assert_eq!(Colour::new(6573123).g(), 76);

fn b(&self) -> u8

Returns the blue RGB component of this Colour.

Examples

use serenity::utils::Colour;

assert_eq!(Colour::new(6573123).b(), 67);

fn tuple(&self) -> (u8, u8, u8)

Returns a tuple of the red, green, and blue components of this Colour.

This is equivalent to creating a tuple with the return values of r, g, and b.

Examples

use serenity::utils::Colour;

assert_eq!(Colour::new(6573123).tuple(), (100, 76, 67));

impl Colour

fn blitz_blue() -> Colour

Creates a new Colour, setting its RGB value to (111, 198, 226).

fn blue() -> Colour

Creates a new Colour, setting its RGB value to (52, 152, 219).

fn blurple() -> Colour

Creates a new Colour, setting its RGB value to (114, 137, 218).

fn dark_blue() -> Colour

Creates a new Colour, setting its RGB value to (32, 102, 148).

fn dark_gold() -> Colour

Creates a new Colour, setting its RGB value to (194, 124, 14).

fn dark_green() -> Colour

Creates a new Colour, setting its RGB value to (31, 139, 76).

fn dark_grey() -> Colour

Creates a new Colour, setting its RGB value to (96, 125, 139).

fn dark_magenta() -> Colour

Creates a new Colour, setting its RGB value to (173, 20, 87).

fn dark_orange() -> Colour

Creates a new Colour, setting its RGB value to (168, 67, 0).

fn dark_purple() -> Colour

Creates a new Colour, setting its RGB value to (113, 54, 138).

fn dark_red() -> Colour

Creates a new Colour, setting its RGB value to (153, 45, 34).

fn dark_teal() -> Colour

Creates a new Colour, setting its RGB value to (17, 128, 106).

fn darker_grey() -> Colour

Creates a new Colour, setting its RGB value to (84, 110, 122).

fn fabled_pink() -> Colour

Creates a new Colour, setting its RGB value to (250, 177, 237).

fn faded_purple() -> Colour

Creates a new Colour, setting its RGB value to (136, 130, 196).`

fn fooyoo() -> Colour

Creates a new Colour, setting its RGB value to (17, 202, 128).

fn gold() -> Colour

Creates a new Colour, setting its RGB value to (241, 196, 15).

fn kerbal() -> Colour

Creates a new Colour, setting its RGB value to (186, 218, 85).

fn light_grey() -> Colour

Creates a new Colour, setting its RGB value to (151, 156, 159).

fn lighter_grey() -> Colour

Creates a new Colour, setting its RGB value to (149, 165, 166).

fn magenta() -> Colour

Creates a new Colour, setting its RGB value to (233, 30, 99).

fn meibe_pink() -> Colour

Creates a new Colour, setting its RGB value to (230, 131, 151).

fn orange() -> Colour

Creates a new Colour, setting its RGB value to (230, 126, 34).

fn purple() -> Colour

Creates a new Colour, setting its RGB value to (155, 89, 182).

fn red() -> Colour

Creates a new Colour, setting its RGB value to (231, 76, 60).

fn rohrkatze_blue() -> Colour

Creates a new Colour, setting its RGB value to (117, 150, 255).

fn rosewater() -> Colour

Creates a new Colour, setting its RGB value to (246, 219, 216).

fn teal() -> Colour

Creates a new Colour, setting its RGB value to (26, 188, 156).