serenity::utils
Trait EmbedMessageBuilding

pub trait EmbedMessageBuilding {
    // Required methods
    fn push_named_link<'a>(
        self,
        name: impl Into<Content<'a>>,
        url: impl Into<Content<'a>>,
    ) -> Self;
    fn push_named_link_safe<'a>(
        self,
        name: impl Into<Content<'a>>,
        url: impl Into<Content<'a>>,
    ) -> Self;
}

Available on crate feature utils only.

A trait with additional functionality over the MessageBuilder for creating content with additional functionality available only in embeds.

Namely, this allows you to create named links via the non-escaping Self::push_named_link method and the escaping Self::push_named_link_safe method.

Examples

Make a named link to Rust’s GitHub organization:

use serenity::utils::{EmbedMessageBuilding, MessageBuilder};

let msg = MessageBuilder::new()
    .push_named_link("Rust's GitHub", "https://github.com/rust-lang")
    .build();

assert_eq!(msg, "[Rust's GitHub](https://github.com/rust-lang)");

Required Methods

fn push_named_link<'a>( self, name: impl Into<Content<'a>>, url: impl Into<Content<'a>>, ) -> Self

Pushes a named link to a message, intended for use in embeds.

Examples

Make a simple link to Rust’s homepage for use in an embed:

use serenity::utils::{EmbedMessageBuilding, MessageBuilder};

let content = MessageBuilder::new()
    .push("Rust's website: ")
    .push_named_link("Homepage", "https://rust-lang.org")
    .build();

assert_eq!(content, "Rust's website: [Homepage](https://rust-lang.org)");

fn push_named_link_safe<'a>( self, name: impl Into<Content<'a>>, url: impl Into<Content<'a>>, ) -> Self

Pushes a named link intended for use in an embed, but with a normalized name to avoid escaping issues.

Refer to Self::push_named_link for more information.

Examples

use serenity::utils::{EmbedMessageBuilding, MessageBuilder};

let content = MessageBuilder::new()
    .push("A weird website name: ")
    .push_named_link_safe("Try to ] break links (](", "https://rust-lang.org")
    .build();

assert_eq!(content, "A weird website name: [Try to   break links ( (](https://rust-lang.org)");

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.