Trait serenity::utils::Parse

pub trait Parse: Sized {
    type Err;
    fn parse<'life0, 'life1, 'life2, 'async_trait>(
        ctx: &'life0 Context, 
        msg: &'life1 Message, 
        s: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Err>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

👎 Deprecated:
Superseded by ArgumentConvert trait

Associated Types

type Err

👎 Deprecated:
Superseded by ArgumentConvert trait

The associated error which can be returned from parsing.

Required methods

fn parse<'life0, 'life1, 'life2, 'async_trait>(
ctx: &'life0 Context,
msg: &'life1 Message,
s: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<Self, Self::Err>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

👎 Deprecated:
Superseded by ArgumentConvert trait

Parses a string s as a command parameter of this type.