Module serenity::all

Special module that re-exports most public items from this crate.

Useful, because you don’t have to remember the full paths of serenity items.

Not exported:

crate::json: it’s a general-purpose JSON wrapper, not intrinsic to serenity

crate::framework::standard: has many standard_framework-specific items that may collide with items from the rest of serenity

Re-exports

pub use crate::utils::token::validate as validate_token;

pub use crate::utils::token::InvalidToken;

pub use crate::async_trait;

pub use crate::builder::*;

builder

pub use crate::cache::*;

cache

pub use crate::client::*;

client

pub use crate::collector::*;

collector

pub use crate::constants::close_codes::*;

pub use crate::constants::*;

pub use crate::framework::*;

framework

pub use crate::gateway::*;

gateway

pub use crate::http::*;

http

pub use crate::interactions_endpoint::*;

interactions_endpoint

pub use crate::utils::*;

utils

pub use crate::model::prelude::*;

pub use crate::*;