Module serenity::builder

Available on crate feature builder only.

A set of builders used to make using methods on certain structs simpler to use.

These are used when not all parameters are required, all parameters are optional, and/or sane default values for required parameters can be applied by a builder.

Re-exports

pub use create_poll::CreatePoll;

pub use create_poll::CreatePollAnswer;

Modules

create_poll

Structs

AddMember

A builder to add parameters when using GuildId::add_member.

AutocompleteChoice

CreateAllowedMentions

A builder to manage the allowed mentions on a message, used by the ChannelId::send_message and ChannelId::edit_message methods.

CreateAttachment

A builder for creating a new attachment from a file path, file data, or URL.

CreateAutocompleteResponse

Discord docs

CreateBotAuthParameters

A builder for constructing an invite link with custom OAuth2 scopes.

CreateButton

A builder for creating a button component in a message

CreateChannel

A builder for creating a new GuildChannel in a Guild.

CreateCommand

A builder for creating a new Command.

CreateCommandOption

A builder for creating a new CommandOption.

CreateCommandPermission

A builder for creating an CommandPermission.

CreateEmbed

A builder to create an embed in a message

CreateEmbedAuthor

A builder to create the author data of an embed. See CreateEmbed::author

CreateEmbedFooter

A builder to create the footer data for an embed. See CreateEmbed::footer

CreateForumPost

Discord docs.

CreateForumTag

Discord docs

CreateGuildWelcomeChannel

A builder for creating a GuildWelcomeChannel.

CreateInputText

A builder for creating an input text component in a modal

CreateInteractionResponseFollowup

Discord docs

CreateInteractionResponseMessage

Discord docs.

CreateInvite

A builder to create a RichInvite for use via GuildChannel::create_invite.

CreateMessage

A builder to specify the contents of an send message request, primarily meant for use through ChannelId::send_message.

CreateModal

Discord docs.

CreateScheduledEvent

Discord docs

CreateSelectMenu

A builder for creating a select menu component in a message

CreateSelectMenuOption

A builder for creating an option of a select menu component in a message

CreateStageInstance

Builder for creating a stage instance

CreateSticker

A builder to create a guild sticker

CreateThread

Discord docs:

CreateWebhook

Discord docs

EditAttachments

You can add new attachments and edit existing ones using this builder.

EditAutoModRule

A builder for creating or editing guild AutoMod rules.

EditChannel

A builder to edit a GuildChannel for use via GuildChannel::edit.

EditCommandPermissions

A builder for creating several CommandPermission.

EditGuild

A builder to optionally edit certain fields of a guild

EditGuildWelcomeScreen

A builder to edit the welcome screen of a guild

EditGuildWidget

A builder to specify the fields to edit in a GuildWidget.

EditInteractionResponse

Discord docs

EditMember

A builder which edits the properties of a Member, to be used in conjunction with Member::edit.

EditMessage

A builder to specify the fields to edit in an existing message.

EditProfile

A builder to edit the current user’s settings, to be used in conjunction with CurrentUser::edit.

EditRole

A builder to create or edit a Role for use via a number of model methods.

EditScheduledEvent

Discord docs

EditStageInstance

Edits a StageInstance.

EditSticker

A builder to create or edit a Sticker for use via a number of model methods.

EditThread

Discord docs.

EditVoiceState

A builder which edits a user’s voice state, to be used in conjunction with GuildChannel::edit_voice_state.

EditWebhook

Discord docs

EditWebhookMessage

A builder to specify the fields to edit in an existing Webhook’s message.

ExecuteWebhook

A builder to create the content for a Webhook’s execution.

GetMessages

Builds a request to the API to retrieve messages.

Enums

CreateActionRow

A builder for creating a components action row in a message.

CreateInteractionResponse

Discord docs.

CreateSelectMenuKind

Discord docs.

Traits

Builderhttp

Common trait for all HTTP request builders in this module.