Module serenity::builder

A set of builders used to make using methods on certain structs simpler to use.

These are used when not all parameters are required, all parameters are optional, and/or sane default values for required parameters can be applied by a builder.

Structs

CreateAllowedMentions
A builder to manage the allowed mentions on a message, used by the ChannelId::send_message method.

CreateChannel
A builder for creating a new GuildChannel in a Guild.

CreateEmbed
A builder to create a fake Embed object, for use with the ChannelId::send_message and ExecuteWebhook::embeds methods.

CreateEmbedAuthor
A builder to create a fake Embed object’s author, for use with the CreateEmbed::author method.

CreateEmbedFooter
A builder to create a fake Embed object’s footer, for use with the CreateEmbed::footer method.

CreateInteraction
A builder for creating a new ApplicationCommand.

CreateInteractionOption
A builder for creating a new ApplicationCommandInteractionDataOption.

CreateInteractionResponse

CreateInteractionResponseData

CreateInteractionResponseFollowup

CreateInvite
A builder to create a RichInvite for use via GuildChannel::create_invite.

CreateMessage
A builder to specify the contents of an Http::send_message request, primarily meant for use through ChannelId::send_message.

EditChannel
A builder to edit a GuildChannel for use via GuildChannel::edit

EditGuild
A builder to optionally edit certain fields of a Guild. This is meant for usage with Guild::edit.

EditInteractionResponse

EditMember
A builder which edits the properties of a Member, to be used in conjunction with Member::edit.

EditMessage
A builder to specify the fields to edit in an existing message.

EditProfile
A builder to edit the current user’s settings, to be used in conjunction with CurrentUser::edit.

EditRole
A builder to create or edit a Role for use via a number of model methods.

ExecuteWebhook
A builder to create the inner content of a Webhook’s execution.

GetMessages
Builds a request for a request to the API to retrieve messages.

Timestamp

Enums

ParseValue