Struct serenity::builder::CreateEmbed

pub struct CreateEmbed { /* private fields */ }

Available on crate feature builder only.

A builder to create a fake Embed object, for use with the ChannelId::send_message and ExecuteWebhook::embeds methods.

Implementations

impl CreateEmbed

pub fn new() -> Self

Equivalent to Self::default.

pub fn author(self, author: CreateEmbedAuthor) -> Self

Set the author of the embed.

Refer to the documentation for CreateEmbedAuthor for more information.

pub fn color<C: Into<Colour>>(self, colour: C) -> Self

Set the colour of the left-hand side of the embed.

This is an alias of Self::colour.

pub fn colour<C: Into<Colour>>(self, colour: C) -> Self

Set the colour of the left-hand side of the embed.

pub fn description(self, description: impl Into<String>) -> Self

Set the description of the embed.

Note: This can’t be longer than 4096 characters.

pub fn field( self, name: impl Into<String>, value: impl Into<String>, inline: bool ) -> Self

Set a field. Note that this will not overwrite other fields, and will add to them.

Note: Maximum amount of characters you can put is 256 in a field name and 1024 in a field value.

pub fn fields<N, V>( self, fields: impl IntoIterator<Item = (N, V, bool)> ) -> Selfwhere N: Into<String>, V: Into<String>,

Adds multiple fields at once.

This is sugar to reduce the need of calling Self::field manually multiple times.

pub fn footer(self, footer: CreateEmbedFooter) -> Self

Set the footer of the embed.

Refer to the documentation for CreateEmbedFooter for more information.

pub fn image(self, url: impl Into<String>) -> Self

Set the image associated with the embed. This only supports HTTP(S).

pub fn thumbnail(self, url: impl Into<String>) -> Self

Set the thumbnail of the embed. This only supports HTTP(S).

pub fn timestamp<T: Into<Timestamp>>(self, timestamp: T) -> Self

Set the timestamp.

See the documentation of Timestamp for more information.

Examples

Passing a string timestamp:

let timestamp: Timestamp = "2004-06-08T16:04:23Z".parse().expect("Invalid timestamp!");
let embed = CreateEmbed::new().title("hello").timestamp(timestamp);

Creating a join-log:

Note: this example isn’t efficient and is for demonstrative purposes.

use serenity::builder::{CreateEmbed, CreateEmbedAuthor, CreateMessage};
use serenity::model::guild::Member;
use serenity::model::id::GuildId;
use serenity::prelude::*;

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn guild_member_addition(&self, context: Context, member: Member) {
        let guild_id = member.guild_id;
        if let Ok(guild) = guild_id.to_partial_guild(&context).await {
            let channels = guild.channels(&context).await.unwrap();

            let channel_search = channels.values().find(|c| c.name == "join-log");

            if let Some(channel) = channel_search {
                let icon_url = member.user.face();
                let author = CreateEmbedAuthor::new(member.user.name).icon_url(icon_url);
                let mut embed = CreateEmbed::new().title("Member Join").author(author);
                if let Some(joined_at) = member.joined_at {
                    embed = embed.timestamp(joined_at);
                }
                let builder = CreateMessage::new().embed(embed);
                let _ = channel.send_message(&context, builder).await;
            }
        }
    }
}

let mut client =
    Client::builder("token", GatewayIntents::default()).event_handler(Handler).await?;

client.start().await?;

pub fn title(self, title: impl Into<String>) -> Self

Set the title of the embed.

pub fn url(self, url: impl Into<String>) -> Self

Set the URL to direct to when clicking on the title.

pub fn attachment(self, filename: impl Into<String>) -> Self

Same as calling Self::image with “attachment://filename.(jpg, png)”.

Note however, you have to be sure you set an attachment (with ChannelId::send_files) with the provided filename. Or else this won’t work.