Struct serenity::builder::CreateEmbed

pub struct CreateEmbed(pub HashMap<&'static str, Value>);

A builder to create a fake Embed object, for use with the ChannelId::send_message and ExecuteWebhook::embeds methods.

Tuple Fields
0: HashMap<&'static str, Value>
Implementations

impl CreateEmbed

pub fn author<F>(&mut self, f: F) -> &mut Selfwhere F: FnOnce(&mut CreateEmbedAuthor) -> &mut CreateEmbedAuthor,

Build the author of the embed.

Refer to the documentation for CreateEmbedAuthor for more information.

pub fn set_author(&mut self, author: CreateEmbedAuthor) -> &mut Self

Set the author of the embed.

pub fn color<C: Into<Colour>>(&mut self, colour: C) -> &mut Self

Set the colour of the left-hand side of the embed.

This is an alias of Self::colour.

pub fn colour<C: Into<Colour>>(&mut self, colour: C) -> &mut Self

Set the colour of the left-hand side of the embed.

pub fn description<D: ToString>(&mut self, description: D) -> &mut Self

Set the description of the embed.

Note: This can’t be longer than 4096 characters.

pub fn field<T, U>(&mut self, name: T, value: U, inline: bool) -> &mut Selfwhere T: ToString, U: ToString,

Set a field. Note that this will not overwrite other fields, and will add to them.

Note: Maximum amount of characters you can put is 256 in a field name and 1024 in a field value.

pub fn fields<T, U, It>(&mut self, fields: It) -> &mut Selfwhere It: IntoIterator<Item = (T, U, bool)>, T: ToString, U: ToString,

Adds multiple fields at once.

This is sugar to reduce the need of calling Self::field manually multiple times.

pub fn footer<F>(&mut self, f: F) -> &mut Selfwhere F: FnOnce(&mut CreateEmbedFooter) -> &mut CreateEmbedFooter,

Build the footer of the embed.

Refer to the documentation for CreateEmbedFooter for more information.

pub fn set_footer( &mut self, create_embed_footer: CreateEmbedFooter ) -> &mut Self

Set the footer of the embed.

pub fn image<S: ToString>(&mut self, url: S) -> &mut Self

Set the image associated with the embed. This only supports HTTP(S).

pub fn thumbnail<S: ToString>(&mut self, url: S) -> &mut Self

Set the thumbnail of the embed. This only supports HTTP(S).

pub fn timestamp<T: Into<Timestamp>>(&mut self, timestamp: T) -> &mut Self

Set the timestamp.

You can pass a Timestamp or anything that converts into it ( chrono::DateTime or &str). If giving a string, it must be in RFC 3339 format:

2017-01-03T23:00:00Z

2004-06-08T16:04:23Z

2004-06-08T16:04:23Z

Examples

Passing a string timestamp:

let mut embed = CreateEmbed::default();
embed.title("hello").timestamp("2004-06-08T16:04:23Z");

Creating a join-log:

Note: this example isn’t efficient and is for demonstrative purposes.

use serenity::model::guild::Member;
use serenity::model::id::GuildId;
use serenity::prelude::*;

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn guild_member_addition(&self, context: Context, member: Member) {
        let guild_id = member.guild_id;
        if let Ok(guild) = guild_id.to_partial_guild(&context).await {
            let channels = guild.channels(&context).await.unwrap();

            let channel_search = channels.values().find(|c| c.name == "join-log");

            if let Some(channel) = channel_search {
                let user = &member.user;

                let _ = channel
                    .send_message(&context, |m| {
                        m.embed(|e| {
                            if let Some(ref joined_at) = member.joined_at {
                                e.timestamp(joined_at);
                            }
                            e.author(|a| a.icon_url(&user.face()).name(&user.name))
                                .title("Member Join")
                        })
                    })
                    .await;
            }
        }
    }
}

let mut client =
    Client::builder("token", GatewayIntents::default()).event_handler(Handler).await?;

client.start().await?;

pub fn title<D: ToString>(&mut self, title: D) -> &mut Self

Set the title of the embed.

pub fn url<S: ToString>(&mut self, url: S) -> &mut Self

Set the URL to direct to when clicking on the title.

pub fn attachment<S: ToString>(&mut self, filename: S) -> &mut Self

Same as calling Self::image with “attachment://filename.(jpg, png)”.

Note however, you have to be sure you set an attachment (with ChannelId::send_files) with the provided filename. Or else this won’t work.