Struct serenity::builder::CreateInteractionResponseFollowup

pub struct CreateInteractionResponseFollowup { /* private fields */ }

Available on crate feature builder only.

Implementations

impl CreateInteractionResponseFollowup

pub fn new() -> Self

Equivalent to Self::default.

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn username(self, username: impl Into<String>) -> Self

Override the default username of the webhook

pub fn avatar(self, avatar_url: impl Into<String>) -> Self

Override the default avatar of the webhook

pub fn tts(self, tts: bool) -> Self

Set whether the message is text-to-speech.

Think carefully before setting this to true.

Defaults to false.

pub fn add_file(self, file: CreateAttachment) -> Self

Appends a file to the message.

pub fn add_files( self, files: impl IntoIterator<Item = CreateAttachment> ) -> Self

Appends a list of files to the message.

pub fn files(self, files: impl IntoIterator<Item = CreateAttachment>) -> Self

Sets a list of files to include in the message.

Calling this multiple times will overwrite the file list. To append files, call Self::add_file or Self::add_files instead.

pub fn add_embed(self, embed: CreateEmbed) -> Self

Adds an embed to the message.

pub fn add_embeds(self, embeds: Vec<CreateEmbed>) -> Self

Adds multiple embeds to the message.

pub fn embed(self, embed: CreateEmbed) -> Self

Sets a single embed to include in the message.

Calling this will overwrite the embed list. To append embeds, call Self::add_embed instead.

pub fn embeds(self, embeds: Vec<CreateEmbed>) -> Self

Sets a list of embeds to include in the message.

Calling this multiple times will overwrite the embed list. To append embeds, call Self::add_embeds instead.

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

pub fn flags(self, flags: MessageFlags) -> Self

Sets the flags for the response.

pub fn ephemeral(self, ephemeral: bool) -> Self

Adds or removes the ephemeral flag

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Sets the components of this message.

pub fn button(self, button: CreateButton) -> Self

Adds a clickable button to this message.

Convenience method that wraps Self::components. Arranges buttons in action rows automatically.

pub fn select_menu(self, select_menu: CreateSelectMenu) -> Self

Adds an interactive select menu to this message.

Convenience method that wraps Self::components.