Struct serenity::builder::CreateMessage

pub struct CreateMessage(pub Map<String, Value>, pub Option<Vec<ReactionType>>);

A builder to specify the contents of an http::send_message request, primarily meant for use through ChannelId::send_message.

There are two situations where different field requirements are present:

When sending an embed, no other field is required;

Otherwise, content is the only required field that is required to be set.

Note that if you only need to send the content of a message, without specifying other fields, then ChannelId::say may be a more preferable option.

Examples

Sending a message with a content of "test" and applying text-to-speech:

use serenity::model::ChannelId;

let channel_id = ChannelId(7);

let _ = channel_id.send_message(|m| m
    .content("test")
    .tts(true)
    .embed(|e| e
        .title("This is an embed")
        .description("With a description")));

Methods

impl CreateMessage

fn content<D: Display>(self, content: D) -> Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

fn embed<F>(self, f: F) -> Self where
F: FnOnce(CreateEmbed) -> CreateEmbed,

Set an embed for the message.

fn tts(self, tts: bool) -> Self

Set whether the message is text-to-speech.

Think carefully before setting this to true.

Defaults to false.

fn reactions<R: Into<ReactionType>>(self, reactions: Vec<R>) -> Self

Adds a list of reactions to create after the message's sent.