Struct serenity::builder::CreateMessage

pub struct CreateMessage<'a>(pub HashMap<&'static str, Value>, pub Option<Vec<ReactionType>>, pub Vec<AttachmentType<'a>>);

This is supported on crate feature builder only.

A builder to specify the contents of an Http::send_message request, primarily meant for use through ChannelId::send_message.

There are three situations where different field requirements are present:

When sending a message without embeds or stickers, Self::content is the only required field that is required to be set.

When sending an Self::embed, no other field is required.

When sending stickers with Self::sticker_id or other sticker methods, no other field is required.

Note that if you only need to send the content of a message, without specifying other fields, then ChannelId::say may be a more preferable option.

Examples

Sending a message with a content of "test" and applying text-to-speech:

use serenity::model::id::ChannelId;

let channel_id = ChannelId(7);

let _ = channel_id.send_message(&http, |m| {
    m.content("test")
        .tts(true)
        .embed(|e| e.title("This is an embed").description("With a description"))
});

Tuple Fields
0: HashMap<&'static str, Value>1: Option<Vec<ReactionType>>2: Vec<AttachmentType<'a>>

This is supported on crate feature model only.

Implementations

impl<'a> CreateMessage<'a>

pub fn content<D: ToString>(&mut self, content: D) -> &mut Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn add_embed<F>(&mut self, f: F) -> &mut Self where
F: FnOnce(&mut CreateEmbed) -> &mut CreateEmbed,

Add an embed for the message.

Note: This will keep all existing embeds. Use Self::set_embed() to replace existing embeds.

pub fn add_embeds(&mut self, embeds: Vec<CreateEmbed>) -> &mut Self

Add multiple embeds for the message.

Note: This will keep all existing embeds. Use Self::set_embeds() to replace existing embeds.

pub fn embed<F>(&mut self, f: F) -> &mut Self where
F: FnOnce(&mut CreateEmbed) -> &mut CreateEmbed,

Set an embed for the message.

Equivalent to Self::set_embed().

Note: This will replace all existing embeds. Use Self::add_embed() to add an additional embed.

pub fn set_embed(&mut self, embed: CreateEmbed) -> &mut Self

Set an embed for the message.

Equivalent to Self::embed().

Note: This will replace all existing embeds. Use Self::add_embed() to add an additional embed.

pub fn set_embeds(&mut self, embeds: Vec<CreateEmbed>) -> &mut Self

Set multiple embeds for the message.

Note: This will replace all existing embeds. Use Self::add_embeds() to keep existing embeds.

pub fn tts(&mut self, tts: bool) -> &mut Self

Set whether the message is text-to-speech.

Think carefully before setting this to true.

Defaults to false.

pub fn reactions<R: Into<ReactionType>, It: IntoIterator<Item = R>>(
&mut self,
reactions: It
) -> &mut Self

Adds a list of reactions to create after the message’s sent.

pub fn add_file<T: Into<AttachmentType<'a>>>(&mut self, file: T) -> &mut Self

This is supported on crate feature model only.

Appends a file to the message.

pub fn add_files<T: Into<AttachmentType<'a>>, It: IntoIterator<Item = T>>(
&mut self,
files: It
) -> &mut Self

This is supported on crate feature model only.

Appends a list of files to the message.

pub fn files<T: Into<AttachmentType<'a>>, It: IntoIterator<Item = T>>(
&mut self,
files: It
) -> &mut Self

This is supported on crate feature model only.

Sets a list of files to include in the message.

Calling this multiple times will overwrite the file list. To append files, call Self::add_file or Self::add_files instead.

pub fn allowed_mentions<F>(&mut self, f: F) -> &mut Self where
F: FnOnce(&mut CreateAllowedMentions) -> &mut CreateAllowedMentions,

Set the allowed mentions for the message.

pub fn reference_message(
&mut self,
reference: impl Into<MessageReference>
) -> &mut Self

Set the reference message this message is a reply to.

pub fn components<F>(&mut self, f: F) -> &mut Self where
F: FnOnce(&mut CreateComponents) -> &mut CreateComponents,

Creates components for this message.

pub fn set_components(&mut self, components: CreateComponents) -> &mut Self

Sets the components of this message.

pub fn flags(&mut self, flags: MessageFlags) -> &mut Self

Sets the flags for the message.

pub fn sticker_id(&mut self, sticker_id: impl Into<StickerId>) -> &mut Self

Sets a single sticker ID to include in the message.

Note: This will replace all existing stickers. Use Self::add_sticker_id() to add an additional sticker.

pub fn add_sticker_id(&mut self, sticker_id: impl Into<StickerId>) -> &mut Self

Add a sticker ID for the message.

Note: There can be a maximum of 3 stickers in a message.

Note: This will keep all existing stickers. Use Self::set_sticker_ids() to replace existing stickers.

pub fn add_sticker_ids<T: Into<StickerId>, It: IntoIterator<Item = T>>(
&mut self,
sticker_ids: It
) -> &mut Self

Add multiple sticker IDs for the message.

Note: There can be a maximum of 3 stickers in a message.

Note: This will keep all existing stickers. Use Self::set_sticker_ids() to replace existing stickers.

pub fn set_sticker_ids<T: Into<StickerId>, It: IntoIterator<Item = T>>(
&mut self,
sticker_ids: It
) -> &mut Self

Sets a list of sticker IDs to include in the message.

Note: There can be a maximum of 3 stickers in a message.

Note: This will replace all existing stickers. Use Self::add_sticker_id() or Self::add_sticker_ids() to keep existing stickers.