Struct serenity::builder::CreateMessage

pub struct CreateMessage { /* private fields */ }

Available on crate feature builder only.

A builder to specify the contents of an Http::send_message request, primarily meant for use through ChannelId::send_message.

There are three situations where different field requirements are present:

When sending a message without embeds or stickers, Self::content is the only required field that is required to be set.

When sending an Self::embed, no other field is required.

When sending stickers with Self::sticker_id or other sticker methods, no other field is required.

Note that if you only need to send the content of a message, without specifying other fields, then ChannelId::say may be a more preferable option.

Examples

Sending a message with a content of "test" and applying text-to-speech:

use serenity::builder::{CreateEmbed, CreateMessage};
use serenity::model::id::ChannelId;
let embed = CreateEmbed::new().title("This is an embed").description("With a description");
let builder = CreateMessage::new().content("test").tts(true).embed(embed);
let _ = channel_id.send_message(&http, builder).await;

Implementations

impl CreateMessage

pub fn new() -> Self

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn add_embed(self, embed: CreateEmbed) -> Self

Add an embed for the message.

Note: This will keep all existing embeds. Use Self::embed() to replace existing embeds.

pub fn add_embeds(self, embeds: Vec<CreateEmbed>) -> Self

Add multiple embeds for the message.

Note: This will keep all existing embeds. Use Self::embeds() to replace existing embeds.

pub fn embed(self, embed: CreateEmbed) -> Self

Set an embed for the message.

Note: This will replace all existing embeds. Use Self::add_embed() to keep existing embeds.

pub fn embeds(self, embeds: Vec<CreateEmbed>) -> Self

Set multiple embeds for the message.

Note: This will replace all existing embeds. Use Self::add_embeds() to keep existing embeds.

pub fn tts(self, tts: bool) -> Self

Set whether the message is text-to-speech.

Think carefully before setting this to true.

Defaults to false.

pub fn reactions<R: Into<ReactionType>, It: IntoIterator<Item = R>>( self, reactions: It ) -> Self

Adds a list of reactions to create after the message’s sent.

pub fn add_file(self, file: CreateAttachment) -> Self

Appends a file to the message.

Note: Requires the Attach Files permission.

pub fn add_files( self, files: impl IntoIterator<Item = CreateAttachment> ) -> Self

Appends a list of files to the message.

Note: Requires the Attach Files permission.

pub fn files(self, files: impl IntoIterator<Item = CreateAttachment>) -> Self

Sets a list of files to include in the message.

Calling this multiple times will overwrite the file list. To append files, call Self::add_file or Self::add_files instead.

Note: Requires the Attach Files permission.

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

pub fn reference_message(self, reference: impl Into<MessageReference>) -> Self

Set the reference message this message is a reply to.

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Sets the components of this message.

pub fn button(self, button: CreateButton) -> Self

Adds a clickable button to this message.

Convenience method that wraps Self::components. Arranges buttons in action rows automatically.

pub fn select_menu(self, select_menu: CreateSelectMenu) -> Self

Adds an interactive select menu to this message.

Convenience method that wraps Self::components.

pub fn flags(self, flags: MessageFlags) -> Self

Sets the flags for the message.

pub fn sticker_id(self, sticker_id: impl Into<StickerId>) -> Self

Sets a single sticker ID to include in the message.

Note: This will replace all existing stickers. Use Self::add_sticker_id() to keep existing stickers.

pub fn sticker_ids<T: Into<StickerId>, It: IntoIterator<Item = T>>( self, sticker_ids: It ) -> Self

Sets a list of sticker IDs to include in the message.

Note: There can be a maximum of 3 stickers in a message.

Note: This will replace all existing stickers. Use Self::add_sticker_id() or Self::add_sticker_ids() to keep existing stickers.

pub fn add_sticker_id(self, sticker_id: impl Into<StickerId>) -> Self

Add a sticker ID for the message.

Note: There can be a maximum of 3 stickers in a message.

Note: This will keep all existing stickers. Use Self::sticker_id() to replace existing sticker.

pub fn add_sticker_ids<T: Into<StickerId>, It: IntoIterator<Item = T>>( self, sticker_ids: It ) -> Self

Add multiple sticker IDs for the message.

Note: There can be a maximum of 3 stickers in a message.

Note: This will keep all existing stickers. Use Self::sticker_ids() to replace existing stickers.