Struct serenity::builder::CreateQuickModal

pub struct CreateQuickModal { /* private fields */ }

Available on crate features collector and builder only.

Convenience builder to create a modal, wait for the user to submit and parse the response.

let modal = CreateQuickModal::new("About you")
    .timeout(std::time::Duration::from_secs(600))
    .short_field("First name")
    .short_field("Last name")
    .paragraph_field("Hobbies and interests");
let response = interaction.quick_modal(ctx, modal).await?;
let inputs = response.unwrap().inputs;
let (first_name, last_name, hobbies) = (&inputs[0], &inputs[1], &inputs[2]);

Implementations

impl CreateQuickModal

pub fn new(title: impl Into<String>) -> Self

pub fn timeout(self, timeout: Duration) -> Self

Sets a timeout when waiting for the modal response.

You should almost always set a timeout here. Otherwise, if the user exits the modal, you will wait forever.

pub fn field(self, input_text: CreateInputText) -> Self

Adds an input text field.

As the custom_id field of CreateInputText, just supply an empty string. All custom IDs are overwritten by CreateQuickModal when sending the modal.

pub fn short_field(self, label: impl Into<String>) -> Self

Convenience method to add a single-line input text field.

Wraps Self::field.

pub fn paragraph_field(self, label: impl Into<String>) -> Self

Convenience method to add a multi-line input text field.

Wraps Self::field.