Struct serenity::builder::EditChannel

pub struct EditChannel(pub Map<String, Value>);

A builder to edit a GuildChannel for use via GuildChannel::edit

Defaults are not directly provided by the builder itself.

Examples

Edit a channel, providing a new name and topic:

// assuming a channel has already been bound
if let Err(why) = channel::edit(|c| c.name("new name").topic("a test topic")) {
    // properly handle the error
}

Methods

impl EditChannel

fn bitrate(self, bitrate: u64) -> Self

The bitrate of the channel in bits.

This is for voice channels only.

fn name(self, name: &str) -> Self

The name of the channel.

Must be between 2 and 100 characters long.

fn position(self, position: u64) -> Self

The position of the channel in the channel list.

fn topic(self, topic: &str) -> Self

The topic of the channel. Can be empty.

Must be between 0 and 1024 characters long.

This is for text channels only.

fn user_limit(self, user_limit: u64) -> Self

The number of users that may be in the channel simultaneously.

This is for voice channels only.