Struct serenity::builder::EditChannel

pub struct EditChannel<'a> { /* private fields */ }

Available on crate feature builder only.

A builder to edit a GuildChannel for use via GuildChannel::edit.

Examples

Edit a channel, providing a new name and topic:

let builder = EditChannel::new().name("new name").topic("a test topic");
if let Err(why) = channel.edit(&http, builder).await {
    // properly handle the error
}

Implementations

impl<'a> EditChannel<'a>

pub fn new() -> Self

Equivalent to Self::default.

pub fn bitrate(self, bitrate: u32) -> Self

The bitrate of the channel in bits.

This is for voice channels only.

pub fn video_quality_mode(self, quality: VideoQualityMode) -> Self

The camera video quality mode of the channel.

This is for voice channels only.

pub fn voice_region(self, id: Option<String>) -> Self

The voice region of the channel. It is automatic when None.

This is for voice channels only.

pub fn name(self, name: impl Into<String>) -> Self

The name of the channel.

Must be between 2 and 100 characters long.

pub fn position(self, position: u32) -> Self

The position of the channel in the channel list.

pub fn topic(self, topic: impl Into<String>) -> Self

The topic of the channel. Can be empty.

Must be between 0 and 1024 characters long.

This is for text channels only.

pub fn nsfw(self, nsfw: bool) -> Self

Is the channel inappropriate for work?

This is for text channels only.

pub fn user_limit(self, user_limit: u32) -> Self

The number of users that may be in the channel simultaneously.

This is for voice channels only.

pub fn category<C: Into<Option<ChannelId>>>(self, category: C) -> Self

The parent category of the channel.

This is for text and voice channels only.

pub fn rate_limit_per_user(self, seconds: u64) -> Self

How many seconds must a user wait before sending another message.

Bots, or users with the MANAGE_MESSAGES and/or MANAGE_CHANNELS permissions are exempt from this restriction.

Note: Must be between 0 and 21600 seconds (360 minutes or 6 hours).

pub fn permissions<I>(self, perms: I) -> Selfwhere I: IntoIterator<Item = PermissionOverwrite>,

A set of overwrites defining what a user or a user carrying a certain role can or can’t do.

Example

Inheriting permissions from an existing channel:

use serenity::model::channel::{PermissionOverwrite, PermissionOverwriteType};
use serenity::model::id::UserId;
use serenity::model::permissions::Permissions;

// Assuming a channel has already been bound.
let permissions = vec![PermissionOverwrite {
    allow: Permissions::VIEW_CHANNEL,
    deny: Permissions::SEND_TTS_MESSAGES,
    kind: PermissionOverwriteType::Member(UserId::new(1234)),
}];

let builder = EditChannel::new().name("my_edited_cool_channel").permissions(permissions);
channel.edit(http, builder).await?;

pub fn audit_log_reason(self, reason: &'a str) -> Self

Sets the request’s audit log reason.