Struct serenity::builder::EditChannel

pub struct EditChannel(pub HashMap<&'static str, Value>);

A builder to edit a GuildChannel for use via GuildChannel::edit

Defaults are not directly provided by the builder itself.

Examples

Edit a channel, providing a new name and topic:

// assuming a channel has already been bound
if let Err(why) = channel.edit(&http, |c| c.name("new name").topic("a test topic")).await {
    // properly handle the error
}

Implementations

impl EditChannel

pub fn bitrate(&mut self, bitrate: u64) -> &mut Self

The bitrate of the channel in bits.

This is for voice channels only.

pub fn video_quality_mode(&mut self, quality: VideoQualityMode) -> &mut Self

The camera video quality mode of the channel.

This is for voice channels only.

pub fn voice_region(&mut self, id: Option<String>) -> &mut Self

The voice region of the channel. It is automatic when None.

This is for voice channels only.

pub fn name<S: ToString>(&mut self, name: S) -> &mut Self

The name of the channel.

Must be between 2 and 100 characters long.

pub fn position(&mut self, position: u64) -> &mut Self

The position of the channel in the channel list.

pub fn topic<S: ToString>(&mut self, topic: S) -> &mut Self

The topic of the channel. Can be empty.

Must be between 0 and 1024 characters long.

This is for text channels only.

pub fn nsfw(&mut self, nsfw: bool) -> &mut Self

Is the channel inappropriate for work?

This is for text channels only.

pub fn user_limit(&mut self, user_limit: u64) -> &mut Self

The number of users that may be in the channel simultaneously.

This is for voice channels only.

pub fn category<C: Into<Option<ChannelId>>>(&mut self, category: C) -> &mut Self

The parent category of the channel.

This is for text and voice channels only.

pub fn slow_mode_rate(&mut self, seconds: u64) -> &mut Self

The seconds a user has to wait before sending another message.

Info: Only values from 0 to 120 are valid.

pub fn permissions<I>(&mut self, perms: I) -> &mut Self where
I: IntoIterator<Item = PermissionOverwrite>,

A set of overwrites defining what a user or a user carrying a certain role can and cannot do.

Example

Inheriting permissions from an exisiting channel:

use serenity::model::channel::{PermissionOverwrite, PermissionOverwriteType};
use serenity::model::id::UserId;
use serenity::model::permissions::Permissions;

// Assuming a channel has already been bound.
let permissions = vec![PermissionOverwrite {
    allow: Permissions::READ_MESSAGES,
    deny: Permissions::SEND_TTS_MESSAGES,
    kind: PermissionOverwriteType::Member(UserId(1234)),
}];

channel.edit(http, |c| {
    c.name("my_edited_cool_channel")
    .permissions(permissions)
})
.await?;