Struct serenity::builder::EditGuild

pub struct EditGuild<'a> { /* private fields */ }

Available on crate feature builder only.

A builder to optionally edit certain fields of a Guild.

Implementations

impl<'a> EditGuild<'a>

pub fn new() -> Self

Equivalent to Self::default.

pub fn afk_channel(self, channel: Option<ChannelId>) -> Self

Set the “AFK voice channel” that users are to move to if they have been AFK for an amount of time, configurable by Self::afk_timeout. Pass None to unset the current value.

pub fn afk_timeout(self, timeout: u64) -> Self

Set the amount of time a user is to be moved to the AFK channel - configured via Self::afk_channel - after being AFK.

pub fn icon(self, icon: Option<String>) -> Self

Set the icon of the guild. Pass None to remove the icon.

Examples

Using the utility builder - CreateAttachment - to read an image and encode it in base64, to then set as the guild icon.

let base64_icon = CreateAttachment::path("./guild_icon.png").await?.to_base64();

// assuming a `guild` has already been bound
let builder = EditGuild::new().icon(Some(base64_icon));
guild.edit(&http, builder).await?;

pub fn delete_icon(self) -> Self

Clear the current guild icon, resetting it to the default logo.

pub fn name(self, name: impl Into<String>) -> Self

Set the name of the guild.

Note: Must be between (and including) 2-100 characters.

pub fn description(self, name: impl Into<String>) -> Self

Set the description of the guild.

Note: Requires that the guild have the DISCOVERABLE feature enabled. You can check this through a guild’s features list.

pub fn features(self, features: Vec<String>) -> Self

Set the features of the guild.

Note: Requires that the guild have the DISCOVERABLE feature enabled. You can check this through a guild’s features list.

pub fn owner(self, user_id: impl Into<UserId>) -> Self

Transfers the ownership of the guild to another user by Id.

Note: The current user must be the owner of the guild.

pub fn splash(self, splash: Option<String>) -> Self

Set the splash image of the guild on the invitation page.

The splash must be base64-encoded 1024x1024 png/jpeg/gif image-data.

Requires that the guild have the INVITE_SPLASH feature enabled. You can check this through a guild’s features list.

pub fn discovery_splash(self, splash: Option<String>) -> Self

Set the splash image of the guild on the discovery page.

The splash must be base64-encoded 1024x1024 png/jpeg/gif image-data.

Requires that the guild have the DISCOVERABLE feature enabled. You can check this through a guild’s features list.

pub fn banner(self, banner: Option<String>) -> Self

Set the banner image of the guild, it appears on the left side-bar.

The banner must be base64-encoded 16:9 png/jpeg image data.

Requires that the guild have the BANNER feature enabled. You can check this through a guild’s features list.

pub fn system_channel_id(self, channel_id: Option<ChannelId>) -> Self

Set the channel ID where welcome messages and boost events will be posted.

pub fn rules_channel_id(self, channel_id: Option<ChannelId>) -> Self

Set the channel ID of the rules and guidelines channel.

Note: This feature is for Community guilds only.

pub fn public_updates_channel_id(self, channel_id: Option<ChannelId>) -> Self

Set the channel ID where admins and moderators receive update messages from Discord.

Note: This feature is for Community guilds only.

pub fn preferred_locale(self, locale: Option<String>) -> Self

Set the preferred locale used in Server Discovery and update messages from Discord.

If this is not set, the locale will default to “en-US”;

Note: This feature is for Community guilds only.

pub fn explicit_content_filter( self, level: Option<ExplicitContentFilter> ) -> Self

Set the content filter level.

pub fn default_message_notifications( self, level: Option<DefaultMessageNotificationLevel> ) -> Self

Set the default message notification level.

pub fn verification_level( self, verification_level: impl Into<VerificationLevel> ) -> Self

Set the verification level of the guild. This can restrict what a user must have prior to being able to send messages in a guild.

Refer to the documentation for VerificationLevel for more information on each variant.

Examples

Setting the verification level to High:

use serenity::model::guild::VerificationLevel;

let builder = EditGuild::new().verification_level(VerificationLevel::High);

// assuming a `guild` has already been bound
let edit = guild.edit(&http, builder).await;
if let Err(why) = edit {
    println!("Error setting verification level: {:?}", why);
}

pub fn system_channel_flags( self, system_channel_flags: SystemChannelFlags ) -> Self

Modifies the notifications that are sent by discord to the configured system channel.

use serenity::model::guild::SystemChannelFlags;

let builder = EditGuild::new().system_channel_flags(
    SystemChannelFlags::SUPPRESS_JOIN_NOTIFICATIONS
        | SystemChannelFlags::SUPPRESS_GUILD_REMINDER_NOTIFICATIONS,
);

// assuming a `guild` has already been bound
let edit = guild.edit(&http, builder).await;
if let Err(why) = edit {
    println!("Error setting system channel flags: {:?}", why);
}

pub fn audit_log_reason(self, reason: &'a str) -> Self

Sets the request’s audit log reason.