Struct serenity::builder::EditInteractionResponse

pub struct EditInteractionResponse(/* private fields */);

Available on crate feature builder only.

Discord docs

Implementations

impl EditInteractionResponse

pub fn new() -> Self

Equivalent to Self::default.

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

Note: Message contents must be under 2000 unicode code points.

pub fn add_embed(self, embed: CreateEmbed) -> Self

Adds an embed for the message.

Embeds from the original message are reset when adding new embeds and must be re-added.

pub fn add_embeds(self, embeds: Vec<CreateEmbed>) -> Self

Adds multiple embeds to the message.

Embeds from the original message are reset when adding new embeds and must be re-added.

pub fn embed(self, embed: CreateEmbed) -> Self

Sets a single embed to include in the message

Calling this will overwrite the embed list. To append embeds, call Self::add_embed instead.

pub fn embeds(self, embeds: Vec<CreateEmbed>) -> Self

Sets the embeds for the message.

Note: You can only have up to 10 embeds per message.

Calling this will overwrite the embed list. To append embeds, call Self::add_embeds instead.

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Sets the components of this message.

pub fn button(self, button: CreateButton) -> Self

Adds a clickable button to this message.

Convenience method that wraps Self::components. Arranges buttons in action rows automatically.

pub fn select_menu(self, select_menu: CreateSelectMenu) -> Self

Adds an interactive select menu to this message.

Convenience method that wraps Self::components.

pub fn attachments(self, attachments: EditAttachments) -> Self

Sets attachments, see EditAttachments for more details.

pub fn new_attachment(self, attachment: CreateAttachment) -> Self

Adds a new attachment to the message.

Resets existing attachments. See the documentation for EditAttachments for details.

pub fn keep_existing_attachment(self, id: AttachmentId) -> Self

Shorthand for EditAttachments::keep.

pub fn clear_attachments(self) -> Self

Shorthand for calling Self::attachments with EditAttachments::new.