Struct serenity::builder::GetMessages

pub struct GetMessages { /* private fields */ }

Available on crate feature builder only.

Builds a request to the API to retrieve messages.

This accepts 2 types of parameters. The first type filters messages based on Id, and is set by one of the following:

Self::after

Self::around

Self::before

These are mutually exclusive, and override each other if called sequentially. If one is not specified, messages are simply sorted by most recent.

The other parameter specifies the number of messages fo retrieve. This is optional, and defaults to 50 if not specified.

See GuildChannel::messages for more examples.

Examples

Creating a GetMessages builder to retrieve the first 25 messages after the message with an Id of 158339864557912064:

use serenity::builder::GetMessages;
use serenity::model::id::{ChannelId, MessageId};

// you can then pass it into a function which retrieves messages:
let channel_id = ChannelId::new(81384788765712384);

let builder = GetMessages::new().after(MessageId::new(158339864557912064)).limit(25);
let _messages = channel_id.messages(&http, builder).await?;

Implementations

impl GetMessages

pub fn new() -> Self

Equivalent to Self::default.

pub fn after(self, message_id: impl Into<MessageId>) -> Self

Indicates to retrieve the messages after a specific message, given its Id.

pub fn around(self, message_id: impl Into<MessageId>) -> Self

Indicates to retrieve the messages around a specific message, in other words in either direction from the message in time.

pub fn before(self, message_id: impl Into<MessageId>) -> Self

Indicates to retrieve the messages before a specific message, given its Id.

pub fn limit(self, limit: u8) -> Self

The maximum number of messages to retrieve for the query.

If this is not specified, a default value of 50 is used.

Note: This field is capped to 100 messages due to a Discord limitation. If an amount larger than 100 is supplied, it will be truncated.