Struct serenity::cache::Cache

pub struct Cache {
    pub channels: HashMap<ChannelId, Arc<RwLock<GuildChannel>>>,
    pub categories: HashMap<ChannelId, Arc<RwLock<ChannelCategory>>>,
    pub groups: HashMap<ChannelId, Arc<RwLock<Group>>>,
    pub guilds: HashMap<GuildId, Arc<RwLock<Guild>>>,
    pub notes: HashMap<UserId, String>,
    pub presences: HashMap<UserId, Presence>,
    pub private_channels: HashMap<ChannelId, Arc<RwLock<PrivateChannel>>>,
    pub shard_count: u64,
    pub unavailable_guilds: HashSet<GuildId>,
    pub user: CurrentUser,
    pub users: HashMap<UserId, Arc<RwLock<User>>>,
}

A cache of all events received over a Shard, where storing at least some data from the event is possible.

This acts as a cache, to avoid making requests over the REST API through the http module where possible. All fields are public, and do not have getters, to allow you more flexibility with the stored data. However, this allows data to be "corrupted", and may or may not cause misfunctions within the library. Mutate data at your own discretion.

Fields
channels: HashMap<ChannelId, Arc<RwLock<GuildChannel>>>

A map of channels in Guilds that the current user has received data for.

When a Event::GuildDelete or Event::GuildUnavailable is received and processed by the cache, the relevant channels are also removed from this map.

categories: HashMap<ChannelId, Arc<RwLock<ChannelCategory>>>

A map of channel categories.

groups: HashMap<ChannelId, Arc<RwLock<Group>>>

A map of the groups that the current user is in.

For bot users this will always be empty, except for in special cases.

guilds: HashMap<GuildId, Arc<RwLock<Guild>>>

A map of guilds with full data available. This includes data like Roles and Emojis that are not available through the REST API.

notes: HashMap<UserId, String>

A map of notes that a user has made for individual users.

An empty note is equivalent to having no note, and creating an empty note is equivalent to deleting a note.

This will always be empty for bot users.

presences: HashMap<UserId, Presence>

A map of users' presences. This is updated in real-time. Note that status updates are often "eaten" by the gateway, and this should not be treated as being entirely 100% accurate.

private_channels: HashMap<ChannelId, Arc<RwLock<PrivateChannel>>>

A map of direct message channels that the current user has open with other users.

shard_count: u64

The total number of shards being used by the bot.

unavailable_guilds: HashSet<GuildId>

A list of guilds which are "unavailable". Refer to the documentation for Event::GuildUnavailable for more information on when this can occur.

Additionally, guilds are always unavailable for bot users when a Ready is received. Guilds are "sent in" over time through the receiving of Event::GuildCreates.

user: CurrentUser

The current user "logged in" and for which events are being received for.

The current user contains information that a regular User does not, such as whether it is a bot, whether the user is verified, etc.

Refer to the documentation for CurrentUser for more information.

users: HashMap<UserId, Arc<RwLock<User>>>

A map of users that the current user sees.

Users are added to - and updated from - this map via the following received events:

ChannelRecipientAdd

GuildMemberAdd

GuildMemberRemove

GuildMembersChunk

GuildSync

PresenceUpdate

Ready

Note, however, that users are not removed from the map on removal events such as GuildMemberRemove, as other structs such as members or recipients may still exist.

Methods

impl Cache

fn unknown_members(&self) -> u64

Fetches the number of Members that have not had data received.

The important detail to note here is that this is the number of _member_s that have not had data received. A single User may have multiple associated member objects that have not been received.

This can be used in combination with Shard::chunk_guilds, and can be used to determine how many members have not yet been received.

use serenity::client::CACHE;
use std::thread;
use std::time::Duration;

struct Handler;

impl EventHandler for Handler {
    fn on_ready(&self, ctx: Context, _: Ready) {
         // Wait some time for guilds to be received.
        //
        // You should keep track of this in a better fashion by tracking how
        // many guilds each `ready` has, and incrementing a counter on
        // GUILD_CREATEs. Once the number is equal, print the number of
        // unknown members.
        //
        // For demonstrative purposes we're just sleeping the thread for 5
        // seconds.
        thread::sleep(Duration::from_secs(5));

        println!("{} unknown members", CACHE.read().unwrap().unknown_members());
    }
}

let mut client = Client::new("token", Handler); client.start().unwrap();

fn all_private_channels(&self) -> Vec<&ChannelId>

Fetches a vector of all PrivateChannel and Group Ids that are stored in the cache.

Examples

If there are 6 private channels and 2 groups in the cache, then 8 Ids will be returned.

Printing the count of all private channels and groups:

use serenity::client::CACHE;

let amount = CACHE.read().unwrap().all_private_channels().len();

println!("There are {} private channels", amount);

fn all_guilds(&self) -> Vec<&GuildId>

Fetches a vector of all Guilds' Ids that are stored in the cache.

Note that if you are utilizing multiple Shards, then the guilds retrieved over all shards are included in this count -- not just the current Context's shard, if accessing from one.

Examples

Print all of the Ids of guilds in the Cache:

use serenity::client::CACHE;

struct Handler;
impl EventHandler for Handler {
    fn on_ready(&self, _: Context, _: Ready) {
        println!("Guilds in the Cache: {:?}", CACHE.read().unwrap().all_guilds());
    }
}
let mut client = Client::new("token", Handler);

fn channel<C: Into<ChannelId>>(&self, id: C) -> Option<Channel>

Retrieves a Channel from the cache based on the given Id.

This will search the channels map, the private_channels map, and then the map of groups to find the channel.

If you know what type of channel you're looking for, you should instead manually retrieve from one of the respective maps or methods:

[GuildChannel]: guild_channel or channels

[PrivateChannel]: private_channel or private_channels

Group: group or groups

fn guild<G: Into<GuildId>>(&self, id: G) -> Option<Arc<RwLock<Guild>>>

Retrieves a guild from the cache based on the given Id.

The only advantage of this method is that you can pass in anything that is indirectly a GuildId.

Examples

Retrieve a guild from the cache and print its name:

use serenity::client::CACHE;

let cache = CACHE.read()?;

if let Some(guild) = cache.guild(7) {
    println!("Guild name: {}", guild.read().unwrap().name);
}

fn guild_channel<C: Into<ChannelId>>(
&self,
id: C
) -> Option<Arc<RwLock<GuildChannel>>>

Retrieves a reference to a Guild's channel. Unlike channel, this will only search guilds for the given channel.

The only advantage of this method is that you can pass in anything that is indirectly a ChannelId.

Examples

Getting a guild's channel via the Id of the message received through a Client::on_message event dispatch:

use serenity::client::CACHE;

struct Handler;

impl EventHandler for Handler {
    fn on_message(&self, ctx: Context, message: Message) {
        let cache = CACHE.read().unwrap();

        let channel = match cache.guild_channel(message.channel_id) {
            Some(channel) => channel,
            None => {
if let Err(why) = message.channel_id.say("Could not find guild's
channel data") {
                    println!("Error sending message: {:?}", why);
                }

                return;
            },
        };
    }
}

let mut client = Client::new("token", Handler); client.start().unwrap();

fn group<C: Into<ChannelId>>(&self, id: C) -> Option<Arc<RwLock<Group>>>

Retrieves a reference to a Group from the cache based on the given associated channel Id.

The only advantage of this method is that you can pass in anything that is indirectly a ChannelId.

Examples

Retrieve a group from the cache and print its owner's id:

use serenity::client::CACHE;

let cache = CACHE.read()?;

if let Some(group) = cache.group(7) {
    println!("Owner Id: {}", group.read().unwrap().owner_id);
}

fn member<G, U>(&self, guild_id: G, user_id: U) -> Option<Member> where
G: Into<GuildId>,
U: Into<UserId>,

Retrieves a Guild's member from the cache based on the guild's and user's given Ids.

Note: This will clone the entire member. Instead, retrieve the guild and retrieve from the guild's members map to avoid this.

Examples

Retrieving the member object of the user that posted a message, in a Client::on_message context:

use serenity::CACHE;

let cache = CACHE.read().unwrap();
let member = {
    let channel = match cache.guild_channel(message.channel_id) {
        Some(channel) => channel,
        None => {
            if let Err(why) = message.channel_id.say("Error finding channel data") {
                println!("Error sending message: {:?}", why);
            }
        },
    };

    match cache.member(channel.guild_id, message.author.id) {
        Some(member) => member,
        None => {
            if let Err(why) = message.channel_id.say("Error finding member data") {
                println!("Error sending message: {:?}", why);
            }
        },
    }
};

let msg = format!("You have {} roles", member.roles.len());

if let Err(why) = message.channel_id.say(&msg) {
    println!("Error sending message: {:?}", why);
}

fn private_channel<C: Into<ChannelId>>(
&self,
channel_id: C
) -> Option<Arc<RwLock<PrivateChannel>>>

Retrieves a [PrivateChannel] from the cache's [private_channels] map, if it exists.

The only advantage of this method is that you can pass in anything that is indirectly a [ChannelId].

Examples

Retrieve a private channel from the cache and send a message:

use serenity::client::CACHE;

let cache = CACHE.read()?;

if let Some(channel) = cache.private_channel(7) {
    channel.read().unwrap().say("Hello there!");
}

fn role<G, R>(&self, guild_id: G, role_id: R) -> Option<Role> where
G: Into<GuildId>,
R: Into<RoleId>,

Retrieves a Guild's role by their Ids.

Note: This will clone the entire role. Instead, retrieve the guild and retrieve from the guild's roles map to avoid this.

Examples

Retrieve a role from the cache and print its name:

use serenity::client::CACHE;

let cache = CACHE.read()?;

if let Some(role) = cache.role(7, 77) {
    println!("Role with Id 77 is called {}", role.name);
}

fn user<U: Into<UserId>>(&self, user_id: U) -> Option<Arc<RwLock<User>>>

Retrieves a User from the cache's users map, if it exists.

The only advantage of this method is that you can pass in anything that is indirectly a UserId.

Examples

Retrieve a user from the cache and print their name:

use serenity::client::CACHE;

let cache = CACHE.read()?;

if let Some(user) = cache.user(7) {
    println!("User with Id 7 is currently named {}", user.read().unwrap().name);
}

fn categories<C: Into<ChannelId>>(
&self,
channel_id: C
) -> Option<Arc<RwLock<ChannelCategory>>>