Struct serenity::cache::Cache

#[non_exhaustive]
pub struct Cache { /* private fields */ }

Available on crate feature cache only.

A cache containing data received from Shards.

Using the cache allows to avoid REST API requests via the http module where possible. Issuing too many requests will lead to ratelimits.

This is the list of cached resources and the events that populate them:

channels: ChannelCreateEvent, ChannelUpdateEvent, GuildCreateEvent

private_channels: ChannelCreateEvent

guilds: GuildCreateEvent

unavailable_guilds: ReadyEvent, GuildDeleteEvent

users: GuildMemberAddEvent, GuildMemberRemoveEvent, GuildMembersChunkEvent, PresenceUpdateEvent, ReadyEvent

presences: PresenceUpdateEvent, ReadyEvent

messages: MessageCreateEvent

The documentation of each event contains the required gateway intents.

Implementations

impl Cache

pub fn new() -> Self

Creates a new cache.

pub fn new_with_settings(settings: Settings) -> Self

Creates a new cache instance with settings applied.

Examples

use serenity::cache::{Cache, Settings};

let mut settings = Settings::default();
settings.max_messages = 10;

let cache = Cache::new_with_settings(settings);

pub fn unknown_members(&self) -> u64

Fetches the number of Members that have not had data received.

The important detail to note here is that this is the number of _member_s that have not had data received. A single User may have multiple associated member objects that have not been received.

This can be used in combination with Shard::chunk_guild, and can be used to determine how many members have not yet been received.

use std::thread;
use std::time::Duration;

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn cache_ready(&self, ctx: Context, _: Vec<GuildId>) {
        println!("{} unknown members", ctx.cache.unknown_members());
    }
}

let mut client =
    Client::builder("token", GatewayIntents::default()).event_handler(Handler).await?;

client.start().await?;

pub fn private_channels(&self) -> ReadOnlyMapRef<'_, ChannelId, PrivateChannel>

Fetches a vector of all PrivateChannel Ids that are stored in the cache.

Examples

If there are 6 private channels and 2 groups in the cache, then 8 Ids will be returned.

Printing the count of all private channels and groups:

let amount = cache.private_channels().len();

println!("There are {} private channels", amount);

pub fn guilds(&self) -> Vec<GuildId>

Fetches a vector of all Guilds’ Ids that are stored in the cache.

Note that if you are utilizing multiple Shards, then the guilds retrieved over all shards are included in this count – not just the current Context’s shard, if accessing from one.

Examples

Print all of the Ids of guilds in the Cache:

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn ready(&self, context: Context, _: Ready) {
        let guilds = context.cache.guilds().len();

        println!("Guilds in the Cache: {}", guilds);
    }
}

pub fn channel<C: Into<ChannelId>>(&self, id: C) -> Option<Channel>

Retrieves a Channel from the cache based on the given Id.

This will search the channels map, then the Self::private_channels map.

If you know what type of channel you’re looking for, you should instead manually retrieve from one of the respective methods:

GuildChannel: Self::guild_channel or Self::guild_channels

PrivateChannel: Self::private_channel or Self::private_channels

pub fn channel_messages( &self, channel_id: impl Into<ChannelId> ) -> Option<ChannelMessagesRef<'_>>

Get a reference to the cached messages for a channel based on the given Id.

Examples

Find all messages by user ID 8 in channel ID 7:

let messages_in_channel = cache.channel_messages(7);
let messages_by_user = messages_in_channel
    .as_ref()
    .map(|msgs| msgs.values().filter(|m| m.author.id == 8).collect::<Vec<_>>());

pub fn guild<G: Into<GuildId>>(&self, id: G) -> Option<GuildRef<'_>>

Gets a reference to a guild from the cache based on the given id.

Examples

Retrieve a guild from the cache and print its name:

// assuming the cache is in scope, e.g. via `Context`
if let Some(guild) = cache.guild(7) {
    println!("Guild name: {}", guild.name);
};

pub fn guild_count(&self) -> usize

Returns the number of cached guilds.

pub fn guild_channel<C: Into<ChannelId>>( &self, id: C ) -> Option<GuildChannelRef<'_>>

Retrieves a reference to a Guild’s channel. Unlike Self::channel, this will only search guilds for the given channel.

The only advantage of this method is that you can pass in anything that is indirectly a ChannelId.

Examples

Getting a guild’s channel via the Id of the message received through a EventHandler::message event dispatch:

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, context: Context, message: Message) {
        let channel_opt = context.cache.guild_channel(message.channel_id).as_deref().cloned();
        let channel = match channel_opt {
            Some(channel) => channel,
            None => {
                let result = message
                    .channel_id
                    .say(&context, "Could not find guild's channel data")
                    .await;
                if let Err(why) = result {
                    println!("Error sending message: {:?}", why);
                }

                return;
            },
        };
    }
}

let mut client =
    Client::builder("token", GatewayIntents::default()).event_handler(Handler).await?;

client.start().await?;

pub fn member<G, U>(&self, guild_id: G, user_id: U) -> Option<Member>where G: Into<GuildId>, U: Into<UserId>,

Retrieves a Guild’s member from the cache based on the guild’s and user’s given Ids.

Note: This will clone the entire member. Instead, retrieve the guild and retrieve from the guild’s members map to avoid this.

Examples

Retrieving the member object of the user that posted a message, in a EventHandler::message context:

let member = {
    let channel = match cache.guild_channel(message.channel_id) {
        Some(channel) => channel,
        None => {
            if let Err(why) = message.channel_id.say(http, "Error finding channel data").await {
                println!("Error sending message: {:?}", why);
            }
            return;
        },
    };

    match cache.member(channel.guild_id, message.author.id) {
        Some(member) => member,
        None => {
            if let Err(why) = message.channel_id.say(&http, "Error finding member data").await {
                println!("Error sending message: {:?}", why);
            }
            return;
        },
    }
};

let msg = format!("You have {} roles", member.roles.len());

if let Err(why) = message.channel_id.say(&http, &msg).await {
    println!("Error sending message: {:?}", why);
}

pub fn member_field<Ret, Fun>( &self, guild_id: impl Into<GuildId>, user_id: impl Into<UserId>, field_selector: Fun ) -> Option<Ret>where Fun: FnOnce(&Member) -> Ret,

This method allows to only clone a field of a member instead of the entire member by providing a field_selector-closure picking what you want to clone.

// We clone only the `name` instead of the entire channel.
if let Some(Some(nick)) = cache.member_field(7, 8, |member| member.nick.clone()) {
    println!("Member's nick: {}", nick);
}

pub fn guild_roles( &self, guild_id: impl Into<GuildId> ) -> Option<HashMap<RoleId, Role>>

pub fn unavailable_guilds(&self) -> ReadOnlyMapRef<'_, GuildId, ()>

This method clones and returns all unavailable guilds.

pub fn guild_channels( &self, guild_id: impl Into<GuildId> ) -> Option<DashMap<ChannelId, GuildChannel, BuildHasher>>

This method returns all channels from a guild of with the given guild_id.

pub fn guild_channel_count(&self) -> usize

Returns the number of guild channels in the cache.

pub fn shard_count(&self) -> u32

Returns the number of shards.

pub fn message<C, M>(&self, channel_id: C, message_id: M) -> Option<Message>where C: Into<ChannelId>, M: Into<MessageId>,

Retrieves a Channel’s message from the cache based on the channel’s and message’s given Ids.

Note: This will clone the entire message.

Examples

Retrieving the message object from a channel, in a EventHandler::message context:

match cache.message(message.channel_id, message.id) {
    Some(m) => assert_eq!(message.content, m.content),
    None => println!("No message found in cache."),
};

pub fn private_channel( &self, channel_id: impl Into<ChannelId> ) -> Option<PrivateChannelRef<'_>>

Retrieves a PrivateChannel from the cache’s Self::private_channels map, if it exists.

The only advantage of this method is that you can pass in anything that is indirectly a ChannelId.

Examples

Retrieve a private channel from the cache and print its recipient’s name:

// assuming the cache has been unlocked

if let Some(channel) = cache.private_channel(7) {
    println!("The recipient is {}", channel.recipient);
};

pub fn role<G, R>(&self, guild_id: G, role_id: R) -> Option<Role>where G: Into<GuildId>, R: Into<RoleId>,

Retrieves a Guild’s role by their Ids.

Note: This will clone the entire role. Instead, retrieve the guild and retrieve from the guild’s roles map to avoid this.

Examples

Retrieve a role from the cache and print its name:

// assuming the cache is in scope, e.g. via `Context`
if let Some(role) = cache.role(7, 77) {
    println!("Role with Id 77 is called {}", role.name);
}

pub fn settings(&self) -> Settings

Returns the settings.

Examples

Printing the maximum number of messages in a channel to be cached:

use serenity::cache::Cache;

let mut cache = Cache::new();
println!("Max settings: {}", cache.settings().max_messages);

pub fn set_max_messages(&self, max: usize)

Sets the maximum amount of messages per channel to cache.

By default, no messages will be cached.

pub fn user<U: Into<UserId>>(&self, user_id: U) -> Option<UserRef<'_>>

Retrieves a User from the cache’s Self::users map, if it exists.

The only advantage of this method is that you can pass in anything that is indirectly a UserId.

Examples

Retrieve a user from the cache and print their name:

if let Some(user) = context.cache.user(7) {
    println!("User with Id 7 is currently named {}", user.name);
}

pub fn users(&self) -> ReadOnlyMapRef<'_, UserId, User>

Clones all users and returns them.

pub fn user_count(&self) -> usize

Returns the amount of cached users.

pub fn current_user(&self) -> CurrentUserRef<'_>

This method provides a reference to the user used by the bot.

pub fn update<E: CacheUpdate>(&self, e: &mut E) -> Option<E::Output>

Updates the cache with the update implementation for an event or other custom update implementation.

Refer to the documentation for CacheUpdate for more information.

Examples

Refer to the CacheUpdate examples.