serenity::cache

Trait CacheUpdate

pub trait CacheUpdate {
    type Output;

    // Required method
    fn update(&self, _: &Cache) -> Option<Self::Output>;
}

Available on crate feature cache only.

Trait used for updating the cache with a type.

This may be implemented on a type and used to update the cache via Cache::update.

Info: You may not access the fields of the cache, as they are public for the crate only.

Required Associated Types

type Output

The return type of an update.

If there is nothing to return, specify this type to be the unit ().

Required Methods

fn update(&self, _: &Cache) -> Option<Self::Output>

Updates the cache with the implementation.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".