Struct serenity::client::ClientBuilder

pub struct ClientBuilder { /* private fields */ }

Available on crate features client and gateway only.

A builder implementing IntoFuture building a Client to interact with Discord.

Implementations

impl ClientBuilder

pub fn new(token: impl AsRef<str>, intents: GatewayIntents) -> Self

Construct a new builder to call methods on for the client construction. The token will automatically be prefixed “Bot “ if not already.

Panic: If you have enabled the framework-feature (on by default), you must specify a framework via the Self::framework method, otherwise awaiting the builder will cause a panic.

pub fn new_with_http(http: Http, intents: GatewayIntents) -> Self

Construct a new builder with a Http instance to calls methods on for the client construction.

Panic: If you have enabled the framework-feature (on by default), you must specify a framework via the Self::framework method, otherwise awaiting the builder will cause a panic.

pub fn token(self, token: impl AsRef<str>) -> Self

Sets a token for the bot. If the token is not prefixed “Bot “, this method will automatically do so.

pub fn get_token(&self) -> &str

Gets the current token used for the Http client.

pub fn application_id(self, application_id: ApplicationId) -> Self

Sets the application id.

pub fn get_application_id(&self) -> Option<ApplicationId>

Gets the application ID, if already initialized. See Self::application_id for more info.

pub fn type_map(self, type_map: TypeMap) -> Self

Sets the entire TypeMap that will be available in Contexts. A TypeMap must not be constructed manually: Self::type_map_insert can be used to insert one type at a time.

pub fn get_type_map(&self) -> &TypeMap

Gets the type map. See Self::type_map for more info.

pub fn type_map_insert<T: TypeMapKey>(self, value: T::Value) -> Self

Insert a single value into the internal TypeMap that will be available in Context::data. This method can be called multiple times in order to populate the TypeMap with values.

pub fn cache_settings<F>(self, f: F) -> Selfwhere F: FnOnce(&mut CacheSettings) -> &mut CacheSettings,

Available on crate feature cache only.

Sets the settings of the cache. Refer to Settings for more information.

pub fn get_cache_settings(&self) -> &CacheSettings

Available on crate feature cache only.

Gets the cache settings. See Self::cache_settings for more info.

pub fn framework<F>(self, framework: F) -> Selfwhere F: Framework + 'static,

Available on crate feature framework only.

Sets the command framework to be used. It will receive messages sent over the gateway and then consider - based on its settings - whether to dispatch a command.

Info: If a reference to the framework is required for manual dispatch, you can implement Framework on Arc<YourFrameworkType> instead of YourFrameworkType.

pub fn get_framework(&self) -> Option<&dyn Framework>

Available on crate feature framework only.

Gets the framework, if already initialized. See Self::framework for more info.

pub fn voice_manager<V>(self, voice_manager: V) -> Selfwhere V: VoiceGatewayManager + 'static,

Available on crate feature voice only.

Sets the voice gateway handler to be used. It will receive voice events sent over the gateway and then consider - based on its settings - whether to dispatch a command.

Info: If a reference to the voice_manager is required for manual dispatch, use the Self::voice_manager_arc-method instead.

pub fn voice_manager_arc( self, voice_manager: Arc<dyn VoiceGatewayManager + 'static> ) -> Self

Available on crate feature voice only.

This method allows to pass an Arc’ed voice_manager - this step is done for you in the voice_manager-method, if you don’t need the extra control. You can provide a clone and keep the original to manually dispatch.

pub fn get_voice_manager(&self) -> Option<Arc<dyn VoiceGatewayManager>>

Available on crate feature voice only.

Gets the voice manager, if already initialized. See Self::voice_manager for more info.

pub fn intents(self, intents: GatewayIntents) -> Self

Sets all intents directly, replacing already set intents. Intents are a bitflag, you can combine them by performing the |-operator.

What are Intents

A gateway intent sets the types of gateway events (e.g. member joins, guild integrations, guild emoji updates, …) the bot shall receive. Carefully picking the needed intents greatly helps the bot to scale, as less intents will result in less events to be received hence less processed by the bot.

Privileged Intents

The intents GatewayIntents::GUILD_PRESENCES, GatewayIntents::GUILD_MEMBERS and GatewayIntents::MESSAGE_CONTENT are privileged. Privileged intents need to be enabled in the developer portal. Once the bot is in 100 guilds or more, the bot must be verified in order to use privileged intents.

pub fn get_intents(&self) -> GatewayIntents

Gets the intents. See Self::intents for more info.

pub fn event_handler<H: EventHandler + 'static>(self, event_handler: H) -> Self

Adds an event handler with multiple methods for each possible event.

pub fn event_handler_arc<H: EventHandler + 'static>( self, event_handler_arc: Arc<H> ) -> Self

Adds an event handler with multiple methods for each possible event. Passed by Arc.

pub fn get_event_handlers(&self) -> &[Arc<dyn EventHandler>]

Gets the added event handlers. See Self::event_handler for more info.

pub fn raw_event_handler<H: RawEventHandler + 'static>( self, raw_event_handler: H ) -> Self

Adds an event handler with a single method where all received gateway events will be dispatched.

pub fn get_raw_event_handlers(&self) -> &[Arc<dyn RawEventHandler>]

Gets the added raw event handlers. See Self::raw_event_handler for more info.

pub fn activity(self, activity: ActivityData) -> Self

Sets the initial activity.

pub fn status(self, status: OnlineStatus) -> Self

Sets the initial status.

pub fn get_presence(&self) -> &PresenceData

Gets the initial presence. See Self::activity and Self::status for more info.