Struct serenity::client::ClientBuilder

pub struct ClientBuilder<'a> { /* fields omitted */ }

A builder implementing Future building a Client to interact with Discord.

Implementations

impl<'a> ClientBuilder<'a>

pub fn new(token: impl AsRef<str>) -> Self

Construct a new builder to call methods on for the client construction. The token will automatically be prefixed “Bot “ if not already.

Panic: If you have enabled the framework-feature (on by default), you must specify a framework via the Self::framework or Self::framework_arc method, otherwise awaiting the builder will cause a panic.

pub fn new_with_http(http: Http) -> Self

Construct a new builder with a Http instance to calls methods on for the client construction.

Panic: If you have enabled the framework-feature (on by default), you must specify a framework via the Self::framework or Self::framework_arc method, otherwise awaiting the builder will cause a panic.

pub fn token(self, token: impl AsRef<str>) -> Self

Sets a token for the bot. If the token is not prefixed “Bot “, this method will automatically do so.

pub fn application_id(self, application_id: u64) -> Self

Sets the application id.

pub fn type_map(self, type_map: TypeMap) -> Self

Sets the entire TypeMap that will be available in Contexts. A TypeMap must not be constructed manually: Self::type_map_insert can be used to insert one type at a time.

pub fn type_map_insert<T: TypeMapKey>(self, value: T::Value) -> Self

Insert a single value into the internal TypeMap that will be available in Context::data. This method can be called multiple times in order to populate the TypeMap with values.

pub fn cache_update_timeout(self, timeout: Duration) -> Self

Sets how long - if wanted to begin with - a cache update shall be attempted for. After the timeout ran out, the update will be skipped.

By default, a cache update will never timeout and potentially cause a deadlock. A timeout however, will invalidate the cache.

pub fn framework<F>(self, framework: F) -> Self where
F: Framework + Send + Sync + 'static,

Sets the command framework to be used. It will receive messages sent over the gateway and then consider - based on its settings - whether to dispatch a command.

Info: If a reference to the framework is required for manual dispatch, use the Self::framework_arc-method instead.

pub fn framework_arc<T: Framework + Send + Sync + 'static>(
self,
framework: Arc<T>
) -> Self

This method allows to pass an Arc’ed framework - this step is done for you in the Self::framework-method, if you don’t need the extra control. You can provide a clone and keep the original to manually dispatch.

pub fn voice_manager<V>(self, voice_manager: V) -> Self where
V: VoiceGatewayManager + Send + Sync + 'static,

Sets the voice gateway handler to be used. It will receive voice events sent over the gateway and then consider - based on its settings - whether to dispatch a command.

Info: If a reference to the voice_manager is required for manual dispatch, use the Self::voice_manager_arc-method instead.

pub fn voice_manager_arc(
self,
voice_manager: Arc<dyn VoiceGatewayManager + Send + Sync + 'static>
) -> Self

This method allows to pass an Arc’ed voice_manager - this step is done for you in the voice_manager-method, if you don’t need the extra control. You can provide a clone and keep the original to manually dispatch.

pub fn intents(self, intents: GatewayIntents) -> Self

Sets all intents directly, replacing already set intents. Intents are a bitflag, you can combine them by performing the |-operator.

What are Intents

A gateway intent sets the types of gateway events (e.g. member joins, guild integrations, guild emoji updates, …) the bot shall receive. Carefully picking the needed intents greatly helps the bot to scale, as less intents will result in less events to be received hence less processed by the bot.

Privileged Intents

The intents GatewayIntents::GUILD_PRESENCES and GatewayIntents::GUILD_MEMBERS are privileged. Privileged intents need to be enabled in the developer portal. Once the bot is in 100 guilds or more, the bot must be verified in order to use privileged intents.

pub fn event_handler<H: EventHandler + 'static>(self, event_handler: H) -> Self

Sets an event handler with multiple methods for each possible event.

pub fn raw_event_handler<H: RawEventHandler + 'static>(
self,
raw_event_handler: H
) -> Self

Sets an event handler with a single method where all received gateway events will be dispatched.