Module serenity::collector

Available on crate feature collector only.

Collectors will receive events from the contextual shard, check if the filter lets them pass, and collects if the receive, collect, or time limits are not reached yet.

Re-exports

pub use component_interaction_collector::*;

pub use event_collector::*;

pub use message_collector::*;

pub use modal_interaction_collector::*;

pub use reaction_collector::*;

Modules

component_interaction_collector

event_collector

message_collector

modal_interaction_collector

reaction_collector

Structs

CollectorBuilder

CommonFilterOptions

Filter

FilterFn

LazyArc

Wraps a &T and clones the value into an Arc lazily. Used with collectors to allow inspecting the value in filters while only cloning values that actually match.

Enums

CollectorError

An error that occurred while working with a collector.

Traits

Collectable

CollectableඞLazyItem

LazyItem