Module serenity::collector

Available on crate feature collector only.

Structs

ComponentInteractionCollector

A ComponentInteractionCollector receives ComponentInteraction’s match the given filters for a set duration.

EventCollectorDeprecated

A EventCollector receives Event’s match the given filters for a set duration.

MessageCollector

A MessageCollector receives Message’s match the given filters for a set duration.

ModalInteractionCollector

A ModalInteractionCollector receives ModalInteraction’s match the given filters for a set duration.

ReactionCollector

A ReactionCollector receives Reaction’s match the given filters for a set duration.

Functions

collect

Fundamental collector function. All collector types in this module are just wrappers around this function.