Struct serenity::collector::MessageCollector

pub struct MessageCollector { /* private fields */ }

Available on crate feature collector only.

A MessageCollector receives Message’s match the given filters for a set duration.

Implementations

impl MessageCollector

pub fn new(shard: impl AsRef<ShardMessenger>) -> Self

Creates a new collector without any filters configured.

pub fn timeout(self, duration: Duration) -> Self

Sets a duration for how long the collector shall receive interactions.

pub fn filter( self, filter: impl Fn(&Message) -> bool + Send + Sync + 'static ) -> Self

Sets a generic filter function.

pub fn author_id(self, author_id: UserId) -> Self

Filters Message’s by a specific UserId.

pub fn channel_id(self, channel_id: ChannelId) -> Self

Filters Message’s by a specific ChannelId.

pub fn guild_id(self, guild_id: GuildId) -> Self

Filters Message’s by a specific GuildId.

pub fn stream(self) -> impl Stream<Item = Message>

Returns a [Stream] over all collected Message.

pub fn build(self) -> impl Stream<Item = Message>

👎Deprecated: use .stream() instead

Deprecated, use Self::stream() instead.

pub async fn next(self) -> Option<Message>

Returns the next Message which passes the filters. You can also call .await on the MessageCollector directly.