Module serenity::constants

A set of constants used by the library.

Modules

close_codes

Enums

OpCode
Enum to map gateway opcodes.

VoiceOpCode
Enum to map voice opcodes.

Constants

EMBED_MAX_LENGTH
The maximum length of the textual size of an embed.

GATEWAY_VERSION
The gateway version used by the library. The gateway URI is retrieved via the REST API.

LARGE_THRESHOLD
The large threshold to send on identify.

MESSAGE_CODE_LIMIT
The maximum unicode code points allowed within a message by Discord.

USER_AGENT
The UserAgent sent along with every request.

Statics

JOIN_MESSAGES
List of messages Discord shows on member join.