Module serenity::constants

A set of constants used by the library.

Modules

close_codes

Enums

OpCode

Enum to map gateway opcodes.

Constants

EMBED_MAX_COUNT

The maximum number of embeds in a message.

EMBED_MAX_LENGTH

The maximum length of the textual size of an embed.

GATEWAY_VERSION

The gateway version used by the library. The gateway URL is retrieved via the REST API.

LARGE_THRESHOLD

The large threshold to send on identify.

MEMBER_FETCH_LIMIT

The maximum number of members the bot can fetch at once

MESSAGE_CODE_LIMIT

The maximum unicode code points allowed within a message by Discord.

STICKER_MAX_COUNT

The maximum number of stickers in a message.

USER_AGENT

The UserAgent sent along with every request.

Statics

JOIN_MESSAGES

List of messages Discord shows on member join.