Module serenity::constants::close_codes

Constants

ALREADY_AUTHENTICATED

More than one identify payload was sent.

AUTHENTICATION_FAILED

The account token sent with the identify payload was incorrect.

DECODE_ERROR

An invalid payload was sent.

DISALLOWED_GATEWAY_INTENTS

Disallowed gateway intents have been provided.

INVALID_GATEWAY_INTENTS

Undocumented gateway intents have been provided.

INVALID_SEQUENCE

The sequence sent when resuming the session was invalid.

INVALID_SHARD

An invalid shard when identifying was sent.

NOT_AUTHENTICATED

A payload was sent prior to identifying.

RATE_LIMITED

Payloads were being sent too quickly.

SESSION_TIMEOUT

A session timed out.

SHARDING_REQUIRED

The session would have handled too many guilds.

UNKNOWN_ERROR

Unknown error; try reconnecting?

UNKNOWN_OPCODE

Invalid Gateway OP Code.