Enum serenity::Error

#[non_exhaustive]pub enum Error {
    Decode(&'static str, Value),
    Format(FormatError),
    Io(IoError),
    Json(JsonError),
    Model(ModelError),
    Num(ParseIntError),
    ExceededLimit(String, u32),
    NotInRange(&'static str, u64, u64, u64),
    Other(&'static str),
    Url(String),
    Client(ClientError),
    Gateway(GatewayError),
    Http(Box<HttpError>),
    Rustls(RustlsError),
    Tungstenite(TungsteniteError),
}

A common error enum returned by most of the library’s functionality within a custom Result.

The most common error types, the ClientError and GatewayError enums, are both wrapped around this in the form of the Client and Gateway variants.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Decode(&'static str, Value)

An error while decoding a payload.

Format(FormatError)

There was an error with a format.

Io(IoError)

An std::io error.

Json(JsonError)

An error from the serde_json crate.

Model(ModelError)

An error from the model module.

Num(ParseIntError)

An error occurred while parsing an integer.

ExceededLimit(String, u32)

Input exceeded a limit. Providing the input and the limit that’s not supposed to be exceeded.

This only exists for the GuildId::ban and Member::ban functions. For their cases, it’s the “reason”.

NotInRange(&'static str, u64, u64, u64)

The input is not in the specified range. Returned by GuildId::members, Guild::members and PartialGuild::members

(param_name, value, range_min, range_max)

Other(&'static str)

Some other error. This is only used for “Expected value ” errors, when a more detailed error can not be easily provided via the Error::Decode variant.

Url(String)

An error from the url crate.

Client(ClientError)

A client error.

Gateway(GatewayError)

An error from the gateway module.

Http(Box<HttpError>)

An error from the http module.

Rustls(RustlsError)

An error occuring in rustls

Tungstenite(TungsteniteError)

An error from the tungstenite crate.