Module serenity::framework

Available on crate feature framework only.

The framework is a customizable method of separating commands.

This is used in combination with ClientBuilder::framework.

Traits

Framework

A trait for defining your own framework for serenity to use.