Module serenity::framework::standard

Modules

help_commands
A collection of default help commands for the framework.

Structs

Args

Command
Command struct used to store commands internally.

CommandError
An error from a command.

CommandGroup

Configuration
The configuration to use for a Framework associated with a Client instance.

CreateCommand

CreateGroup
Used to create command groups

StandardFramework
A utility for easily managing dispatches to commands.

Enums

ArgError
Defines how an operation on an Args method failed.

CommandOrAlias

CommandType
Command function type. Allows to access internal framework things inside your commands.

DispatchError
A enum representing all possible fail conditions under which a command won't be executed.