Module serenity::framework::standard

Modules

buckets

help_commands
A collection of default help commands for the framework.

macros

Structs

Args
A utility struct for handling “arguments” of a command.

BucketBuilder

Check
A check can be part of a command or group and will be executed to determine whether a user is permitted to use related item.

Command

CommandGroup

CommandOptions

Configuration
The configuration to use for a StandardFramework associated with a Client instance.

GroupOptions

HelpCommand

HelpOptions

Iter
Parse each argument individually, as an iterator.

RawArguments
Access to all of the arguments, as an iterator.

StandardFramework
A utility for easily managing dispatches to commands.

WithWhiteSpace
A configuration struct for deciding whether the framework should allow optional whitespace between prefixes, group prefixes and command names.

Enums

ArgError
Defines how an operation on an Args method failed.

Delimiter
Dictates how Args should split arguments, if by one character, or a string.

DispatchError
An enum representing all possible fail conditions under which a command won’t be executed.

HelpBehaviour
Describes the behaviour the help-command shall execute once it encounters a command which the user or command fails to meet following criteria : Lacking required permissions to execute the command. Lacking required roles to execute the command. The command can’t be used in the current channel (as in DM only or guild only).

OnlyIn

Reason
This type describes why a check has failed.

Traits

CommonOptions

Type Definitions

CheckFunction

CommandError

CommandFn

CommandResult

HelpCommandFn