serenity::framework::standard
Module buckets

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

Available on crate features framework and standard_framework only.

Structs

BucketBuilderDeprecated

RateLimitInfoDeprecated

Contains information about a rate limit.

RevertBucketDeprecated

An error struct that can be returned from a command to set the bucket one step back.

Enums

LimitedForDeprecated

Decides what a bucket will use to collect tickets for.

RateLimitActionDeprecated

Action taken for the command invocation.