Module serenity::framework::standard::buckets

Available on crate features framework and standard_framework only.

Structs

BucketBuilder

RateLimitInfo

Contains information about a rate limit.

RevertBucket

An error struct that can be returned from a command to set the bucket one step back.

Enums

LimitedFor

Decides what a bucket will use to collect tickets for.

RateLimitAction

Action taken for the command invocation.