Module serenity::framework::standard::help_commands

A collection of default help commands for the framework.

Example

Using the with_embeds function to have the framework's help message use embeds:

use serenity::framework::standard::help_commands;
use serenity::Client;
use std::env;

let mut client = Client::new(&env::var("DISCORD_TOKEN").unwrap());
use serenity::framework::StandardFramework;

client.with_framework(StandardFramework::new()
    .command("help", |c| c.exec_help(help_commands::with_embeds)));

The same can be accomplished with no embeds by substituting with_embeds with the plain function.

Functions

plain
Posts formatted text displaying each individual command group and its commands.

with_embeds
Posts an embed showing each individual command group and its commands.