serenity::framework::standard
Module macros

👎Deprecated: The standard framework is deprecated, and will be removed in 0.13. Please migrate to poise for command handling

Available on crate features framework and standard_framework only.

Attribute Macros

check

A macro for marking a function as a condition checker to groups and commands.

command

The heart of the attribute-based framework.

group

Create a grouping of commands.

help

A brother macro to command, but for the help command.

hook

A macro that transforms async functions (and closures) into plain functions, whose return type is a boxed Future.