Module serenity::framework::standard::macros

This is supported on crate features framework and standard_framework only.

Attribute Macros

check

A macro for marking a function as a condition checker to groups and commands.

command

The heart of the attribute-based framework.

group

Create a grouping of commands.

help

A brother macro to command, but for the help command. An interface for simple browsing of all the available commands the bot provides, and reading through specific information regarding a command.

hook

A macro that transforms async functions (and closures) into plain functions, whose return type is a boxed Future.