Attribute Macro serenity::framework::standard::macros::help

#[help]

Available on crate features framework and standard_framework only.

A brother macro to command, but for the help command. An interface for simple browsing of all the available commands the bot provides, and reading through specific information regarding a command.

As such, the options here will pertain in the help command’s layout than its functionality.

Options

Syntax	Description	Argument explanation

#[suggestion_text(s)]
#[suggestion_text = s]	When suggesting a command’s name	s is a string

#[no_help_available_text(s)]
#[no_help_available_text = s]	When help is unavailable for a command.	s is a string

#[usage_label(s)]
#[usage_label = s]	How should the command be used.	s is a string

#[usage_sample_label(s)]
#[usage_sample_label = s]	Actual sample label.	s is a string

#[ungrouped_label(s)]
#[ungrouped_label = s]	Ungrouped commands label.	s is a string

#[grouped_label(s)]
#[grouped_label = s]	Grouped commands label.	s is a string

#[sub_commands_label(s)]
#[sub_commands_label = s]	Sub commands label.	s is a string

#[description_label(s)]
#[description_label = s]	Label at the start of the description.	s is a string

#[aliases_label(s)]
#[aliases_label= s]	Label for a command’s aliases.	s is a string

#[guild_only_text(s)]
#[guild_only_text = s]	When a command is specific to guilds only.	s is a string

#[checks_label(s)]
#[checks_label = s]	The header text when showing checks in the help command.	s is a string

#[dm_only_text(s)]
#[dm_only_text = s]	When a command is specific to dms only.	s is a string

#[dm_and_guild_text(s)]
#[dm_and_guild_text = s]	When a command is usable in both guilds and dms.	s is a string

#[available_text(s)]
#[available_text = s]	When a command is available.	s is a string

#[command_not_found_text(s)]
#[command_not_found_text = s]	When a command wasn’t found.	s is a string

#[individual_command_tip(s)]
#[individual_command_tip = s]	How the user should access a command’s details.	s is a string

#[strikethrough_commands_tip_in_dm(s)]
#[strikethrough_commands_tip_in_dm = s]	Reasoning behind strikethrough-commands.
Only used in dms.	s is a string. If not provided, default text will be used instead.

#[strikethrough_commands_tip_in_guild(s)]
#[strikethrough_commands_tip_in_guild = s]	Reasoning behind strikethrough-commands.
Only used in guilds.	s is a string. If not provided, default text will be used instead.

#[group_prefix(s)]
#[group_prefix = s]	For introducing a group’s prefix	s is a string

#[lacking_role(s)]
#[lacking_role = s]	If a user lacks required roles, this will treat how commands will be displayed.	s is a string. Accepts strike (strikethroughs), hide (will not be listed) or nothing(leave be).

#[lacking_ownership(s)]
#[lacking_ownership = s]	If a user lacks ownership, this will treat how these commands will be displayed.	s is a string. Accepts strike (strikethroughs), hide (will not be listed) or nothing(leave be).

#[lacking_permissions(s)]
#[lacking_permissions = s]	If a user lacks permissions, this will treat how commands will be displayed.	s is a string. Accepts strike (strikethroughs), hide (will not be listed) or nothing(leave be).

#[lacking_conditions(s)]
#[lacking_conditions = s]	If conditions (of a check) may be lacking by the user, this will treat how these commands will be displayed.	s is a string. Accepts strike (strikethroughs), hide (will not be listed) or nothing(leave be).

#[wrong_channel(s)]
#[wrong_channel = s]	If a user is using the help-command in a channel where a command is not available, this behaviour will be executed.	s is a string. Accepts strike (strikethroughs), hide (will not be listed) or nothing(leave be).

#[embed_error_colour(n)]	Colour that the help-embed will use upon an error.	n is a name to one of the provided constants of the Colour struct or an RGB value #RRGGBB.

#[embed_success_colour(n)]	Colour that the help-embed will use normally.	n is a name to one of the provided constants of the Colour struct or an RGB value #RRGGBB.

#[max_levenshtein_distance(n)]	How much should the help command search for a similar name.
Indicator for a nested guild. The prefix will be repeated based on what kind of level the item sits. A sub-group would be level two, a sub-sub-group would be level three.	n is a 64-bit, unsigned integer.

#[indention_prefix(s)]
#[indention_prefix = s]	The prefix used to express how deeply nested a command or group is.	s is a string