Struct serenity::framework::standard::HelpOptions

pub struct HelpOptions {

    pub names: &'static [&'static str],
    pub suggestion_text: &'static str,
    pub no_help_available_text: &'static str,
    pub usage_label: &'static str,
    pub usage_sample_label: &'static str,
    pub ungrouped_label: &'static str,
    pub description_label: &'static str,
    pub grouped_label: &'static str,
    pub aliases_label: &'static str,
    pub guild_only_text: &'static str,
    pub checks_label: &'static str,
    pub sub_commands_label: &'static str,
    pub dm_only_text: &'static str,
    pub dm_and_guild_text: &'static str,
    pub available_text: &'static str,
    pub command_not_found_text: &'static str,
    pub individual_command_tip: &'static str,
    pub strikethrough_commands_tip_in_dm: Option<&'static str>,
    pub strikethrough_commands_tip_in_guild: Option<&'static str>,
    pub group_prefix: &'static str,
    pub lacking_role: HelpBehaviour,
    pub lacking_permissions: HelpBehaviour,
    pub lacking_ownership: HelpBehaviour,
    pub lacking_conditions: HelpBehaviour,
    pub wrong_channel: HelpBehaviour,
    pub embed_error_colour: Colour,
    pub embed_success_colour: Colour,
    pub max_levenshtein_distance: usize,
    pub indention_prefix: &'static str,

}

Available on crate features framework and standard_framework only.

Fields
names: &'static [&'static str]

Which names should the help command use for dispatching. Defaults to ["help"]

suggestion_text: &'static str

Suggests a command’s name.

no_help_available_text: &'static str

If no help is available, this text will be displayed.

usage_label: &'static str

How to use a command, {usage_label}: {command_name} {args}

usage_sample_label: &'static str

Actual sample label, {usage_sample_label}: {command_name} {args}

ungrouped_label: &'static str

Text labeling ungrouped commands, {ungrouped_label}: ...

description_label: &'static str

Text labeling the start of the description.

grouped_label: &'static str

Text labeling grouped commands, {grouped_label} {group_name}: ...

aliases_label: &'static str

Text labeling a command’s alternative names (aliases).

guild_only_text: &'static str

Text specifying that a command is only usable in a guild.

checks_label: &'static str

Text labelling a command’s names of checks.

sub_commands_label: &'static str

Text labelling a command’s subcommands

dm_only_text: &'static str

Text specifying that a command is only usable in via DM.

dm_and_guild_text: &'static str

Text specifying that a command can be used via DM and in guilds.

available_text: &'static str

Text expressing that a command is available.

command_not_found_text: &'static str

Error-message once a command could not be found.

Output-example (without whitespace between both substitutions: {command_not_found_text}{command_name} {command_name} describes user’s input as in: {prefix}help {command_name}.

individual_command_tip: &'static str

Explains the user on how to use access a single command’s details.

strikethrough_commands_tip_in_dm: Option<&'static str>

Explains reasoning behind strikethrough-commands, see fields requiring HelpBehaviour for further information. If HelpBehaviour::Strike is unused, this field will evaluate to None during creation inside of the help macro.

Note: Text is only used in direct messages.

strikethrough_commands_tip_in_guild: Option<&'static str>

Explains reasoning behind strikethrough-commands, see fields requiring HelpBehaviour for further information. If HelpBehaviour::Strike is unused, this field will evaluate to None during creation inside of the help macro.

Note: Text is only used in guilds.

group_prefix: &'static str

Announcing a group’s prefix as in: {group_prefix} {prefix}.

lacking_role: HelpBehaviour

If a user lacks required roles, this will treat how these commands will be displayed.

lacking_permissions: HelpBehaviour

If a user lacks permissions, this will treat how these commands will be displayed.

lacking_ownership: HelpBehaviour

If a user lacks ownership, this will treat how these commands will be displayed.

lacking_conditions: HelpBehaviour

If conditions (of a check) may be lacking by the user, this will treat how these commands will be displayed.

wrong_channel: HelpBehaviour

If a user is using the help-command in a channel where a command is not available, this behaviour will be executed.

embed_error_colour: Colour

Colour help-embed will use upon encountering an error.

embed_success_colour: Colour

Colour help-embed will use if no error occurred.

max_levenshtein_distance: usize

If not 0, help will check whether a command is similar to searched named.

indention_prefix: &'static str

Help will use this as prefix to express how deeply nested a command or group is.