Trait serenity::framework::Framework

pub trait Framework: Send + Sync {
    fn dispatch<'life0, 'async_trait>(
        &'life0 self,
        event: FullEvent
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn init<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        client: &'life1 Client
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Available on crate feature framework only.

A trait for defining your own framework for serenity to use.

Should you implement this trait, or define a message handler, depends on you. However, using this will benefit you by abstracting the EventHandler away.

Required Methods

fn dispatch<'life0, 'async_trait>(
&'life0 self,
event: FullEvent
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Called on every incoming event.

Provided Methods

fn init<'life0, 'life1, 'async_trait>(
&'life0 mut self,
client: &'life1 Client
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Called directly after the Client is created.