serenity

Module gateway

Available on crate feature gateway only.

Contains the necessary plumping for maintaining a connection with Discord. The primary building blocks are the Client and the Shard.

The Client is a high-level interface that takes care of communicating with Discord’s REST API as well as receiving and dispatching events from the gateway using a WebSocket client.

On the other hand, the Shard is a low-level receiver and sender representing a single connection to Discord. The client will handle shard management automatically for you, so you should only care about using it directly if you really need to. See the sharding module for details and documentation.

Re-exports

pub use self::sharding::*;

Modules

client

A module for Client and supporting types.

sharding

Mechanisms for configuring and managing sharded gateway connections.

Structs

ActivityData

Activity data of the current user.

PresenceData

Presence data of the current user.

WsClient

Enums

ChunkGuildFilter

Discord docs.

GatewayError

An error that occurred while attempting to deal with the gateway.

Traits

VoiceGatewayManagervoice

Interface for any compatible voice plugin.