Module serenity::gateway

Available on crate feature gateway only.

Contains the necessary plumping for maintaining a connection with Discord. The primary building blocks are the Client and the Shard.

The Client is a high-level interface that takes care of communicating with Discord’s REST API as well as receiving and dispatching events from the gateway using a WebSocket client.

On the other hand, the Shard is a low-level receiver and sender representing a single connection to Discord. The client will handle shard management automatically for you, so you should only care about using it directly if you really need to. See the sharding module for details and documentation.

Re-exports

pub use self::sharding::*;

Modules

client

The Client contains information about a bot’s token, as well as event handlers. Dispatching events to handlers and starting sharded gateway connections is handled directly by the client. In addition, the client automatically handles caching via the Cache struct.

sharding

Sharding is a method for load-balancing bots across separate threads or processes. Sharding is enforced on bots by Discord once they reach a certain number of guilds (2500). Once this threshold is reached, a but must be sharded such that at most 2500 guilds are allocated per shard.

Structs

ActivityData

Activity data of the current user.

PresenceData

Presence data of the current user.

WsClient

Enums

ChunkGuildFilter

Discord docs.

GatewayError

An error that occurred while attempting to deal with the gateway.

Traits

VoiceGatewayManagervoice

Interface for any compatible voice plugin.