serenity::gateway
Module sharding

Available on crate feature gateway only.

Module containing types for gateway sharding.

Sharding is a method for load-balancing bots across separate threads or processes. Sharding is enforced on bots by Discord once they reach a certain number of guilds (2500). Once this threshold is reached, a but must be sharded such that at most 2500 guilds are allocated per shard.

The “recommended” number of guilds per shard is around 1000. Sharding allows for bots to be distributed by handing shards off to separate processes or even separate machines in a distributed network (e.g. cloud workers). However, sometimes you may wish for all shards to share some global state. Serenity accomodates both of these usecases.

See Discord’s documentation for more information.

This module also provides some lower-level facilities for performing sharding manually:

ShardManager

The shard manager provides a clean interface for communicating with shard runners either individually or collectively, with functions such as ShardManager::shutdown and ShardManager::restart to manage shards in a fine-grained way.

For most use cases, the ShardManager will fit all your low-level sharding needs.

ShardQueuer

A light wrapper around an mpsc receiver that receives ShardQueuerMessages. It should be run in its own thread so it can receive messages to start shards concurrently in a queue.

ShardRunner

The shard runner is responsible for directly running a single shard and communicating with the gateway through its respective WebSocket client. It performs actions such as identifying, reconnecting, resuming, and sending presence updates to the gateway.

Structs

CollectorCallbackcollector

Newtype around a callback that will be called on every incoming request. As long as this collector should still receive events, it should return true. Once it returns false, it is removed.

Shard

A Shard is a higher-level handler for a websocket connection to Discord’s gateway.

ShardManager

A manager for handling the status of shards by starting them, restarting them, and stopping them when required.

ShardManagerOptions

ShardMessenger

A handle to a ShardRunner.

ShardQueue

A queue of ShardIds that is split up into multiple buckets according to the value of max_concurrency.

ShardQueuer

The shard queuer is a simple loop that runs indefinitely to manage the startup of shards.

ShardRunner

A runner for managing a Shard and its respective WebSocket client.

ShardRunnerInfo

Information about a ShardRunner.

ShardRunnerOptions

Options to be passed to ShardRunner::new.

ShardStageUpdateEvent

An event denoting that a shard’s connection stage was changed.

Enums

ConnectionStage

Indicates the current connection stage of a Shard.

ReconnectType

The type of reconnection that should be performed.

ShardAction

ShardQueuerMessage

A message to be sent to the ShardQueuer.

ShardRunnerMessage

A message to send from a shard over a WebSocket.