Module serenity::http::client

Structs

Http
Note: For all member functions that return a Result, the Error kind will be either Error::Http or Error::Json.

HttpBuilder
A builder implementing Future building a Http client to perform requests to Discord’s HTTP API. If you do not need to use a proxy or do not need to disable the rate limiter, you can use Http::new or Http::new_with_token instead.