Struct serenity::http::client::Http

pub struct Http {
    pub ratelimiter: Ratelimiter,
    pub ratelimiter_disabled: bool,
    pub proxy: Option<Url>,
    pub token: String,
    pub application_id: u64,
    // some fields omitted
}

Note: For all member functions that return a Result, the Error kind will be either Error::Http or Error::Json.

Fields
ratelimiter: Ratelimiterratelimiter_disabled: boolproxy: Option<Url>token: Stringapplication_id: u64
Implementations

impl Http

pub fn new(client: Arc<Client>, token: &str) -> Self

pub fn new_with_application_id(application_id: u64) -> Self

pub fn new_with_token(token: &str) -> Self

pub fn new_with_token_application_id(token: &str, application_id: u64) -> Self

pub async fn add_member_role(
&self,
guild_id: u64,
user_id: u64,
role_id: u64
) -> Result<()>

Adds a single Role to a Member in a Guild.

Note: Requires the Manage Roles permission and respect of role hierarchy.

pub async fn ban_user(
&self,
guild_id: u64,
user_id: u64,
delete_message_days: u8,
reason: &str
) -> Result<()>

Bans a User from a Guild, removing their messages sent in the last X number of days.

Passing a delete_message_days of 0 is equivalent to not removing any messages. Up to 7 days’ worth of messages may be deleted.

Note: Requires that you have the Ban Members permission.

pub async fn broadcast_typing(&self, channel_id: u64) -> Result<()>

Broadcasts that the current user is typing in the given Channel.

This lasts for about 10 seconds, and will then need to be renewed to indicate that the current user is still typing.

This should rarely be used for bots, although it is a good indicator that a long-running command is still being processed.

pub async fn create_channel(
&self,
guild_id: u64,
map: &Map<String, Value>
) -> Result<GuildChannel>

Creates a GuildChannel in the Guild given its Id.

Refer to the Discord’s docs for information on what fields this requires.

Note: Requires the Manage Channels permission.

pub async fn create_emoji(&self, guild_id: u64, map: &Value) -> Result<Emoji>

Creates an emoji in the given Guild with the given data.

View the source code for Guild::create_emoji method to see what fields this requires.

Note: Requires the Manage Emojis permission.

pub async fn create_followup_message(
&self,
interaction_token: &str,
map: &Value
) -> Result<Message>

This is supported on crate feature unstable_discord_api only.

Create a follow-up message for an Interaction.

Functions the same as Self::execute_webhook

pub async fn create_global_application_command(
&self,
map: &Value
) -> Result<ApplicationCommand>

This is supported on crate feature unstable_discord_api only.

Creates a new global command.

New global commands will be available in all guilds after 1 hour.

Refer to Discord’s docs for field information.

Note: Creating a command with the same name as an existing command for your application will overwrite the old command.

pub async fn create_global_application_commands(
&self,
map: &Value
) -> Result<Vec<ApplicationCommand>>

This is supported on crate feature unstable_discord_api only.

Creates new global application commands.

pub async fn create_guild_application_commands(
&self,
guild_id: u64,
map: &Value
) -> Result<Vec<ApplicationCommand>>

This is supported on crate feature unstable_discord_api only.

Creates new guild application commands.

pub async fn create_guild(&self, map: &Value) -> Result<PartialGuild>

Creates a guild with the data provided.

Only a PartialGuild will be immediately returned, and a full Guild will be received over a Shard, if at least one is running.

Note: This endpoint is currently limited to 10 active guilds. The limits are raised for whitelisted GameBridge applications. See the documentation on this endpoint for more info.

Examples

Create a guild called "test" in the US West region:

use serde_json::json;
use serenity::http::Http;

let map = json!({
    "name": "test",
    "region": "us-west",
});

let _result = http.create_guild(&map).await?;

pub async fn create_guild_application_command(
&self,
guild_id: u64,
map: &Value
) -> Result<ApplicationCommand>

This is supported on crate feature unstable_discord_api only.

Creates a new guild command.

New guild commands will be available in the guild immediately.

Refer to Discord’s docs for field information.

pub async fn create_guild_integration(
&self,
guild_id: u64,
integration_id: u64,
map: &Value
) -> Result<()>

Creates an Integration for a Guild.

Refer to Discord’s docs for field information.

Note: Requires the Manage Guild permission.

pub async fn create_interaction_response(
&self,
interaction_id: u64,
interaction_token: &str,
map: &Value
) -> Result<()>

This is supported on crate feature unstable_discord_api only.

Creates a response to an Interaction from the gateway.

Refer to Discord’s docs for the object it takes.

pub async fn create_invite(
&self,
channel_id: u64,
map: &Map<String, Value>
) -> Result<RichInvite>

Creates a RichInvite for the given channel.

Refer to Discord’s docs for field information.

All fields are optional.

Note: Requires the Create Invite permission.

pub async fn create_permission(
&self,
channel_id: u64,
target_id: u64,
map: &Value
) -> Result<()>

Creates a permission override for a member or a role in a channel.

pub async fn create_private_channel(
&self,
map: &Value
) -> Result<PrivateChannel>

Creates a private channel with a user.

pub async fn create_reaction(
&self,
channel_id: u64,
message_id: u64,
reaction_type: &ReactionType
) -> Result<()>

Reacts to a message.

pub async fn create_role(
&self,
guild_id: u64,
map: &Map<String, Value>
) -> Result<Role>

Creates a role.

pub async fn create_webhook(
&self,
channel_id: u64,
map: &Value
) -> Result<Webhook>

Creates a webhook for the given channel’s Id, passing in the given data.

This method requires authentication.

The Value is a map with the values of:

avatar: base64-encoded 128x128 image for the webhook’s default avatar (optional);

name: the name of the webhook, limited to between 2 and 100 characters long.

Examples

Creating a webhook named test:

use serde_json::json;
use serenity::http::Http;

let channel_id = 81384788765712384;
let map = json!({"name": "test"});

let webhook = http.create_webhook(channel_id, &map).await?;

pub async fn delete_channel(&self, channel_id: u64) -> Result<Channel>

Deletes a private channel or a channel in a guild.

pub async fn delete_emoji(&self, guild_id: u64, emoji_id: u64) -> Result<()>

Deletes an emoji from a server.

pub async fn delete_followup_message(
&self,
interaction_token: &str,
message_id: u64
) -> Result<()>

This is supported on crate feature unstable_discord_api only.

Deletes a follow-up message for an interaction.

pub async fn delete_global_application_command(
&self,
command_id: u64
) -> Result<()>

This is supported on crate feature unstable_discord_api only.

Deletes a global command.

pub async fn delete_guild(&self, guild_id: u64) -> Result<PartialGuild>

Deletes a guild, only if connected account owns it.

pub async fn delete_guild_application_command(
&self,
guild_id: u64,
command_id: u64
) -> Result<()>

This is supported on crate feature unstable_discord_api only.

Deletes a guild command.

pub async fn delete_guild_integration(
&self,
guild_id: u64,
integration_id: u64
) -> Result<()>

Removes an integration from a guild.

pub async fn delete_invite(&self, code: &str) -> Result<Invite>

Deletes an invite by code.

pub async fn delete_message(
&self,
channel_id: u64,
message_id: u64
) -> Result<()>

Deletes a message if created by us or we have specific permissions.

pub async fn delete_messages(&self, channel_id: u64, map: &Value) -> Result<()>

Deletes a bunch of messages, only works for bots.

pub async fn delete_message_reactions(
&self,
channel_id: u64,
message_id: u64
) -> Result<()>

Deletes all of the Reactions associated with a Message.

Examples

use serenity::model::id::{ChannelId, MessageId};

let channel_id = ChannelId(7);
let message_id = MessageId(8);

http.delete_message_reactions(channel_id.0, message_id.0).await?;

pub async fn delete_message_reaction_emoji(
&self,
channel_id: u64,
message_id: u64,
reaction_type: &ReactionType
) -> Result<()>

Deletes all the reactions for a given emoji on a message.

pub async fn delete_original_interaction_response(
&self,
interaction_token: &str
) -> Result<()>

This is supported on crate feature unstable_discord_api only.

Deletes the initial interaction response.

pub async fn delete_permission(
&self,
channel_id: u64,
target_id: u64
) -> Result<()>

Deletes a permission override from a role or a member in a channel.

pub async fn delete_reaction(
&self,
channel_id: u64,
message_id: u64,
user_id: Option<u64>,
reaction_type: &ReactionType
) -> Result<()>

Deletes a reaction from a message if owned by us or we have specific permissions.

pub async fn delete_role(&self, guild_id: u64, role_id: u64) -> Result<()>

Deletes a role from a server. Can’t remove the default everyone role.

pub async fn delete_webhook(&self, webhook_id: u64) -> Result<()>

Deletes a Webhook given its Id.

This method requires authentication, whereas Self::delete_webhook_with_token does not.

Examples

Deletes a webhook given its Id:

use serenity::http::Http;

// Due to the `delete_webhook` function requiring you to authenticate, you
// must have set the token first.
let http = Http::default();

http.delete_webhook(245037420704169985).await?;
      Ok(())

pub async fn delete_webhook_with_token(
&self,
webhook_id: u64,
token: &str
) -> Result<()>

Deletes a Webhook given its Id and unique token.

This method does not require authentication.

Examples

Deletes a webhook given its Id and unique token:

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";

http.delete_webhook_with_token(id, token).await?;

pub async fn edit_channel(
&self,
channel_id: u64,
map: &Map<String, Value>
) -> Result<GuildChannel>

Changes channel information.

pub async fn edit_emoji(
&self,
guild_id: u64,
emoji_id: u64,
map: &Value
) -> Result<Emoji>

Changes emoji information.

pub async fn edit_followup_message(
&self,
interaction_token: &str,
message_id: u64,
map: &Value
) -> Result<Message>

This is supported on crate feature unstable_discord_api only.

Edits a follow-up message for an interaction.

Refer to Discord’s docs for Edit Webhook Message for field information.

pub async fn edit_global_application_command(
&self,
command_id: u64,
map: &Value
) -> Result<ApplicationCommand>

This is supported on crate feature unstable_discord_api only.

Edits a global command.

Updates will be available in all guilds after 1 hour.

Refer to Discord’s docs for field information.

pub async fn edit_guild(
&self,
guild_id: u64,
map: &Map<String, Value>
) -> Result<PartialGuild>

Changes guild information.

pub async fn edit_guild_application_command(
&self,
guild_id: u64,
command_id: u64,
map: &Value
) -> Result<ApplicationCommand>

This is supported on crate feature unstable_discord_api only.

Edits a guild command.

Updates for guild commands will be available immediately.

Refer to Discord’s docs for field information.

pub async fn edit_guild_application_command_permissions(
&self,
guild_id: u64,
command_id: u64,
map: &Value
) -> Result<ApplicationCommandPermission>

This is supported on crate feature unstable_discord_api only.

Edits a guild command permissions.

Updates for guild commands will be available immediately.

Refer to Discord’s documentation for field information.

pub async fn edit_guild_application_commands_permissions(
&self,
guild_id: u64,
map: &Value
) -> Result<Vec<ApplicationCommandPermission>>

This is supported on crate feature unstable_discord_api only.

Edits a guild commands permissions.

Updates for guild commands will be available immediately.

Refer to Discord’s documentation for field information.

pub async fn edit_guild_channel_positions(
&self,
guild_id: u64,
value: &Value
) -> Result<()>

Edits the positions of a guild’s channels.

pub async fn edit_guild_widget(
&self,
guild_id: u64,
map: &Value
) -> Result<GuildWidget>

Edits a Guild’s widget.

pub async fn edit_guild_welcome_screen(
&self,
guild_id: u64,
map: &Value
) -> Result<GuildWelcomeScreen>

Edits a guild welcome screen.

pub async fn edit_member(
&self,
guild_id: u64,
user_id: u64,
map: &Map<String, Value>
) -> Result<Member>

Does specific actions to a member.

pub async fn edit_message(
&self,
channel_id: u64,
message_id: u64,
map: &Value
) -> Result<Message>

Edits a message by Id.

Note: Only the author of a message can modify it.

pub async fn crosspost_message(
&self,
channel_id: u64,
message_id: u64
) -> Result<Message>

Crossposts a message by Id.

Note: Only available on announcements channels.

pub async fn edit_member_me(
&self,
guild_id: u64,
map: &Map<String, Value>
) -> Result<Member>

Edits the current member for the provided Guild via its Id.

pub async fn edit_nickname(
&self,
guild_id: u64,
new_nickname: Option<&str>
) -> Result<()>

Edits the current user’s nickname for the provided Guild via its Id.

Pass None to reset the nickname.

pub async fn get_original_interaction_response(
&self,
interaction_token: &str
) -> Result<Message>

This is supported on crate feature unstable_discord_api only.

Gets the initial interaction response.

pub async fn edit_original_interaction_response(
&self,
interaction_token: &str,
map: &Value
) -> Result<Message>

This is supported on crate feature unstable_discord_api only.

Edits the initial interaction response.

Refer to Discord’s docs for Edit Webhook Message for field information.

pub async fn edit_profile(
&self,
map: &Map<String, Value>
) -> Result<CurrentUser>

Edits the current user’s profile settings.

pub async fn edit_role(
&self,
guild_id: u64,
role_id: u64,
map: &Map<String, Value>
) -> Result<Role>

Changes a role in a guild.

pub async fn edit_role_position(
&self,
guild_id: u64,
role_id: u64,
position: u64
) -> Result<Vec<Role>>

Changes the position of a role in a guild.

pub async fn edit_voice_state(
&self,
guild_id: u64,
user_id: u64,
map: &Map<String, Value>
) -> Result<()>

Changes another user’s voice state in a stage channel.

The Value is a map with values of:

channel_id: ID of the channel the user is currently in (required)

supress: Bool which toggles user’s suppressed state. Setting this to false will invite the user to speak.

Example

Suppress a user

use serde_json::json;
use serenity::http::Http;

let guild_id = 187450744427773963;
let user_id = 150443906511667200;
let value = json!({
    "channel_id": "826929611849334784",
    "suppress": true,
});

let map = value.as_object().unwrap();

// Edit state for another user
http.edit_voice_state(guild_id, user_id, &map).await?;

pub async fn edit_voice_state_me(
&self,
guild_id: u64,
map: &Map<String, Value>
) -> Result<()>

Changes the current user’s voice state in a stage channel.

The Value is a map with values of:

channel_id: ID of the channel the user is currently in (required)

supress: Bool which toggles user’s suppressed state. Setting this to false will invite the user to speak.

request_to_speak_timestamp: ISO8601 timestamp to set the user’s request to speak. This can be any present or future time.

Example

Unsuppress the current bot user

use serde_json::json;
use serenity::http::Http;

let guild_id = 187450744427773963;
let value = json!({
    "channel_id": "826929611849334784",
    "suppress": false,
    "request_to_speak_timestamp": "2021-03-31T18:45:31.297561+00:00"
});

let map = value.as_object().unwrap();

// Edit state for current user
http.edit_voice_state_me(guild_id, &map).await?;

pub async fn edit_webhook(
&self,
webhook_id: u64,
map: &Value
) -> Result<Webhook>

Edits a the webhook with the given data.

The Value is a map with optional values of:

avatar: base64-encoded 128x128 image for the webhook’s default avatar (optional);

name: the name of the webhook, limited to between 2 and 100 characters long.

Note that, unlike with Self::create_webhook, all values are optional.

This method requires authentication, whereas Self::edit_webhook_with_token does not.

Examples

Edit the image of a webhook given its Id and unique token:

use serde_json::json;
use serenity::http::Http;

let id = 245037420704169985;
let image = serenity::utils::read_image("./webhook_img.png")?;
let map = json!({
    "avatar": image,
});

let edited = http.edit_webhook(id, &map).await?;

pub async fn edit_webhook_with_token(
&self,
webhook_id: u64,
token: &str,
map: &Map<String, Value>
) -> Result<Webhook>

Edits the webhook with the given data.

Refer to the documentation for Self::edit_webhook for more information.

This method does not require authentication.

Examples

Edit the name of a webhook given its Id and unique token:

use serde_json::json;
use serenity::http::Http;

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let value = json!({"name": "new name"});
let map = value.as_object().unwrap();

let edited = http.edit_webhook_with_token(id, token, map).await?;

pub async fn execute_webhook(
&self,
webhook_id: u64,
token: &str,
wait: bool,
map: &Map<String, Value>
) -> Result<Option<Message>>

Executes a webhook, posting a Message in the webhook’s associated Channel.

This method does not require authentication.

Pass true to wait to wait for server confirmation of the message sending before receiving a response. From the Discord docs:

waits for server confirmation of message send before response, and returns the created message body (defaults to false; when false a message that is not saved does not return an error)

The map can optionally contain the following data:

avatar_url: Override the default avatar of the webhook with a URL.

tts: Whether this is a text-to-speech message (defaults to false).

username: Override the default username of the webhook.

Additionally, at least one of the following must be given:

content: The content of the message.

embeds: An array of rich embeds.

Note: For embed objects, all fields are registered by Discord except for height, provider, proxy_url, type (it will always be rich), video, and width. The rest will be determined by Discord.

Examples

Sending a webhook with message content of test:

use serde_json::json;
use serenity::http::Http;

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let value = json!({"content": "test"});
let map = value.as_object().unwrap();

let message = http.execute_webhook(id, token, true, map).await?;

pub async fn execute_webhook_with_files<'a, T, It: IntoIterator<Item = T>>(
&self,
webhook_id: u64,
token: &str,
wait: bool,
files: It,
map: Map<String, Value>
) -> Result<Option<Message>> where
T: Into<AttachmentType<'a>>,

Send file(s) over a webhook.

Errors

Returns an HttpError::UnsuccessfulRequest(ErrorResponse) if the files are too large to send.

pub async fn edit_webhook_message(
&self,
webhook_id: u64,
token: &str,
message_id: u64,
map: &Map<String, Value>
) -> Result<Message>

Edits a webhook’s message by Id.

pub async fn delete_webhook_message(
&self,
webhook_id: u64,
token: &str,
message_id: u64
) -> Result<()>

Deletes a webhook’s messsage by Id.

pub async fn get_active_maintenances(&self) -> Result<Vec<Maintenance>>

Gets the active maintenances from Discord’s Status API.

Does not require authentication.

pub async fn get_bans(&self, guild_id: u64) -> Result<Vec<Ban>>

Gets all the users that are banned in specific guild.

pub async fn get_audit_logs(
&self,
guild_id: u64,
action_type: Option<u8>,
user_id: Option<u64>,
before: Option<u64>,
limit: Option<u8>
) -> Result<AuditLogs>

Gets all audit logs in a specific guild.

pub async fn get_bot_gateway(&self) -> Result<BotGateway>

Gets current bot gateway.

pub async fn get_channel_invites(
&self,
channel_id: u64
) -> Result<Vec<RichInvite>>

Gets all invites for a channel.

pub async fn get_channel_webhooks(
&self,
channel_id: u64
) -> Result<Vec<Webhook>>

Retrieves the webhooks for the given channel’s Id.

This method requires authentication.

Examples

Retrieve all of the webhooks owned by a channel:

let channel_id = 81384788765712384;

let webhooks = http.get_channel_webhooks(channel_id).await?;

pub async fn get_channel(&self, channel_id: u64) -> Result<Channel>

Gets channel information.

pub async fn get_channels(&self, guild_id: u64) -> Result<Vec<GuildChannel>>

Gets all channels in a guild.

pub async fn get_current_application_info(
&self
) -> Result<CurrentApplicationInfo>

Gets information about the current application.

Note: Only applications may use this endpoint.

pub async fn get_current_user(&self) -> Result<CurrentUser>

Gets information about the user we’re connected with.

pub async fn get_emojis(&self, guild_id: u64) -> Result<Vec<Emoji>>

Gets all emojis of a guild.

pub async fn get_emoji(&self, guild_id: u64, emoji_id: u64) -> Result<Emoji>

Gets information about an emoji in a guild.

pub async fn get_gateway(&self) -> Result<Gateway>

Gets current gateway.

pub async fn get_global_application_commands(
&self
) -> Result<Vec<ApplicationCommand>>

This is supported on crate feature unstable_discord_api only.

Fetches all of the global commands for your application.

pub async fn get_global_application_command(
&self,
command_id: u64
) -> Result<ApplicationCommand>

This is supported on crate feature unstable_discord_api only.

Fetches a global commands for your application by its Id.

pub async fn get_guild(&self, guild_id: u64) -> Result<PartialGuild>

Gets guild information.

pub async fn get_guild_with_counts(&self, guild_id: u64) -> Result<PartialGuild>

Gets guild information with counts.

pub async fn get_guild_application_commands(
&self,
guild_id: u64
) -> Result<Vec<ApplicationCommand>>

This is supported on crate feature unstable_discord_api only.

Fetches all of the guild commands for your application for a specific guild.

pub async fn get_guild_application_command(
&self,
guild_id: u64,
command_id: u64
) -> Result<ApplicationCommand>

This is supported on crate feature unstable_discord_api only.

Fetches a guild command by its Id.

pub async fn get_guild_application_commands_permissions(
&self,
guild_id: u64
) -> Result<Vec<ApplicationCommandPermission>>

This is supported on crate feature unstable_discord_api only.

Fetches all of the guild commands permissions for your application for a specific guild.

pub async fn get_guild_application_command_permissions(
&self,
guild_id: u64,
command_id: u64
) -> Result<ApplicationCommandPermission>

This is supported on crate feature unstable_discord_api only.

Gives the guild command permission for your application for a specific guild.

pub async fn get_guild_embed(&self, guild_id: u64) -> Result<GuildEmbed>

👎 Deprecated:
get_guild_embed was renamed to get_guild_widget

Gets a guild embed information.

pub async fn get_guild_widget(&self, guild_id: u64) -> Result<GuildWidget>

Gets a guild widget information.

pub async fn get_guild_preview(&self, guild_id: u64) -> Result<GuildPreview>

Gets a guild preview.

pub async fn get_guild_welcome_screen(
&self,
guild_id: u64
) -> Result<GuildWelcomeScreen>

Gets a guild welcome screen information.

pub async fn get_guild_integrations(
&self,
guild_id: u64
) -> Result<Vec<Integration>>

Gets integrations that a guild has.

pub async fn get_guild_invites(&self, guild_id: u64) -> Result<Vec<RichInvite>>

Gets all invites to a guild.

pub async fn get_guild_vanity_url(&self, guild_id: u64) -> Result<String>

Gets a guild’s vanity URL if it has one.

pub async fn get_guild_members(
&self,
guild_id: u64,
limit: Option<u64>,
after: Option<u64>
) -> Result<Vec<Member>>

Gets the members of a guild. Optionally pass a limit and the Id of the user to offset the result by.

pub async fn get_guild_prune_count(
&self,
guild_id: u64,
map: &Value
) -> Result<GuildPrune>

Gets the amount of users that can be pruned.

pub async fn get_guild_regions(&self, guild_id: u64) -> Result<Vec<VoiceRegion>>

Gets regions that a guild can use. If a guild has the VIP_REGIONS feature enabled, then additional VIP-only regions are returned.

pub async fn get_guild_roles(&self, guild_id: u64) -> Result<Vec<Role>>

Retrieves a list of roles in a Guild.

pub async fn get_guild_webhooks(&self, guild_id: u64) -> Result<Vec<Webhook>>

Retrieves the webhooks for the given guild’s Id.

This method requires authentication.

Examples

Retrieve all of the webhooks owned by a guild:

let guild_id = 81384788765712384;

let webhooks = http.get_guild_webhooks(guild_id).await?;

pub async fn get_guilds(
&self,
target: &GuildPagination,
limit: u64
) -> Result<Vec<GuildInfo>>

Gets a paginated list of the current user’s guilds.

The limit has a maximum value of 100.

Discord’s documentation

Examples

Get the first 10 guilds after a certain guild’s Id:

use serenity::{http::GuildPagination, model::id::GuildId};

let guild_id = GuildId(81384788765712384);

let guilds = http.get_guilds(&GuildPagination::After(guild_id), 10).await?;

pub async fn get_invite(&self, code: &str, stats: bool) -> Result<Invite>

Gets information about a specific invite.

pub async fn get_member(&self, guild_id: u64, user_id: u64) -> Result<Member>

Gets member of a guild.

pub async fn get_message(
&self,
channel_id: u64,
message_id: u64
) -> Result<Message>

Gets a message by an Id, bots only.

pub async fn get_messages(
&self,
channel_id: u64,
query: &str
) -> Result<Vec<Message>>

Gets X messages from a channel.

pub async fn get_pins(&self, channel_id: u64) -> Result<Vec<Message>>

Gets all pins of a channel.

pub async fn get_reaction_users(
&self,
channel_id: u64,
message_id: u64,
reaction_type: &ReactionType,
limit: u8,
after: Option<u64>
) -> Result<Vec<User>>

Gets user Ids based on their reaction to a message. This endpoint is dumb.

pub async fn get_unresolved_incidents(&self) -> Result<Vec<Incident>>

Gets the current unresolved incidents from Discord’s Status API.

Does not require authentication.

pub async fn get_upcoming_maintenances(&self) -> Result<Vec<Maintenance>>

Gets the upcoming (planned) maintenances from Discord’s Status API.

Does not require authentication.

pub async fn get_user(&self, user_id: u64) -> Result<User>

Gets a user by Id.

pub async fn get_user_connections(&self) -> Result<Vec<Connection>>

Gets the current user’s third party connections.

This method only works for user tokens with the connections OAuth2 scope.

pub async fn get_user_dm_channels(&self) -> Result<Vec<PrivateChannel>>

Gets our DM channels.

pub async fn get_voice_regions(&self) -> Result<Vec<VoiceRegion>>

Gets all voice regions.

pub async fn get_webhook(&self, webhook_id: u64) -> Result<Webhook>

Retrieves a webhook given its Id.

This method requires authentication, whereas Self::get_webhook_with_token does not.

Examples

Retrieve a webhook by Id:

let id = 245037420704169985;
let webhook = http.get_webhook(id).await?;

pub async fn get_webhook_with_token(
&self,
webhook_id: u64,
token: &str
) -> Result<Webhook>

Retrieves a webhook given its Id and unique token.

This method does not require authentication.

Examples

Retrieve a webhook by Id and its unique token:

let id = 245037420704169985;
let token = "ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";

let webhook = http.get_webhook_with_token(id, token).await?;

pub async fn kick_member(&self, guild_id: u64, user_id: u64) -> Result<()>

Kicks a member from a guild.

pub async fn kick_member_with_reason(
&self,
guild_id: u64,
user_id: u64,
reason: &str
) -> Result<()>

Kicks a member from a guild with a provided reason.

pub async fn leave_guild(&self, guild_id: u64) -> Result<()>

Leaves a guild.

pub async fn send_files<'a, T, It: IntoIterator<Item = T>>(
&self,
channel_id: u64,
files: It,
map: Map<String, Value>
) -> Result<Message> where
T: Into<AttachmentType<'a>>,

Sends file(s) to a channel.

Errors

Returns an HttpError::UnsuccessfulRequest(ErrorResponse) if the files are too large to send.

pub async fn send_message(
&self,
channel_id: u64,
map: &Value
) -> Result<Message>

Sends a message to a channel.

pub async fn pin_message(&self, channel_id: u64, message_id: u64) -> Result<()>

Pins a message in a channel.

pub async fn remove_ban(&self, guild_id: u64, user_id: u64) -> Result<()>

Unbans a user from a guild.

pub async fn remove_member_role(
&self,
guild_id: u64,
user_id: u64,
role_id: u64
) -> Result<()>

Deletes a single Role from a Member in a Guild.

Note: Requires the Manage Roles permission and respect of role hierarchy.

pub async fn start_guild_prune(
&self,
guild_id: u64,
map: &Value
) -> Result<GuildPrune>

Starts removing some members from a guild based on the last time they’ve been online.

pub async fn start_integration_sync(
&self,
guild_id: u64,
integration_id: u64
) -> Result<()>

Starts syncing an integration with a guild.

pub fn start_typing(self: &Arc<Self>, channel_id: u64) -> Result<Typing>

Starts typing in the specified Channel for an indefinite period of time.

Returns Typing that is used to trigger the typing. Typing::stop must be called on the returned struct to stop typing. Note that on some clients, typing may persist for a few seconds after Typing::stop is called. Typing is also stopped when the struct is dropped.

If a message is sent while typing is triggered, the user will stop typing for a brief period of time and then resume again until either Typing::stop is called or the struct is dropped.

This should rarely be used for bots, although it is a good indicator that a long-running command is still being processed.

Examples

// Initiate typing (assuming http is `Arc<Http>`)
let typing = http.start_typing(7)?;

// Run some long-running process
long_process();

// Stop typing
typing.stop();

pub async fn unpin_message(
&self,
channel_id: u64,
message_id: u64
) -> Result<()>

Unpins a message from a channel.

pub async fn fire<T: DeserializeOwned>(&self, req: Request<'_>) -> Result<T>

Fires off a request, deserializing the response reader via the given type bound.

If you don’t need to deserialize the response and want the response instance itself, use Self::request.

Examples

Create a new message via the RouteInfo::CreateMessage endpoint and deserialize the response into a Message:

use serenity::{
    http::{
        routing::RouteInfo,
        request::RequestBuilder,
    },
    model::channel::Message,
};

let bytes = vec![
    // payload bytes here
];
let channel_id = 381880193700069377;
let route_info = RouteInfo::CreateMessage { channel_id };

let mut request = RequestBuilder::new(route_info);
request.body(Some(&bytes));

let message = http.fire::<Message>(request.build()).await?;

println!("Message content: {}", message.content);

Errors

If there is an error, it will be either Error::Http or Error::Json.

pub async fn request(&self, req: Request<'_>) -> Result<ReqwestResponse>

Performs a request, ratelimiting it if necessary.

Returns the raw reqwest Response. Use Self::fire to deserialize the response into some type.

Examples

Send a body of bytes over the RouteInfo::CreateMessage endpoint:

use serenity::http::{
    request::RequestBuilder,
    routing::RouteInfo,
};

let bytes = vec![
    // payload bytes here
];
let channel_id = 381880193700069377;
let route_info = RouteInfo::CreateMessage { channel_id };

let mut request = RequestBuilder::new(route_info);
request.body(Some(&bytes));

let response = http.request(request.build()).await?;

println!("Response successful?: {}", response.status().is_success());