Enum serenity::http::routing::Route

#[non_exhaustive]
pub enum Route {

    ChannelsId(u64),
    ChannelsIdInvites(u64),
    ChannelsIdMessages(u64),
    ChannelsIdMessagesBulkDelete(u64),
    ChannelsIdMessagesId(LightMethod, u64),
    ChannelsIdMessagesIdAck(u64),
    ChannelsIdMessagesIdReactions(u64),
    ChannelsIdMessagesIdReactionsUserIdType(u64),
    ChannelsIdPermissionsOverwriteId(u64),
    ChannelsIdPins(u64),
    ChannelsIdPinsMessageId(u64),
    ChannelsIdCrosspostsMessageId(u64),
    ChannelsIdTyping(u64),
    ChannelsIdWebhooks(u64),
    Gateway,
    GatewayBot,
    Guilds,
    GuildsId(u64),
    GuildsIdBans(u64),
    GuildsIdAuditLogs(u64),
    GuildsIdBansUserId(u64),
    GuildsIdChannels(u64),
    GuildsIdWidget(u64),
    GuildsIdPreview(u64),
    GuildsIdEmojis(u64),
    GuildsIdEmojisId(u64),
    GuildsIdIntegrations(u64),
    GuildsIdIntegrationsId(u64),
    GuildsIdIntegrationsIdSync(u64),
    GuildsIdInvites(u64),
    GuildsIdMembers(u64),
    GuildsIdMembersId(u64),
    GuildsIdMembersIdRolesId(u64),
    GuildsIdMembersMeNick(u64),
    GuildsIdPrune(u64),
    GuildsIdRegions(u64),
    GuildsIdRoles(u64),
    GuildsIdRolesId(u64),
    GuildsIdVanityUrl(u64),
    GuildsIdVoiceStates(u64),
    GuildsIdVoiceStatesMe(u64),
    GuildsIdWebhooks(u64),
    GuildsIdWelcomeScreen(u64),
    InvitesCode,
    UsersId,
    UsersMe,
    UsersMeChannels,
    UsersMeGuilds,
    UsersMeGuildsId,
    VoiceRegions,
    WebhooksId(u64),
    WebhooksIdMessagesId(u64),
    WebhooksApplicationId(u64),
    InteractionsId(u64),
    ApplicationsIdCommands(u64),
    ApplicationsIdCommandsId(u64),
    ApplicationsIdGuildsIdCommands(u64),
    ApplicationsIdGuildsIdCommandsPermissions(u64),
    ApplicationsIdGuildsIdCommandIdPermissions(u64),
    ApplicationsIdGuildsIdCommandsId(u64),
    None,

}

A representation of all routes registered within the library. These are safe and memory-efficient representations of each path that functions exist for in the http module.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ChannelsId(u64)

Route for the /channels/:channel_id path.

The data is the relevant ChannelId.

ChannelsIdInvites(u64)

Route for the /channels/:channel_id/invites path.

The data is the relevant ChannelId.

ChannelsIdMessages(u64)

Route for the /channels/:channel_id/messages path.

The data is the relevant ChannelId.

ChannelsIdMessagesBulkDelete(u64)

Route for the /channels/:channel_id/messages/bulk-delete path.

The data is the relevant ChannelId.

ChannelsIdMessagesId(LightMethod, u64)

Route for the /channels/:channel_id/messages/:message_id path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdAck(u64)

Route for the /channels/:channel_id/messages/:message_id/ack path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdReactions(u64)

Route for the /channels/:channel_id/messages/:message_id/reactions path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdReactionsUserIdType(u64)

Route for the /channels/:channel_id/messages/:message_id/reactions/:reaction/@me path.

The data is the relevant ChannelId.

ChannelsIdPermissionsOverwriteId(u64)

Route for the /channels/:channel_id/permissions/:target_id path.

The data is the relevant ChannelId.

ChannelsIdPins(u64)

Route for the /channels/:channel_id/pins path.

The data is the relevant ChannelId.

ChannelsIdPinsMessageId(u64)

Route for the /channels/:channel_id/pins/:message_id path.

The data is the relevant ChannelId.

ChannelsIdCrosspostsMessageId(u64)

Route for the /channels/:channel_id/message/:message_id/crosspost path.

The data is the relevant ChannelId.

ChannelsIdTyping(u64)

Route for the /channels/:channel_id/typing path.

The data is the relevant ChannelId.

ChannelsIdWebhooks(u64)

Route for the /channels/:channel_id/webhooks path.

The data is the relevant ChannelId.

Gateway

Route for the /gateway path.

GatewayBot

Route for the /gateway/bot path.

Guilds

Route for the /guilds path.

GuildsId(u64)

Route for the /guilds/:guild_id path.

The data is the relevant GuildId.

GuildsIdBans(u64)

Route for the /guilds/:guild_id/bans path.

The data is the relevant GuildId.

GuildsIdAuditLogs(u64)

Route for the /guilds/:guild_id/audit-logs path. The data is the relevant GuildId.

GuildsIdBansUserId(u64)

Route for the /guilds/:guild_id/bans/:user_id path.

The data is the relevant GuildId.

GuildsIdChannels(u64)

Route for the /guilds/:guild_id/channels/:channel_id path.

The data is the relevant GuildId.

GuildsIdWidget(u64)

Route for the /guilds/:guild_id/widget path.

The data is the relevant GuildId.

GuildsIdPreview(u64)

Route for the /guilds/:guild_id/preview path.

The data is the relevant GuildPreview.

GuildsIdEmojis(u64)

Route for the /guilds/:guild_id/emojis path.

The data is the relevant GuildId.

GuildsIdEmojisId(u64)

Route for the /guilds/:guild_id/emojis/:emoji_id path.

The data is the relevant GuildId.

GuildsIdIntegrations(u64)

Route for the /guilds/:guild_id/integrations path.

The data is the relevant GuildId.

GuildsIdIntegrationsId(u64)

Route for the /guilds/:guild_id/integrations/:integration_id path.

The data is the relevant GuildId.

GuildsIdIntegrationsIdSync(u64)

Route for the /guilds/:guild_id/integrations/:integration_id/sync path.

The data is the relevant GuildId.

GuildsIdInvites(u64)

Route for the /guilds/:guild_id/invites path.

The data is the relevant GuildId.

GuildsIdMembers(u64)

Route for the /guilds/:guild_id/members path.

The data is the relevant GuildId.

GuildsIdMembersId(u64)

Route for the /guilds/:guild_id/members/:user_id path.

The data is the relevant GuildId.

GuildsIdMembersIdRolesId(u64)

Route for the /guilds/:guild_id/members/:user_id/roles/:role_id path.

The data is the relevant GuildId.

GuildsIdMembersMeNick(u64)

Route for the /guilds/:guild_id/members/@me/nick path.

The data is the relevant GuildId.

GuildsIdPrune(u64)

Route for the /guilds/:guild_id/prune path.

The data is the relevant GuildId.

GuildsIdRegions(u64)

Route for the /guilds/:guild_id/regions path.

The data is the relevant GuildId.

GuildsIdRoles(u64)

Route for the /guilds/:guild_id/roles path.

The data is the relevant GuildId.

GuildsIdRolesId(u64)

Route for the /guilds/:guild_id/roles/:role_id path.

The data is the relevant GuildId.

GuildsIdVanityUrl(u64)

Route for the /guilds/:guild_id/vanity-url path.

The data is the relevant GuildId.

GuildsIdVoiceStates(u64)

Route for the /guilds/:guild_id/voice-states/:user_id path.

The data is the relevant GuildId.

GuildsIdVoiceStatesMe(u64)

Route for the /guilds/:guild_id/voice-states/@me path.

The data is the relevant GuildId.

GuildsIdWebhooks(u64)

Route for the /guilds/:guild_id/webhooks path.

The data is the relevant GuildId.

GuildsIdWelcomeScreen(u64)

Route for the /guilds/:guild_id/welcome-screen path.

The data is the relevant GuildId.

InvitesCode

Route for the /invites/:code path.

UsersId

Route for the /users/:user_id path.

UsersMe

Route for the /users/@me path.

UsersMeChannels

Route for the /users/@me/channels path.

UsersMeGuilds

Route for the /users/@me/guilds path.

UsersMeGuildsId

Route for the /users/@me/guilds/:guild_id path.

VoiceRegions

Route for the /voice/regions path.

WebhooksId(u64)

Route for the /webhooks/:webhook_id path.

WebhooksIdMessagesId(u64)

Route for the /webhooks/:webhook_id/:token/messages/:message_id path.

The data is the relevant WebhookId.

WebhooksApplicationId(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /webhooks/:application_id path.

The data is the relevant ApplicationId.

InteractionsId(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /interactions/:interaction_id path.

The data is the relevant InteractionId.

ApplicationsIdCommands(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /applications/:application_id path.

The data is the relevant ApplicationId.

ApplicationsIdCommandsId(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /applications/:application_id/commands/:command_id path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommands(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /applications/:application_id/guilds/:guild_id path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommandsPermissions(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /applications/:application_id/guilds/:guild_id/commands/permissions path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommandIdPermissions(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /applications/:application_id/guilds/:guild_id/commands/:command_id/permissions path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommandsId(u64)

This is supported on crate feature unstable_discord_api only.

Route for the /applications/:application_id/guilds/:guild_id path.

The data is the relevant ApplicationId.

None

Route where no ratelimit headers are in place (i.e. user account-only routes).

This is a special case, in that if the route is None then pre- and post-hooks are not executed.

Implementations

impl Route

pub fn channel(channel_id: u64) -> String

pub fn channel_invites(channel_id: u64) -> String

pub fn channel_message(channel_id: u64, message_id: u64) -> String

pub fn channel_message_crosspost(channel_id: u64, message_id: u64) -> String

pub fn channel_message_reaction<D, T>(
channel_id: u64,
message_id: u64,
user_id: D,
reaction_type: T
) -> String where
D: Display,
T: Display,

pub fn channel_message_reaction_emoji<T>(
channel_id: u64,
message_id: u64,
reaction_type: T
) -> String where
T: Display,

pub fn channel_message_reactions(channel_id: u64, message_id: u64) -> String

pub fn channel_message_reactions_list(
channel_id: u64,
message_id: u64,
reaction: &str,
limit: u8,
after: Option<u64>
) -> String

pub fn channel_messages(channel_id: u64, query: Option<&str>) -> String

pub fn channel_messages_bulk_delete(channel_id: u64) -> String

pub fn channel_permission(channel_id: u64, target_id: u64) -> String

pub fn channel_pin(channel_id: u64, message_id: u64) -> String

pub fn channel_pins(channel_id: u64) -> String

pub fn channel_typing(channel_id: u64) -> String

pub fn channel_webhooks(channel_id: u64) -> String

pub fn gateway() -> &'static str

pub fn gateway_bot() -> &'static str

pub fn guild(guild_id: u64) -> String

pub fn guild_with_counts(guild_id: u64) -> String

pub fn guild_audit_logs(
guild_id: u64,
action_type: Option<u8>,
user_id: Option<u64>,
before: Option<u64>,
limit: Option<u8>
) -> String

pub fn guild_ban(guild_id: u64, user_id: u64) -> String

pub fn guild_ban_optioned(
guild_id: u64,
user_id: u64,
delete_message_days: u8,
reason: &str
) -> String

pub fn guild_kick_optioned(guild_id: u64, user_id: u64, reason: &str) -> String

pub fn guild_bans(guild_id: u64) -> String

pub fn guild_channels(guild_id: u64) -> String

pub fn guild_widget(guild_id: u64) -> String

pub fn guild_preview(guild_id: u64) -> String

pub fn guild_emojis(guild_id: u64) -> String

pub fn guild_emoji(guild_id: u64, emoji_id: u64) -> String

pub fn guild_integration(guild_id: u64, integration_id: u64) -> String

pub fn guild_integration_sync(guild_id: u64, integration_id: u64) -> String

pub fn guild_integrations(guild_id: u64) -> String

pub fn guild_invites(guild_id: u64) -> String

pub fn guild_member(guild_id: u64, user_id: u64) -> String

pub fn guild_member_role(guild_id: u64, user_id: u64, role_id: u64) -> String

pub fn guild_members(guild_id: u64) -> String

pub fn guild_members_optioned(
guild_id: u64,
after: Option<u64>,
limit: Option<u64>
) -> String

pub fn guild_nickname(guild_id: u64) -> String

pub fn guild_prune(guild_id: u64, days: u64) -> String

pub fn guild_regions(guild_id: u64) -> String

pub fn guild_role(guild_id: u64, role_id: u64) -> String

pub fn guild_roles(guild_id: u64) -> String

pub fn guild_vanity_url(guild_id: u64) -> String

pub fn guild_voice_states(guild_id: u64, user_id: u64) -> String

pub fn guild_voice_states_me(guild_id: u64) -> String

pub fn guild_webhooks(guild_id: u64) -> String

pub fn guild_welcome_screen(guild_id: u64) -> String

pub fn guilds() -> &'static str

pub fn invite(code: &str) -> String

pub fn invite_optioned(code: &str, stats: bool) -> String

pub fn oauth2_application_current() -> &'static str

pub fn private_channel() -> &'static str

pub fn status_incidents_unresolved() -> &'static str

pub fn status_maintenances_active() -> &'static str

pub fn status_maintenances_upcoming() -> &'static str

pub fn user<D: Display>(target: D) -> String

pub fn user_dm_channels<D: Display>(target: D) -> String

pub fn user_guild<D: Display>(target: D, guild_id: u64) -> String

pub fn user_guilds<D: Display>(target: D) -> String

pub fn user_guilds_optioned<D: Display>(
target: D,
after: Option<u64>,
before: Option<u64>,
limit: u64
) -> String

pub fn voice_regions() -> &'static str

pub fn webhook(webhook_id: u64) -> String

pub fn webhook_with_token<D>(webhook_id: u64, token: D) -> String where
D: Display,

pub fn webhook_with_token_optioned<D>(
webhook_id: u64,
token: D,
wait: bool
) -> String where
D: Display,

pub fn webhook_message<D>(webhook_id: u64, token: D, message_id: u64) -> String where
D: Display,

pub fn webhook_original_interaction_response<D: Display>(
application_id: u64,
token: D
) -> String

This is supported on crate feature unstable_discord_api only.

pub fn webhook_followup_message<D: Display>(
application_id: u64,
token: D,
message_id: u64
) -> String

This is supported on crate feature unstable_discord_api only.

pub fn webhook_followup_messages<D: Display>(
application_id: u64,
token: D
) -> String

This is supported on crate feature unstable_discord_api only.

pub fn interaction_response<D: Display>(application_id: u64, token: D) -> String

This is supported on crate feature unstable_discord_api only.

pub fn application_command(application_id: u64, command_id: u64) -> String

This is supported on crate feature unstable_discord_api only.

pub fn application_commands(application_id: u64) -> String

This is supported on crate feature unstable_discord_api only.

pub fn application_guild_command(
application_id: u64,
guild_id: u64,
command_id: u64
) -> String

This is supported on crate feature unstable_discord_api only.

pub fn application_guild_command_permissions(
application_id: u64,
guild_id: u64,
command_id: u64
) -> String

This is supported on crate feature unstable_discord_api only.

pub fn application_guild_commands(application_id: u64, guild_id: u64) -> String

This is supported on crate feature unstable_discord_api only.

pub fn application_guild_commands_permissions(
application_id: u64,
guild_id: u64
) -> String

This is supported on crate feature unstable_discord_api only.