Enum serenity::http::routing::Route

#[non_exhaustive]
pub enum Route {

    ChannelsId(u64),
    ChannelsIdInvites(u64),
    ChannelsIdMessages(u64),
    ChannelsIdMessagesBulkDelete(u64),
    ChannelsIdMessagesId(LightMethod, u64),
    ChannelsIdMessagesIdAck(u64),
    ChannelsIdMessagesIdReactions(u64),
    ChannelsIdMessagesIdReactionsUserIdType(u64),
    ChannelsIdPermissionsOverwriteId(u64),
    ChannelsIdPins(u64),
    ChannelsIdPinsMessageId(u64),
    ChannelsIdCrosspostsMessageId(u64),
    ChannelsIdTyping(u64),
    ChannelsIdWebhooks(u64),
    ChannelsIdMessagesIdThreads(u64),
    ChannelsIdThreads(u64),
    ChannelsIdThreadMembersMe(u64),
    ChannelsIdThreadMembersUserId(u64),
    ChannelsIdThreadMembers(u64),
    ChannelsIdArchivedPublicThreads(u64),
    ChannelsIdArchivedPrivateThreads(u64),
    ChannelsIdMeJoindedArchivedPrivateThreads(u64),
    FollowNewsChannel(u64),
    Gateway,
    GatewayBot,
    Guilds,
    GuildsId(u64),
    GuildsIdAutoModRules(u64),
    GuildsIdAutoModRulesId(u64),
    GuildsIdBans(u64),
    GuildsIdAuditLogs(u64),
    GuildsIdBansUserId(u64),
    GuildsIdChannels(u64),
    GuildsIdWidget(u64),
    GuildsIdPreview(u64),
    GuildsIdEmojis(u64),
    GuildsIdEmojisId(u64),
    GuildsIdIntegrations(u64),
    GuildsIdIntegrationsId(u64),
    GuildsIdIntegrationsIdSync(u64),
    GuildsIdInvites(u64),
    GuildsIdMembers(u64),
    GuildsIdMembersId(u64),
    GuildsIdMembersIdRolesId(u64),
    GuildsIdMembersMe(u64),
    GuildsIdMembersMeNick(u64),
    GuildsIdMembersSearch(u64),
    GuildsIdPrune(u64),
    GuildsIdRegions(u64),
    GuildsIdRoles(u64),
    GuildsIdRolesId(u64),
    GuildsIdScheduledEvents(u64),
    GuildsIdScheduledEventsId(u64),
    GuildsIdScheduledEventsIdUsers(u64),
    GuildsIdStickers(u64),
    GuildsIdStickersId(u64),
    GuildsIdVanityUrl(u64),
    GuildsIdVoiceStates(u64),
    GuildsIdVoiceStatesMe(u64),
    GuildsIdWebhooks(u64),
    GuildsIdWelcomeScreen(u64),
    GuildsIdThreadsActive,
    InvitesCode,
    StickerPacks,
    StickersId,
    UsersId,
    UsersMe,
    UsersMeChannels,
    UsersMeConnections,
    UsersMeGuilds,
    UsersMeGuildsId,
    VoiceRegions,
    WebhooksId(u64),
    WebhooksIdMessagesId(u64),
    WebhooksApplicationId(u64),
    InteractionsId(u64),
    ApplicationsIdCommands(u64),
    ApplicationsIdCommandsId(u64),
    ApplicationsIdGuildsIdCommands(u64),
    ApplicationsIdGuildsIdCommandsPermissions(u64),
    ApplicationsIdGuildsIdCommandIdPermissions(u64),
    ApplicationsIdGuildsIdCommandsId(u64),
    StageInstances,
    StageInstancesChannelId(u64),
    None,

}

A representation of all routes registered within the library. These are safe and memory-efficient representations of each path that functions exist for in the http module.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ChannelsId(u64)

Route for the /channels/:channel_id path.

The data is the relevant ChannelId.

ChannelsIdInvites(u64)

Route for the /channels/:channel_id/invites path.

The data is the relevant ChannelId.

ChannelsIdMessages(u64)

Route for the /channels/:channel_id/messages path.

The data is the relevant ChannelId.

ChannelsIdMessagesBulkDelete(u64)

Route for the /channels/:channel_id/messages/bulk-delete path.

The data is the relevant ChannelId.

ChannelsIdMessagesId(LightMethod, u64)

Route for the /channels/:channel_id/messages/:message_id path.

The data is the relevant ChannelId.

This route is a unique case. The ratelimit for message deletions is different than the overall route ratelimit.

Refer to the docs on Rate Limits in the yellow warning section.

ChannelsIdMessagesIdAck(u64)

Route for the /channels/:channel_id/messages/:message_id/ack path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdReactions(u64)

Route for the /channels/:channel_id/messages/:message_id/reactions path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdReactionsUserIdType(u64)

Route for the /channels/:channel_id/messages/:message_id/reactions/:reaction/@me path.

The data is the relevant ChannelId.

ChannelsIdPermissionsOverwriteId(u64)

Route for the /channels/:channel_id/permissions/:target_id path.

The data is the relevant ChannelId.

ChannelsIdPins(u64)

Route for the /channels/:channel_id/pins path.

The data is the relevant ChannelId.

ChannelsIdPinsMessageId(u64)

Route for the /channels/:channel_id/pins/:message_id path.

The data is the relevant ChannelId.

ChannelsIdCrosspostsMessageId(u64)

Route for the /channels/:channel_id/message/:message_id/crosspost path.

The data is the relevant ChannelId.

ChannelsIdTyping(u64)

Route for the /channels/:channel_id/typing path.

The data is the relevant ChannelId.

ChannelsIdWebhooks(u64)

Route for the /channels/:channel_id/webhooks path.

The data is the relevant ChannelId.

ChannelsIdMessagesIdThreads(u64)

Route for the /channels/:channel_id/messages/:message_id/threads path.

The data is the relevant ChannelId.

ChannelsIdThreads(u64)

Route for the /channels/:channel_id/threads path.

The data is the relevant ChannelId.

ChannelsIdThreadMembersMe(u64)

Route for the /channels/:channel_id/thread-members/@me path.

The data is the relevant ChannelId.

ChannelsIdThreadMembersUserId(u64)

Route for the /channels/:channel_id/thread-members/:user_id path.

The data is the relevant ChannelId.

ChannelsIdThreadMembers(u64)

Route for the /channels/channel_id/thread-members path.

The data is the relevant ChannelId.

ChannelsIdArchivedPublicThreads(u64)

Route for the /channels/:channel_id/threads/archived/public path.

The data is the relevant ChannelId.

ChannelsIdArchivedPrivateThreads(u64)

Route for the /channels/:channel_id/threads/archived/private path.

The data is the relevant ChannelId.

ChannelsIdMeJoindedArchivedPrivateThreads(u64)

Route for the /channels/:channel_id/users/@me/threads/archived/private path.

The data is the relevant ChannelId.

FollowNewsChannel(u64)

Route for the /channels/{channel.id}/followers path.

The data is the relevant ChannelId.

Gateway

Route for the /gateway path.

GatewayBot

Route for the /gateway/bot path.

Guilds

Route for the /guilds path.

GuildsId(u64)

Route for the /guilds/:guild_id path.

The data is the relevant GuildId.

GuildsIdAutoModRules(u64)

Route for the /guilds/:guild_id/auto-moderation/rules path.

GuildsIdAutoModRulesId(u64)

Route for the /guilds/:guild_id/auto-moderation/rules/:rule_id path.

GuildsIdBans(u64)

Route for the /guilds/:guild_id/bans path.

The data is the relevant GuildId.

GuildsIdAuditLogs(u64)

Route for the /guilds/:guild_id/audit-logs path. The data is the relevant GuildId.

GuildsIdBansUserId(u64)

Route for the /guilds/:guild_id/bans/:user_id path.

The data is the relevant GuildId.

GuildsIdChannels(u64)

Route for the /guilds/:guild_id/channels/:channel_id path.

The data is the relevant GuildId.

GuildsIdWidget(u64)

Route for the /guilds/:guild_id/widget path.

The data is the relevant GuildId.

GuildsIdPreview(u64)

Route for the /guilds/:guild_id/preview path.

The data is the relevant GuildPreview.

GuildsIdEmojis(u64)

Route for the /guilds/:guild_id/emojis path.

The data is the relevant GuildId.

GuildsIdEmojisId(u64)

Route for the /guilds/:guild_id/emojis/:emoji_id path.

The data is the relevant GuildId.

GuildsIdIntegrations(u64)

Route for the /guilds/:guild_id/integrations path.

The data is the relevant GuildId.

GuildsIdIntegrationsId(u64)

Route for the /guilds/:guild_id/integrations/:integration_id path.

The data is the relevant GuildId.

GuildsIdIntegrationsIdSync(u64)

Route for the /guilds/:guild_id/integrations/:integration_id/sync path.

The data is the relevant GuildId.

GuildsIdInvites(u64)

Route for the /guilds/:guild_id/invites path.

The data is the relevant GuildId.

GuildsIdMembers(u64)

Route for the /guilds/:guild_id/members path.

The data is the relevant GuildId.

GuildsIdMembersId(u64)

Route for the /guilds/:guild_id/members/:user_id path.

The data is the relevant GuildId.

GuildsIdMembersIdRolesId(u64)

Route for the /guilds/:guild_id/members/:user_id/roles/:role_id path.

The data is the relevant GuildId.

GuildsIdMembersMe(u64)

Route for the /guilds/:guild_id/members/@me path.

The data is the relevant GuildId.

GuildsIdMembersMeNick(u64)

Route for the /guilds/:guild_id/members/@me/nick path.

The data is the relevant GuildId.

GuildsIdMembersSearch(u64)

Route for the /guilds/:guild_id/members/search path.

The data is the relevant GuildId.

GuildsIdPrune(u64)

Route for the /guilds/:guild_id/prune path.

The data is the relevant GuildId.

GuildsIdRegions(u64)

Route for the /guilds/:guild_id/regions path.

The data is the relevant GuildId.

GuildsIdRoles(u64)

Route for the /guilds/:guild_id/roles path.

The data is the relevant GuildId.

GuildsIdRolesId(u64)

Route for the /guilds/:guild_id/roles/:role_id path.

The data is the relevant GuildId.

GuildsIdScheduledEvents(u64)

Route for the /guilds/:guild_id/scheduled-events path.

The data is the relevant GuildId.

GuildsIdScheduledEventsId(u64)

Route for the /guilds/:guild_id/scheduled-events/:event_id path.

The data is the relevant GuildId.

GuildsIdScheduledEventsIdUsers(u64)

Route for the /guilds/:guild_id/scheduled-events/:event_id/users path.

The data is the relevant GuildId.

GuildsIdStickers(u64)

Route for the /guilds/:guild_id/stickers path.

The data is the relevant GuildId.

GuildsIdStickersId(u64)

Route for the /guilds/:guild_id/stickers/:sticker_id path.

The data is the relevant GuildId.

GuildsIdVanityUrl(u64)

Route for the /guilds/:guild_id/vanity-url path.

The data is the relevant GuildId.

GuildsIdVoiceStates(u64)

Route for the /guilds/:guild_id/voice-states/:user_id path.

The data is the relevant GuildId.

GuildsIdVoiceStatesMe(u64)

Route for the /guilds/:guild_id/voice-states/@me path.

The data is the relevant GuildId.

GuildsIdWebhooks(u64)

Route for the /guilds/:guild_id/webhooks path.

The data is the relevant GuildId.

GuildsIdWelcomeScreen(u64)

Route for the /guilds/:guild_id/welcome-screen path.

The data is the relevant GuildId.

GuildsIdThreadsActive

Route for the /guilds/:guild_id/threads/active path.

The data is the relevant GuildId.

InvitesCode

Route for the /invites/:code path.

StickerPacks

Route for the /sticker-packs path.

StickersId

Route for the /stickers/:sticker_id path.

UsersId

Route for the /users/:user_id path.

UsersMe

Route for the /users/@me path.

UsersMeChannels

Route for the /users/@me/channels path.

UsersMeConnections

Route for the /users/@me/connections path.

UsersMeGuilds

Route for the /users/@me/guilds path.

UsersMeGuildsId

Route for the /users/@me/guilds/:guild_id path.

VoiceRegions

Route for the /voice/regions path.

WebhooksId(u64)

Route for the /webhooks/:webhook_id path.

WebhooksIdMessagesId(u64)

Route for the /webhooks/:webhook_id/:token/messages/:message_id path.

The data is the relevant WebhookId.

WebhooksApplicationId(u64)

Route for the /webhooks/:application_id path.

The data is the relevant ApplicationId.

InteractionsId(u64)

Route for the /interactions/:interaction_id path.

The data is the relevant InteractionId.

ApplicationsIdCommands(u64)

Route for the /applications/:application_id path.

The data is the relevant ApplicationId.

ApplicationsIdCommandsId(u64)

Route for the /applications/:application_id/commands/:command_id path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommands(u64)

Route for the /applications/:application_id/guilds/:guild_id path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommandsPermissions(u64)

Route for the /applications/:application_id/guilds/:guild_id/commands/permissions path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommandIdPermissions(u64)

Route for the /applications/:application_id/guilds/:guild_id/commands/:command_id/permissions path.

The data is the relevant ApplicationId.

ApplicationsIdGuildsIdCommandsId(u64)

Route for the /applications/:application_id/guilds/:guild_id path.

The data is the relevant ApplicationId.

StageInstances

Route for the /stage-instances path.

The data is the relevant ChannelId.

StageInstancesChannelId(u64)

Route for the /stage-instances/:channel_id path.

The data is the relevant ChannelId.

None

Route where no ratelimit headers are in place (i.e. user account-only routes).

This is a special case, in that if the route is None then pre- and post-hooks are not executed.

Implementations

impl Route

pub fn channel(channel_id: u64) -> String

pub fn channel_invites(channel_id: u64) -> String

pub fn channel_message(channel_id: u64, message_id: u64) -> String

pub fn channel_message_crosspost(channel_id: u64, message_id: u64) -> String

pub fn channel_message_reaction<D, T>( channel_id: u64, message_id: u64, user_id: D, reaction_type: T ) -> Stringwhere D: Display, T: Display,

pub fn channel_message_reaction_emoji<T>( channel_id: u64, message_id: u64, reaction_type: T ) -> Stringwhere T: Display,

pub fn channel_message_reactions(channel_id: u64, message_id: u64) -> String

pub fn channel_message_reactions_list( channel_id: u64, message_id: u64, reaction: &str, limit: u8, after: Option<u64> ) -> String

pub fn channel_messages(channel_id: u64, query: Option<&str>) -> String

pub fn channel_messages_bulk_delete(channel_id: u64) -> String

pub fn channel_follow_news(channel_id: u64) -> String

pub fn channel_permission(channel_id: u64, target_id: u64) -> String

pub fn channel_pin(channel_id: u64, message_id: u64) -> String

pub fn channel_pins(channel_id: u64) -> String

pub fn channel_typing(channel_id: u64) -> String

pub fn channel_webhooks(channel_id: u64) -> String

pub fn channel_public_threads(channel_id: u64, message_id: u64) -> String

pub fn channel_private_threads(channel_id: u64) -> String

pub fn channel_thread_member(channel_id: u64, user_id: u64) -> String

pub fn channel_thread_member_me(channel_id: u64) -> String

pub fn channel_thread_members(channel_id: u64) -> String

pub fn channel_archived_public_threads( channel_id: u64, before: Option<u64>, limit: Option<u64> ) -> String

pub fn channel_archived_private_threads( channel_id: u64, before: Option<u64>, limit: Option<u64> ) -> String

pub fn channel_joined_private_threads( channel_id: u64, before: Option<u64>, limit: Option<u64> ) -> String

pub fn gateway() -> &'static str

pub fn gateway_bot() -> &'static str

pub fn guild(guild_id: u64) -> String

pub fn guild_with_counts(guild_id: u64) -> String

pub fn guild_audit_logs( guild_id: u64, action_type: Option<u8>, user_id: Option<u64>, before: Option<u64>, limit: Option<u8> ) -> String

pub fn guild_automod_rule(guild_id: u64, rule_id: u64) -> String

pub fn guild_automod_rules(guild_id: u64) -> String

pub fn guild_ban(guild_id: u64, user_id: u64) -> String

pub fn guild_ban_optioned( guild_id: u64, user_id: u64, delete_message_days: u8 ) -> String

pub fn guild_kick_optioned(guild_id: u64, user_id: u64) -> String

pub fn guild_bans(guild_id: u64) -> String

pub fn guild_channels(guild_id: u64) -> String

pub fn guild_widget(guild_id: u64) -> String

pub fn guild_preview(guild_id: u64) -> String

pub fn guild_emojis(guild_id: u64) -> String

pub fn guild_emoji(guild_id: u64, emoji_id: u64) -> String

pub fn guild_integration(guild_id: u64, integration_id: u64) -> String

pub fn guild_integration_sync(guild_id: u64, integration_id: u64) -> String

pub fn guild_integrations(guild_id: u64) -> String

pub fn guild_invites(guild_id: u64) -> String

pub fn guild_member(guild_id: u64, user_id: u64) -> String

pub fn guild_member_role(guild_id: u64, user_id: u64, role_id: u64) -> String

pub fn guild_members(guild_id: u64) -> String

pub fn guild_members_search( guild_id: u64, query: &str, limit: Option<u64> ) -> String

pub fn guild_members_optioned( guild_id: u64, after: Option<u64>, limit: Option<u64> ) -> String

pub fn guild_member_me(guild_id: u64) -> String

pub fn guild_nickname(guild_id: u64) -> String

pub fn guild_prune(guild_id: u64, days: u64) -> String

pub fn guild_regions(guild_id: u64) -> String

pub fn guild_role(guild_id: u64, role_id: u64) -> String

pub fn guild_roles(guild_id: u64) -> String

pub fn guild_scheduled_event( guild_id: u64, event_id: u64, with_user_count: Option<bool> ) -> String

pub fn guild_scheduled_events( guild_id: u64, with_user_count: Option<bool> ) -> String

pub fn guild_scheduled_event_users( guild_id: u64, event_id: u64, after: Option<u64>, before: Option<u64>, limit: Option<u64>, with_member: Option<bool> ) -> String

pub fn guild_sticker(guild_id: u64, sticker_id: u64) -> String

pub fn guild_stickers(guild_id: u64) -> String

pub fn guild_vanity_url(guild_id: u64) -> String

pub fn guild_voice_states(guild_id: u64, user_id: u64) -> String

pub fn guild_voice_states_me(guild_id: u64) -> String

pub fn guild_webhooks(guild_id: u64) -> String

pub fn guild_welcome_screen(guild_id: u64) -> String

pub fn guild_threads_active(guild_id: u64) -> String

pub fn guilds() -> &'static str

pub fn invite(code: &str) -> String

pub fn invite_optioned( code: &str, member_counts: bool, expiration: bool, event_id: Option<u64> ) -> String

pub fn oauth2_application_current() -> &'static str

pub fn private_channel() -> &'static str

pub fn status_incidents_unresolved() -> &'static str

pub fn status_maintenances_active() -> &'static str

pub fn status_maintenances_upcoming() -> &'static str

pub fn sticker(sticker_id: u64) -> String

pub fn sticker_packs() -> &'static str

pub fn user<D: Display>(target: D) -> String

pub fn user_me_connections() -> &'static str

pub fn user_dm_channels<D: Display>(target: D) -> String

pub fn user_guild<D: Display>(target: D, guild_id: u64) -> String

pub fn user_guilds<D: Display>(target: D) -> String

pub fn user_guilds_optioned<D: Display>( target: D, after: Option<u64>, before: Option<u64>, limit: Option<u64> ) -> String

pub fn voice_regions() -> &'static str

pub fn webhook(webhook_id: u64) -> String

pub fn webhook_with_token<D>(webhook_id: u64, token: D) -> Stringwhere D: Display,

pub fn webhook_with_token_optioned<D>( webhook_id: u64, token: D, wait: bool ) -> Stringwhere D: Display,

pub fn webhook_message<D>(webhook_id: u64, token: D, message_id: u64) -> Stringwhere D: Display,

pub fn webhook_original_interaction_response<D: Display>( application_id: u64, token: D ) -> String

pub fn webhook_followup_message<D: Display>( application_id: u64, token: D, message_id: u64 ) -> String

pub fn webhook_followup_messages<D: Display>( application_id: u64, token: D ) -> String

pub fn interaction_response<D: Display>(application_id: u64, token: D) -> String

pub fn application_command(application_id: u64, command_id: u64) -> String

pub fn application_commands(application_id: u64) -> String

pub fn application_commands_optioned( application_id: u64, with_localizations: bool ) -> String

pub fn application_guild_command( application_id: u64, guild_id: u64, command_id: u64 ) -> String

pub fn application_guild_command_permissions( application_id: u64, guild_id: u64, command_id: u64 ) -> String

pub fn application_guild_commands(application_id: u64, guild_id: u64) -> String

pub fn application_guild_commands_optioned( application_id: u64, guild_id: u64, with_localizations: bool ) -> String

pub fn application_guild_commands_permissions( application_id: u64, guild_id: u64 ) -> String

pub fn stage_instances() -> &'static str

pub fn stage_instance(channel_id: u64) -> String