Module serenity::interactions_endpoint

Available on crate feature interactions_endpoint only.

Contains tools related to Discord’s Interactions Endpoint URL feature.

“You can optionally configure an interactions endpoint to receive interactions via HTTP POSTs rather than over Gateway with a bot user.”

https://discord.com/developers/docs/tutorials/upgrading-to-application-commands#adding-an-interactions-endpoint-url

See Verifier for example usage.

Structs

InvalidKey

The byte array couldn’t be parsed into a valid cryptographic public key.

Verifier

Used to cryptographically verify incoming interactions HTTP request for authenticity.