Module serenity::json

This module exports different types for JSON interactions. It encapsulates the differences between serde_json and simd-json to allow ignoring those in the rest of the codebase.

Macros

json

Construct a serde_json::Value from a JSON literal.

Structs

JsonError

This type represents all possible errors that can occur when serializing or deserializing JSON data.

Constants

NULLNon-simd_json

Functions

from_reader

Deserialize an instance of type T from bytes of JSON text.

from_slice

Deserialize an instance of type T from bytes of JSON text.

from_str

Deserialize an instance of type T from a string of JSON text.

from_value

Interpret a Value as an instance of type T.

hashmap_to_json_map

Converts a HashMap into a final JsonMap representation.

to_string

Serialize the given data structure as a String of JSON.

to_string_pretty

Serialize the given data structure as a pretty-printed String of JSON.

to_value

Convert a T into a Value which is an enum that can represent any valid JSON data.

to_vec

Serialize the given data structure as a JSON byte vector.

to_vec_pretty

Serialize the given data structure as a pretty-printed JSON byte vector.

Type Aliases

JsonMapNon-simd_json

ValueNon-simd_json