Module serenity::json

This module exports different types for JSON interactions. It encapsulates the differences between serde_json and simd-json to allow ignoring those in the rest of the codebase.

Modules

prelude

Macros

jsonNon-simd-json

Construct a serde_json::Value from a JSON literal.

Structs

JsonErrorNon-simd-json

This type represents all possible errors that can occur when serializing or deserializing JSON data.

Constants

NULLNon-simd-json

Traits

ToNumber

Functions

hashmap_to_json_map

Converts a HashMap into a final JsonMap representation.

Type Definitions

JsonMapNon-simd-json

ValueNon-simd-json