serenity::model

Module application

Models about OAuth2 applications.

Structs

ActionRow

An action row.

ApplicationFlags

The flags of the application.

AuthorizingIntegrationOwners

AutocompleteOption

The focused option for autocomplete interactions return by CommandData::autocomplete.

Button

A button component.

ButtonStyle

The style of a button.

Checkbox

An interactive component that allows users to use a checkbox in modals.

CheckboxGroup

An interactive component that allows users to group together multiple checkboxes in modals.

Command

The base command model that belongs to an application.

CommandData

The command data payload.

CommandDataOption

A set of a parameter and a value from the user.

CommandDataResolved

The resolved data of a command data interaction payload. It contains the objects of CommandDataOptions.

CommandGeneratedOriginal

The base command model that belongs to an application.

CommandInteraction

An interaction when a user invokes a slash command.

CommandOption

The parameters for an Command.

CommandOptionChoice

The only valid values a user can pick in an CommandOption.

CommandOptionGeneratedOriginal

The parameters for an Command.

CommandOptionType

The type of an CommandOption.

CommandPermission

The CommandPermission data.

CommandPermissionType

The type of a CommandPermission.

CommandPermissions

An Command permission.

CommandType

The type of an application command.

ComponentInteraction

An interaction triggered by a message component.

ComponentInteractionData

A message component interaction data, provided by ComponentInteraction::data

ComponentType

The type of a component

Container

A container component, similar to an embed but without all the functionality.

CurrentApplicationInfo

Information about the current application and its owner.

EntryPointHandlerType

Signifies how the invocation of a command of type PrimaryEntryPoint should be handled.

FileComponent

A file component, will not render a text preview to the user.

FileUpload

An interactive component that allows users to upload files in modals.

InputText

An input text component for modal interactions

InputTextStyle

The style of the input text

InstallParams

Settings for the application’s default in-app authorization link

InstallationContext

An enum representing the [installation contexts].

InstallationContextConfig

Information about how the CurrentApplicationInfo is installed.

InteractionContext

An enum representing the different [interaction contexts].

InteractionType

The type of an Interaction.

Label

A layout component that wraps modal components with a label and optional description.

MediaGallery

A Media Gallery is a component that allows you to display media attachments in an organized gallery format.

MediaGalleryItem

An individual media gallery item.

MembershipState

Discord docs.

MessageCommandInteractionMetadata

MessageComponentInteractionMetadata

MessageInteractionNon-unstable

Sent when a Message is a response to an Interaction.

MessageModalSubmitInteractionMetadata

ModalInteraction

An interaction triggered by a modal submit.

ModalInteractionData

A modal submit interaction data, provided by ModalInteraction::data

PartialCurrentApplicationInfo

Partial information about the given application.

PingInteraction

A ping interaction, which can only be received through an endpoint url.

RadioGroup

An interactive component that allows users to select a radio button in modals.

ResolvedOption

Section

A component that is a container for up to 3 text display components and an accessory.

SelectMenu

A select menu component.

SelectMenuOption

A select menu component options.

Separator

A component that adds vertical padding and visual division between other components.

SeparatorSpacingSize

The size of a separator component.

Team

Information about the Team group of the application.

TeamMember

Information about a Member on a Team.

TextDisplay

A component that allows you to add text to your message, similiar to the content field of a message.

Thumbnail

A section component’s thumbnail.

UnfurledMediaItem

A url or attachment.

Enums

ActionRowComponent

A component which can be inside of an ActionRow.

AuthorizingIntegrationOwner

A cleaned up enum for determining the authorizing owner for an Interaction.

ButtonKind

CommandDataOptionValue

The value of an CommandDataOption.

Component

Represents top-level Discord components, a part of messages that are usually interactable.

ComponentInteractionDataKind

ContainerComponent

A child component encapsulated within a container.

Interaction

Discord docs

LabelComponent

MessageInteractionMetadata

Metadata about the interaction, including the source of the interaction relevant server and user IDs.

ModalComponent

A component which can appear inside a modal form submission.

ResolvedTarget

The resolved value of a CommandData::target_id.

ResolvedValue

The resolved value of a CommandDataOption.

Scope

The available OAuth2 Scopes.

SectionAccessory

A component that is contextually associated to the content of a section.

SectionComponent

A child component representing the content of a section.

SelectMenuKind

TeamMemberRole

Unresolved

Option value variants that couldn’t be resolved by CommandData::options().