serenity::model

Module channel

Models relating to channels and types within channels.

Structs

Attachment

A file uploaded with a message. Not to be confused with Embeds.

AttachmentFlags

Flags for an attachment.

AutoArchiveDuration

See ThreadMetadata::auto_archive_duration.

BaseGuildChannel

Represents the shared fields between GuildChannel and GuildThread.

BaseInteractionChannel

Represents the shared fields between a InteractionChannel and a InteractionGuildThread.

BaseTheme

The mode of the shared client theme.

ChannelFlags

Describes extra features of the channel.

ChannelInfoChannel

Discord docs.

ChannelMention

Discord docs.

ChannelType

A representation of a type of channel.

CountDetails

A representation of reaction count details.

Embed

Represents a rich embed which allows using richer markdown, multiple fields and more. This was heavily inspired by slack’s attachments.

EmbedAuthor

An author object in an embed.

EmbedField

A field object in an embed.

EmbedFooter

Footer information for an embed.

EmbedImage

An image object in an embed.

EmbedMediaFlags

Flags for embed media.

EmbedProvider

The provider of an embed.

EmbedVideo

Video information for an embed.

FollowedChannel

A container for the IDs returned by following a news channel.

ForumLayoutType

See GuildChannel::default_forum_layout.

ForumTag

An object that represents a tag able to be applied to a thread in a GUILD_FORUM channel.

GuildChannel

Represents a channel in a Guild, excluding thread information.

GuildThread

InteractionChannel

Represents a partial channel from an interaction.

InteractionGuildThread

Represents a partial thread from an interaction.

Message

A representation of a message over a guild’s text channel, a group, or a private channel.

MessageActivity

Rich Presence activity information.

MessageActivityKind

Discord docs.

MessageApplication

Rich Presence application information.

MessageCall

Information about a call in a private channel.

MessageFlags

Describes extra features of the message.

MessageGeneratedOriginal

A representation of a message over a guild’s text channel, a group, or a private channel.

MessageLinkmodel

Uniquely identifies a message.

MessagePin

A pinned message returned as part of a paginated Get Channel Pins query.

MessagePinsPage

The response data for a paginated Get Channel Pins query.

MessageReaction

A representation of a reaction to a message.

MessageReference

Reference data sent with crossposted messages.

MessageReferenceKind

Message Reference Type information

MessageSnapshot

Discord docs

MessageType

Differentiates between regular and different types of system messages.

MessagesItermodel

A helper class returned by GenericChannelId::messages_iter

PartialGuildThread

A partial guild thread.

PartialThreadMember

PermissionOverwrite

A channel-specific permission overwrite for a member or role.

Poll

A poll that has been attached to a Message.

PollAnswer

A possible answer for a Poll.

PollAnswerCount

The count of a single PollAnswer’s results.

PollLayoutType

Represents the different layouts that a Poll may have.

PollMedia

A piece of data used in mutliple parts of the Poll structure.

PollResults

The model for the results of a Poll.

PrivateChannel

A Direct Message text channel with another user.

Reaction

An emoji reaction to a message.

ReactionConversionError

ReactionTypes

A list of types a reaction can be.

RoleSubscriptionData

Discord docs

SharedClientTheme

A shared client theme that is attached to Message.

SortOrder

The sort order for threads in a forum.

StageInstance

Discord docs.

StageInstancePrivacyLevel

See StageInstance::privacy_level.

ThreadMember

A model representing a user in a Guild Thread.

ThreadMemberFlags

Describes extra features of the message.

ThreadMetadata

A thread data.

ThreadMetadataGeneratedOriginal

A thread data.

ThreadsData

A response to getting several threads channels.

VideoQualityMode

The video quality mode for a voice channel.

Enums

Channel

A container for any channel.

ChannelInfoField

Discord docs.

ForumEmoji

An object that specifies the emoji to use for Forum related emoji parameters.

GenericGuildChannelRef

A container for a reference to any Guild channel.

GenericInteractionChannel

Nonce

PermissionOverwriteType

The type of edit being made to a Channel’s permissions.

PollMediaEmoji

The “Partial Emoji” attached to a PollMedia model.

ReactionType

The type of a Reaction sent.