Module serenity::model::channel

Models relating to channels and types within channels.

Structs

Attachment
A file uploaded with a message. Not to be confused with Embeds.

ChannelCategory
A category of GuildChannels.

ChannelMention
Channel Mention Object

Embed
Represents a rich embed which allows using richer markdown, multiple fields and more. This was heavily inspired by slack’s attachments.

EmbedAuthor
An author object in an embed.

EmbedField
A field object in an embed.

EmbedFooter
Footer information for an embed.

EmbedImage
An image object in an embed.

EmbedProvider
The provider of an embed.

EmbedThumbnail
The dimensions and URL of an embed thumbnail.

EmbedVideo
Video information for an embed.

GuildChannel
Represents a guild’s text, news, or voice channel. Some methods are available only for voice channels and some are only available for text channels. News channels are a subset of text channels and lack slow mode hence slow_mode_rate will be None.

Message
A representation of a message over a guild’s text channel, a group, or a private channel.

MessageActivity
Rich Presence activity information.

MessageApplication
Rich Presence application information.

MessageFlags
Describes extra features of the message.

MessageReaction
A representation of a reaction to a message.

MessageReference
Reference data sent with crossposted messages.

MessagesIter
A helper class returned by ChannelId::messages_iter

PermissionOverwrite
A channel-specific permission overwrite for a member or role.

PrivateChannel
A Direct Message text channel with another user.

Reaction
An emoji reaction to a message.

ReactionConversionError

Sticker
A sticker sent with a message.

Enums

Channel
A container for any channel.

ChannelType
A representation of a type of channel.

MessageActivityKind

MessageType
Differentiates between regular and different types of system messages.

NeverFails

PermissionOverwriteType
The type of edit being made to a Channel’s permissions.

ReactionType
The type of a Reaction sent.

StickerFormatType
Differentiates between sticker formats.

VideoQualityMode
The video quality mode for a voice channel.