Enum serenity::model::channel::ReactionType

#[non_exhaustive]
pub enum ReactionType {
    Custom {
        animated: bool,
        id: EmojiId,
        name: Option<String>,
    },
    Unicode(String),
}

The type of a Reaction sent.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Custom

Fields

animated: bool

Whether the emoji is animated.

id: EmojiId

The Id of the custom Emoji.

name: Option<String>

The name of the custom emoji. This is primarily used for decoration and distinguishing the emoji client-side.

A reaction with a Guilds custom Emoji, which is unique to the guild.

Unicode(String)

A reaction with a twemoji.

Implementations

impl ReactionType

pub fn as_data(&self) -> String

This is supported on crate feature model only.

Creates a data-esque display of the type. This is not very useful for displaying, as the primary client can not render it, but can be useful for debugging.

Note: This is mainly for use internally. There is otherwise most likely little use for it.

pub fn unicode_eq(&self, other: &str) -> bool

This is supported on crate feature model only.

Helper function to allow testing equality of unicode emojis without having to perform any allocation. Will always return false if the reaction was not a unicode reaction.

pub fn unicode_partial_cmp(&self, other: &str) -> Option<Ordering>

This is supported on crate feature model only.

Helper function to allow comparing unicode emojis without having to perform any allocation. Will return None if the reaction was not a unicode reaction.