serenity::model::channel
Struct GuildChannel

#[non_exhaustive]
pub struct GuildChannel {

    pub id: ChannelId,
    pub bitrate: Option<NonMaxU32>,
    pub parent_id: Option<ChannelId>,
    pub guild_id: GuildId,
    pub kind: ChannelType,
    pub owner_id: Option<UserId>,
    pub last_message_id: Option<MessageId>,
    pub last_pin_timestamp: Option<Timestamp>,
    pub name: FixedString<u16>,
    pub permission_overwrites: FixedArray<PermissionOverwrite>,
    pub position: u16,
    pub topic: Option<FixedString<u16>>,
    pub user_limit: Option<NonMaxU16>,
    pub nsfw: bool,
    pub rate_limit_per_user: Option<NonMaxU16>,
    pub rtc_region: Option<FixedString<u8>>,
    pub video_quality_mode: Option<VideoQualityMode>,
    pub message_count: Option<NonMaxU32>,
    pub member_count: Option<NonMaxU8>,
    pub thread_metadata: Option<ThreadMetadata>,
    pub member: Option<PartialThreadMember>,
    pub default_auto_archive_duration: Option<AutoArchiveDuration>,
    pub permissions: Option<Permissions>,
    pub flags: ChannelFlags,
    pub total_message_sent: Option<NonMaxU32>,
    pub available_tags: FixedArray<ForumTag>,
    pub applied_tags: FixedArray<ForumTagId>,
    pub default_reaction_emoji: Option<ForumEmoji>,
    pub default_thread_rate_limit_per_user: Option<NonMaxU16>,
    pub status: Option<FixedString<u16>>,
    pub default_sort_order: Option<SortOrder>,
    pub default_forum_layout: Option<ForumLayoutType>,

}

Represents a guild’s text, news, or voice channel.

Some methods are available only for voice channels and some are only available for text channels. News channels are a subset of text channels and lack slow mode hence Self::rate_limit_per_user will be None.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: ChannelId

The unique Id of the channel.

bitrate: Option<NonMaxU32>

The bitrate of the channel.

Note: This is only available for voice and stage channels.

parent_id: Option<ChannelId>

The Id of the parent category for a channel, or of the parent text channel for a thread.

Note: This is only available for channels in a category and thread channels.

guild_id: GuildId

The Id of the guild the channel is located in.

The original voice channel has an Id equal to the guild’s Id, incremented by one.

kind: ChannelType

The type of the channel.

owner_id: Option<UserId>

The Id of the user who created this channel

Note: This is only available for threads and forum posts

last_message_id: Option<MessageId>

The Id of the last message sent in the channel.

Note: This is only available for text channels.

last_pin_timestamp: Option<Timestamp>

The timestamp of the time a pin was most recently made.

Note: This is only available for text channels.

name: FixedString<u16>

The name of the channel. (1-100 characters)

permission_overwrites: FixedArray<PermissionOverwrite>

Permission overwrites for Members and for Roles.

position: u16

The position of the channel.

The default text channel will almost always have a position of 0.

topic: Option<FixedString<u16>>

The topic of the channel.

Note: This is only available for text, forum and stage channels.

user_limit: Option<NonMaxU16>

The maximum number of members allowed in the channel.

This is max 99 for voice channels and 10,000 for stage channels (0 refers to no limit).

nsfw: bool

Used to tell if the channel is not safe for work.

rate_limit_per_user: Option<NonMaxU16>

A rate limit that applies per user and excludes bots.

Note: This is only available for text channels excluding news channels.

rtc_region: Option<FixedString<u8>>

The region override.

Note: This is only available for voice and stage channels. None for voice and stage channels means automatic region selection.

video_quality_mode: Option<VideoQualityMode>

The video quality mode for a voice channel.

message_count: Option<NonMaxU32>

An approximate count of messages in the thread.

Note: This is only available on thread channels.

member_count: Option<NonMaxU8>

An approximate count of users in a thread, stops counting at 50.

Note: This is only available on thread channels.

thread_metadata: Option<ThreadMetadata>

The thread metadata.

Note: This is only available on thread channels.

member: Option<PartialThreadMember>

Thread member object for the current user, if they have joined the thread, only included on certain API endpoints.

default_auto_archive_duration: Option<AutoArchiveDuration>

Default duration for newly created threads, in minutes, to automatically archive the thread after recent activity.

permissions: Option<Permissions>

Computed permissions for the invoking user in the channel, including overwrites.

Only included inside CommandDataResolved.

flags: ChannelFlags

Extra information about the channel

Note: This is only available in forum channels.

total_message_sent: Option<NonMaxU32>

The number of messages ever sent in a thread, it’s similar to message_count on message creation, but will not decrement the number when a message is deleted.

available_tags: FixedArray<ForumTag>

The set of available tags.

Note: This is only available in forum channels.

applied_tags: FixedArray<ForumTagId>

The set of applied tags.

Note: This is only available in a thread in a forum.

default_reaction_emoji: Option<ForumEmoji>

The emoji to show in the add reaction button

Note: This is only available in a forum.

default_thread_rate_limit_per_user: Option<NonMaxU16>

The initial rate_limit_per_user to set on newly created threads in a channel. This field is copied to the thread at creation time and does not live update.

Note: This is only available in a forum or text channel.

status: Option<FixedString<u16>>

The status of a voice channel.

Note: This is only available in voice channels.

default_sort_order: Option<SortOrder>

The default sort order type used to order posts

Note: This is only available in a forum.

default_forum_layout: Option<ForumLayoutType>

The default forum layout view used to display posts in a forum. Defaults to 0, which indicates a layout view has not been set by a channel admin.

Note: This is only available in a forum.

Implementations

impl GuildChannel

pub fn is_text_based(&self) -> bool

Available on crate feature model only.

Whether or not this channel is text-based, meaning that it is possible to send messages.

pub async fn delete( &self, http: &Http, reason: Option<&str>, ) -> Result<GuildChannel>

Available on crate feature model only.

Deletes this channel, returning the channel on a successful deletion.

Note: Requires the Manage Channels permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn edit( &mut self, http: &Http, builder: EditChannel<'_>, ) -> Result<()>

Available on crate feature model only.

Edits the channel’s settings.

Refer to the documentation for EditChannel for a full list of methods.

Note: Requires the Manage Channels permission. Modifying permissions via EditChannel::permissions also requires the Manage Roles permission.

Examples

Change a voice channels name and bitrate:

let builder = EditChannel::new().name("test").bitrate(86400);
channel.edit(&http, builder).await;

Errors

Returns Error::Http if the current user lacks permission or if invalid data is given.

pub async fn edit_thread( &mut self, http: &Http, builder: EditThread<'_>, ) -> Result<()>

Available on crate feature model only.

Edits a thread.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn edit_voice_state( &self, http: &Http, user_id: UserId, builder: EditVoiceState, ) -> Result<()>

Available on crate feature model only.

Edits the voice state of a given user in a stage channel.

Note: Requires the Request to Speak permission. Also requires the Mute Members permission to suppress another user or unsuppress the current user. This is not required if suppressing the current user.

Example

Invite a user to speak.

use serenity::builder::EditVoiceState;
use serenity::model::ModelError;

let channel = {
    let guild = cache.guild(guild_id).ok_or(ModelError::ItemMissing)?;
    guild.channels.get(&channel_id).ok_or(ModelError::ItemMissing)?.clone()
};

let builder = EditVoiceState::new().suppress(false);
channel.edit_voice_state(&http, user_id, builder).await?;

Errors

Returns a ModelError::InvalidChannelType if the channel is not a stage channel.

Returns Error::Http if the user lacks permission, or if invalid data is given.

pub async fn edit_own_voice_state( &self, http: &Http, builder: EditVoiceState, ) -> Result<()>

Available on crate feature model only.

Edits the current user’s voice state in a stage channel.

Note: Requires the Request to Speak permission. The Mute Members permission is not required.

Example

Send a request to speak, then clear the request.

use serenity::builder::EditVoiceState;
use serenity::model::ModelError;

let channel = {
    let guild = cache.guild(guild_id).ok_or(ModelError::ItemMissing)?;
    guild.channels.get(&channel_id).ok_or(ModelError::ItemMissing)?.clone()
};

// Send a request to speak
let builder = EditVoiceState::new().request_to_speak(true);
channel.edit_own_voice_state(&http, builder.clone()).await?;

// Clear own request to speak
let builder = builder.request_to_speak(false);
channel.edit_own_voice_state(&http, builder).await?;

Errors

If the cache is enabled, returns a ModelError::InvalidChannelType if the channel is not a stage channel.

Returns Error::Http if the user lacks permission, or if invalid data is given.

pub fn guild<'a>(&self, cache: &'a Cache) -> Option<GuildRef<'a>>

Available on crate features model and cache only.

Attempts to find this channel’s guild in the Cache.

pub async fn send_message( &self, http: &Http, builder: CreateMessage<'_>, ) -> Result<Message>

Available on crate feature model only.

Sends a message to the channel.

Refer to the documentation for CreateMessage for information regarding content restrictions and requirements.

Errors

See CreateMessage::execute for a list of possible errors, and their corresponding reasons.

pub fn members(&self, cache: &Cache) -> Result<Vec<Member>>

Available on crate features model and cache only.

Retrieves Members from the current channel.

ChannelType::Voice and ChannelType::Stage returns Members using the channel.

ChannelType::Text and ChannelType::News return Members that can read the channel.

Errors

Other ChannelTypes lack the concept of Members and will return: ModelError::InvalidChannelType.

pub async fn create_webhook( &self, http: &Http, builder: CreateWebhook<'_>, ) -> Result<Webhook>

Available on crate feature model only.

Creates a webhook in the channel.

Errors

Returns ModelError::InvalidChannelType if the corresponding channel is not of type ChannelType::Text or ChannelType::News.

See CreateWebhook::execute for a detailed list of other possible errors,

pub async fn get_stage_instance(&self, http: &Http) -> Result<StageInstance>

Available on crate feature model only.

Gets a stage instance.

Errors

Returns ModelError::InvalidChannelType if the channel is not a stage channel.

Returns Error::Http if there is no stage instance currently.

pub async fn create_stage_instance( &self, http: &Http, builder: CreateStageInstance<'_>, ) -> Result<StageInstance>

Available on crate feature model only.

Creates a stage instance.

Errors

Returns ModelError::InvalidChannelType if the channel is not a stage channel.

Returns Error::Http if there is already a stage instance currently.

pub async fn edit_stage_instance( &self, http: &Http, builder: EditStageInstance<'_>, ) -> Result<StageInstance>

Available on crate feature model only.

Edits the stage instance

Errors

Returns ModelError::InvalidChannelType if the channel is not a stage channel.

Returns Error::Http if the channel is not a stage channel, or there is no stage instance currently.

pub async fn delete_stage_instance( &self, http: &Http, reason: Option<&str>, ) -> Result<()>

Available on crate feature model only.

Deletes a stage instance.

Errors

Returns ModelError::InvalidChannelType if the channel is not a stage channel.

Returns Error::Http if there is no stage instance currently.