serenity::model

Module connection

Models for user connections.

Structs

Connection

Information about a connection between the current user and a third party service.

ConnectionGeneratedOriginal

Information about a connection between the current user and a third party service.

ConnectionVisibility

The visibility of a user connection on a user’s profile.