Module serenity::model::event

All the events this library handles.

Structs

ChannelCreateEvent

Event data for the channel creation event.

ChannelDeleteEvent

ChannelPinsUpdateEvent

ChannelUpdateEvent

GuildBanAddEvent

GuildBanRemoveEvent

GuildCreateEvent

GuildDeleteEvent

GuildEmojisUpdateEvent

GuildIntegrationsUpdateEvent

GuildMemberAddEvent

GuildMemberRemoveEvent

GuildMemberUpdateEvent

GuildMembersChunkEvent

GuildRoleCreateEvent

GuildRoleDeleteEvent

GuildRoleUpdateEvent

GuildStickersUpdateEvent

GuildUnavailableEvent

GuildUpdateEvent

IntegrationCreateEvent

IntegrationDeleteEvent

IntegrationUpdateEvent

InteractionCreateEvent

InviteCreateEvent

InviteDeleteEvent

MessageCreateEvent

MessageDeleteBulkEvent

MessageDeleteEvent

MessageUpdateEvent

PresenceUpdateEvent

PresencesReplaceEvent

ReactionAddEvent

ReactionRemoveAllEvent

ReactionRemoveEvent

ReadyEvent

The “Ready” event, containing initial ready cache

RelatedIdsForEventType

Defines the related IDs that may exist for an EventType.

ResumedEvent

StageInstanceCreateEvent

StageInstanceDeleteEvent

StageInstanceUpdateEvent

ThreadCreateEvent

ThreadDeleteEvent

ThreadListSyncEvent

ThreadMemberUpdateEvent

ThreadMembersUpdateEvent

ThreadUpdateEvent

TypingStartEvent

UnknownEvent

UserUpdateEvent

VoiceServerUpdateEvent

VoiceStateUpdateEvent

WebhookUpdateEvent

Enums

Event

Event received over a websocket connection

EventType

The type of event dispatch received from the gateway.

GatewayEvent

RelatedId

Similar to Option, but with additional variants relevant to Event’s id methods (such as Event::user_id).

Functions

deserialize_event_with_type

Deserializes a serde_json::Value into an Event.