Module serenity::model::event

All the events this library handles.

Every event includes the gateway intent required to receive it, as well as a link to the Discord documentation for the event.

Structs

AutoModActionExecutionEvent

Requires GatewayIntents::AUTO_MODERATION_EXECUTION.

AutoModRuleCreateEvent

Requires GatewayIntents::AUTO_MODERATION_CONFIGURATION.

AutoModRuleDeleteEvent

Requires GatewayIntents::AUTO_MODERATION_CONFIGURATION.

AutoModRuleUpdateEvent

Requires GatewayIntents::AUTO_MODERATION_CONFIGURATION.

ChannelCreateEvent

Event data for the channel creation event.

ChannelDeleteEvent

Requires GatewayIntents::GUILDS.

ChannelPinsUpdateEvent

Requires GatewayIntents::GUILDS or GatewayIntents::DIRECT_MESSAGES.

ChannelUpdateEvent

Requires GatewayIntents::GUILDS.

CommandPermissionsUpdateEvent

Requires no gateway intents.

GuildAuditLogEntryCreateEvent

Requires GatewayIntents::GUILD_MODERATION and Permissions::VIEW_AUDIT_LOG.

GuildBanAddEvent

Requires GatewayIntents::GUILD_BANS.

GuildBanRemoveEvent

Requires GatewayIntents::GUILD_BANS.

GuildCreateEvent

Requires GatewayIntents::GUILDS.

GuildDeleteEvent

Requires GatewayIntents::GUILDS.

GuildEmojisUpdateEvent

Requires GatewayIntents::GUILD_EMOJIS_AND_STICKERS.

GuildIntegrationsUpdateEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

GuildMemberAddEvent

Requires GatewayIntents::GUILD_MEMBERS.

GuildMemberRemoveEvent

Requires GatewayIntents::GUILD_MEMBERS.

GuildMemberUpdateEvent

Requires GatewayIntents::GUILD_MEMBERS.

GuildMembersChunkEvent

Requires no gateway intents.

GuildRoleCreateEvent

Requires GatewayIntents::GUILDS.

GuildRoleDeleteEvent

Requires GatewayIntents::GUILDS.

GuildRoleUpdateEvent

Requires GatewayIntents::GUILDS.

GuildScheduledEventCreateEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventDeleteEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventUpdateEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventUserAddEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventUserRemoveEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildStickersUpdateEvent

Requires GatewayIntents::GUILD_EMOJIS_AND_STICKERS.

GuildUpdateEvent

Requires GatewayIntents::GUILDS.

IntegrationCreateEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

IntegrationDeleteEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

IntegrationUpdateEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

InteractionCreateEvent

Requires no gateway intents.

InviteCreateEvent

Requires GatewayIntents::GUILD_INVITES and [`Permissions::MANAGE_CHANNELS´] permission.

InviteDeleteEvent

Requires GatewayIntents::GUILD_INVITES and [`Permissions::MANAGE_CHANNELS´] permission.

MessageCreateEvent

Requires GatewayIntents::GUILD_MESSAGES or GatewayIntents::DIRECT_MESSAGES.

MessageDeleteBulkEvent

Requires GatewayIntents::GUILD_MESSAGES or GatewayIntents::DIRECT_MESSAGES.

MessageDeleteEvent

Requires GatewayIntents::GUILD_MESSAGES or GatewayIntents::DIRECT_MESSAGES.

MessageUpdateEvent

Requires GatewayIntents::GUILD_MESSAGES.

PresenceUpdateEvent

Requires GatewayIntents::GUILD_PRESENCES.

PresencesReplaceEvent

Not officially documented.

ReactionAddEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReactionRemoveAllEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReactionRemoveEmojiEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReactionRemoveEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReadyEvent

The “Ready” event, containing initial ready cache

ResumedEvent

Requires no gateway intents.

StageInstanceCreateEvent

Requires GatewayIntents::GUILDS.

StageInstanceDeleteEvent

Requires GatewayIntents::GUILDS.

StageInstanceUpdateEvent

Requires GatewayIntents::GUILDS.

ThreadCreateEvent

Requires GatewayIntents::GUILDS.

ThreadDeleteEvent

Requires GatewayIntents::GUILDS.

ThreadListSyncEvent

Requires GatewayIntents::GUILDS.

ThreadMemberUpdateEvent

Requires GatewayIntents::GUILDS, and, to receive this event for other users, GatewayIntents::GUILD_MEMBERS.

ThreadMembersUpdateEvent

Requires GatewayIntents::GUILD_MEMBERS.

ThreadUpdateEvent

Requires GatewayIntents::GUILDS.

TypingStartEvent

Requires GatewayIntents::GUILD_MESSAGE_TYPING or GatewayIntents::DIRECT_MESSAGE_TYPING.

UnknownEvent

UserUpdateEvent

Sent when properties about the current bot’s user change.

VoiceChannelStatusUpdateEvent

Requires GatewayIntents::GUILDS.

VoiceServerUpdateEvent

Requires no gateway intents.

VoiceStateUpdateEvent

Requires GatewayIntents::GUILD_VOICE_STATES.

WebhookUpdateEvent

Requires GatewayIntents::GUILD_WEBHOOKS.

Enums

Event

Event received over a websocket connection

GatewayEvent

Discord docs.