Module serenity::model::gateway

Models pertaining to the gateway.

Structs

Activity
Representation of an activity that a User is performing.

ActivityAssets
The assets for an activity.

ActivityEmoji
Representation of an emoji used in a custom status

ActivityFlags
A set of flags defining what is in an activity’s payload.

ActivityParty
Information about an activity’s party.

ActivitySecrets
Secrets for an activity.

ActivityTimestamps
Timestamps of when a user started and/or is ending their activity.

BotGateway
A representation of the data retrieved from the bot gateway endpoint.

ClientStatus
Information detailing the current active status of a User.

Gateway
A representation of the data retrieved from the gateway endpoint.

Presence
Information detailing the current online status of a User.

Ready
An initial set of information given after IDENTIFYing to the gateway.

SessionStartLimit
Information describing how many gateway sessions you can initiate within a ratelimit period.

Enums

ActivityType