Module serenity::model::guild

Models relating to guilds and types that it owns.

Modules

audit_log

Audit log types for administrative actions within guilds.

automod

Auto moderation types

Structs

Ban

A representation of a banning of a user.

Emoji

Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

Guild

Information about a Discord guild, such as channels, emojis, etc.

GuildInfo

Basic information about a guild.

GuildPreview

Preview Guild information.

GuildPrune

Representation of the number of members that would be pruned by a guild prune operation.

GuildWelcomeChannel

A channel shown in the GuildWelcomeScreen.

GuildWelcomeScreen

Information relating to a guild’s welcome screen.

GuildWidget

A Guild widget.

Integration

Various information about integrations.

IntegrationAccount

Integration account object.

IntegrationApplication

Integration application object.

Member

Information about a member of a guild.

MembersItermodel

A helper class returned by GuildId::members_iter

PartialGuild

Partial information about a Guild. This does not include information like member data.

PartialMember

A partial amount of data for a member.

Role

Information about a role within a guild. A role represents a set of permissions, and can be attached to one or multiple users. A role has various miscellaneous configurations, such as being assigned a colour. Roles are unique per guild and do not cross over to other guilds in any way, and can have channel-specific permission overrides in addition to guild-level permissions.

RoleTags

The tags of a Role.

ScheduledEvent

Information about a guild scheduled event.

ScheduledEventMetadata

Discord docs.

ScheduledEventUser

Discord docs.

SystemChannelFlags

Describes a system channel flags.

ThreadMember

Discord docs.

ThreadMemberFlags

Describes extra features of the message.

UnavailableGuild

Data for an unavailable guild.

Enums

DefaultMessageNotificationLevel

Default message notification level for a guild.

ExplicitContentFilter

Setting used to filter explicit messages from members.

GuildWelcomeChannelEmoji

A GuildWelcomeScreen emoji.

GuildWidgetStyle

IntegrationExpireBehaviour

The behavior once the integration expires.

MfaLevel

Multi-Factor Authentication level for guild moderators.

NsfwLevel

The Guild nsfw level.

PremiumTier

The guild’s premium tier, depends on the amount of users boosting the guild currently

RoleParseErrorcache and model and utils

ScheduledEventStatus

Discord docs.

ScheduledEventType

Discord docs.

VerificationLevel

The level to set as criteria prior to a user being able to send messages in a Guild.