Module serenity::model::guild

Models relating to guilds and types that it owns.

Structs

AuditLogEntry

AuditLogs

Ban
A representation of a banning of a user.

Change

Emoji
Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

Guild
Information about a Discord guild, such as channels, emojis, etc.

GuildEmbed	Deprecated
A Guild embed.

GuildInfo
Basic information about a guild.

GuildPreview
Preview Guild information.

GuildPrune
Representation of the number of members that would be pruned by a guild prune operation.

GuildUnavailable
Data for an unavailable guild.

GuildWelcomeChannel
A channel shown in the GuildWelcomeScreen.

GuildWelcomeScreen
Information relating to a guild’s welcome screen.

GuildWidget
A Guild widget.

Integration
Various information about integrations.

IntegrationAccount
Integration account object.

IntegrationApplication
Integration application object.

Member
Information about a member of a guild.

MembersIter
A helper class returned by GuildId::members_iter

Options

PartialGuild
Partial information about a Guild. This does not include information like member data.

PartialMember
A partial amount of data for a member.

Role
Information about a role within a guild. A role represents a set of permissions, and can be attached to one or multiple users. A role has various miscellaneous configurations, such as being assigned a colour. Roles are unique per guild and do not cross over to other guilds in any way, and can have channel-specific permission overrides in addition to guild-level permissions.

RoleTags
The tags of a Role.

SystemChannelFlags
Describes a system channel flags.

Enums

Action
Determines the action that was done on a target.

ActionChannel

ActionChannelOverwrite

ActionEmoji

ActionIntegration

ActionInvite

ActionMember

ActionMessage

ActionRole

ActionWebhook

DefaultMessageNotificationLevel
Default message notification level for a guild.

ExplicitContentFilter
Setting used to filter explicit messages from members.

GuildContainer
A container for guilds.

GuildStatus

GuildWelcomeScreenEmoji
A GuildWelcomeScreen emoji.

GuildWidgetStyle

IntegrationExpireBehaviour
The behavior once the integration expires.

MfaLevel
Multi-Factor Authentication level for guild moderators.

NsfwLevel
The Guild nsfw level.

PremiumTier
The guild’s premium tier, depends on the amount of users boosting the guild currently

Region
The name of a region that a voice server can be located in.

Target
Determines to what entity an action was used on.

VerificationLevel
The level to set as criteria prior to a user being able to send messages in a Guild.