Module serenity::model::guild::automod

Auto moderation types

Discord docs

Structs

ActionExecution

Gateway event payload sent when a rule is triggered and an action is executed (e.g. message is blocked).

Rule

Configured auto moderation rule.

TriggerMetadata

Individual change for trigger metadata within an audit log entry.

Enums

Action

An action which will execute whenever a rule is triggered.

ActionType

Type of Action.

EventType

Indicates in what event context a rule should be checked.

KeywordPresetType

Internally pre-defined wordsets which will be searched for in content.

Trigger

Characterizes the type of content which can trigger the rule.

TriggerType

Type of Trigger.