serenity::model::guild
Struct Guild

#[non_exhaustive]
pub struct Guild {

    pub id: GuildId,
    pub name: FixedString,
    pub icon: Option<ImageHash>,
    pub icon_hash: Option<ImageHash>,
    pub splash: Option<ImageHash>,
    pub discovery_splash: Option<ImageHash>,
    pub owner_id: UserId,
    pub afk_metadata: Option<AfkMetadata>,
    pub widget_channel_id: Option<ChannelId>,
    pub verification_level: VerificationLevel,
    pub default_message_notifications: DefaultMessageNotificationLevel,
    pub explicit_content_filter: ExplicitContentFilter,
    pub roles: ExtractMap<RoleId, Role>,
    pub emojis: ExtractMap<EmojiId, Emoji>,
    pub features: FixedArray<FixedString>,
    pub mfa_level: MfaLevel,
    pub application_id: Option<ApplicationId>,
    pub system_channel_id: Option<ChannelId>,
    pub system_channel_flags: SystemChannelFlags,
    pub rules_channel_id: Option<ChannelId>,
    pub max_presences: Option<NonMaxU64>,
    pub max_members: Option<NonMaxU64>,
    pub vanity_url_code: Option<FixedString>,
    pub description: Option<FixedString>,
    pub banner: Option<FixedString>,
    pub premium_tier: PremiumTier,
    pub premium_subscription_count: Option<NonMaxU64>,
    pub preferred_locale: FixedString,
    pub public_updates_channel_id: Option<ChannelId>,
    pub max_video_channel_users: Option<NonMaxU64>,
    pub max_stage_video_channel_users: Option<NonMaxU64>,
    pub approximate_member_count: Option<NonMaxU64>,
    pub approximate_presence_count: Option<NonMaxU64>,
    pub welcome_screen: Option<GuildWelcomeScreen>,
    pub nsfw_level: NsfwLevel,
    pub stickers: ExtractMap<StickerId, Sticker>,
    pub joined_at: Timestamp,
    pub member_count: u64,
    pub voice_states: ExtractMap<UserId, VoiceState>,
    pub members: ExtractMap<UserId, Member>,
    pub channels: ExtractMap<ChannelId, GuildChannel>,
    pub threads: FixedArray<GuildChannel>,
    pub presences: ExtractMap<UserId, Presence>,
    pub stage_instances: FixedArray<StageInstance>,
    pub scheduled_events: FixedArray<ScheduledEvent>,
    /* private fields */

}

Information about a Discord guild, such as channels, emojis, etc.

Discord docs plus extension.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: GuildId

The unique Id identifying the guild.

This is equivalent to the Id of the default role (@everyone).

name: FixedString

The name of the guild.

icon: Option<ImageHash>

The hash of the icon used by the guild.

In the client, this appears on the guild list on the left-hand side.

icon_hash: Option<ImageHash>

Icon hash, returned when in the template object

splash: Option<ImageHash>

An identifying hash of the guild’s splash icon.

If the InviteSplash feature is enabled, this can be used to generate a URL to a splash image.

discovery_splash: Option<ImageHash>

An identifying hash of the guild discovery’s splash icon.

Note: Only present for guilds with the DISCOVERABLE feature.

owner_id: UserId

The Id of the User who owns the guild.

afk_metadata: Option<AfkMetadata>

Information about the voice afk channel.

widget_channel_id: Option<ChannelId>

The channel id that the widget will generate an invite to, or null if set to no invite

verification_level: VerificationLevel

Indicator of the current verification level of the guild.

default_message_notifications: DefaultMessageNotificationLevel

Indicator of whether notifications for all messages are enabled by default in the guild.

explicit_content_filter: ExplicitContentFilter

Default explicit content filter level.

roles: ExtractMap<RoleId, Role>

A mapping of the guild’s roles.

emojis: ExtractMap<EmojiId, Emoji>

All of the guild’s custom emojis.

features: FixedArray<FixedString>

The guild features. More information available at discord documentation.

The following is a list of known features:

ANIMATED_ICON

BANNER

COMMERCE

COMMUNITY

DISCOVERABLE

FEATURABLE

INVITE_SPLASH

MEMBER_VERIFICATION_GATE_ENABLED

MONETIZATION_ENABLED

MORE_STICKERS

NEWS

PARTNERED

PREVIEW_ENABLED

PRIVATE_THREADS

ROLE_ICONS

SEVEN_DAY_THREAD_ARCHIVE

THREE_DAY_THREAD_ARCHIVE

TICKETED_EVENTS_ENABLED

VANITY_URL

VERIFIED

VIP_REGIONS

WELCOME_SCREEN_ENABLED

THREE_DAY_THREAD_ARCHIVE

SEVEN_DAY_THREAD_ARCHIVE

PRIVATE_THREADS

mfa_level: MfaLevel

Indicator of whether the guild requires multi-factor authentication for Roles or Users with moderation permissions.

application_id: Option<ApplicationId>

Application ID of the guild creator if it is bot-created.

system_channel_id: Option<ChannelId>

The ID of the channel to which system messages are sent.

system_channel_flags: SystemChannelFlags

System channel flags.

rules_channel_id: Option<ChannelId>

The id of the channel where rules and/or guidelines are displayed.

Note: Only available on COMMUNITY guild, see Self::features.

max_presences: Option<NonMaxU64>

The maximum number of presences for the guild. The default value is currently 25000.

Note: It is in effect when it is None.

max_members: Option<NonMaxU64>

The maximum number of members for the guild.

vanity_url_code: Option<FixedString>

The vanity url code for the guild, if it has one.

description: Option<FixedString>

The server’s description, if it has one.

banner: Option<FixedString>

The guild’s banner, if it has one.

premium_tier: PremiumTier

The server’s premium boosting level.

premium_subscription_count: Option<NonMaxU64>

The total number of users currently boosting this server.

preferred_locale: FixedString

The preferred locale of this guild only set if guild has the “DISCOVERABLE” feature, defaults to en-US.

public_updates_channel_id: Option<ChannelId>

The id of the channel where admins and moderators of Community guilds receive notices from Discord.

Note: Only available on COMMUNITY guild, see Self::features.

max_video_channel_users: Option<NonMaxU64>

The maximum amount of users in a video channel.

max_stage_video_channel_users: Option<NonMaxU64>

The maximum amount of users in a stage video channel

approximate_member_count: Option<NonMaxU64>

Approximate number of members in this guild.

approximate_presence_count: Option<NonMaxU64>

Approximate number of non-offline members in this guild.

welcome_screen: Option<GuildWelcomeScreen>

The welcome screen of the guild.

Note: Only available on COMMUNITY guild, see Self::features.

nsfw_level: NsfwLevel

The guild NSFW state. See discord support article.

stickers: ExtractMap<StickerId, Sticker>

All of the guild’s custom stickers.

joined_at: Timestamp

The date that the current user joined the guild.

member_count: u64

The number of members in the guild.

voice_states: ExtractMap<UserId, VoiceState>

A mapping of Users to their current voice state.

members: ExtractMap<UserId, Member>

Users who are members of the guild.

Members might not all be available when the ReadyEvent is received if the Self::member_count is greater than the LARGE_THRESHOLD set by the library.

channels: ExtractMap<ChannelId, GuildChannel>

All voice and text channels contained within a guild.

This contains all channels regardless of permissions (i.e. the ability of the bot to read from or connect to them).

threads: FixedArray<GuildChannel>

All active threads in this guild that current user has permission to view.

A thread is guaranteed (for errors, not for panics) to be cached if a MESSAGE_CREATE event is fired in said thread, however an INTERACTION_CREATE may not have a private thread in cache.

presences: ExtractMap<UserId, Presence>

A mapping of Users’ Ids to their current presences.

Note: This will be empty unless the “guild presences” privileged intent is enabled.

stage_instances: FixedArray<StageInstance>

The stage instances in this guild.

scheduled_events: FixedArray<ScheduledEvent>

The stage instances in this guild.

Implementations

impl Guild

pub fn widget_enabled(&self) -> Option<bool>

Whether or not the guild widget is enabled.

pub fn set_widget_enabled(&mut self, value: Option<bool>)

Sets the widget_enabled to the value provided.

pub fn premium_progress_bar_enabled(&self) -> bool

Whether the guild has the boost progress bar enabled

pub fn set_premium_progress_bar_enabled(&mut self, value: bool)

Sets the premium_progress_bar_enabled to the value provided.

pub fn large(&self) -> bool

Indicator of whether the guild is considered “large” by Discord.

pub fn set_large(&mut self, value: bool)

Sets the large to the value provided.

pub fn unavailable(&self) -> bool

Whether this guild is unavailable due to an outage.

pub fn set_unavailable(&mut self, value: bool)

Sets the unavailable to the value provided.

impl Guild

pub fn default_channel(&self, uid: UserId) -> Option<&GuildChannel>

Available on crate feature model only.

Returns the “default” channel of the guild for the passed user id. (This returns the first channel that can be read by the user, if there isn’t one, returns None)

pub fn default_channel_guaranteed(&self) -> Option<&GuildChannel>

Available on crate feature model only.

Returns the guaranteed “default” channel of the guild. (This returns the first channel that can be read by everyone, if there isn’t one, returns None)

Note: This is very costly if used in a server with lots of channels, members, or both.

pub fn banner_url(&self) -> Option<String>

Available on crate feature model only.

Returns the formatted URL of the guild’s banner image, if one exists.

pub async fn create( http: &Http, name: &str, icon: Option<ImageHash>, ) -> Result<PartialGuild>

Available on crate feature model only.

Creates a guild with the data provided.

Only a PartialGuild will be immediately returned, and a full Guild will be received over a Shard.

Note: This endpoint is usually only available for user accounts. Refer to Discord’s information for the endpoint here for more information. If you require this as a bot, re-think what you are doing and if it really needs to be doing this.

Examples

Create a guild called "test" in the [US West region] with no icon:

use serenity::model::guild::Guild;
let guild = Guild::create(&http, "test", None).await;

Errors

Returns Error::Http if the current user cannot create a Guild.

pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<()>

Available on crate feature model only.

Deletes the current guild if the current user is the owner of the guild.

Note: Requires the current user to be the owner of the guild.

Errors

If the cache is enabled, then returns a ModelError::InvalidUser if the current user is not the guild owner.

Otherwise returns Error::Http if the current user is not the owner of the guild.

pub async fn edit(&mut self, http: &Http, builder: EditGuild<'_>) -> Result<()>

Available on crate feature model only.

Edits the current guild with new data where specified.

Note: Requires the Manage Guild permission.

Examples

Change a guild’s icon using a file named “icon.png”:

let icon = CreateAttachment::path("./icon.png").await?;

// assuming a `guild` has already been bound
let builder = EditGuild::new().icon(Some(&icon));
guild.edit(&http, builder).await?;

Errors

Returns Error::Http if the current user lacks permission or if invalid data is given.

pub async fn get( cache_http: impl CacheHttp, guild_id: GuildId, ) -> Result<PartialGuild>

Available on crate feature model only.

Gets a partial amount of guild data by its Id.

Note: This will not be a Guild, as the REST API does not send all data with a guild retrieval.

Errors

Returns an Error::Http if the current user is not in the guild.

pub fn member_highest_role(&self, member: &Member) -> Option<&Role>

Available on crate feature model only.

Gets the highest role a Member of this Guild has.

Returns None if the member has no roles or the member from this guild.

pub fn greater_member_hierarchy( &self, lhs_id: UserId, rhs_id: UserId, ) -> Option<UserId>

Available on crate feature model only.

Returns which of two Users has a higher Member hierarchy.

Hierarchy is essentially who has the Role with the highest position.

Returns None if at least one of the given users’ member instances is not present. Returns None if the users have the same hierarchy, as neither are greater than the other.

If both user IDs are the same, None is returned. If one of the users is the guild owner, their ID is returned.

pub fn icon_url(&self) -> Option<String>

Available on crate feature model only.

Returns the formatted URL of the guild’s icon, if one exists.

This will produce a WEBP image URL, or GIF if the guild has a GIF icon.

pub async fn member( &self, http: &Http, user_id: UserId, ) -> Result<Cow<'_, Member>>

Available on crate feature model only.

Gets a user’s Member for the guild by Id.

If the cache feature is enabled Self::members will be checked first, if so, a reference to the member will be returned.

Errors

Returns an Error::Http if the user is not in the guild or if the guild is otherwise unavailable.

pub fn members_with_status( &self, status: OnlineStatus, ) -> impl Iterator<Item = &Member>

Available on crate feature model only.

Gets a list of all the members (satisfying the status provided to the function) in this guild.

pub fn member_named(&self, name: &str) -> Option<&Member>

Available on crate feature model only.

Retrieves the first Member found that matches the name - with an optional discriminator

provided.

Searching with a discriminator given is the most precise form of lookup, as no two people can share the same username and discriminator.

If a member can not be found by username or username#discriminator, then a search will be done for the nickname. When searching by nickname, the hash (#) and everything after it is included in the search.

The following are valid types of searches:

username: “zey”

username and discriminator: “zey#5479”

Note: This will only search members that are cached. If you want to search all members in the guild via the Http API, use GuildId::search_members.

pub fn members_starting_with( &self, prefix: &str, case_sensitive: bool, sorted: bool, ) -> Vec<(&Member, &str)>

Available on crate feature model only.

Retrieves all Member that start with a given String.

sorted decides whether the best early match of the prefix should be the criteria to sort the result.

For the prefix “zey” and the unsorted result:

“zeya”, “zeyaa”, “zeyla”, “zeyzey”, “zeyzeyzey”

It would be sorted:

“zeya”, “zeyaa”, “zeyla”, “zeyzey”, “zeyzeyzey”

Note: This will only search members that are cached. If you want to search all members in the guild via the Http API, use GuildId::search_members.

pub fn members_containing( &self, substring: &str, case_sensitive: bool, sorted: bool, ) -> Vec<(&Member, String)>

Available on crate feature model only.

Retrieves all Member containing a given String as either username or nick, with a priority on username.

If the substring is “yla”, following results are possible:

“zeyla”, “meiyla”, “yladenisyla”

If ‘case_sensitive’ is false, the following are not found:

“zeYLa”, “meiyLa”, “LYAdenislyA”

sorted decides whether the best early match of the search-term should be the criteria to sort the result. It will look at the account name first, if that does not fit the search-criteria substring, the display-name will be considered.

For the substring “zey” and the unsorted result:

“azey”, “zey”, “zeyla”, “zeylaa”, “zeyzeyzey”

It would be sorted:

“zey”, “azey”, “zeyla”, “zeylaa”, “zeyzeyzey”

Note: Due to two fields of a Member being candidates for the searched field, setting sorted to true will result in an overhead, as both fields have to be considered again for sorting.

Note: This will only search members that are cached. If you want to search all members in the guild via the Http API, use GuildId::search_members.

pub fn members_username_containing( &self, substring: &str, case_sensitive: bool, sorted: bool, ) -> Vec<(&Member, String)>

Available on crate feature model only.

Retrieves a tuple of Members containing a given String in their username as the first field and the name used for sorting as the second field.

If the substring is “yla”, following results are possible:

“zeyla”, “meiyla”, “yladenisyla”

If ‘case_sensitive’ is false, the following are not found:

“zeYLa”, “meiyLa”, “LYAdenislyA”

sort decides whether the best early match of the search-term should be the criteria to sort the result.

For the substring “zey” and the unsorted result:

“azey”, “zey”, “zeyla”, “zeylaa”, “zeyzeyzey”

It would be sorted:

“zey”, “azey”, “zeyla”, “zeylaa”, “zeyzeyzey”

Note: This will only search members that are cached. If you want to search all members in the guild via the Http API, use GuildId::search_members.

pub fn members_nick_containing( &self, substring: &str, case_sensitive: bool, sorted: bool, ) -> Vec<(&Member, String)>

Available on crate feature model only.

Retrieves all Member containing a given String in their nick.

If the substring is “yla”, following results are possible:

“zeyla”, “meiyla”, “yladenisyla”

If ‘case_sensitive’ is false, the following are not found:

“zeYLa”, “meiyLa”, “LYAdenislyA”

sort decides whether the best early match of the search-term should be the criteria to sort the result.

For the substring “zey” and the unsorted result:

“azey”, “zey”, “zeyla”, “zeylaa”, “zeyzeyzey”

It would be sorted:

“zey”, “azey”, “zeyla”, “zeylaa”, “zeyzeyzey”

Note: Instead of panicking, when sorting does not find a nick, the username will be used (this should never happen).

Note: This will only search members that are cached. If you want to search all members in the guild via the Http API, use GuildId::search_members.

pub fn user_permissions_in( &self, channel: &GuildChannel, member: &Member, ) -> Permissions

Available on crate feature model only.

Calculate a Member’s permissions in a given channel in the guild.

pub fn partial_member_permissions_in( &self, channel: &GuildChannel, member_id: UserId, member: &PartialMember, ) -> Permissions

Available on crate feature model only.

Calculate a PartialMember’s permissions in a given channel in a guild.

Panics

Panics if the passed UserId does not match the PartialMember id, if user is Some.

pub fn splash_url(&self) -> Option<String>

Available on crate feature model only.

Returns the formatted URL of the guild’s splash image, if one exists.

pub fn role_by_name(&self, role_name: &str) -> Option<&Role>

Available on crate feature model only.

Obtain a reference to a role by its name.

Note: If two or more roles have the same name, obtained reference will be one of them.

Examples

Obtain a reference to a Role by its name.

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if let Some(guild_id) = msg.guild_id {
            if let Some(guild) = guild_id.to_guild_cached(&ctx.cache) {
                if let Some(role) = guild.role_by_name("role_name") {
                    println!("{:?}", role);
                }
            }
        }
    }
}