Struct serenity::model::guild::PartialGuild

#[non_exhaustive]pub struct PartialGuild {
    pub id: GuildId,
    pub afk_channel_id: Option<ChannelId>,
    pub afk_timeout: u64,
    pub default_message_notifications: DefaultMessageNotificationLevel,
    pub widget_channel_id: Option<ChannelId>,
    pub widget_enabled: bool,
    pub emojis: HashMap<EmojiId, Emoji>,
    pub features: Vec<String>,
    pub icon: Option<String>,
    pub mfa_level: MfaLevel,
    pub name: String,
    pub owner_id: UserId,
    pub region: String,
    pub roles: HashMap<RoleId, Role>,
    pub splash: Option<String>,
    pub verification_level: VerificationLevel,
    pub description: Option<String>,
    pub premium_tier: PremiumTier,
    pub premium_subscription_count: u64,
    pub banner: Option<String>,
    pub vanity_url_code: Option<String>,
}

Partial information about a Guild. This does not include information like member data.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct {{ .. }} syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
id: GuildIdafk_channel_id: Option<ChannelId>afk_timeout: u64default_message_notifications: DefaultMessageNotificationLevelwidget_channel_id: Option<ChannelId>widget_enabled: boolemojis: HashMap<EmojiId, Emoji>features: Vec<String>

Features enabled for the guild.

Refer to Guild::features for more information.

icon: Option<String>mfa_level: MfaLevelname: Stringowner_id: UserIdregion: Stringroles: HashMap<RoleId, Role>splash: Option<String>verification_level: VerificationLeveldescription: Option<String>premium_tier: PremiumTierpremium_subscription_count: u64banner: Option<String>vanity_url_code: Option<String>
Implementations

impl PartialGuild

pub async fn ban(
&self,
http: impl AsRef<Http>,
user: impl Into<UserId>,
dmd: u8
) -> Result<()>

Ban a User from the guild, deleting a number of days’ worth of messages (dmd) between the range 0 and 7.

Note: Requires the Ban Members permission.

Examples

Ban a member and remove all messages they’ve sent in the last 4 days:

// assumes a `user` and `guild` have already been bound
let _ = guild.ban(user, 4);

Errors

Returns a ModelError::DeleteMessageDaysAmount if the number of days’ worth of messages to delete is over the maximum.

Also may return Error::Http if the current user lacks permission.

pub async fn ban_with_reason(
&self,
http: impl AsRef<Http>,
user: impl Into<UserId>,
dmd: u8,
reason: impl AsRef<str>
) -> Result<()>

Ban a User from the guild with a reason. Refer to ban to further documentation.

Errors

In addition to the reasons ban may return an error, can also return an error if the reason is too long.

pub async fn bans(&self, http: impl AsRef<Http>) -> Result<Vec<Ban>>

Gets a list of the guild’s bans.

Requires the Ban Members permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn channels(
&self,
http: impl AsRef<Http>
) -> Result<HashMap<ChannelId, GuildChannel>>

Gets all of the guild’s channels over the REST API.

Errors

Returns Error::Http if the current user is not in the guild or if the guild is otherwise unavailable.

pub async fn create_channel(
&self,
http: impl AsRef<Http>,
f: impl FnOnce(&mut CreateChannel) -> &mut CreateChannel
) -> Result<GuildChannel>

Creates a GuildChannel in the guild.

Refer to Http::create_channel for more information.

Requires the Manage Channels permission.

Examples

Create a voice channel in a guild with the name test:

use serenity::model::ChannelType;

guild.create_channel(|c| c.name("test").kind(ChannelType::Voice));

Errors

Returns Error::Http if the current user lacks permission, or if invalid data was given, such as the channel name being too long.

pub async fn create_emoji(
&self,
http: impl AsRef<Http>,
name: &str,
image: &str
) -> Result<Emoji>

Creates an emoji in the guild with a name and base64-encoded image.

Refer to the documentation for Guild::create_emoji for more information.

Requires the Manage Emojis permission.

Examples

See the EditProfile::avatar example for an in-depth example as to how to read an image from the filesystem and encode it as base64. Most of the example can be applied similarly for this method.

Errors

Returns Error::Http if the current user lacks permission, if the emoji name is too long, or if the image is too large.

pub async fn create_integration(
&self,
http: impl AsRef<Http>,
integration_id: impl Into<IntegrationId>,
kind: &str
) -> Result<()>

Creates an integration for the guild.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn create_application_command<F>(
&self,
http: impl AsRef<Http>,
application_id: u64,
f: F
) -> Result<ApplicationCommand> where
F: FnOnce(&mut CreateInteraction) -> &mut CreateInteraction,

This is supported on crate feature unstable_discord_api only.

Creates a new ApplicationCommand for the guild.

See the documentation for Interaction::create_global_application_command on how to use this.

Note: application_id is usually the bot’s id, unless it’s a very old bot.

pub async fn create_role<F>(&self, http: impl AsRef<Http>, f: F) -> Result<Role> where
F: FnOnce(&mut EditRole) -> &mut EditRole,

Creates a new role in the guild with the data set, if any.

See the documentation for Guild::create_role on how to use this.

Note: Requires the Manage Roles permission.

Errors

Returns Error::Http if the current user lacks permission, or if an invalid value was set.

pub async fn delete(&self, http: impl AsRef<Http>) -> Result<PartialGuild>

Deletes the current guild if the current user is the owner of the guild.

Note: Requires the current user to be the owner of the guild.

Errors

Returns Error::Http if the current user is not the owner of the guild.

pub async fn delete_emoji(
&self,
http: impl AsRef<Http>,
emoji_id: impl Into<EmojiId>
) -> Result<()>

Deletes an Emoji from the guild.

Requires the Manage Emojis permission.

Errors

Returns Error::Http if the current user lacks permission, or if an emoji with that Id does not exist in the guild.

pub async fn delete_integration(
&self,
http: impl AsRef<Http>,
integration_id: impl Into<IntegrationId>
) -> Result<()>

Deletes an integration by Id from the guild.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission, or if an integration with that Id does not exist in the guild.

pub async fn delete_role(
&self,
http: impl AsRef<Http>,
role_id: impl Into<RoleId>
) -> Result<()>

Deletes a Role by Id from the guild.

Also see Role::delete if you have the cache and model features enabled.

Requires the Manage Roles permission.

Errors

Returns Error::Http if the current user lacks permission, or if a Role with that Id does not exist in the Guild.

pub async fn edit<F>(&mut self, http: impl AsRef<Http>, f: F) -> Result<()> where
F: FnOnce(&mut EditGuild) -> &mut EditGuild,

Edits the current guild with new data where specified.

Note: Requires the current user to have the Manage Guild permission.

Errors

Returns Error::Http if an invalid value is set, or if the current user lacks permission to edit the guild.

pub async fn edit_emoji(
&self,
http: impl AsRef<Http>,
emoji_id: impl Into<EmojiId>,
name: &str
) -> Result<Emoji>

Edits an Emoji’s name in the guild.

Also see Emoji::edit if you have the cache and methods features enabled.

Requires the Manage Emojis permission.

Errors

Returns Error::Http if the current user lacks permission, or if an emoji with that Id does not exist in the guild.

pub async fn edit_member<F>(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
f: F
) -> Result<Member> where
F: FnOnce(&mut EditMember) -> &mut EditMember,

Edits the properties of member of the guild, such as muting or nicknaming them.

Refer to EditMember’s documentation for a full list of methods and permission restrictions.

Examples

Mute a member and set their roles to just one role with a predefined Id:

use serenity::model::GuildId;

GuildId(7).edit_member(user_id, |m| m.mute(true).roles(&vec![role_id])).await;

Errors

Returns Error::Http if the current user lacks the necessary permissions.

pub async fn edit_nickname(
&self,
http: impl AsRef<Http>,
new_nickname: Option<&str>
) -> Result<()>

Edits the current user’s nickname for the guild.

Pass None to reset the nickname.

Note: Requires the Change Nickname permission.

Errors

Returns Error::Http if the current user lacks permission to change their nickname.

pub async fn get(
http: impl AsRef<Http>,
guild_id: impl Into<GuildId>
) -> Result<PartialGuild>

Gets a partial amount of guild data by its Id.

Errors

Returns Error::Http if the current user is not in the guild.

pub async fn kick(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<()>

Kicks a Member from the guild.

Requires the Kick Members permission.

Errors

Returns Error::Http if the member cannot be kicked by the current user.

pub async fn kick_with_reason(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
reason: &str
) -> Result<()>

Errors

In addition to the reasons kick may return an error, can also return an error if the reason is too long.

pub fn icon_url(&self) -> Option<String>

Returns a formatted URL of the guild’s icon, if the guild has an icon.

pub async fn emojis(&self, http: impl AsRef<Http>) -> Result<Vec<Emoji>>

Gets all Emojis of this guild via HTTP.

Errors

Returns Error::Http if the guild is unavailable.

pub async fn emoji(
&self,
http: impl AsRef<Http>,
emoji_id: EmojiId
) -> Result<Emoji>

Gets an Emoji of this guild by its ID via HTTP.

Errors

Returns Error::Http if an Emoji with the given Id does not exist for the guild.

pub async fn integrations(
&self,
http: impl AsRef<Http>
) -> Result<Vec<Integration>>

Gets all integration of the guild.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn invites(&self, http: impl AsRef<Http>) -> Result<Vec<RichInvite>>

Gets all of the guild’s invites.

Requires the Manage Guild permission.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn leave(&self, http: impl AsRef<Http>) -> Result<()>

Leaves the guild.

Errors

Returns Error::Http if the current user is unable to leave the Guild, or currently is not in the guild.

pub async fn member(
&self,
cache_http: impl CacheHttp,
user_id: impl Into<UserId>
) -> Result<Member>

Gets a user’s Member for the guild by Id.

Errors

Returns Error::Http if the member is not in the Guild, or if the Guild is otherwise unavailable.

pub async fn members(
&self,
http: impl AsRef<Http>,
limit: Option<u64>,
after: impl Into<Option<UserId>>
) -> Result<Vec<Member>>

Gets a list of the guild’s members.

Optionally pass in the limit to limit the number of results. Minimum value is 1, maximum and default value is 1000.

Optionally pass in after to offset the results by a User’s Id.

Errors

Returns an Error::Http if the API returns an error, may also return Error::NotInRange if the input is not within range.

pub async fn move_member(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>,
channel_id: impl Into<ChannelId>
) -> Result<Member>

Moves a member to a specific voice channel.

Requires the Move Members permission.

Errors

Returns an Error::Http if the current user lacks permission, or if the member is not currently in a voice channel for this Guild.

pub fn user_permissions_in(
&self,
channel: &GuildChannel,
member: &Member
) -> Result<Permissions>

Calculate a Member’s permissions in a given channel in the guild.

Errors

Returns Error::Model if the Member has a non-existent Role for some reason.

pub fn role_permissions_in(
&self,
channel: &GuildChannel,
role: &Role
) -> Result<Permissions>

Calculate a Role’s permissions in a given channel in the guild.

Errors

Returns Error::Model if the Role or Channel is not from this Guild.

pub async fn prune_count(
&self,
http: impl AsRef<Http>,
days: u16
) -> Result<GuildPrune>

Gets the number of Members that would be pruned with the given number of days.

Requires the Kick Members permission.

See Guild::prune_count.

Errors

Returns Error::Http if the current user lacks permission.

pub async fn shard_id(&self, cache: impl AsRef<Cache>) -> u64

Returns the Id of the shard associated with the guild.

When the cache is enabled this will automatically retrieve the total number of shards.

Note: When the cache is enabled, this function unlocks the cache to retrieve the total number of shards in use. If you already have the total, consider using utils::shard_id.

pub fn splash_url(&self) -> Option<String>

Returns the formatted URL of the guild’s splash image, if one exists.

pub async fn start_integration_sync(
&self,
http: impl AsRef<Http>,
integration_id: impl Into<IntegrationId>
) -> Result<()>

Starts an integration sync for the given integration Id.

Requires the Manage Guild permission.

Errors

See Guild::start_integration_sync.

pub async fn unban(
&self,
http: impl AsRef<Http>,
user_id: impl Into<UserId>
) -> Result<()>

Unbans a User from the guild.

Requires the Ban Members permission.

Errors

See Guild::unban.

pub async fn vanity_url(&self, http: impl AsRef<Http>) -> Result<String>

Retrieve’s the guild’s vanity URL.

Note: Requires the Manage Guild permission.

Errors

See Guild::vanity_url.

pub async fn webhooks(&self, http: impl AsRef<Http>) -> Result<Vec<Webhook>>

Retrieves the guild’s webhooks.

Note: Requires the Manage Webhooks permission.

Errors

See Guild::webhooks.

pub fn role_by_name(&self, role_name: &str) -> Option<&Role>

Obtain a reference to a role by its name.

Note: If two or more roles have the same name, obtained reference will be one of them.

Examples

Obtain a reference to a Role by its name.

use serenity::model::prelude::*;
use serenity::prelude::*;

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, context: Context, msg: Message) {
        if let Some(guild_id) = msg.guild_id {
            if let Some(guild) = guild_id.to_guild_cached(&context).await {
                if let Some(role) = guild.role_by_name("role_name") {
                    println!("Obtained role's reference: {:?}", role);
                }
            }
        }
    }
}

let mut client =Client::builder("token").event_handler(Handler).await?;

client.start().await?;

pub fn await_reply<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> CollectReply<'a>

Returns a future that will await one message sent in this guild.

pub fn await_replies<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> MessageCollectorBuilder<'a>

Returns a stream builder which can be awaited to obtain a stream of messages in this guild.

pub fn await_reaction<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> CollectReaction<'a>

Await a single reaction in this guild.

pub fn await_reactions<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> ReactionCollectorBuilder<'a>

Returns a stream builder which can be awaited to obtain a stream of reactions sent in this guild.