Module serenity::model::interactions::application_command

Structs

ApplicationCommand

The base command model that belongs to an application.

ApplicationCommandInteraction

An interaction when a user invokes a slash command.

ApplicationCommandInteractionData

The command data payload.

ApplicationCommandInteractionDataOption

A set of a parameter and a value from the user.

ApplicationCommandInteractionDataResolved

The resolved data of a command data interaction payload. It contains the objects of ApplicationCommandInteractionDataOptions.

ApplicationCommandOption

The parameters for an ApplicationCommand.

ApplicationCommandOptionChoice

The only valid values a user can pick in an ApplicationCommandOption.

ApplicationCommandPermission

An ApplicationCommand permission.

ApplicationCommandPermissionData

The ApplicationCommandPermission data.

Enums

ApplicationCommandInteractionDataOptionValue

The resolved value of an ApplicationCommandInteractionDataOption.

ApplicationCommandOptionType

The type of an ApplicationCommandOption.

ApplicationCommandPermissionType

The type of an ApplicationCommandPermissionData.

ApplicationCommandType

The type of an application command.

ResolvedTarget

The resolved value of a ApplicationCommandInteractionData::target_id.