Module serenity::model::invite

Models for server and channel invites.

Structs

Invite

Information about an invite code.

InviteChannel

A minimal amount of information about the channel an invite points to.

InviteGuild

A minimal amount of information about the guild an invite points to.

RichInvite

Detailed information about an invite. This information can only be retrieved by anyone with the Manage Guild permission. Otherwise, a minimal amount of information can be retrieved via the Invite struct.