Module serenity::model::misc

Miscellaneous helper traits, enums, and structs for models.

Structs

EmojiIdentifier

A version of an emoji used only when solely the animated state, Id, and name are known.

EmojiIdentifierParseErrormodel and utils

ImageHash

An image hash returned from the Discord API.

Incident

An incident retrieved from the Discord status page.

IncidentUpdate

An update to an incident from the Discord status page.

Maintenance

A Discord status maintenance message. This can be either for active maintenances or for scheduled maintenances.

Enums

ImageHashParseError

An error returned when ImageHash is passed an erronous value.