Module serenity::model::misc

Miscellaneous helper traits, enums, and structs for models.

Structs

AffectedComponent
A component that was affected during a service incident.

EmojiIdentifier
A version of an emoji used only when solely the animated state, Id, and name are known.

Incident
An incident retrieved from the Discord status page.

IncidentUpdate
An update to an incident from the Discord status page.

Maintenance
A Discord status maintenance message. This can be either for active maintenances or for scheduled maintenances.

Mention
A struct that represents some way to insert a notification, link, or emoji into a message.

Enums

ChannelIdParseError

ChannelParseError

IncidentStatus
The type of status update during a service incident.

RoleIdParseError

RoleParseError

UserIdParseError

UserParseError

Traits

Mentionable
Allows something - such as a channel or role - to be mentioned in a message.