Module serenity::model::monetization

Structs

Entitlement

Represents that a user or guild has access to a premium offering in the application.

Sku

A premium offering that can be made available to an application’s users and guilds.

SkuFlags

Differentates between user and server subscriptions.

Enums

EntitlementKind

Differentiates between Entitlement types.

EntitlementOwner

SkuKind

Differentiates between SKU classes.