Struct serenity::model::permissions::Permissions

pub struct Permissions { /* private fields */ }

A set of permissions that can be assigned to Users and Roles via PermissionOverwrites, roles globally in a Guild, and to GuildChannels.

Implementations

impl Permissions

pub const CREATE_INSTANT_INVITE: Self

Allows for the creation of RichInvites.

pub const KICK_MEMBERS: Self

Allows for the kicking of guild members.

pub const BAN_MEMBERS: Self

Allows the banning of guild members.

pub const ADMINISTRATOR: Self

Allows all permissions, bypassing channel permission overwrites.

pub const MANAGE_CHANNELS: Self

Allows management and editing of guild channels.

pub const MANAGE_GUILD: Self

Allows management and editing of the guild.

pub const ADD_REACTIONS: Self

Members with this permission can add new Reactions to a Message. Members can still react using reactions already added to messages without this permission.

pub const VIEW_AUDIT_LOG: Self

Allows viewing a guild’s audit logs.

pub const PRIORITY_SPEAKER: Self

Allows the use of priority speaking in voice channels.

pub const STREAM: Self

pub const VIEW_CHANNEL: Self

Allows guild members to view a channel, which includes reading messages in text channels and joining voice channels.

pub const SEND_MESSAGES: Self

Allows sending messages in a guild channel.

pub const SEND_TTS_MESSAGES: Self

Allows the sending of text-to-speech messages in a channel.

pub const MANAGE_MESSAGES: Self

Allows the deleting of other messages in a guild channel.

Note: This does not allow the editing of other messages.

pub const EMBED_LINKS: Self

Allows links from this user - or users of this role - to be embedded, with potential data such as a thumbnail, description, and page name.

pub const ATTACH_FILES: Self

Allows uploading of files.

pub const READ_MESSAGE_HISTORY: Self

Allows the reading of a channel’s message history.

pub const MENTION_EVERYONE: Self

Allows the usage of the @everyone mention, which will notify all users in a channel. The @here mention will also be available, and can be used to mention all non-offline users.

Note: You probably want this to be disabled for most roles and users.

pub const USE_EXTERNAL_EMOJIS: Self

Allows the usage of custom emojis from other guilds.

This does not dictate whether custom emojis in this guild can be used in other guilds.

pub const VIEW_GUILD_INSIGHTS: Self

Allows for viewing guild insights.

pub const CONNECT: Self

Allows the joining of a voice channel.

pub const SPEAK: Self

Allows the user to speak in a voice channel.

pub const MUTE_MEMBERS: Self

Allows the muting of members in a voice channel.

pub const DEAFEN_MEMBERS: Self

Allows the deafening of members in a voice channel.

pub const MOVE_MEMBERS: Self

Allows the moving of members from one voice channel to another.

pub const USE_VAD: Self

Allows the usage of voice-activity-detection in a voice channel.

If this is disabled, then Members must use push-to-talk.

pub const CHANGE_NICKNAME: Self

Allows members to change their own nickname in the guild.

pub const MANAGE_NICKNAMES: Self

Allows members to change other members’ nicknames.

pub const MANAGE_ROLES: Self

Allows management and editing of roles below their own.

pub const MANAGE_WEBHOOKS: Self

Allows management of webhooks.

pub const MANAGE_EMOJIS_AND_STICKERS: Self

Allows management of emojis and stickers created without the use of an Integration.

pub const USE_SLASH_COMMANDS: Self

Allows using slash commands.

pub const REQUEST_TO_SPEAK: Self

Allows for requesting to speak in stage channels.

pub const MANAGE_THREADS: Self

Allows for deleting and archiving threads, and viewing all private threads.

pub const CREATE_PUBLIC_THREADS: Self

Allows for creating threads.

pub const CREATE_PRIVATE_THREADS: Self

Allows for creating private threads.

pub const USE_EXTERNAL_STICKERS: Self

Allows the usage of custom stickers from other servers.

pub const SEND_MESSAGES_IN_THREADS: Self

Allows for sending messages in threads

pub const USE_EMBEDDED_ACTIVITIES: Self

Allows for launching activities in a voice channel

pub const MODERATE_MEMBERS: Self

Allows for timing out users to prevent them from sending or reacting to messages in chat and threads, and from speaking in voice and stage channels.

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl Permissions

pub fn get_permission_names(self) -> Vec<&'static str>

Returns a list of names of all contained permissions.

impl Permissions

pub fn add_reactions(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Add Reactions permission.

pub fn administrator(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Administrator permission.

pub fn attach_files(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Attach Files permission.

pub fn ban_members(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Ban Members permission.

pub fn change_nickname(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Change Nickname permission.

pub fn connect(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Connect permission.

pub fn view_audit_log(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the View Audit Log permission.

pub fn view_channel(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the View Channel permission.

pub fn view_guild_insights(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the View Guild Insights permission.

pub fn priority_speaker(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permission contains the Priority Speaker permission.

pub fn stream(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permission contains the Stream permission.

pub fn create_instant_invite(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Create Instant Invite permission.

pub fn create_private_threads(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Create Private Threads permission.

pub fn create_public_threads(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Create Public Threads permission.

pub fn deafen_members(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Deafen Members permission.

pub fn embed_links(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Embed Links permission.

pub fn external_emojis(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Use External Emojis permission.

pub fn kick_members(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Kick Members permission.

pub fn manage_channels(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Channels permission.

pub fn manage_emojis_and_stickers(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Emojis and Stickers permission.

pub fn manage_guild(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Guild permission.

pub fn manage_messages(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Messages permission.

pub fn manage_nicknames(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Nicknames permission.

pub fn manage_roles(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Roles permission.

pub fn manage_threads(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Threads permission.

pub fn manage_webhooks(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Manage Webhooks permission.

pub fn mention_everyone(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Mention Everyone permission.

pub fn moderate_members(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Moderate Members permission.

pub fn move_members(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Move Members permission.

pub fn mute_members(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Mute Members permission.

pub fn read_message_history(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Read Message History permission.

pub fn send_messages(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Send Messages permission.

pub fn send_messages_in_threads(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Send Messages in Threads permission.

pub fn send_tts_messages(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Send TTS Messages permission.

pub fn speak(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Speak permission.

pub fn request_to_speak(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Request To Speak permission.

pub fn use_embedded_activities(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Use Embedded Activities permission.

pub fn use_external_emojis(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Use External Emojis permission.

pub fn use_external_stickers(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Use External Stickers permission.

pub fn use_slash_commands(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Use Slash Commands permission.

pub fn use_vad(self) -> bool

This is supported on crate feature model only.

Shorthand for checking that the set of permissions contains the Use VAD permission.