Struct serenity::model::permissions::Permissions

pub struct Permissions {
    pub bits: u64,
}

A set of permissions that can be assigned to Users and Roles via PermissionOverwrites, roles globally in a Guild, and to GuildChannels.

Fields
bits: u64

The flags making up the permissions.

Note

Do not modify this yourself; use the provided methods. Do the same when creating, unless you’re absolutely certain that you’re giving valid permission flags.

Implementations

impl Permissions

pub const CREATE_INVITE: Permissions

Allows for the creation of RichInvites.

pub const KICK_MEMBERS: Permissions

Allows for the kicking of guild members.

pub const BAN_MEMBERS: Permissions

Allows the banning of guild members.

pub const ADMINISTRATOR: Permissions

Allows all permissions, bypassing channel permission overwrites.

pub const MANAGE_CHANNELS: Permissions

Allows management and editing of guild channels.

pub const MANAGE_GUILD: Permissions

Allows management and editing of the guild.

pub const ADD_REACTIONS: Permissions

Members with this permission can add new Reactions to a Message. Members can still react using reactions already added to messages without this permission.

pub const VIEW_AUDIT_LOG: Permissions

Allows viewing a guild’s audit logs.

pub const PRIORITY_SPEAKER: Permissions

Allows the use of priority speaking in voice channels.

pub const STREAM: Permissions

pub const READ_MESSAGES: Permissions

Allows reading messages in a guild channel. If a user does not have this permission, then they will not be able to see the channel.

pub const SEND_MESSAGES: Permissions

Allows sending messages in a guild channel.

pub const SEND_TTS_MESSAGES: Permissions

Allows the sending of text-to-speech messages in a channel.

pub const MANAGE_MESSAGES: Permissions

Allows the deleting of other messages in a guild channel.

Note: This does not allow the editing of other messages.

pub const EMBED_LINKS: Permissions

Allows links from this user - or users of this role - to be embedded, with potential data such as a thumbnail, description, and page name.

pub const ATTACH_FILES: Permissions

Allows uploading of files.

pub const READ_MESSAGE_HISTORY: Permissions

Allows the reading of a channel’s message history.

pub const MENTION_EVERYONE: Permissions

Allows the usage of the @everyone mention, which will notify all users in a channel. The @here mention will also be available, and can be used to mention all non-offline users.

Note: You probably want this to be disabled for most roles and users.

pub const USE_EXTERNAL_EMOJIS: Permissions

Allows the usage of custom emojis from other guilds.

This does not dictate whether custom emojis in this guild can be used in other guilds.

pub const CONNECT: Permissions

Allows the joining of a voice channel.

pub const SPEAK: Permissions

Allows the user to speak in a voice channel.

pub const MUTE_MEMBERS: Permissions

Allows the muting of members in a voice channel.

pub const DEAFEN_MEMBERS: Permissions

Allows the deafening of members in a voice channel.

pub const MOVE_MEMBERS: Permissions

Allows the moving of members from one voice channel to another.

pub const USE_VAD: Permissions

Allows the usage of voice-activity-detection in a voice channel.

If this is disabled, then Members must use push-to-talk.

pub const CHANGE_NICKNAME: Permissions

Allows members to change their own nickname in the guild.

pub const MANAGE_NICKNAMES: Permissions

Allows members to change other members’ nicknames.

pub const MANAGE_ROLES: Permissions

Allows management and editing of roles below their own.

pub const MANAGE_WEBHOOKS: Permissions

Allows management of webhooks.

pub const MANAGE_EMOJIS: Permissions

Allows management of emojis created without the use of an Integration.

pub const REQUEST_TO_SPEAK: Permissions

Allows for requesting to speak in stage channels.

pub const USE_SLASH_COMMANDS: Permissions

Allows using slash commands.

pub const fn empty() -> Permissions

Returns an empty set of flags

pub const fn all() -> Permissions

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u64) -> Option<Permissions>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Permissions

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Permissions

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Permissions) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Permissions) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: Permissions)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Permissions)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Permissions)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Permissions, value: bool)

Inserts or removes the specified flags depending on the passed value.

impl Permissions

pub fn get_permission_names(self) -> Vec<&'static str>

Returns a list of names of all contained permissions.

impl Permissions

pub fn add_reactions(self) -> bool

Shorthand for checking that the set of permissions contains the Add Reactions permission.

pub fn administrator(self) -> bool

Shorthand for checking that the set of permissions contains the Administrator permission.

pub fn attach_files(self) -> bool

Shorthand for checking that the set of permissions contains the Attach Files permission.

pub fn ban_members(self) -> bool

Shorthand for checking that the set of permissions contains the Ban Members permission.

pub fn change_nickname(self) -> bool

Shorthand for checking that the set of permissions contains the Change Nickname permission.

pub fn connect(self) -> bool

Shorthand for checking that the set of permissions contains the Connect permission.

pub fn view_audit_log(self) -> bool

Shorthand for checking that the set of permissions contains the View Audit Log permission.

pub fn priority_speaker(self) -> bool

Shorthand for checking that the set of permission contains the Priority Speaker permission.

pub fn stream(self) -> bool

Shorthand for checking that the set of permission contains the Stream permission.

pub fn create_invite(self) -> bool

Shorthand for checking that the set of permissions contains the Create Invite permission.

pub fn deafen_members(self) -> bool

Shorthand for checking that the set of permissions contains the Deafen Members permission.

pub fn embed_links(self) -> bool

Shorthand for checking that the set of permissions contains the Embed Links permission.

pub fn external_emojis(self) -> bool

Shorthand for checking that the set of permissions contains the Use External Emojis permission.

pub fn kick_members(self) -> bool

Shorthand for checking that the set of permissions contains the Kick Members permission.

pub fn manage_channels(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Channels permission.

pub fn manage_emojis(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Emojis permission.

pub fn manage_guild(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Guild permission.

pub fn manage_messages(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Messages permission.

pub fn manage_nicknames(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Nicknames permission.

pub fn manage_roles(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Roles permission.

pub fn manage_webhooks(self) -> bool

Shorthand for checking that the set of permissions contains the Manage Webhooks permission.

pub fn mention_everyone(self) -> bool

Shorthand for checking that the set of permissions contains the Mention Everyone permission.

pub fn move_members(self) -> bool

Shorthand for checking that the set of permissions contains the Move Members permission.

pub fn mute_members(self) -> bool

Shorthand for checking that the set of permissions contains the Mute Members permission.

pub fn read_message_history(self) -> bool

Shorthand for checking that the set of permissions contains the Read Message History permission.

pub fn read_messages(self) -> bool

Shorthand for checking that the set of permissions contains the Read Messages permission.

pub fn send_messages(self) -> bool

Shorthand for checking that the set of permissions contains the Send Messages permission.

pub fn send_tts_messages(self) -> bool

Shorthand for checking that the set of permissions contains the Send TTS Messages permission.

pub fn speak(self) -> bool

Shorthand for checking that the set of permissions contains the Speak permission.

pub fn request_to_speak(self) -> bool

Shorthand for checking that the set of permissions contains the Request To Speak permission.

pub fn use_external_emojis(self) -> bool

Shorthand for checking that the set of permissions contains the Use External Emojis permission.

pub fn use_slash_commands(self) -> bool

Shorthand for checking that the set of permissions contains the Use Slash Commands permission.

pub fn use_vad(self) -> bool

Shorthand for checking that the set of permissions contains the Use VAD permission.