Module serenity::model::prelude::event

All the events this library handles.

Structs

ApplicationCommandPermissionsUpdateEvent

Discord docs.

AutoModerationActionExecutionEvent

Discord docs.

AutoModerationRuleCreateEvent

Discord docs.

AutoModerationRuleDeleteEvent

Discord docs.

AutoModerationRuleUpdateEvent

Discord docs.

ChannelCreateEvent

Event data for the channel creation event.

ChannelDeleteEvent

Discord docs.

ChannelPinsUpdateEvent

Discord docs.

ChannelUpdateEvent

Discord docs.

GuildBanAddEvent

Discord docs.

GuildBanRemoveEvent

Discord docs.

GuildCreateEvent

Discord docs.

GuildDeleteEvent

Discord docs.

GuildEmojisUpdateEvent

Discord docs.

GuildIntegrationsUpdateEvent

Discord docs.

GuildMemberAddEvent

Discord docs.

GuildMemberRemoveEvent

Discord docs.

GuildMemberUpdateEvent

Discord docs.

GuildMembersChunkEvent

Discord docs.

GuildRoleCreateEvent

Discord docs.

GuildRoleDeleteEvent

Discord docs.

GuildRoleUpdateEvent

Discord docs.

GuildScheduledEventCreateEvent

Discord docs.

GuildScheduledEventDeleteEvent

Discord docs.

GuildScheduledEventUpdateEvent

Discord docs.

GuildScheduledEventUserAddEvent

Discord docs.

GuildScheduledEventUserRemoveEvent

Discord docs.

GuildStickersUpdateEvent

Discord docs.

GuildUnavailableEvent

GuildUpdateEvent

Discord docs.

IntegrationCreateEvent

Discord docs.

IntegrationDeleteEvent

Discord docs.

IntegrationUpdateEvent

Discord docs.

InteractionCreateEvent

Discord docs.

InviteCreateEvent

Discord docs.

InviteDeleteEvent

Discord docs.

MessageCreateEvent

Discord docs.

MessageDeleteBulkEvent

Discord docs.

MessageDeleteEvent

Discord docs.

MessageUpdateEvent

Discord docs.

PresenceUpdateEvent

Discord docs.

PresencesReplaceEvent

ReactionAddEvent

Discord docs.

ReactionRemoveAllEvent

Discord docs.

ReactionRemoveEvent

Discord docs.

ReadyEvent

The “Ready” event, containing initial ready cache

RelatedIdsForEventType

Defines the related IDs that may exist for an EventType.

ResumedEvent

Discord docs.

StageInstanceCreateEvent

Discord docs.

StageInstanceDeleteEvent

Discord docs.

StageInstanceUpdateEvent

Discord docs.

ThreadCreateEvent

Discord docs.

ThreadDeleteEvent

Discord docs.

ThreadListSyncEvent

Discord docs.

ThreadMemberUpdateEvent

Discord docs.

ThreadMembersUpdateEvent

Discord docs.

ThreadUpdateEvent

Discord docs.

TypingStartEvent

Discord docs.

UnknownEvent

UserUpdateEvent

Discord docs.

VoiceServerUpdateEvent

Discord docs.

VoiceStateUpdateEvent

Discord docs.

WebhookUpdateEvent

Discord docs.

Enums

Event

Event received over a websocket connection

EventType

The type of event dispatch received from the gateway.

GatewayEvent

Discord docs.

RelatedId

Similar to Option, but with additional variants relevant to Event’s id methods (such as Event::user_id).

Functions

deserialize_event_with_type

Deserializes a serde_json::Value into an Event.