Struct serenity::model::timestamp::Timestamp

pub struct Timestamp(_);

Implementations

impl Timestamp

pub fn from_millis(millis: i64) -> Result<Self, InvalidTimestamp>

Creates a new Timestamp from the number of milliseconds since 1970.

Errors

Returns Err if the value is invalid.

pub fn now() -> Self

Create a new Timestamp with the current date and time in UTC.

pub fn from_unix_timestamp(secs: i64) -> Result<Self, InvalidTimestamp>

Creates a new Timestamp from a UNIX timestamp (seconds since 1970).

Errors

Returns Err if the value is invalid.

pub fn unix_timestamp(&self) -> i64

Returns the number of non-leap seconds since January 1, 1970 0:00:00 UTC

pub fn parse(input: &str) -> Result<Timestamp, ParseError>

Parse a timestamp from an RFC 3339 date and time string.

Examples

let timestamp = Timestamp::parse("2016-04-30T11:18:25Z").unwrap();
let timestamp = Timestamp::parse("2016-04-30T11:18:25+00:00").unwrap();
let timestamp = Timestamp::parse("2016-04-30T11:18:25.796Z").unwrap();

assert!(Timestamp::parse("2016-04-30T11:18:25").is_err());
assert!(Timestamp::parse("2016-04-30T11:18").is_err());

Errors

Returns Err if the string is not a valid RFC 3339 date and time string.

pub fn to_rfc3339(&self) -> Option<String>