Struct serenity::model::user::User

#[non_exhaustive]pub struct User {
    pub id: UserId,
    pub avatar: Option<String>,
    pub bot: bool,
    pub discriminator: u16,
    pub name: String,
    pub public_flags: Option<UserPublicFlags>,
}

Information about a user.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct {{ .. }} syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
id: UserId

The unique Id of the user. Can be used to calculate the account’s creation date.

avatar: Option<String>

Optional avatar hash.

bot: bool

Indicator of whether the user is a bot.

discriminator: u16

The account’s discriminator to differentiate the user from others with the same name. The name+discriminator pair is always unique.

name: String

The account’s username. Changing username will trigger a discriminator change if the username+discriminator pair becomes non-unique.

public_flags: Option<UserPublicFlags>

the public flags on a user’s account

Implementations

impl User

pub fn avatar_url(&self) -> Option<String>

Returns the formatted URL of the user’s icon, if one exists.

This will produce a WEBP image URL, or GIF if the user has a GIF avatar.

pub async fn create_dm_channel(
&self,
cache_http: impl CacheHttp
) -> Result<PrivateChannel>

Creates a direct message channel between the current user and the user. This can also retrieve the channel if one already exists.

Errors

See UserId::create_dm_channel for what errors may be returned.

pub fn created_at(&self) -> DateTime<Utc>

Retrieves the time that this user was created at.

pub fn default_avatar_url(&self) -> String

Returns the formatted URL to the user’s default avatar URL.

This will produce a PNG URL.

pub async fn direct_message<F>(
&self,
cache_http: impl CacheHttp,
f: F
) -> Result<Message> where
F: FnOnce(&'b mut CreateMessage<'a>) -> &'b mut CreateMessage<'a>,

Sends a message to a user through a direct message channel. This is a channel that can only be accessed by you and the recipient.

Examples

When a user sends a message with a content of "~help", DM the author a help message, and then react with '👌' to verify message sending:

use serenity::model::Permissions;

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, ctx: Context, msg: Message) {
        if msg.content == "~help" {
            let url = match ctx.cache.current_user().await.invite_url(&ctx, Permissions::empty()).await {
                Ok(v) => v,
                Err(why) => {
                    println!("Error creating invite url: {:?}", why);

                    return;
                },
            };

            let help = format!(
                "Helpful info here. Invite me with this link: <{}>",
                url,
            );

            let dm = msg.author.direct_message(&ctx, |m| {
                m.content(&help)
            })
            .await;

            match dm {
                Ok(_) => {
                    let _ = msg.react(&ctx, '👌').await;
                },
                Err(why) => {
                    println!("Err sending help: {:?}", why);

                    let _ = msg.reply(&ctx, "There was an error DMing you help.").await;
                },
            };
        }
    }
}

let mut client =Client::builder("token").event_handler(Handler).await?;

Errors

Returns a ModelError::MessagingBot if the user being direct messaged is a bot user.

May also return an Error::Http if the message was illformed, or if the user cannot be sent a direct message.

Error::Json can also be returned if there is an error deserializing the API response.

pub async fn dm<F>(&self, cache_http: impl CacheHttp, f: F) -> Result<Message> where
F: FnOnce(&'b mut CreateMessage<'a>) -> &'b mut CreateMessage<'a>,

This is an alias of direct_message.

pub fn face(&self) -> String

Retrieves the URL to the user’s avatar, falling back to the default avatar if needed.

This will call avatar_url first, and if that returns None, it then falls back to default_avatar_url.

pub async fn has_role(
&self,
cache_http: impl CacheHttp,
guild: impl Into<GuildContainer>,
role: impl Into<RoleId>
) -> Result<bool>

Check if a user has a Role. This will retrieve the Guild from the Cache if it is available, and then check if that guild has the given Role.

Three forms of data may be passed in to the guild parameter: either a PartialGuild, a GuildId, or a u64.

Examples

Check if a guild has a Role by Id:

// Assumes a 'guild_id' and `role_id` have already been bound
let _ = message.author.has_role(guild_id, role_id);

Errors

Returns an Error::Http if the given Guild is unavailable, if that Role does not exist in the given Guild, or if the given User is not in that Guild.

May also return an Error::Json if there is an error in deserializing the API response.

pub async fn refresh(&mut self, cache_http: impl CacheHttp) -> Result<()>

Refreshes the information about the user.

Replaces the instance with the data retrieved over the REST API.

Errors

See UserId::to_user for what errors may be returned.

pub fn static_avatar_url(&self) -> Option<String>

Returns a static formatted URL of the user’s icon, if one exists.

This will always produce a WEBP image URL.

pub fn tag(&self) -> String

Returns the “tag” for the user.

The “tag” is defined as “username#discriminator”, such as “zeyla#5479”.

Examples

Make a command to tell the user what their tag is:

use serenity::utils::MessageBuilder;
use serenity::utils::ContentModifier::Bold;

struct Handler;

#[serenity::async_trait]
impl EventHandler for Handler {
    async fn message(&self, context: Context, msg: Message) {
        if msg.content == "!mytag" {
            let content = MessageBuilder::new()
                .push("Your tag is ")
                .push(Bold + msg.author.tag())
                .build();

            let _ = msg.channel_id.say(&context.http, &content).await;
        }
    }
}
let mut client =Client::builder("token").event_handler(Handler).await?;

client.start().await?;

pub async fn nick_in(
&self,
cache_http: impl CacheHttp,
guild_id: impl Into<GuildId>
) -> Option<String>

Returns the user’s nickname in the given guild_id.

If none is used, it returns None.

pub fn await_reply<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> CollectReply<'a>

This is supported on crate feature collector only.

Returns a future that will await one message by this user.

pub fn await_replies<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> MessageCollectorBuilder<'a>

This is supported on crate feature collector only.

Returns a stream builder which can be awaited to obtain a stream of messages sent by this user.

pub fn await_reaction<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> CollectReaction<'a>

This is supported on crate feature collector only.

Await a single reaction by this user.

pub fn await_reactions<'a>(
&self,
shard_messenger: &'a impl AsRef<ShardMessenger>
) -> ReactionCollectorBuilder<'a>

This is supported on crate feature collector only.

Returns a stream builder which can be awaited to obtain a stream of reactions sent by this user.