Struct serenity::model::webhook::Webhook

#[non_exhaustive]
pub struct Webhook {
    pub id: WebhookId,
    pub kind: WebhookType,
    pub avatar: Option<String>,
    pub channel_id: Option<ChannelId>,
    pub guild_id: Option<GuildId>,
    pub name: Option<String>,
    pub token: Option<String>,
    pub user: Option<User>,
}

A representation of a webhook, which is a low-effort way to post messages to channels. They do not necessarily require a bot user or authentication to use.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: WebhookId

The unique Id.

Can be used to calculate the creation date of the webhook.

kind: WebhookType

The type of the webhook.

avatar: Option<String>

The default avatar.

This can be modified via ExecuteWebhook::avatar_url.

channel_id: Option<ChannelId>

The Id of the channel that owns the webhook.

guild_id: Option<GuildId>

The Id of the guild that owns the webhook.

name: Option<String>

The default name of the webhook.

This can be modified via ExecuteWebhook::username.

token: Option<String>

The webhook’s secure token.

user: Option<User>

The user that created the webhook.

Note: This is not received when getting a webhook by its token.

Implementations

impl Webhook

pub async fn delete(&self, http: impl AsRef<Http>) -> Result<()>

This is supported on crate feature model only.

Deletes the webhook.

As this calls the Http::delete_webhook_with_token function, authentication is not required.

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the webhook does not exist, the token is invalid, or if the webhook could not otherwise be deleted.

pub async fn edit_name(
&mut self,
http: impl AsRef<Http>,
name: &str
) -> Result<()>

This is supported on crate feature model only.

Edits the name of a webhook.

Refer to Http::edit_webhook for restrictions on webhook names.

Does not require authentication, as this calls Http::edit_webhook_with_token internally.

Examples

let url = "https://discord.com/api/webhooks/245037420704169985/ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let mut webhook = http.get_webhook_from_url(url).await?;

webhook.edit_name(&http, "new name").await?;

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the content is malformed, or if the token is invalid.

Or may return an Error::Json if there is an error in deserialising Discord’s response.

pub async fn edit_avatar<'a>(
&mut self,
http: impl AsRef<Http>,
avatar: impl Into<AttachmentType<'a>>
) -> Result<()>

This is supported on crate feature model only.

Edits a webhook’s avatar.

Refer to Http::edit_webhook for restrictions on wehbhook avatars.

Does not require authentication, as it calls Http::edit_webhook_with_token internally.

Examples

let url = "https://discord.com/api/webhooks/245037420704169985/ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let mut webhook = http.get_webhook_from_url(url).await?;

webhook.edit_avatar(&http, "./webhook_img.png").await?;

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the content is malformed, or if the token is invalid.

Or may return an Error::Json if there is an error in deserialising Discord’s response.

pub async fn delete_avatar(&mut self, http: impl AsRef<Http>) -> Result<()>

This is supported on crate feature model only.

Deletes a webhook’s avatar, resetting it to the default logo.

Does not require authentication, as it calls Http::edit_webhook_with_token internally.

Examples

let url = "https://discord.com/api/webhooks/245037420704169985/ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let mut webhook = http.get_webhook_from_url(url).await?;

webhook.delete_avatar(&http).await?;

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the content is malformed, or if the token is invalid.

Or may return an Error::Json if there is an error in deserialising Discord’s response.

pub async fn execute<'a, F>(
&self,
http: impl AsRef<Http>,
wait: bool,
f: F
) -> Result<Option<Message>> where
for<'b> F: FnOnce(&'b mut ExecuteWebhook<'a>) -> &'b mut ExecuteWebhook<'a>,

This is supported on crate feature model only.

Executes a webhook with the fields set via the given builder.

The builder provides a method of setting only the fields you need, without needing to pass a long set of arguments.

Examples

Execute a webhook with message content of test:

let url = "https://discord.com/api/webhooks/245037420704169985/ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let mut webhook = http.get_webhook_from_url(url).await?;

webhook.execute(&http, false, |w| w.content("test")).await?;

Execute a webhook with message content of test, overriding the username to serenity, and sending an embed:

use serenity::model::channel::Embed;

let url = "https://discord.com/api/webhooks/245037420704169985/ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV";
let mut webhook = http.get_webhook_from_url(url).await?;

let embed = Embed::fake(|e| {
    e.title("Rust's website")
        .description(
            "Rust is a systems programming language that runs
                   blazingly fast, prevents segfaults, and guarantees
                   thread safety.",
        )
        .url("https://rust-lang.org")
});

webhook
    .execute(&http, false, |w| w.content("test").username("serenity").embeds(vec![embed]))
    .await?;

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the content is malformed, or if the webhook’s token is invalid.

Or may return an Error::Json if there is an error deserialising Discord’s response.

pub async fn get_message(
&self,
http: impl AsRef<Http>,
message_id: MessageId
) -> Result<Message>

This is supported on crate feature model only.

Gets a previously sent message from the webhook.

Errors

Returns an Error::Model if the Self::token is None.

May also return Error::Http if the webhook’s token is invalid, or the given message Id does not belong to the current webhook.

Or may return an Error::Json if there is an error deserialising Discord’s response.

pub async fn edit_message<F>(
&self,
http: impl AsRef<Http>,
message_id: MessageId,
f: F
) -> Result<Message> where
F: FnOnce(&mut EditWebhookMessage) -> &mut EditWebhookMessage,

This is supported on crate feature model only.

Edits a webhook message with the fields set via the given builder.

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the content is malformed, the webhook’s token is invalid, or the given message Id does not belong to the current webhook.

Or may return an Error::Json if there is an error deserialising Discord’s response.

pub async fn delete_message(
&self,
http: impl AsRef<Http>,
message_id: MessageId
) -> Result<()>

This is supported on crate feature model only.

Deletes a webhook message.

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the webhook’s token is invalid or the given message Id does not belong to the current webhook.

pub async fn refresh(&mut self, http: impl AsRef<Http>) -> Result<()>

This is supported on crate feature model only.

Retrieves the latest information about the webhook, editing the webhook in-place.

As this calls the Http::get_webhook_with_token function, authentication is not required.

Errors

Returns an Error::Model if the Self::token is None.

May also return an Error::Http if the http client errors or if Discord returns an error. Such as if the Webhook was deleted.

Or may return an Error::Json if there is an error deserialising Discord’s response.

pub fn url(&self) -> Result<String>

This is supported on crate feature model only.

Returns the url of the webhook.

assert_eq!(hook.url(), "https://discord.com/api/webhooks/245037420704169985/ig5AO-wdVWpCBtUUMxmgsWryqgsW3DChbKYOINftJ4DCrUbnkedoYZD0VOH1QLr-S3sV")

Errors

Returns an Error::Model if the Self::token is None.