Module serenity::prelude

A set of exports which can be helpful to use.

Note that the SerenityError re-export is equivalent to serenity::Error, although is re-exported as a separate name to remove likely ambiguity with other crate error enums.

Examples

Import all of the exports:

use serenity::prelude::*;

Reexports

pub use client::Client;

Structs

Context
The context is a general utility struct provided on event dispatches, which helps with dealing with the current "context" of the event dispatch. The context also acts as a general high-level interface over the associated Shard which received the event, or the low-level http module.

Enums

ClientError
An error returned from the Client.

GatewayError
An error that occurred while attempting to deal with the gateway.

HttpError

ModelError
An error returned from the model module.

SerenityError
A common error enum returned by most of the library's functionality within a custom Result.

Traits

EventHandler

Mentionable
Allows something - such as a channel or role - to be mentioned in a message.