Module serenity::prelude

A set of exports which can be helpful to use.

Note that the SerenityError re-export is equivalent to serenity::Error, although is re-exported as a separate name to remove likely ambiguity with other crate error enums.

Examples

Import all of the exports:

use serenity::prelude::*;

Re-exports

pub use crate::client::Client;

gateway

pub use crate::http::CacheHttp;

pub use crate::model::mention::Mentionable;

pub use crate::model::gateway::GatewayIntents;

pub use crate::model::ModelError;

pub use crate::CacheAndHttp;

client

Structs

Contextclient

The context is a general utility struct provided on event dispatches, which helps with dealing with the current “context” of the event dispatch. The context also acts as a general high-level interface over the associated Shard which received the event, or the low-level http module.

Mutex

An asynchronous Mutex-like type.

RwLock

An asynchronous reader-writer lock.

TypeMapclient

TypeMap is a simple abstraction around the standard library’s HashMap type, where types are its keys. This allows for statically-checked value retrieval.

Enums

ClientErrorclient and gateway

An error returned from the Client.

GatewayErrorgateway

An error that occurred while attempting to deal with the gateway.

HttpErrorhttp

SerenityError

A common error enum returned by most of the library’s functionality within a custom Result.

Traits

EventHandlergateway and client

The core trait for handling events by serenity.

RawEventHandlergateway and client

This core trait for handling raw events

TypeMapKeyclient

TypeMapKey is used to declare key types that are eligible for use with TypeMap.