Trait serenity::prelude::TypeMapKey

pub trait TypeMapKey: Any {
    type Value: Any + Send + Sync;
}

Available on crate feature client only.

TypeMapKey is used to declare key types that are eligible for use with TypeMap.

Required Associated Types

type Value: Any + Send + Sync

Defines the value type that corresponds to this TypeMapKey.