Trait serenity::prelude::TypeMapKey

pub trait TypeMapKey: Any {
    type Value: Send + Sync;
}

This is supported on crate feature client only.

TypeMapKey is used to declare key types that are eligible for use with TypeMap.

Associated Types

type Value: Send + Sync

Defines the value type that corresponds to this TypeMapKey.