Type Definition serenity::Result

pub type Result<T> = StdResult<T, Error>;

The common result type between most library functions.

The library exposes functions which, for a result type, exposes only one type, rather than the usual 2 (Result<T, Error>). This is because all functions that return a result return serenity’s Error, so this is implied, and a “simpler” result is used.