Enum serenity::utils::ContentModifier

#[non_exhaustive]
pub enum ContentModifier {
    Italic,
    Bold,
    Strikethrough,
    Code,
    Underline,
    Spoiler,
}

Formatting modifiers for MessageBuilder content pushes

Provides an enum of formatting modifiers for a string, for combination with string types and Content types.

Examples

Create a new Content type which describes a bold-italic “text”:

use serenity::utils::ContentModifier::{Bold, Italic};
use serenity::utils::Content;
let content: Content = Bold + Italic + "text";

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Italic

Bold

Strikethrough

Code

Underline

Spoiler