Trait serenity::utils::ArgumentConvert

pub trait ArgumentConvert: Sized {
    type Err;

    // Required method
    fn convert<'life0, 'async_trait>(
        ctx: impl 'async_trait + CacheHttp,
        guild_id: Option<GuildId>,
        channel_id: Option<ChannelId>,
        s: &'life0 str
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Err>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Available on crate features client and utils only.

Parse a value from a string in context of a received message.

This trait is a superset of std::str::FromStr. The difference is that this trait aims to support serenity-specific Discord types like Member or Message.

Trait implementations may do network requests as part of their parsing procedure.

Useful for implementing argument parsing in command frameworks.

Required Associated Types

type Err

The associated error which can be returned from parsing.

Required Methods

fn convert<'life0, 'async_trait>( ctx: impl 'async_trait + CacheHttp, guild_id: Option<GuildId>, channel_id: Option<ChannelId>, s: &'life0 str ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Err>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Parses a string s as a command parameter of this type.