Trait serenity::utils::Parse

pub trait Parse: Sized {
    type Err;
#[must_use]    fn parse<'life0, 'life1, 'life2, 'async_trait>(
        ctx: &'life0 Context, 
        msg: &'life1 Message, 
        s: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Err>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

Parse a value from a string in context of a received message.

This trait is a superset of std::str::FromStr. The difference is that this trait aims to support serenity-specific Discord types like Member or Message.

Trait implementations may do network requests as part of their parsing procedure.

Useful for implementing argument parsing in command frameworks.

Associated Types

type Err

The associated error which can be returned from parsing.

Loading content...
Required methods

#[must_use]fn parse<'life0, 'life1, 'life2, 'async_trait>(
ctx: &'life0 Context,
msg: &'life1 Message,
s: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<Self, Self::Err>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Parses a string s as a command parameter of this type.

Loading content...